Function core::any::type_name_of_val

pub fn type_name_of_val<T: ?Sized>(_val: &T) -> &'static str

🔬 This is a nightly-only experimental API. (type_name_of_val #66359)

以字符串切片的形式返回指向值的类型名称。 这与 `type_name::()` 相同，但可以在变量类型不易获得的情况下使用。

Returns the name of the type of the pointed-to value as a string slice. This is the same as type_name::<T>(), but can be used where the type of a variable is not easily available.

Note

这用于诊断用途。 没有指定字符串的确切内容和格式，只是对类型的尽力描述。 例如，`type_name_of_val::<>>(None)` 可以返回 `"Option"` 或 `"std::option::Option"`，但是 不是`"foobar"`。 此外，输出可能会在编译器版本之间发生变化。

This is intended for diagnostic use. The exact contents and format of the string are not specified, other than being a best-effort description of the type. For example, type_name_of_val::<Option<String>>(None) could return "Option<String>" or "std::option::Option<std::string::String>", but not "foobar". In addition, the output may change between versions of the compiler.

此函数不解析 trait 对象，这意味着 `type_name_of_val(&7u32 as &dyn Debug)` 可能返回 `"dyn Debug"`，但不能返回 `"u32"`。

This function does not resolve trait objects, meaning that type_name_of_val(&7u32 as &dyn Debug) may return "dyn Debug", but not "u32".

类型名称不应被视为类型的唯一标识符； 多个类型可以共享相同的类型名称。

The type name should not be considered a unique identifier of a type; multiple types may share the same type name.

当前实现使用与编译器诊断和调试信息相同的基础结构，但这不能保证。

The current implementation uses the same infrastructure as compiler diagnostics and debuginfo, but this is not guaranteed.

Examples

打印默认整数和浮点类型。

Prints the default integer and float types.

#![feature(type_name_of_val)]
use std::any::type_name_of_val;

let x = 1;
println!("{}", type_name_of_val(&x));
let y = 1.0;
println!("{}", type_name_of_val(&y));
Run