Module core::arch

🔬 This is a nightly-only experimental API. (stdsimd)

Platform dependent vendor intrinsics.

This documentation is for the version of this module in the coresimd crate, but you probably want to use the stdsimd crate which should have more complete documentation.

Also note that while this module may appear to contains the intrinsics for only one platform it actually contains intrinsics for multiple platforms compiled in conditionally. For other platforms of stdsimd see:

x86

x86_64

arm

aarch64

Modules

aarch64 	 [
Experimental
] [
AArch64
]
Platform-specific intrinsics for the aarch64 platform.

arm 	 [
Experimental
] [
ARM
]
Platform-specific intrinsics for the arm platform.

x86 	 [
Experimental
] [
x86
]
Platform-specific intrinsics for the x86 platform.

x86_64 	 [
Experimental
] [
x86-64
]
Platform-specific intrinsics for the x86_64 platform.