core::arch::aarch64

Function svstnt1h_scatter_u64base_offset_s64

pub unsafe fn svstnt1h_scatter_u64base_offset_s64(
    pg: svbool_t,
    bases: svuint64_t,
    offset: i64,
    data: svint64_t,
)

🔬This is a nightly-only experimental API. (stdarch_aarch64_sve #145052)

Available on (AArch64 or ARM64EC) and target feature sve and target feature sve2 only.

Truncate to 16 bits and store, non-temporal Arm’s documentation

Safety

pointer::offset safety constraints must be met for the address calculation for each active element (governed by pg).

This dereferences and accesses the calculated address for each active element (governed by pg).

Addresses passed in bases lack provenance, so this is similar to using a usize as ptr cast (or core::ptr::with_exposed_provenance) on each lane before using it.

Non-temporal accesses have special memory ordering rules, and explicit barriers may be required for some applications.