Module core::arch::arm

🔬 This is a nightly-only experimental API. (stdsimd #27731)

This is supported on ARM only.

Platform-specific intrinsics for the arm platform.

See the module documentation for more details.

Structs

float32x2_t	ExperimentalARM
ARM-specific 64-bit wide vector of two packed f32.

float32x4_t	ExperimentalARM
ARM-specific 128-bit wide vector of four packed f32.

int16x2_t	ExperimentalARM
ARM-specific 32-bit wide vector of two packed i16.

int16x4_t	ExperimentalARM
ARM-specific 64-bit wide vector of four packed i16.

int16x8_t	ExperimentalARM
ARM-specific 128-bit wide vector of eight packed i16.

int32x2_t	ExperimentalARM
ARM-specific 64-bit wide vector of two packed i32.

int32x4_t	ExperimentalARM
ARM-specific 128-bit wide vector of four packed i32.

int64x1_t	ExperimentalARM
ARM-specific 64-bit wide vector of one packed i64.

int64x2_t	ExperimentalARM
ARM-specific 128-bit wide vector of two packed i64.

int8x4_t	ExperimentalARM
ARM-specific 32-bit wide vector of four packed i8.

int8x8_t	ExperimentalARM
ARM-specific 64-bit wide vector of eight packed i8.

int8x16_t	ExperimentalARM
ARM-specific 128-bit wide vector of sixteen packed i8.

int8x8x2_t	ExperimentalARM
ARM-specific type containing two int8x8_t vectors.

int8x8x3_t	ExperimentalARM
ARM-specific type containing three int8x8_t vectors.

int8x8x4_t	ExperimentalARM
ARM-specific type containing four int8x8_t vectors.

poly16x4_t	ExperimentalARM
ARM-specific 64-bit wide vector of four packed u16.

poly16x8_t	ExperimentalARM
ARM-specific 128-bit wide vector of eight packed u16.

poly8x8_t	ExperimentalARM
ARM-specific 64-bit wide polynomial vector of eight packed u8.

poly8x16_t	ExperimentalARM
ARM-specific 128-bit wide vector of sixteen packed u8.

poly8x8x2_t	ExperimentalARM
ARM-specific type containing two poly8x8_t vectors.

poly8x8x3_t	ExperimentalARM
ARM-specific type containing three poly8x8_t vectors.

poly8x8x4_t	ExperimentalARM
ARM-specific type containing four poly8x8_t vectors.

uint16x2_t	ExperimentalARM
ARM-specific 32-bit wide vector of two packed u16.

uint16x4_t	ExperimentalARM
ARM-specific 64-bit wide vector of four packed u16.

uint16x8_t	ExperimentalARM
ARM-specific 128-bit wide vector of eight packed u16.

uint32x2_t	ExperimentalARM
ARM-specific 64-bit wide vector of two packed u32.

uint32x4_t	ExperimentalARM
ARM-specific 128-bit wide vector of four packed u32.

uint64x1_t	ExperimentalARM
ARM-specific 64-bit wide vector of one packed u64.

uint64x2_t	ExperimentalARM
ARM-specific 128-bit wide vector of two packed u64.

uint8x4_t	ExperimentalARM
ARM-specific 32-bit wide vector of four packed u8.

uint8x8_t	ExperimentalARM
ARM-specific 64-bit wide vector of eight packed u8.

uint8x16_t	ExperimentalARM
ARM-specific 128-bit wide vector of sixteen packed u8.

uint8x8x2_t	ExperimentalARM
ARM-specific type containing two uint8x8_t vectors.

uint8x8x3_t	ExperimentalARM
ARM-specific type containing three uint8x8_t vectors.

uint8x8x4_t	ExperimentalARM
ARM-specific type containing four uint8x8_t vectors.

Functions

__DMB⚠	ExperimentalARM and mclass
Data Memory Barrier

__DSB⚠	ExperimentalARM and mclass
Data Synchronization Barrier

__ISB⚠	ExperimentalARM and mclass
Instruction Synchronization Barrier

__NOP⚠	ExperimentalARM and mclass
No Operation

__SEV⚠	ExperimentalARM and mclass
Send Event

__WFE⚠	ExperimentalARM and mclass
Wait For Event

__WFI⚠	ExperimentalARM and mclass
Wait For Interrupt

__breakpoint⚠	ExperimentalARM
Inserts a breakpoint instruction.

__disable_fault_irq⚠	ExperimentalARM and mclass
Disable FIQ

__disable_irq⚠	ExperimentalARM and mclass
Disable IRQ Interrupts

__enable_fault_irq⚠	ExperimentalARM and mclass
Enable FIQ

__enable_irq⚠	ExperimentalARM and mclass
Enable IRQ Interrupts

__get_APSR⚠	ExperimentalARM and mclass
Get APSR Register

__get_BASEPRI⚠	ExperimentalARM and mclass
Get Base Priority

__get_CONTROL⚠	ExperimentalARM and mclass
Get Control Register

__get_FAULTMASK⚠	ExperimentalARM and mclass
Get Fault Mask

__get_IPSR⚠	ExperimentalARM and mclass
Get IPSR Register

__get_MSP⚠	ExperimentalARM and mclass
Get Main Stack Pointer

__get_PRIMASK⚠	ExperimentalARM and mclass
Get Priority Mask

__get_PSP⚠	ExperimentalARM and mclass
Get Process Stack Pointer

__get_xPSR⚠	ExperimentalARM and mclass
Get xPSR Register

__set_BASEPRI⚠	ExperimentalARM and mclass
Set Base Priority

__set_BASEPRI_MAX⚠	ExperimentalARM and mclass
Set Base Priority with condition

__set_CONTROL⚠	ExperimentalARM and mclass
Set Control Register

__set_FAULTMASK⚠	ExperimentalARM and mclass
Set Fault Mask

__set_MSP⚠	ExperimentalARM and mclass
Set Main Stack Pointer

__set_PRIMASK⚠	ExperimentalARM and mclass
Set Priority Mask

__set_PSP⚠	ExperimentalARM and mclass
Set Process Stack Pointer

_rev_u16⚠	ExperimentalARM
Reverse the order of the bytes.

_rev_u32⚠	ExperimentalARM
Reverse the order of the bytes.

qadd⚠	ExperimentalARM
Signed saturating addition

qadd8⚠	ExperimentalARM
Saturating four 8-bit integer additions

qadd16⚠	ExperimentalARM
Saturating two 16-bit integer additions

qasx⚠	ExperimentalARM
Returns the 16-bit signed saturated equivalent of

qsax⚠	ExperimentalARM
Returns the 16-bit signed saturated equivalent of

qsub⚠	ExperimentalARM
Signed saturating subtraction

qsub8⚠	ExperimentalARM
Saturating two 8-bit integer subtraction

qsub16⚠	ExperimentalARM
Saturating two 16-bit integer subtraction

sadd8⚠	ExperimentalARM
Returns the 8-bit signed saturated equivalent of

sadd16⚠	ExperimentalARM
Returns the 16-bit signed saturated equivalent of

sasx⚠	ExperimentalARM
Returns the 16-bit signed equivalent of

sel⚠	ExperimentalARM
Select bytes from each operand according to APSR GE flags

shadd8⚠	ExperimentalARM
Signed halving parallel byte-wise addition.

shadd16⚠	ExperimentalARM
Signed halving parallel halfword-wise addition.

shsub8⚠	ExperimentalARM
Signed halving parallel byte-wise subtraction.

shsub16⚠	ExperimentalARM
Signed halving parallel halfword-wise subtraction.

smlad⚠	ExperimentalARM
Dual 16-bit Signed Multiply with Addition of products and 32-bit accumulation.

smlsd⚠	ExperimentalARM
Dual 16-bit Signed Multiply with Subtraction of products and 32-bit accumulation and overflow detection.

smuad⚠	ExperimentalARM
Signed Dual Multiply Add.

smuadx⚠	ExperimentalARM
Signed Dual Multiply Add Reversed.

smusd⚠	ExperimentalARM
Signed Dual Multiply Subtract.

smusdx⚠	ExperimentalARM
Signed Dual Multiply Subtract Reversed.

usad8⚠	ExperimentalARM
Sum of 8-bit absolute differences.

usad8a⚠	ExperimentalARM
Sum of 8-bit absolute differences and constant.

vadd_f32⚠	ExperimentalARM and neon
Vector add.

vadd_s8⚠	ExperimentalARM and neon
Vector add.

vadd_s16⚠	ExperimentalARM and neon
Vector add.

vadd_s32⚠	ExperimentalARM and neon
Vector add.

vadd_u8⚠	ExperimentalARM and neon
Vector add.

vadd_u16⚠	ExperimentalARM and neon
Vector add.

vadd_u32⚠	ExperimentalARM and neon
Vector add.

vaddl_s8⚠	ExperimentalARM and neon
Vector long add.

vaddl_s16⚠	ExperimentalARM and neon
Vector long add.

vaddl_s32⚠	ExperimentalARM and neon
Vector long add.

vaddl_u8⚠	ExperimentalARM and neon
Vector long add.

vaddl_u16⚠	ExperimentalARM and neon
Vector long add.

vaddl_u32⚠	ExperimentalARM and neon
Vector long add.

vaddq_f32⚠	ExperimentalARM and neon
Vector add.

vaddq_s8⚠	ExperimentalARM and neon
Vector add.

vaddq_s16⚠	ExperimentalARM and neon
Vector add.

vaddq_s32⚠	ExperimentalARM and neon
Vector add.

vaddq_s64⚠	ExperimentalARM and neon
Vector add.

vaddq_u8⚠	ExperimentalARM and neon
Vector add.

vaddq_u16⚠	ExperimentalARM and neon
Vector add.

vaddq_u32⚠	ExperimentalARM and neon
Vector add.

vaddq_u64⚠	ExperimentalARM and neon
Vector add.

vmovl_s8⚠	ExperimentalARM and neon
Vector long move.

vmovl_s16⚠	ExperimentalARM and neon
Vector long move.

vmovl_s32⚠	ExperimentalARM and neon
Vector long move.

vmovl_u8⚠	ExperimentalARM and neon
Vector long move.

vmovl_u16⚠	ExperimentalARM and neon
Vector long move.

vmovl_u32⚠	ExperimentalARM and neon
Vector long move.

vmovn_s16⚠	ExperimentalARM and neon
Vector narrow integer.

vmovn_s32⚠	ExperimentalARM and neon
Vector narrow integer.

vmovn_s64⚠	ExperimentalARM and neon
Vector narrow integer.

vmovn_u16⚠	ExperimentalARM and neon
Vector narrow integer.

vmovn_u32⚠	ExperimentalARM and neon
Vector narrow integer.

vmovn_u64⚠	ExperimentalARM and neon
Vector narrow integer.

vpmax_f32⚠	ExperimentalARM and neon
Folding maximum of adjacent pairs

vpmax_s8⚠	ExperimentalARM and neon
Folding maximum of adjacent pairs

vpmax_s16⚠	ExperimentalARM and neon
Folding maximum of adjacent pairs

vpmax_s32⚠	ExperimentalARM and neon
Folding maximum of adjacent pairs

vpmax_u8⚠	ExperimentalARM and neon
Folding maximum of adjacent pairs

vpmax_u16⚠	ExperimentalARM and neon
Folding maximum of adjacent pairs

vpmax_u32⚠	ExperimentalARM and neon
Folding maximum of adjacent pairs

vpmin_f32⚠	ExperimentalARM and neon
Folding minimum of adjacent pairs

vpmin_s8⚠	ExperimentalARM and neon
Folding minimum of adjacent pairs

vpmin_s16⚠	ExperimentalARM and neon
Folding minimum of adjacent pairs

vpmin_s32⚠	ExperimentalARM and neon
Folding minimum of adjacent pairs

vpmin_u8⚠	ExperimentalARM and neon
Folding minimum of adjacent pairs

vpmin_u16⚠	ExperimentalARM and neon
Folding minimum of adjacent pairs

vpmin_u32⚠	ExperimentalARM and neon
Folding minimum of adjacent pairs

vrsqrte_f32⚠	ExperimentalARM and neon
Reciprocal square-root estimate.