Function core::arch::arm::__dsb

pub unsafe fn __dsb<A>(arg: A) where
    A: Dsb, 

🔬 This is a nightly-only experimental API. (stdsimd #48556)

This is supported on ARM only.

Generates a DSB (data synchronization barrier) instruction or equivalent CP15 instruction.

DSB ensures the completion of memory accesses. A DSB behaves as the equivalent DMB and has additional properties. After a DSB instruction completes, all memory accesses of the specified type issued before the DSB are guaranteed to have completed.

The __dsb() intrinsic also acts as a compiler memory barrier of the appropriate type.