core::arch::hexagon

Module v64

🔬This is a nightly-only experimental API. (stdarch_hexagon #151523)

Available on Hexagon only.

HVX intrinsics for 64-byte vector mode (512-bit vectors) Hexagon HVX 64-byte vector mode intrinsics

This module provides intrinsics for the Hexagon Vector Extensions (HVX) in 64-byte vector mode (512-bit vectors).

HVX is a wide vector extension designed for high-performance signal processing. Hexagon HVX Programmer’s Reference Manual

Vector Types

In 64-byte mode:

HvxVector is 512 bits (64 bytes) containing 16 x 32-bit values

HvxVectorPair is 1024 bits (128 bytes)

HvxVectorPred is 512 bits (64 bytes) for predicate operations

To use this module, compile with -C target-feature=+hvx-length64b.

Architecture Versions

Different intrinsics require different HVX architecture versions. Use the appropriate target feature to enable the required version:

HVX v60: -C target-feature=+hvxv60 (basic HVX operations)

HVX v62: -C target-feature=+hvxv62

HVX v65: -C target-feature=+hvxv65 (includes floating-point support)

HVX v66: -C target-feature=+hvxv66

HVX v68: -C target-feature=+hvxv68

HVX v69: -C target-feature=+hvxv69

HVX v73: -C target-feature=+hvxv73

HVX v79: -C target-feature=+hvxv79

Each version includes all features from previous versions.

Structs

HvxVectorExperimental

HVX vector type (512 bits / 64 bytes)

HvxVectorPairExperimental

HVX vector pair type (1024 bits / 128 bytes)

HvxVectorPredExperimental

HVX vector predicate type (512 bits / 64 bytes)

Functions

q6_q_and_qq⚠Experimental

Qd4=and(Qs4,Qt4)

q6_q_and_qqn⚠Experimental

Qd4=and(Qs4,!Qt4)

q6_q_not_q⚠Experimental

Qd4=not(Qs4)

q6_q_or_qq⚠Experimental

Qd4=or(Qs4,Qt4)

q6_q_or_qqn⚠Experimental

Qd4=or(Qs4,!Qt4)

q6_q_vand_vr⚠Experimental

Qd4=vand(Vu32,Rt32)

q6_q_vandor_qvr⚠Experimental

Qx4|=vand(Vu32,Rt32)

q6_q_vcmp_eq_vbvb⚠Experimental

Qd4=vcmp.eq(Vu32.b,Vv32.b)

q6_q_vcmp_eq_vhvh⚠Experimental

Qd4=vcmp.eq(Vu32.h,Vv32.h)

q6_q_vcmp_eq_vwvw⚠Experimental

Qd4=vcmp.eq(Vu32.w,Vv32.w)

q6_q_vcmp_eqand_qvbvb⚠Experimental

Qx4&=vcmp.eq(Vu32.b,Vv32.b)

q6_q_vcmp_eqand_qvhvh⚠Experimental

Qx4&=vcmp.eq(Vu32.h,Vv32.h)

q6_q_vcmp_eqand_qvwvw⚠Experimental

Qx4&=vcmp.eq(Vu32.w,Vv32.w)

q6_q_vcmp_eqor_qvbvb⚠Experimental

Qx4|=vcmp.eq(Vu32.b,Vv32.b)

q6_q_vcmp_eqor_qvhvh⚠Experimental

Qx4|=vcmp.eq(Vu32.h,Vv32.h)

q6_q_vcmp_eqor_qvwvw⚠Experimental

Qx4|=vcmp.eq(Vu32.w,Vv32.w)

q6_q_vcmp_eqxacc_qvbvb⚠Experimental

Qx4^=vcmp.eq(Vu32.b,Vv32.b)

q6_q_vcmp_eqxacc_qvhvh⚠Experimental

Qx4^=vcmp.eq(Vu32.h,Vv32.h)

q6_q_vcmp_eqxacc_qvwvw⚠Experimental

Qx4^=vcmp.eq(Vu32.w,Vv32.w)

q6_q_vcmp_gt_vbvb⚠Experimental

Qd4=vcmp.gt(Vu32.b,Vv32.b)

q6_q_vcmp_gt_vhfvhf⚠Experimental

Qd4=vcmp.gt(Vu32.hf,Vv32.hf)

q6_q_vcmp_gt_vhvh⚠Experimental

Qd4=vcmp.gt(Vu32.h,Vv32.h)

q6_q_vcmp_gt_vsfvsf⚠Experimental

Qd4=vcmp.gt(Vu32.sf,Vv32.sf)

q6_q_vcmp_gt_vubvub⚠Experimental

Qd4=vcmp.gt(Vu32.ub,Vv32.ub)

q6_q_vcmp_gt_vuhvuh⚠Experimental

Qd4=vcmp.gt(Vu32.uh,Vv32.uh)

q6_q_vcmp_gt_vuwvuw⚠Experimental

Qd4=vcmp.gt(Vu32.uw,Vv32.uw)

q6_q_vcmp_gt_vwvw⚠Experimental

Qd4=vcmp.gt(Vu32.w,Vv32.w)

q6_q_vcmp_gtand_qvbvb⚠Experimental

Qx4&=vcmp.gt(Vu32.b,Vv32.b)

q6_q_vcmp_gtand_qvhfvhf⚠Experimental

Qx4&=vcmp.gt(Vu32.hf,Vv32.hf)

q6_q_vcmp_gtand_qvhvh⚠Experimental

Qx4&=vcmp.gt(Vu32.h,Vv32.h)

q6_q_vcmp_gtand_qvsfvsf⚠Experimental

Qx4&=vcmp.gt(Vu32.sf,Vv32.sf)

q6_q_vcmp_gtand_qvubvub⚠Experimental

Qx4&=vcmp.gt(Vu32.ub,Vv32.ub)

q6_q_vcmp_gtand_qvuhvuh⚠Experimental

Qx4&=vcmp.gt(Vu32.uh,Vv32.uh)

q6_q_vcmp_gtand_qvuwvuw⚠Experimental

Qx4&=vcmp.gt(Vu32.uw,Vv32.uw)

q6_q_vcmp_gtand_qvwvw⚠Experimental

Qx4&=vcmp.gt(Vu32.w,Vv32.w)

q6_q_vcmp_gtor_qvbvb⚠Experimental

Qx4|=vcmp.gt(Vu32.b,Vv32.b)

q6_q_vcmp_gtor_qvhfvhf⚠Experimental

Qx4|=vcmp.gt(Vu32.hf,Vv32.hf)

q6_q_vcmp_gtor_qvhvh⚠Experimental

Qx4|=vcmp.gt(Vu32.h,Vv32.h)

q6_q_vcmp_gtor_qvsfvsf⚠Experimental

Qx4|=vcmp.gt(Vu32.sf,Vv32.sf)

q6_q_vcmp_gtor_qvubvub⚠Experimental

Qx4|=vcmp.gt(Vu32.ub,Vv32.ub)

q6_q_vcmp_gtor_qvuhvuh⚠Experimental

Qx4|=vcmp.gt(Vu32.uh,Vv32.uh)

q6_q_vcmp_gtor_qvuwvuw⚠Experimental

Qx4|=vcmp.gt(Vu32.uw,Vv32.uw)

q6_q_vcmp_gtor_qvwvw⚠Experimental

Qx4|=vcmp.gt(Vu32.w,Vv32.w)

q6_q_vcmp_gtxacc_qvbvb⚠Experimental

Qx4^=vcmp.gt(Vu32.b,Vv32.b)

q6_q_vcmp_gtxacc_qvhfvhf⚠Experimental

Qx4^=vcmp.gt(Vu32.hf,Vv32.hf)

q6_q_vcmp_gtxacc_qvhvh⚠Experimental

Qx4^=vcmp.gt(Vu32.h,Vv32.h)

q6_q_vcmp_gtxacc_qvsfvsf⚠Experimental

Qx4^=vcmp.gt(Vu32.sf,Vv32.sf)

q6_q_vcmp_gtxacc_qvubvub⚠Experimental

Qx4^=vcmp.gt(Vu32.ub,Vv32.ub)

q6_q_vcmp_gtxacc_qvuhvuh⚠Experimental

Qx4^=vcmp.gt(Vu32.uh,Vv32.uh)

q6_q_vcmp_gtxacc_qvuwvuw⚠Experimental

Qx4^=vcmp.gt(Vu32.uw,Vv32.uw)

q6_q_vcmp_gtxacc_qvwvw⚠Experimental

Qx4^=vcmp.gt(Vu32.w,Vv32.w)

q6_q_vsetq2_r⚠Experimental

Qd4=vsetq2(Rt32)

q6_q_vsetq_r⚠Experimental

Qd4=vsetq(Rt32)

q6_q_xor_qq⚠Experimental

Qd4=xor(Qs4,Qt4)

q6_qb_vshuffe_qhqh⚠Experimental

Qd4.b=vshuffe(Qs4.h,Qt4.h)

q6_qh_vshuffe_qwqw⚠Experimental

Qd4.h=vshuffe(Qs4.w,Qt4.w)

q6_r_vextract_vr⚠Experimental

Rd32=vextract(Vu32,Rs32)

q6_v_equals_v⚠Experimental

Vd32=Vu32

q6_v_hi_w⚠Experimental

Vd32=hi(Vss32)

q6_v_lo_w⚠Experimental

Vd32=lo(Vss32)

q6_v_vabs_v⚠Experimental

Vd32.f8=vabs(Vu32.f8)

q6_v_valign_vvi⚠Experimental

Vd32=valign(Vu32,Vv32,#u3)

q6_v_valign_vvr⚠Experimental

Vd32=valign(Vu32,Vv32,Rt8)

q6_v_vand_qnr⚠Experimental

Vd32=vand(!Qu4,Rt32)

q6_v_vand_qnv⚠Experimental

Vd32=vand(!Qv4,Vu32)

q6_v_vand_qr⚠Experimental

Vd32=vand(Qu4,Rt32)

q6_v_vand_qv⚠Experimental

Vd32=vand(Qv4,Vu32)

q6_v_vand_vv⚠Experimental

Vd32=vand(Vu32,Vv32)

q6_v_vandor_vqnr⚠Experimental

Vx32|=vand(!Qu4,Rt32)

q6_v_vandor_vqr⚠Experimental

Vx32|=vand(Qu4,Rt32)

q6_v_vdelta_vv⚠Experimental

Vd32=vdelta(Vu32,Vv32)

q6_v_vfmax_vv⚠Experimental

Vd32.f8=vfmax(Vu32.f8,Vv32.f8)

q6_v_vfmin_vv⚠Experimental

Vd32.f8=vfmin(Vu32.f8,Vv32.f8)

q6_v_vfneg_v⚠Experimental

Vd32.f8=vfneg(Vu32.f8)

q6_v_vgetqfext_vr⚠Experimental

Vd32=vgetqfext(Vu32.x,Rt32)

q6_v_vlalign_vvi⚠Experimental

Vd32=vlalign(Vu32,Vv32,#u3)

q6_v_vlalign_vvr⚠Experimental

Vd32=vlalign(Vu32,Vv32,Rt8)

q6_v_vmux_qvv⚠Experimental

Vd32=vmux(Qt4,Vu32,Vv32)

q6_v_vnot_v⚠Experimental

Vd32=vnot(Vu32)

q6_v_vor_vv⚠Experimental

Vd32=vor(Vu32,Vv32)

q6_v_vrdelta_vv⚠Experimental

Vd32=vrdelta(Vu32,Vv32)

q6_v_vror_vr⚠Experimental

Vd32=vror(Vu32,Rt32)

q6_v_vsetqfext_vr⚠Experimental

Vd32.x=vsetqfext(Vu32,Rt32)

q6_v_vsplat_r⚠Experimental

Vd32=vsplat(Rt32)

q6_v_vxor_vv⚠Experimental

Vd32=vxor(Vu32,Vv32)

q6_v_vzero⚠Experimental

Vd32=#0

q6_vb_condacc_qnvbvb⚠Experimental

if (!Qv4) Vx32.b+=Vu32.b

q6_vb_condacc_qvbvb⚠Experimental

if (Qv4) Vx32.b+=Vu32.b

q6_vb_condnac_qnvbvb⚠Experimental

if (!Qv4) Vx32.b-=Vu32.b

q6_vb_condnac_qvbvb⚠Experimental

if (Qv4) Vx32.b-=Vu32.b

q6_vb_prefixsum_q⚠Experimental

Vd32.b=prefixsum(Qv4)

q6_vb_vabs_vb⚠Experimental

Vd32.b=vabs(Vu32.b)

q6_vb_vabs_vb_sat⚠Experimental

Vd32.b=vabs(Vu32.b):sat

q6_vb_vadd_vbvb⚠Experimental

Vd32.b=vadd(Vu32.b,Vv32.b)

q6_vb_vadd_vbvb_sat⚠Experimental

Vd32.b=vadd(Vu32.b,Vv32.b):sat

q6_vb_vasr_vhvhr_rnd_sat⚠Experimental

Vd32.b=vasr(Vu32.h,Vv32.h,Rt8):rnd:sat

q6_vb_vasr_vhvhr_sat⚠Experimental

Vd32.b=vasr(Vu32.h,Vv32.h,Rt8):sat

q6_vb_vavg_vbvb⚠Experimental

Vd32.b=vavg(Vu32.b,Vv32.b)

q6_vb_vavg_vbvb_rnd⚠Experimental

Vd32.b=vavg(Vu32.b,Vv32.b):rnd

q6_vb_vcvt_vhfvhf⚠Experimental

Vd32.b=vcvt(Vu32.hf,Vv32.hf)

q6_vb_vdeal_vb⚠Experimental

Vd32.b=vdeal(Vu32.b)

q6_vb_vdeale_vbvb⚠Experimental

Vd32.b=vdeale(Vu32.b,Vv32.b)

q6_vb_vlut32_vbvbi⚠Experimental

Vd32.b=vlut32(Vu32.b,Vv32.b,#u3)

q6_vb_vlut32_vbvbr⚠Experimental

Vd32.b=vlut32(Vu32.b,Vv32.b,Rt8)

q6_vb_vlut32_vbvbr_nomatch⚠Experimental

Vd32.b=vlut32(Vu32.b,Vv32.b,Rt8):nomatch

q6_vb_vlut32or_vbvbvbi⚠Experimental

Vx32.b|=vlut32(Vu32.b,Vv32.b,#u3)

q6_vb_vlut32or_vbvbvbr⚠Experimental

Vx32.b|=vlut32(Vu32.b,Vv32.b,Rt8)

q6_vb_vmax_vbvb⚠Experimental

Vd32.b=vmax(Vu32.b,Vv32.b)

q6_vb_vmin_vbvb⚠Experimental

Vd32.b=vmin(Vu32.b,Vv32.b)

q6_vb_vnavg_vbvb⚠Experimental

Vd32.b=vnavg(Vu32.b,Vv32.b)

q6_vb_vnavg_vubvub⚠Experimental

Vd32.b=vnavg(Vu32.ub,Vv32.ub)

q6_vb_vpack_vhvh_sat⚠Experimental

Vd32.b=vpack(Vu32.h,Vv32.h):sat

q6_vb_vpacke_vhvh⚠Experimental

Vd32.b=vpacke(Vu32.h,Vv32.h)

q6_vb_vpacko_vhvh⚠Experimental

Vd32.b=vpacko(Vu32.h,Vv32.h)

q6_vb_vround_vhvh_sat⚠Experimental

Vd32.b=vround(Vu32.h,Vv32.h):sat

q6_vb_vshuff_vb⚠Experimental

Vd32.b=vshuff(Vu32.b)

q6_vb_vshuffe_vbvb⚠Experimental

Vd32.b=vshuffe(Vu32.b,Vv32.b)

q6_vb_vshuffo_vbvb⚠Experimental

Vd32.b=vshuffo(Vu32.b,Vv32.b)

q6_vb_vsplat_r⚠Experimental

Vd32.b=vsplat(Rt32)

q6_vb_vsub_vbvb⚠Experimental

Vd32.b=vsub(Vu32.b,Vv32.b)

q6_vb_vsub_vbvb_sat⚠Experimental

Vd32.b=vsub(Vu32.b,Vv32.b):sat

q6_vgather_aqrmvh⚠Experimental

if (Qs4) vtmp.h=vgather(Rt32,Mu2,Vv32.h).h

q6_vgather_aqrmvw⚠Experimental

if (Qs4) vtmp.w=vgather(Rt32,Mu2,Vv32.w).w

q6_vgather_aqrmww⚠Experimental

if (Qs4) vtmp.h=vgather(Rt32,Mu2,Vvv32.w).h

q6_vgather_armvh⚠Experimental

vtmp.h=vgather(Rt32,Mu2,Vv32.h).h

q6_vgather_armvw⚠Experimental

vtmp.w=vgather(Rt32,Mu2,Vv32.w).w

q6_vgather_armww⚠Experimental

vtmp.h=vgather(Rt32,Mu2,Vvv32.w).h

q6_vh_condacc_qnvhvh⚠Experimental

if (!Qv4) Vx32.h+=Vu32.h

q6_vh_condacc_qvhvh⚠Experimental

if (Qv4) Vx32.h+=Vu32.h

q6_vh_condnac_qnvhvh⚠Experimental

if (!Qv4) Vx32.h-=Vu32.h

q6_vh_condnac_qvhvh⚠Experimental

if (Qv4) Vx32.h-=Vu32.h

q6_vh_equals_vhf⚠Experimental

Vd32.h=Vu32.hf

q6_vh_prefixsum_q⚠Experimental

Vd32.h=prefixsum(Qv4)

q6_vh_vabs_vh⚠Experimental

Vd32.h=vabs(Vu32.h)

q6_vh_vabs_vh_sat⚠Experimental

Vd32.h=vabs(Vu32.h):sat

q6_vh_vadd_vclb_vhvh⚠Experimental

Vd32.h=vadd(vclb(Vu32.h),Vv32.h)

q6_vh_vadd_vhvh⚠Experimental

Vd32.h=vadd(Vu32.h,Vv32.h)

q6_vh_vadd_vhvh_sat⚠Experimental

Vd32.h=vadd(Vu32.h,Vv32.h):sat

q6_vh_vasl_vhr⚠Experimental

Vd32.h=vasl(Vu32.h,Rt32)

q6_vh_vasl_vhvh⚠Experimental

Vd32.h=vasl(Vu32.h,Vv32.h)

q6_vh_vaslacc_vhvhr⚠Experimental

Vx32.h+=vasl(Vu32.h,Rt32)

q6_vh_vasr_vhr⚠Experimental

Vd32.h=vasr(Vu32.h,Rt32)

q6_vh_vasr_vhvh⚠Experimental

Vd32.h=vasr(Vu32.h,Vv32.h)

q6_vh_vasr_vwvwr⚠Experimental

Vd32.h=vasr(Vu32.w,Vv32.w,Rt8)

q6_vh_vasr_vwvwr_rnd_sat⚠Experimental

Vd32.h=vasr(Vu32.w,Vv32.w,Rt8):rnd:sat

q6_vh_vasr_vwvwr_sat⚠Experimental

Vd32.h=vasr(Vu32.w,Vv32.w,Rt8):sat

q6_vh_vasracc_vhvhr⚠Experimental

Vx32.h+=vasr(Vu32.h,Rt32)

q6_vh_vavg_vhvh⚠Experimental

Vd32.h=vavg(Vu32.h,Vv32.h)

q6_vh_vavg_vhvh_rnd⚠Experimental

Vd32.h=vavg(Vu32.h,Vv32.h):rnd

q6_vh_vcvt_vhf⚠Experimental

Vd32.h=vcvt(Vu32.hf)

q6_vh_vdeal_vh⚠Experimental

Vd32.h=vdeal(Vu32.h)

q6_vh_vdmpy_vubrb⚠Experimental

Vd32.h=vdmpy(Vu32.ub,Rt32.b)

q6_vh_vdmpyacc_vhvubrb⚠Experimental

Vx32.h+=vdmpy(Vu32.ub,Rt32.b)

q6_vh_vlsr_vhvh⚠Experimental

Vd32.h=vlsr(Vu32.h,Vv32.h)

q6_vh_vmax_vhvh⚠Experimental

Vd32.h=vmax(Vu32.h,Vv32.h)

q6_vh_vmin_vhvh⚠Experimental

Vd32.h=vmin(Vu32.h,Vv32.h)

q6_vh_vmpy_vhrh_s1_rnd_sat⚠Experimental

Vd32.h=vmpy(Vu32.h,Rt32.h):<<1:rnd:sat

q6_vh_vmpy_vhrh_s1_sat⚠Experimental

Vd32.h=vmpy(Vu32.h,Rt32.h):<<1:sat

q6_vh_vmpy_vhvh_s1_rnd_sat⚠Experimental

Vd32.h=vmpy(Vu32.h,Vv32.h):<<1:rnd:sat

q6_vh_vmpyi_vhrb⚠Experimental

Vd32.h=vmpyi(Vu32.h,Rt32.b)

q6_vh_vmpyi_vhvh⚠Experimental

Vd32.h=vmpyi(Vu32.h,Vv32.h)

q6_vh_vmpyiacc_vhvhrb⚠Experimental

Vx32.h+=vmpyi(Vu32.h,Rt32.b)

q6_vh_vmpyiacc_vhvhvh⚠Experimental

Vx32.h+=vmpyi(Vu32.h,Vv32.h)

q6_vh_vnavg_vhvh⚠Experimental

Vd32.h=vnavg(Vu32.h,Vv32.h)

q6_vh_vnormamt_vh⚠Experimental

Vd32.h=vnormamt(Vu32.h)

q6_vh_vpack_vwvw_sat⚠Experimental

Vd32.h=vpack(Vu32.w,Vv32.w):sat

q6_vh_vpacke_vwvw⚠Experimental

Vd32.h=vpacke(Vu32.w,Vv32.w)

q6_vh_vpacko_vwvw⚠Experimental

Vd32.h=vpacko(Vu32.w,Vv32.w)

q6_vh_vpopcount_vh⚠Experimental

Vd32.h=vpopcount(Vu32.h)

q6_vh_vround_vwvw_sat⚠Experimental

Vd32.h=vround(Vu32.w,Vv32.w):sat

q6_vh_vsat_vwvw⚠Experimental

Vd32.h=vsat(Vu32.w,Vv32.w)

q6_vh_vshuff_vh⚠Experimental

Vd32.h=vshuff(Vu32.h)

q6_vh_vshuffe_vhvh⚠Experimental

Vd32.h=vshuffe(Vu32.h,Vv32.h)

q6_vh_vshuffo_vhvh⚠Experimental

Vd32.h=vshuffo(Vu32.h,Vv32.h)

q6_vh_vsplat_r⚠Experimental

Vd32.h=vsplat(Rt32)

q6_vh_vsub_vhvh⚠Experimental

Vd32.h=vsub(Vu32.h,Vv32.h)

q6_vh_vsub_vhvh_sat⚠Experimental

Vd32.h=vsub(Vu32.h,Vv32.h):sat

q6_vhf_equals_vh⚠Experimental

Vd32.hf=Vu32.h

q6_vhf_equals_vqf16⚠Experimental

Vd32.hf=Vu32.qf16

q6_vhf_equals_wqf32⚠Experimental

Vd32.hf=Vuu32.qf32

q6_vhf_vabs_vhf⚠Experimental

Vd32.hf=vabs(Vu32.hf)

q6_vhf_vadd_vhfvhf⚠Experimental

Vd32.hf=vadd(Vu32.hf,Vv32.hf)

q6_vhf_vcvt_vh⚠Experimental

Vd32.hf=vcvt(Vu32.h)

q6_vhf_vcvt_vsfvsf⚠Experimental

Vd32.hf=vcvt(Vu32.sf,Vv32.sf)

q6_vhf_vcvt_vuh⚠Experimental

Vd32.hf=vcvt(Vu32.uh)

q6_vhf_vfmax_vhfvhf⚠Experimental

Vd32.hf=vfmax(Vu32.hf,Vv32.hf)

q6_vhf_vfmin_vhfvhf⚠Experimental

Vd32.hf=vfmin(Vu32.hf,Vv32.hf)

q6_vhf_vfneg_vhf⚠Experimental

Vd32.hf=vfneg(Vu32.hf)

q6_vhf_vmax_vhfvhf⚠Experimental

Vd32.hf=vmax(Vu32.hf,Vv32.hf)

q6_vhf_vmin_vhfvhf⚠Experimental

Vd32.hf=vmin(Vu32.hf,Vv32.hf)

q6_vhf_vmpy_vhfvhf⚠Experimental

Vd32.hf=vmpy(Vu32.hf,Vv32.hf)

q6_vhf_vmpyacc_vhfvhfvhf⚠Experimental

Vx32.hf+=vmpy(Vu32.hf,Vv32.hf)

q6_vhf_vsub_vhfvhf⚠Experimental

Vd32.hf=vsub(Vu32.hf,Vv32.hf)

q6_vmem_qnriv⚠Experimental

if (!Qv4) vmem(Rt32+#s4)=Vs32

q6_vmem_qnriv_nt⚠Experimental

if (!Qv4) vmem(Rt32+#s4):nt=Vs32

q6_vmem_qriv⚠Experimental

if (Qv4) vmem(Rt32+#s4)=Vs32

q6_vmem_qriv_nt⚠Experimental

if (Qv4) vmem(Rt32+#s4):nt=Vs32

q6_vqf16_vadd_vhfvhf⚠Experimental

Vd32.qf16=vadd(Vu32.hf,Vv32.hf)

q6_vqf16_vadd_vqf16vhf⚠Experimental

Vd32.qf16=vadd(Vu32.qf16,Vv32.hf)

q6_vqf16_vadd_vqf16vqf16⚠Experimental

Vd32.qf16=vadd(Vu32.qf16,Vv32.qf16)

q6_vqf16_vmpy_vhfvhf⚠Experimental

Vd32.qf16=vmpy(Vu32.hf,Vv32.hf)

q6_vqf16_vmpy_vqf16vhf⚠Experimental

Vd32.qf16=vmpy(Vu32.qf16,Vv32.hf)

q6_vqf16_vmpy_vqf16vqf16⚠Experimental

Vd32.qf16=vmpy(Vu32.qf16,Vv32.qf16)

q6_vqf16_vsub_vhfvhf⚠Experimental

Vd32.qf16=vsub(Vu32.hf,Vv32.hf)

q6_vqf16_vsub_vqf16vhf⚠Experimental

Vd32.qf16=vsub(Vu32.qf16,Vv32.hf)

q6_vqf16_vsub_vqf16vqf16⚠Experimental

Vd32.qf16=vsub(Vu32.qf16,Vv32.qf16)

q6_vqf32_vadd_vqf32vqf32⚠Experimental

Vd32.qf32=vadd(Vu32.qf32,Vv32.qf32)

q6_vqf32_vadd_vqf32vsf⚠Experimental

Vd32.qf32=vadd(Vu32.qf32,Vv32.sf)

q6_vqf32_vadd_vsfvsf⚠Experimental

Vd32.qf32=vadd(Vu32.sf,Vv32.sf)

q6_vqf32_vmpy_vqf32vqf32⚠Experimental

Vd32.qf32=vmpy(Vu32.qf32,Vv32.qf32)

q6_vqf32_vmpy_vsfvsf⚠Experimental

Vd32.qf32=vmpy(Vu32.sf,Vv32.sf)

q6_vqf32_vsub_vqf32vqf32⚠Experimental

Vd32.qf32=vsub(Vu32.qf32,Vv32.qf32)

q6_vqf32_vsub_vqf32vsf⚠Experimental

Vd32.qf32=vsub(Vu32.qf32,Vv32.sf)

q6_vqf32_vsub_vsfvsf⚠Experimental

Vd32.qf32=vsub(Vu32.sf,Vv32.sf)

q6_vscatter_qrmvhv⚠Experimental

if (Qs4) vscatter(Rt32,Mu2,Vv32.h).h=Vw32

q6_vscatter_qrmvwv⚠Experimental

if (Qs4) vscatter(Rt32,Mu2,Vv32.w).w=Vw32

q6_vscatter_qrmwwv⚠Experimental

if (Qs4) vscatter(Rt32,Mu2,Vvv32.w).h=Vw32

q6_vscatter_rmvhv⚠Experimental

vscatter(Rt32,Mu2,Vv32.h).h=Vw32

q6_vscatter_rmvwv⚠Experimental

vscatter(Rt32,Mu2,Vv32.w).w=Vw32

q6_vscatter_rmwwv⚠Experimental

vscatter(Rt32,Mu2,Vvv32.w).h=Vw32

q6_vscatteracc_rmvhv⚠Experimental

vscatter(Rt32,Mu2,Vv32.h).h+=Vw32

q6_vscatteracc_rmvwv⚠Experimental

vscatter(Rt32,Mu2,Vv32.w).w+=Vw32

q6_vscatteracc_rmwwv⚠Experimental

vscatter(Rt32,Mu2,Vvv32.w).h+=Vw32

q6_vsf_equals_vqf32⚠Experimental

Vd32.sf=Vu32.qf32

q6_vsf_equals_vw⚠Experimental

Vd32.sf=Vu32.w

q6_vsf_vabs_vsf⚠Experimental

Vd32.sf=vabs(Vu32.sf)

q6_vsf_vadd_vsfvsf⚠Experimental

Vd32.sf=vadd(Vu32.sf,Vv32.sf)

q6_vsf_vdmpy_vhfvhf⚠Experimental

Vd32.sf=vdmpy(Vu32.hf,Vv32.hf)

q6_vsf_vdmpyacc_vsfvhfvhf⚠Experimental

Vx32.sf+=vdmpy(Vu32.hf,Vv32.hf)

q6_vsf_vfmax_vsfvsf⚠Experimental

Vd32.sf=vfmax(Vu32.sf,Vv32.sf)

q6_vsf_vfmin_vsfvsf⚠Experimental

Vd32.sf=vfmin(Vu32.sf,Vv32.sf)

q6_vsf_vfneg_vsf⚠Experimental

Vd32.sf=vfneg(Vu32.sf)

q6_vsf_vmax_vsfvsf⚠Experimental

Vd32.sf=vmax(Vu32.sf,Vv32.sf)

q6_vsf_vmin_vsfvsf⚠Experimental

Vd32.sf=vmin(Vu32.sf,Vv32.sf)

q6_vsf_vmpy_vsfvsf⚠Experimental

Vd32.sf=vmpy(Vu32.sf,Vv32.sf)

q6_vsf_vsub_vsfvsf⚠Experimental

Vd32.sf=vsub(Vu32.sf,Vv32.sf)

q6_vub_vabsdiff_vubvub⚠Experimental

Vd32.ub=vabsdiff(Vu32.ub,Vv32.ub)

q6_vub_vadd_vubvb_sat⚠Experimental

Vd32.ub=vadd(Vu32.ub,Vv32.b):sat

q6_vub_vadd_vubvub_sat⚠Experimental

Vd32.ub=vadd(Vu32.ub,Vv32.ub):sat

q6_vub_vasr_vhvhr_rnd_sat⚠Experimental

Vd32.ub=vasr(Vu32.h,Vv32.h,Rt8):rnd:sat

q6_vub_vasr_vhvhr_sat⚠Experimental

Vd32.ub=vasr(Vu32.h,Vv32.h,Rt8):sat

q6_vub_vasr_vuhvuhr_rnd_sat⚠Experimental

Vd32.ub=vasr(Vu32.uh,Vv32.uh,Rt8):rnd:sat

q6_vub_vasr_vuhvuhr_sat⚠Experimental

Vd32.ub=vasr(Vu32.uh,Vv32.uh,Rt8):sat

q6_vub_vasr_wuhvub_rnd_sat⚠Experimental

Vd32.ub=vasr(Vuu32.uh,Vv32.ub):rnd:sat

q6_vub_vasr_wuhvub_sat⚠Experimental

Vd32.ub=vasr(Vuu32.uh,Vv32.ub):sat

q6_vub_vavg_vubvub⚠Experimental

Vd32.ub=vavg(Vu32.ub,Vv32.ub)

q6_vub_vavg_vubvub_rnd⚠Experimental

Vd32.ub=vavg(Vu32.ub,Vv32.ub):rnd

q6_vub_vcvt_vhfvhf⚠Experimental

Vd32.ub=vcvt(Vu32.hf,Vv32.hf)

q6_vub_vlsr_vubr⚠Experimental

Vd32.ub=vlsr(Vu32.ub,Rt32)

q6_vub_vmax_vubvub⚠Experimental

Vd32.ub=vmax(Vu32.ub,Vv32.ub)

q6_vub_vmin_vubvub⚠Experimental

Vd32.ub=vmin(Vu32.ub,Vv32.ub)

q6_vub_vpack_vhvh_sat⚠Experimental

Vd32.ub=vpack(Vu32.h,Vv32.h):sat

q6_vub_vround_vhvh_sat⚠Experimental

Vd32.ub=vround(Vu32.h,Vv32.h):sat

q6_vub_vround_vuhvuh_sat⚠Experimental

Vd32.ub=vround(Vu32.uh,Vv32.uh):sat

q6_vub_vsat_vhvh⚠Experimental

Vd32.ub=vsat(Vu32.h,Vv32.h)

q6_vub_vsub_vubvb_sat⚠Experimental

Vd32.ub=vsub(Vu32.ub,Vv32.b):sat

q6_vub_vsub_vubvub_sat⚠Experimental

Vd32.ub=vsub(Vu32.ub,Vv32.ub):sat

q6_vuh_vabsdiff_vhvh⚠Experimental

Vd32.uh=vabsdiff(Vu32.h,Vv32.h)

q6_vuh_vabsdiff_vuhvuh⚠Experimental

Vd32.uh=vabsdiff(Vu32.uh,Vv32.uh)

q6_vuh_vadd_vuhvuh_sat⚠Experimental

Vd32.uh=vadd(Vu32.uh,Vv32.uh):sat

q6_vuh_vasr_vuwvuwr_rnd_sat⚠Experimental

Vd32.uh=vasr(Vu32.uw,Vv32.uw,Rt8):rnd:sat

q6_vuh_vasr_vuwvuwr_sat⚠Experimental

Vd32.uh=vasr(Vu32.uw,Vv32.uw,Rt8):sat

q6_vuh_vasr_vwvwr_rnd_sat⚠Experimental

Vd32.uh=vasr(Vu32.w,Vv32.w,Rt8):rnd:sat

q6_vuh_vasr_vwvwr_sat⚠Experimental

Vd32.uh=vasr(Vu32.w,Vv32.w,Rt8):sat

q6_vuh_vasr_wwvuh_rnd_sat⚠Experimental

Vd32.uh=vasr(Vuu32.w,Vv32.uh):rnd:sat

q6_vuh_vasr_wwvuh_sat⚠Experimental

Vd32.uh=vasr(Vuu32.w,Vv32.uh):sat

q6_vuh_vavg_vuhvuh⚠Experimental

Vd32.uh=vavg(Vu32.uh,Vv32.uh)

q6_vuh_vavg_vuhvuh_rnd⚠Experimental

Vd32.uh=vavg(Vu32.uh,Vv32.uh):rnd

q6_vuh_vcl0_vuh⚠Experimental

Vd32.uh=vcl0(Vu32.uh)

q6_vuh_vcvt_vhf⚠Experimental

Vd32.uh=vcvt(Vu32.hf)

q6_vuh_vlsr_vuhr⚠Experimental

Vd32.uh=vlsr(Vu32.uh,Rt32)

q6_vuh_vmax_vuhvuh⚠Experimental

Vd32.uh=vmax(Vu32.uh,Vv32.uh)

q6_vuh_vmin_vuhvuh⚠Experimental

Vd32.uh=vmin(Vu32.uh,Vv32.uh)

q6_vuh_vmpy_vuhvuh_rs16⚠Experimental

Vd32.uh=vmpy(Vu32.uh,Vv32.uh):>>16

q6_vuh_vpack_vwvw_sat⚠Experimental

Vd32.uh=vpack(Vu32.w,Vv32.w):sat

q6_vuh_vround_vuwvuw_sat⚠Experimental

Vd32.uh=vround(Vu32.uw,Vv32.uw):sat

q6_vuh_vround_vwvw_sat⚠Experimental

Vd32.uh=vround(Vu32.w,Vv32.w):sat

q6_vuh_vsat_vuwvuw⚠Experimental

Vd32.uh=vsat(Vu32.uw,Vv32.uw)

q6_vuh_vsub_vuhvuh_sat⚠Experimental

Vd32.uh=vsub(Vu32.uh,Vv32.uh):sat

q6_vuw_vabsdiff_vwvw⚠Experimental

Vd32.uw=vabsdiff(Vu32.w,Vv32.w)

q6_vuw_vadd_vuwvuw_sat⚠Experimental

Vd32.uw=vadd(Vu32.uw,Vv32.uw):sat

q6_vuw_vavg_vuwvuw⚠Experimental

Vd32.uw=vavg(Vu32.uw,Vv32.uw)

q6_vuw_vavg_vuwvuw_rnd⚠Experimental

Vd32.uw=vavg(Vu32.uw,Vv32.uw):rnd

q6_vuw_vcl0_vuw⚠Experimental

Vd32.uw=vcl0(Vu32.uw)

q6_vuw_vlsr_vuwr⚠Experimental

Vd32.uw=vlsr(Vu32.uw,Rt32)

q6_vuw_vmpye_vuhruh⚠Experimental

Vd32.uw=vmpye(Vu32.uh,Rt32.uh)

q6_vuw_vmpyeacc_vuwvuhruh⚠Experimental

Vx32.uw+=vmpye(Vu32.uh,Rt32.uh)

q6_vuw_vrmpy_vubrub⚠Experimental

Vd32.uw=vrmpy(Vu32.ub,Rt32.ub)

q6_vuw_vrmpy_vubvub⚠Experimental

Vd32.uw=vrmpy(Vu32.ub,Vv32.ub)

q6_vuw_vrmpyacc_vuwvubrub⚠Experimental

Vx32.uw+=vrmpy(Vu32.ub,Rt32.ub)

q6_vuw_vrmpyacc_vuwvubvub⚠Experimental

Vx32.uw+=vrmpy(Vu32.ub,Vv32.ub)

q6_vuw_vrotr_vuwvuw⚠Experimental

Vd32.uw=vrotr(Vu32.uw,Vv32.uw)

q6_vuw_vsub_vuwvuw_sat⚠Experimental

Vd32.uw=vsub(Vu32.uw,Vv32.uw):sat

q6_vw_condacc_qnvwvw⚠Experimental

if (!Qv4) Vx32.w+=Vu32.w

q6_vw_condacc_qvwvw⚠Experimental

if (Qv4) Vx32.w+=Vu32.w

q6_vw_condnac_qnvwvw⚠Experimental

if (!Qv4) Vx32.w-=Vu32.w

q6_vw_condnac_qvwvw⚠Experimental

if (Qv4) Vx32.w-=Vu32.w

q6_vw_equals_vsf⚠Experimental

Vd32.w=Vu32.sf

q6_vw_prefixsum_q⚠Experimental

Vd32.w=prefixsum(Qv4)

q6_vw_vabs_vw⚠Experimental

Vd32.w=vabs(Vu32.w)

q6_vw_vabs_vw_sat⚠Experimental

Vd32.w=vabs(Vu32.w):sat

q6_vw_vadd_vclb_vwvw⚠Experimental

Vd32.w=vadd(vclb(Vu32.w),Vv32.w)

q6_vw_vadd_vwvw⚠Experimental

Vd32.w=vadd(Vu32.w,Vv32.w)

q6_vw_vadd_vwvw_sat⚠Experimental

Vd32.w=vadd(Vu32.w,Vv32.w):sat

q6_vw_vadd_vwvwq_carry_sat⚠Experimental

Vd32.w=vadd(Vu32.w,Vv32.w,Qs4):carry:sat

q6_vw_vasl_vwr⚠Experimental

Vd32.w=vasl(Vu32.w,Rt32)

q6_vw_vasl_vwvw⚠Experimental

Vd32.w=vasl(Vu32.w,Vv32.w)

q6_vw_vaslacc_vwvwr⚠Experimental

Vx32.w+=vasl(Vu32.w,Rt32)

q6_vw_vasr_vwr⚠Experimental

Vd32.w=vasr(Vu32.w,Rt32)

q6_vw_vasr_vwvw⚠Experimental

Vd32.w=vasr(Vu32.w,Vv32.w)

q6_vw_vasracc_vwvwr⚠Experimental

Vx32.w+=vasr(Vu32.w,Rt32)

q6_vw_vavg_vwvw⚠Experimental

Vd32.w=vavg(Vu32.w,Vv32.w)

q6_vw_vavg_vwvw_rnd⚠Experimental

Vd32.w=vavg(Vu32.w,Vv32.w):rnd

q6_vw_vdmpy_vhrb⚠Experimental

Vd32.w=vdmpy(Vu32.h,Rt32.b)

q6_vw_vdmpy_vhrh_sat⚠Experimental

Vd32.w=vdmpy(Vu32.h,Rt32.h):sat

q6_vw_vdmpy_vhruh_sat⚠Experimental

Vd32.w=vdmpy(Vu32.h,Rt32.uh):sat

q6_vw_vdmpy_vhvh_sat⚠Experimental

Vd32.w=vdmpy(Vu32.h,Vv32.h):sat

q6_vw_vdmpy_whrh_sat⚠Experimental

Vd32.w=vdmpy(Vuu32.h,Rt32.h):sat

q6_vw_vdmpy_whruh_sat⚠Experimental

Vd32.w=vdmpy(Vuu32.h,Rt32.uh,#1):sat

q6_vw_vdmpyacc_vwvhrb⚠Experimental

Vx32.w+=vdmpy(Vu32.h,Rt32.b)

q6_vw_vdmpyacc_vwvhrh_sat⚠Experimental

Vx32.w+=vdmpy(Vu32.h,Rt32.h):sat

q6_vw_vdmpyacc_vwvhruh_sat⚠Experimental

Vx32.w+=vdmpy(Vu32.h,Rt32.uh):sat

q6_vw_vdmpyacc_vwvhvh_sat⚠Experimental

Vx32.w+=vdmpy(Vu32.h,Vv32.h):sat

q6_vw_vdmpyacc_vwwhrh_sat⚠Experimental

Vx32.w+=vdmpy(Vuu32.h,Rt32.h):sat

q6_vw_vdmpyacc_vwwhruh_sat⚠Experimental

Vx32.w+=vdmpy(Vuu32.h,Rt32.uh,#1):sat

q6_vw_vfmv_vw⚠Experimental

Vd32.w=vfmv(Vu32.w)

q6_vw_vinsert_vwr⚠Experimental

Vx32.w=vinsert(Rt32)

q6_vw_vlsr_vwvw⚠Experimental

Vd32.w=vlsr(Vu32.w,Vv32.w)

q6_vw_vmax_vwvw⚠Experimental

Vd32.w=vmax(Vu32.w,Vv32.w)

q6_vw_vmin_vwvw⚠Experimental

Vd32.w=vmin(Vu32.w,Vv32.w)

q6_vw_vmpye_vwvuh⚠Experimental

Vd32.w=vmpye(Vu32.w,Vv32.uh)

q6_vw_vmpyi_vwrb⚠Experimental

Vd32.w=vmpyi(Vu32.w,Rt32.b)

q6_vw_vmpyi_vwrh⚠Experimental

Vd32.w=vmpyi(Vu32.w,Rt32.h)

q6_vw_vmpyi_vwrub⚠Experimental

Vd32.w=vmpyi(Vu32.w,Rt32.ub)

q6_vw_vmpyiacc_vwvwrb⚠Experimental

Vx32.w+=vmpyi(Vu32.w,Rt32.b)

q6_vw_vmpyiacc_vwvwrh⚠Experimental

Vx32.w+=vmpyi(Vu32.w,Rt32.h)

q6_vw_vmpyiacc_vwvwrub⚠Experimental

Vx32.w+=vmpyi(Vu32.w,Rt32.ub)

q6_vw_vmpyie_vwvuh⚠Experimental

Vd32.w=vmpyie(Vu32.w,Vv32.uh)

q6_vw_vmpyieacc_vwvwvh⚠Experimental

Vx32.w+=vmpyie(Vu32.w,Vv32.h)

q6_vw_vmpyieacc_vwvwvuh⚠Experimental

Vx32.w+=vmpyie(Vu32.w,Vv32.uh)

q6_vw_vmpyieo_vhvh⚠Experimental

Vd32.w=vmpyieo(Vu32.h,Vv32.h)

q6_vw_vmpyio_vwvh⚠Experimental

Vd32.w=vmpyio(Vu32.w,Vv32.h)

q6_vw_vmpyo_vwvh_s1_rnd_sat⚠Experimental

Vd32.w=vmpyo(Vu32.w,Vv32.h):<<1:rnd:sat

q6_vw_vmpyo_vwvh_s1_sat⚠Experimental

Vd32.w=vmpyo(Vu32.w,Vv32.h):<<1:sat

q6_vw_vmpyoacc_vwvwvh_s1_rnd_sat_shift⚠Experimental

Vx32.w+=vmpyo(Vu32.w,Vv32.h):<<1:rnd:sat:shift

q6_vw_vmpyoacc_vwvwvh_s1_sat_shift⚠Experimental

Vx32.w+=vmpyo(Vu32.w,Vv32.h):<<1:sat:shift

q6_vw_vnavg_vwvw⚠Experimental

Vd32.w=vnavg(Vu32.w,Vv32.w)

q6_vw_vnormamt_vw⚠Experimental

Vd32.w=vnormamt(Vu32.w)

q6_vw_vrmpy_vbvb⚠Experimental

Vd32.w=vrmpy(Vu32.b,Vv32.b)

q6_vw_vrmpy_vubrb⚠Experimental

Vd32.w=vrmpy(Vu32.ub,Rt32.b)

q6_vw_vrmpy_vubvb⚠Experimental

Vd32.w=vrmpy(Vu32.ub,Vv32.b)

q6_vw_vrmpyacc_vwvbvb⚠Experimental

Vx32.w+=vrmpy(Vu32.b,Vv32.b)

q6_vw_vrmpyacc_vwvubrb⚠Experimental

Vx32.w+=vrmpy(Vu32.ub,Rt32.b)

q6_vw_vrmpyacc_vwvubvb⚠Experimental

Vx32.w+=vrmpy(Vu32.ub,Vv32.b)

q6_vw_vsatdw_vwvw⚠Experimental

Vd32.w=vsatdw(Vu32.w,Vv32.w)

q6_vw_vsub_vwvw⚠Experimental

Vd32.w=vsub(Vu32.w,Vv32.w)

q6_vw_vsub_vwvw_sat⚠Experimental

Vd32.w=vsub(Vu32.w,Vv32.w):sat

q6_w_equals_w⚠Experimental

Vdd32=Vuu32

q6_w_vcombine_vv⚠Experimental

Vdd32=vcombine(Vu32,Vv32)

q6_w_vdeal_vvr⚠Experimental

Vdd32=vdeal(Vu32,Vv32,Rt8)

q6_w_vmpye_vwvuh⚠Experimental

Vdd32=vmpye(Vu32.w,Vv32.uh)

q6_w_vmpyoacc_wvwvh⚠Experimental

Vxx32+=vmpyo(Vu32.w,Vv32.h)

q6_w_vshuff_vvr⚠Experimental

Vdd32=vshuff(Vu32,Vv32,Rt8)

q6_w_vswap_qvv⚠Experimental

Vdd32=vswap(Qt4,Vu32,Vv32)

q6_w_vzero⚠Experimental

Vdd32=#0

q6_wb_vadd_wbwb⚠Experimental

Vdd32.b=vadd(Vuu32.b,Vvv32.b)

q6_wb_vadd_wbwb_sat⚠Experimental

Vdd32.b=vadd(Vuu32.b,Vvv32.b):sat

q6_wb_vshuffoe_vbvb⚠Experimental

Vdd32.b=vshuffoe(Vu32.b,Vv32.b)

q6_wb_vsub_wbwb⚠Experimental

Vdd32.b=vsub(Vuu32.b,Vvv32.b)

q6_wb_vsub_wbwb_sat⚠Experimental

Vdd32.b=vsub(Vuu32.b,Vvv32.b):sat

q6_wh_vadd_vubvub⚠Experimental

Vdd32.h=vadd(Vu32.ub,Vv32.ub)

q6_wh_vadd_whwh⚠Experimental

Vdd32.h=vadd(Vuu32.h,Vvv32.h)

q6_wh_vadd_whwh_sat⚠Experimental

Vdd32.h=vadd(Vuu32.h,Vvv32.h):sat

q6_wh_vaddacc_whvubvub⚠Experimental

Vxx32.h+=vadd(Vu32.ub,Vv32.ub)

q6_wh_vdmpy_wubrb⚠Experimental

Vdd32.h=vdmpy(Vuu32.ub,Rt32.b)

q6_wh_vdmpyacc_whwubrb⚠Experimental

Vxx32.h+=vdmpy(Vuu32.ub,Rt32.b)

q6_wh_vlut16_vbvhi⚠Experimental

Vdd32.h=vlut16(Vu32.b,Vv32.h,#u3)

q6_wh_vlut16_vbvhr⚠Experimental

Vdd32.h=vlut16(Vu32.b,Vv32.h,Rt8)

q6_wh_vlut16_vbvhr_nomatch⚠Experimental

Vdd32.h=vlut16(Vu32.b,Vv32.h,Rt8):nomatch

q6_wh_vlut16or_whvbvhi⚠Experimental

Vxx32.h|=vlut16(Vu32.b,Vv32.h,#u3)

q6_wh_vlut16or_whvbvhr⚠Experimental

Vxx32.h|=vlut16(Vu32.b,Vv32.h,Rt8)

q6_wh_vmpa_wubrb⚠Experimental

Vdd32.h=vmpa(Vuu32.ub,Rt32.b)

q6_wh_vmpa_wubrub⚠Experimental

Vdd32.h=vmpa(Vuu32.ub,Rt32.ub)

q6_wh_vmpa_wubwb⚠Experimental

Vdd32.h=vmpa(Vuu32.ub,Vvv32.b)

q6_wh_vmpa_wubwub⚠Experimental

Vdd32.h=vmpa(Vuu32.ub,Vvv32.ub)

q6_wh_vmpaacc_whwubrb⚠Experimental

Vxx32.h+=vmpa(Vuu32.ub,Rt32.b)

q6_wh_vmpaacc_whwubrub⚠Experimental

Vxx32.h+=vmpa(Vuu32.ub,Rt32.ub)

q6_wh_vmpy_vbvb⚠Experimental

Vdd32.h=vmpy(Vu32.b,Vv32.b)

q6_wh_vmpy_vubrb⚠Experimental

Vdd32.h=vmpy(Vu32.ub,Rt32.b)

q6_wh_vmpy_vubvb⚠Experimental

Vdd32.h=vmpy(Vu32.ub,Vv32.b)

q6_wh_vmpyacc_whvbvb⚠Experimental

Vxx32.h+=vmpy(Vu32.b,Vv32.b)

q6_wh_vmpyacc_whvubrb⚠Experimental

Vxx32.h+=vmpy(Vu32.ub,Rt32.b)

q6_wh_vmpyacc_whvubvb⚠Experimental

Vxx32.h+=vmpy(Vu32.ub,Vv32.b)

q6_wh_vshuffoe_vhvh⚠Experimental

Vdd32.h=vshuffoe(Vu32.h,Vv32.h)

q6_wh_vsub_vubvub⚠Experimental

Vdd32.h=vsub(Vu32.ub,Vv32.ub)

q6_wh_vsub_whwh⚠Experimental

Vdd32.h=vsub(Vuu32.h,Vvv32.h)

q6_wh_vsub_whwh_sat⚠Experimental

Vdd32.h=vsub(Vuu32.h,Vvv32.h):sat

q6_wh_vsxt_vb⚠Experimental

Vdd32.h=vsxt(Vu32.b)

q6_wh_vtmpy_wbrb⚠Experimental

Vdd32.h=vtmpy(Vuu32.b,Rt32.b)

q6_wh_vtmpy_wubrb⚠Experimental

Vdd32.h=vtmpy(Vuu32.ub,Rt32.b)

q6_wh_vtmpyacc_whwbrb⚠Experimental

Vxx32.h+=vtmpy(Vuu32.b,Rt32.b)

q6_wh_vtmpyacc_whwubrb⚠Experimental

Vxx32.h+=vtmpy(Vuu32.ub,Rt32.b)

q6_wh_vunpack_vb⚠Experimental

Vdd32.h=vunpack(Vu32.b)

q6_wh_vunpackoor_whvb⚠Experimental

Vxx32.h|=vunpacko(Vu32.b)

q6_whf_vcvt2_vb⚠Experimental

Vdd32.hf=vcvt2(Vu32.b)

q6_whf_vcvt2_vub⚠Experimental

Vdd32.hf=vcvt2(Vu32.ub)

q6_whf_vcvt_v⚠Experimental

Vdd32.hf=vcvt(Vu32.f8)

q6_whf_vcvt_vb⚠Experimental

Vdd32.hf=vcvt(Vu32.b)

q6_whf_vcvt_vub⚠Experimental

Vdd32.hf=vcvt(Vu32.ub)

q6_wqf32_vmpy_vhfvhf⚠Experimental

Vdd32.qf32=vmpy(Vu32.hf,Vv32.hf)

q6_wqf32_vmpy_vqf16vhf⚠Experimental

Vdd32.qf32=vmpy(Vu32.qf16,Vv32.hf)

q6_wqf32_vmpy_vqf16vqf16⚠Experimental

Vdd32.qf32=vmpy(Vu32.qf16,Vv32.qf16)

q6_wsf_vadd_vhfvhf⚠Experimental

Vdd32.sf=vadd(Vu32.hf,Vv32.hf)

q6_wsf_vcvt_vhf⚠Experimental

Vdd32.sf=vcvt(Vu32.hf)

q6_wsf_vmpy_vhfvhf⚠Experimental

Vdd32.sf=vmpy(Vu32.hf,Vv32.hf)

q6_wsf_vmpyacc_wsfvhfvhf⚠Experimental

Vxx32.sf+=vmpy(Vu32.hf,Vv32.hf)

q6_wsf_vsub_vhfvhf⚠Experimental

Vdd32.sf=vsub(Vu32.hf,Vv32.hf)

q6_wub_vadd_wubwub_sat⚠Experimental

Vdd32.ub=vadd(Vuu32.ub,Vvv32.ub):sat

q6_wub_vsub_wubwub_sat⚠Experimental

Vdd32.ub=vsub(Vuu32.ub,Vvv32.ub):sat

q6_wuh_vadd_wuhwuh_sat⚠Experimental

Vdd32.uh=vadd(Vuu32.uh,Vvv32.uh):sat

q6_wuh_vmpy_vubrub⚠Experimental

Vdd32.uh=vmpy(Vu32.ub,Rt32.ub)

q6_wuh_vmpy_vubvub⚠Experimental

Vdd32.uh=vmpy(Vu32.ub,Vv32.ub)

q6_wuh_vmpyacc_wuhvubrub⚠Experimental

Vxx32.uh+=vmpy(Vu32.ub,Rt32.ub)

q6_wuh_vmpyacc_wuhvubvub⚠Experimental

Vxx32.uh+=vmpy(Vu32.ub,Vv32.ub)

q6_wuh_vsub_wuhwuh_sat⚠Experimental

Vdd32.uh=vsub(Vuu32.uh,Vvv32.uh):sat

q6_wuh_vunpack_vub⚠Experimental

Vdd32.uh=vunpack(Vu32.ub)

q6_wuh_vzxt_vub⚠Experimental

Vdd32.uh=vzxt(Vu32.ub)

q6_wuw_vadd_wuwwuw_sat⚠Experimental

Vdd32.uw=vadd(Vuu32.uw,Vvv32.uw):sat

q6_wuw_vdsad_wuhruh⚠Experimental

Vdd32.uw=vdsad(Vuu32.uh,Rt32.uh)

q6_wuw_vdsadacc_wuwwuhruh⚠Experimental

Vxx32.uw+=vdsad(Vuu32.uh,Rt32.uh)

q6_wuw_vmpy_vuhruh⚠Experimental

Vdd32.uw=vmpy(Vu32.uh,Rt32.uh)

q6_wuw_vmpy_vuhvuh⚠Experimental

Vdd32.uw=vmpy(Vu32.uh,Vv32.uh)

q6_wuw_vmpyacc_wuwvuhruh⚠Experimental

Vxx32.uw+=vmpy(Vu32.uh,Rt32.uh)

q6_wuw_vmpyacc_wuwvuhvuh⚠Experimental

Vxx32.uw+=vmpy(Vu32.uh,Vv32.uh)

q6_wuw_vrmpy_wubrubi⚠Experimental

Vdd32.uw=vrmpy(Vuu32.ub,Rt32.ub,#u1)

q6_wuw_vrmpyacc_wuwwubrubi⚠Experimental

Vxx32.uw+=vrmpy(Vuu32.ub,Rt32.ub,#u1)

q6_wuw_vrsad_wubrubi⚠Experimental

Vdd32.uw=vrsad(Vuu32.ub,Rt32.ub,#u1)

q6_wuw_vrsadacc_wuwwubrubi⚠Experimental

Vxx32.uw+=vrsad(Vuu32.ub,Rt32.ub,#u1)

q6_wuw_vsub_wuwwuw_sat⚠Experimental

Vdd32.uw=vsub(Vuu32.uw,Vvv32.uw):sat

q6_wuw_vunpack_vuh⚠Experimental

Vdd32.uw=vunpack(Vu32.uh)

q6_wuw_vzxt_vuh⚠Experimental

Vdd32.uw=vzxt(Vu32.uh)

q6_ww_v6mpy_wubwbi_h⚠Experimental

Vdd32.w=v6mpy(Vuu32.ub,Vvv32.b,#u2):h

q6_ww_v6mpy_wubwbi_v⚠Experimental

Vdd32.w=v6mpy(Vuu32.ub,Vvv32.b,#u2):v

q6_ww_v6mpyacc_wwwubwbi_h⚠Experimental

Vxx32.w+=v6mpy(Vuu32.ub,Vvv32.b,#u2):h

q6_ww_v6mpyacc_wwwubwbi_v⚠Experimental

Vxx32.w+=v6mpy(Vuu32.ub,Vvv32.b,#u2):v

q6_ww_vadd_vhvh⚠Experimental

Vdd32.w=vadd(Vu32.h,Vv32.h)

q6_ww_vadd_vuhvuh⚠Experimental

Vdd32.w=vadd(Vu32.uh,Vv32.uh)

q6_ww_vadd_wwww⚠Experimental

Vdd32.w=vadd(Vuu32.w,Vvv32.w)

q6_ww_vadd_wwww_sat⚠Experimental

Vdd32.w=vadd(Vuu32.w,Vvv32.w):sat

q6_ww_vaddacc_wwvhvh⚠Experimental

Vxx32.w+=vadd(Vu32.h,Vv32.h)

q6_ww_vaddacc_wwvuhvuh⚠Experimental

Vxx32.w+=vadd(Vu32.uh,Vv32.uh)

q6_ww_vasrinto_wwvwvw⚠Experimental

Vxx32.w=vasrinto(Vu32.w,Vv32.w)

q6_ww_vdmpy_whrb⚠Experimental

Vdd32.w=vdmpy(Vuu32.h,Rt32.b)

q6_ww_vdmpyacc_wwwhrb⚠Experimental

Vxx32.w+=vdmpy(Vuu32.h,Rt32.b)

q6_ww_vmpa_whrb⚠Experimental

Vdd32.w=vmpa(Vuu32.h,Rt32.b)

q6_ww_vmpa_wuhrb⚠Experimental

Vdd32.w=vmpa(Vuu32.uh,Rt32.b)

q6_ww_vmpaacc_wwwhrb⚠Experimental

Vxx32.w+=vmpa(Vuu32.h,Rt32.b)

q6_ww_vmpaacc_wwwuhrb⚠Experimental

Vxx32.w+=vmpa(Vuu32.uh,Rt32.b)

q6_ww_vmpy_vhrh⚠Experimental

Vdd32.w=vmpy(Vu32.h,Rt32.h)

q6_ww_vmpy_vhvh⚠Experimental

Vdd32.w=vmpy(Vu32.h,Vv32.h)

q6_ww_vmpy_vhvuh⚠Experimental

Vdd32.w=vmpy(Vu32.h,Vv32.uh)

q6_ww_vmpyacc_wwvhrh⚠Experimental

Vxx32.w+=vmpy(Vu32.h,Rt32.h)

q6_ww_vmpyacc_wwvhrh_sat⚠Experimental

Vxx32.w+=vmpy(Vu32.h,Rt32.h):sat

q6_ww_vmpyacc_wwvhvh⚠Experimental

Vxx32.w+=vmpy(Vu32.h,Vv32.h)

q6_ww_vmpyacc_wwvhvuh⚠Experimental

Vxx32.w+=vmpy(Vu32.h,Vv32.uh)

q6_ww_vrmpy_wubrbi⚠Experimental

Vdd32.w=vrmpy(Vuu32.ub,Rt32.b,#u1)

q6_ww_vrmpyacc_wwwubrbi⚠Experimental

Vxx32.w+=vrmpy(Vuu32.ub,Rt32.b,#u1)

q6_ww_vsub_vhvh⚠Experimental

Vdd32.w=vsub(Vu32.h,Vv32.h)

q6_ww_vsub_vuhvuh⚠Experimental

Vdd32.w=vsub(Vu32.uh,Vv32.uh)

q6_ww_vsub_wwww⚠Experimental

Vdd32.w=vsub(Vuu32.w,Vvv32.w)

q6_ww_vsub_wwww_sat⚠Experimental

Vdd32.w=vsub(Vuu32.w,Vvv32.w):sat

q6_ww_vsxt_vh⚠Experimental

Vdd32.w=vsxt(Vu32.h)

q6_ww_vtmpy_whrb⚠Experimental

Vdd32.w=vtmpy(Vuu32.h,Rt32.b)

q6_ww_vtmpyacc_wwwhrb⚠Experimental

Vxx32.w+=vtmpy(Vuu32.h,Rt32.b)

q6_ww_vunpack_vh⚠Experimental

Vdd32.w=vunpack(Vu32.h)

q6_ww_vunpackoor_wwvh⚠Experimental

Vxx32.w|=vunpacko(Vu32.h)