Function core::arch::mips64::__msa_msubv_w

pub unsafe fn __msa_msubv_w(a: v4i32, b: v4i32, c: v4i32) -> v4i32

🔬 This is a nightly-only experimental API. (stdsimd #48556)

This is supported on MIPS-64 and target feature msa only.

Vector Multiply and Subtract

The integer elements in vector c (four signed 32-bit integer numbers) are multiplied by integer elements in vector b (four signed 32-bit integer numbers) and subtracted from the integer elements in vector a (four signed 32-bit integer numbers). The most significant half of the multiplication result is discarded.