core::arch

Module nvptx

🔬This is a nightly-only experimental API. (stdarch_nvptx #111199)

Available on target_arch=nvptx64 only.

Platform-specific intrinsics for the NVPTX platform.

See the module documentation for more details.

Structs

f16x2Experimental

PTX-specific 32-bit wide floating point (f16 x 2) vector type

Functions

__assert_fail⚠Experimental

Syscall to be used whenever the assert expression produces a false value.

_block_dim_x⚠Experimental

x-th thread-block dimension.

_block_dim_y⚠Experimental

y-th thread-block dimension.

_block_dim_z⚠Experimental

z-th thread-block dimension.

_block_idx_x⚠Experimental

x-th thread-block index.

_block_idx_y⚠Experimental

y-th thread-block index.

_block_idx_z⚠Experimental

z-th thread-block index.

_grid_dim_x⚠Experimental

x-th block-grid dimension.

_grid_dim_y⚠Experimental

y-th block-grid dimension.

_grid_dim_z⚠Experimental

z-th block-grid dimension.

_syncthreads⚠Experimental

Synchronizes all threads in the block.

_thread_idx_x⚠Experimental

x-th thread index.

_thread_idx_y⚠Experimental

y-th thread index.

_thread_idx_z⚠Experimental

z-th thread index.

f16x2_add⚠Experimental

Add two values, round to nearest even

f16x2_fma⚠Experimental

Fused multiply-add, round to nearest even

f16x2_max⚠Experimental

Find the maximum of two values

f16x2_max_nan⚠Experimental

Find the maximum of two values, NaNs pass through.

f16x2_min⚠Experimental

Find the minimum of two values

f16x2_min_nan⚠Experimental

Find the minimum of two values, NaNs pass through.

f16x2_mul⚠Experimental

Multiply two values, round to nearest even

f16x2_neg⚠Experimental

Arithmetic negate

f16x2_sub⚠Experimental

Subtract two values, round to nearest even

free⚠Experimental

Free previously dynamically allocated memory.

malloc⚠Experimental

Allocate memory dynamically from a fixed-size heap in global memory.

trap⚠Experimental

Generates the trap instruction TRAP

vprintf⚠Experimental

Print formatted output from a kernel to a host-side output stream.