Function core::arch::riscv32::hfence_vvma

pub unsafe fn hfence_vvma(vaddr: usize, asid: usize)

🔬This is a nightly-only experimental API. (stdsimd #48556)

Available on RISC-V RV32 only.

Hypervisor memory management fence for given guest virtual address and guest address space

Guarantees that any previous stores already visible to the current hart are ordered before all implicit reads by that hart done for VS-stage address translation for instructions that:

are subsequent to the HFENCE.VVMA, and

execute when hgatp.VMID has the same setting as it did when HFENCE.VVMA executed.

This fence specifies a single guest virtual address, and a single guest address-space identifier.