core::arch::riscv32

Function hlvx_wu

pub unsafe fn hlvx_wu(src: *const u32) -> u32

🔬This is a nightly-only experimental API. (riscv_ext_intrinsics #114544)

Available on RISC-V RV32 only.

Accesses virtual machine instruction by unsigned word integer

This instruction performs an explicit memory access as though V=1; the memory being read must be executable in both stages of address translation, but read permission is not required.

This function is unsafe for it accesses the virtual supervisor or user via a HLVX.WU instruction which is effectively a dereference to any memory address.