Function core::arch::riscv64::aes64ks1i

pub unsafe fn aes64ks1i(rs1: u64, const RNUM: u8) -> u64

🔬This is a nightly-only experimental API. (riscv_ext_intrinsics #114544)

Available on RISC-V RV64 and target feature zkne and target feature zknd only.

This instruction implements part of the KeySchedule operation for the AES Block cipher involving the SBox operation.

This instruction implements the rotation, SubBytes and Round Constant addition steps of the AES block cipher Key Schedule. This instruction must always be implemented such that its execution latency does not depend on the data being operated on. Note that rnum must be in the range 0x0..0xA. The values 0xB..0xF are reserved.

Source: RISC-V Cryptography Extensions Volume I: Scalar & Entropy Source Instructions

Version: v1.0.1

Section: 3.10

Note

The RNUM parameter is expected to be a constant value inside the range of 0..=10.

Safety

This function is safe to use if the zkne or zknd target feature is present.