core::arch::riscv64

Function hinval_vvma_all

pub unsafe fn hinval_vvma_all()

🔬This is a nightly-only experimental API. (riscv_ext_intrinsics #114544)

Available on RISC-V RV64 only.

Invalidate hypervisor translation cache for all guest address spaces and guest virtual addresses

This instruction invalidates any address-translation cache entries that an HFENCE.VVMA instruction with the same values of vaddr and asid would invalidate.

This fence applies to any guest address spaces and guest virtual addresses.