core::arch::riscv64
Function sha512sig0

pub unsafe fn sha512sig0(rs1: u64) -> u64

🔬This is a nightly-only experimental API. (riscv_ext_intrinsics #114544)

Available on RISC-V RV64 and target feature zknh only.

Implements the Sigma0 transformation function as used in the SHA2-512 hash function [49] (Section 4.1.3).

This instruction is supported for the RV64 base architecture. It implements the Sigma0 transform of the SHA2-512 hash function. [49]. This instruction must always be implemented such that its execution latency does not depend on the data being operated on.

Source: RISC-V Cryptography Extensions Volume I: Scalar & Entropy Source Instructions

Version: v1.0.1

Section: 3.37

Safety

This function is safe to use if the zknh target feature is present.