Function core::arch::wasm32::i16x8_relaxed_dot_i8x16_i7x16

pub fn i16x8_relaxed_dot_i8x16_i7x16(a: v128, b: v128) -> v128

🔬This is a nightly-only experimental API. (stdarch_wasm_relaxed_simd #111196)

Available on target_family="wasm" and target feature relaxed-simd and WebAssembly only.

A relaxed dot-product instruction.

This instruction will perform pairwise products of the 8-bit values in a and b and then accumulate adjacent pairs into 16-bit results producing a final i16x8 vector. The bytes of a are always interpreted as signed and the bytes in b may be interpreted as signed or unsigned. If the top bit in b isn’t set then the value is the same regardless of whether it’s signed or unsigned.

The accumulation into 16-bit values may be saturated on some platforms, and on other platforms it may wrap-around on overflow.