Function core::arch::wasm32::i32x4_trunc_sat_f64x2_zero

pub fn i32x4_trunc_sat_f64x2_zero(a: v128) -> v128

This is supported on WebAssembly and target feature simd128 only.

[帮助纠正😄]使用 IEEE convertToIntegerTowardZero 函数将两个双精度浮点通道饱和转换为两个小整数通道。

Saturating conversion of the two double-precision floating point lanes to two lower integer lanes using the IEEE convertToIntegerTowardZero function.

结果的两个较高通道被初始化为零。 如果任何输入通道是 NaN，则结果通道为 0。如果通道的四舍五入整数值超出目标类型的范围，则结果饱和到最接近的可表示整数值。

The two higher lanes of the result are initialized to zero. If any input lane is a NaN, the resulting lane is 0. If the rounded integer value of a lane is outside the range of the destination type, the result is saturated to the nearest representable integer value.