core::arch::wasm32

Function i64x2_shr
·

pub fn i64x2_shr(a: v128, amt: u32) -> v128

Available on WebAssembly and target feature simd128 only.

Shifts each lane to the right by the specified number of bits, sign extending.

Only the low bits of the shift amount are used if the shift amount is greater than the lane width.