Function core::arch::wasm32::memory_size

pub fn memory_size<const MEM: u32>() -> usize

This is supported on WebAssembly only.

[帮助纠正😄]对应于 wasm 的 memory.size 指令的固有特性

Corresponding intrinsic to wasm’s memory.size instruction

此函数在调用时将以页为单位返回当前内存大小。 当前的 WebAssembly 页面大小为 65536 字节 (64 KB)。

This function, when called, will return the current memory size in units of pages. The current WebAssembly page size is 65536 bytes (64 KB).

参数 MEM 是要返回其大小的内存的数字索引。 请注意，目前 WebAssembly 规范仅支持一个内存，因此需要传入零。该参数的存在是为了与未来的 WebAssembly 修订版向前兼容。 如果将非零参数传递给此函数，它将当前无条件中止。

The argument MEM is the numerical index of which memory to return the size of. Note that currently the WebAssembly specification only supports one memory, so it is required that zero is passed in. The argument is present to be forward-compatible with future WebAssembly revisions. If a nonzero argument is passed to this function it will currently unconditionally abort.