Function core::arch::wasm32::u32x4_trunc_sat_f64x2_zero

pub fn u32x4_trunc_sat_f64x2_zero(a: v128) -> v128

Available on target_family="wasm" and target feature simd128 and WebAssembly only.

Saturating conversion of the two double-precision floating point lanes to two lower integer lanes using the IEEE convertToIntegerTowardZero function.

The two higher lanes of the result are initialized to zero. If any input lane is a NaN, the resulting lane is 0. If the rounded integer value of a lane is outside the range of the destination type, the result is saturated to the nearest representable integer value.

This documentation is an old archive. Please see https://rust.docs.kernel.org instead.