Function core::arch::wasm32::u8x16_shuffle

pub fn u8x16_shuffle<const I0: usize, const I1: usize, const I2: usize, const I3: usize, const I4: usize, const I5: usize, const I6: usize, const I7: usize, const I8: usize, const I9: usize, const I10: usize, const I11: usize, const I12: usize, const I13: usize, const I14: usize, const I15: usize>(
    a: v128, 
    b: v128
) -> v128

This is supported on WebAssembly and target feature simd128 only.

[帮助纠正😄]返回一个新Vector，其通道选自 16 个立即操作数中指定的两个输入Vector $a 和 $b 的通道。

Returns a new vector with lanes selected from the lanes of the two input vectors $a and $b specified in the 16 immediate operands.

$a 和 $b 表达式必须具有 v128 类型，并且此函数生成一个用 16 个字节编码的 wasm 指令，提供要返回的元素的索引。 [0, 15] 范围内的索引 i 选择 a 的第 i 个元素。 [16, 31] 范围内的索引选择 b 的第 i - 16 个元素。

The $a and $b expressions must have type v128, and this function generates a wasm instruction that is encoded with 16 bytes providing the indices of the elements to return. The indices i in range [0, 15] select the i-th element of a. The indices in range [16, 31] select the i - 16-th element of b.

请注意，这是一个宏，因为所有索引表达式 $i* 的代码生成要求必须是常量。 如果任何表达式不是常量，将生成编译器错误。

Note that this is a macro due to the codegen requirements of all of the index expressions $i* must be constant. A compiler error will be generated if any of the expressions are not constant.

所有索引 $i* 必须具有 u32 类型。

All indexes $i* must have the type u32.