Function core::arch::wasm32::v128_store16_lane

pub unsafe fn v128_store16_lane<const L: usize>(v: v128, m: *mut u16)

Available on target_family="wasm" and target feature simd128 and WebAssembly only.

Stores the 16-bit value from lane L of v into m

This intrinsic is provided for completeness and is equivalent to *m = u16x8_extract_lane::<L>(v) (which doesn’t require unsafe).

Unsafety

This intrinsic is unsafe because it takes a raw pointer as an argument, and the pointer must be valid to store 2 bytes to. Note that there is no alignment requirement on this pointer since this intrinsic performs a 1-aligned store.