Function core::arch::wasm64::i64x2_load_extend_u32x2

pub unsafe fn i64x2_load_extend_u32x2(m: *const u32) -> v128

Available on target_family="wasm" and target feature simd128 and WebAssembly only.

Load two 32-bit integers and zero extend each one to a 64-bit lane

Unsafety

This intrinsic is unsafe because it takes a raw pointer as an argument, and the pointer must be valid to load 8 bytes from. Note that there is no alignment requirement on this pointer since this intrinsic performs a 1-aligned load.