Function core::arch::wasm64::v128_load32_lane

pub unsafe fn v128_load32_lane<const L: usize>(v: v128, m: *const u32) -> v128

Available on WebAssembly and target feature simd128 only.

Loads a 32-bit value from m and sets lane L of v to that value.

This intrinsic is provided for completeness and is equivalent to u32x4_replace_lane::<L>(v, *m) (which doesn’t require unsafe).

Unsafety

This intrinsic is unsafe because it takes a raw pointer as an argument, and the pointer must be valid to load 4 bytes from. Note that there is no alignment requirement on this pointer since this intrinsic performs a 1-aligned load.