Function core::arch::x86::_blsfill_u32

pub unsafe fn _blsfill_u32(x: u32) -> u32

Available on (x86 or x86-64) and target feature tbm and x86 only.

Sets all bits of x below the least significant one.

If there is no set bit in x, it sets all the bits.