Function core::arch::x86::_mm256_hadd_pd

pub unsafe fn _mm256_hadd_pd(a: __m256d, b: __m256d) -> __m256d

This is supported on x86 and target feature avx only.

[帮助纠正😄]4 个 64 位浮点 a 和 b 的两个压缩Vector中相邻对的水平相加。 结果，来自 a 的元素总和在偶数位置返回，而来自 b 的元素总和在奇数位置返回。

Horizontal addition of adjacent pairs in the two packed vectors of 4 64-bit floating points a and b. In the result, sums of elements from a are returned in even locations, while sums of elements from b are returned in odd locations.

英特尔的文档

Intel’s documentation