core::arch::x86

Function _mm256_mask_roundscale_ph
·

pub fn _mm256_mask_roundscale_ph(
    src: __m256h,
    k: __mmask16,
    a: __m256h,
    const IMM8: i32,
) -> __m256h

Available on x86 and target feature avx512fp16 and target feature avx512vl only.

Round packed half-precision (16-bit) floating-point elements in a to the number of fraction bits specified by imm8, and store the results in dst using writemask k (elements are copied from src when the corresponding mask bit is not set).

Rounding is done according to the imm8 parameter, which can be one of:

_MM_FROUND_TO_NEAREST_INT : round to nearest

_MM_FROUND_TO_NEG_INF : round down

_MM_FROUND_TO_POS_INF : round up

_MM_FROUND_TO_ZERO : truncate

_MM_FROUND_CUR_DIRECTION : use MXCSR.RC - see _MM_SET_ROUNDING_MODE

Intel’s documentation