Function core::arch::x86::_mm256_mask_ternarylogic_epi64

pub unsafe fn _mm256_mask_ternarylogic_epi64<const IMM8: i32>(
    src: __m256i, 
    k: __mmask8, 
    a: __m256i, 
    b: __m256i
) -> __m256i

🔬 This is a nightly-only experimental API. (stdsimd #48556)

This is supported on x86 and target feature avx512f,avx512vl only.

[帮助纠正😄]按位三进制逻辑，提供实现任何三操作数二进制函数的能力； 具体的二元函数由 imm8 中的值指定。 对于每个打包的 64 位整数中的每个位，来自 src、a 和 b 的对应位用于形成 imm8 的 3 位索引，并使用 writemask 将 imm8 中该位的值写入 dst 中的相应位 k 以 64 位粒度（当相应的掩码位未设置时，从 src 复制 64 位元素）。

Bitwise ternary logic that provides the capability to implement any three-operand binary function; the specific binary function is specified by value in imm8. For each bit in each packed 64-bit integer, the corresponding bit from src, a, and b are used to form a 3 bit index into imm8, and the value at that bit in imm8 is written to the corresponding bit in dst using writemask k at 64-bit granularity (64-bit elements are copied from src when the corresponding mask bit is not set).

英特尔的文档

Intel’s documentation