core::arch::x86

Function _mm256_movepi16_mask
·

pub fn _mm256_movepi16_mask(a: __m256i) -> __mmask16

Available on x86 and target feature avx512bw and target feature avx512vl only.

Set each bit of mask register k based on the most significant bit of the corresponding packed 16-bit integer in a.

Intel’s documentation