Function core::arch::x86::_mm512_getmant_round_pd

pub unsafe fn _mm512_getmant_round_pd<const NORM: _MM_MANTISSA_NORM_ENUM, const SIGN: _MM_MANTISSA_SIGN_ENUM, const SAE: i32>(
    a: __m512d
) -> __m512d

🔬 This is a nightly-only experimental API. (stdsimd #48556)

This is supported on x86 and target feature avx512f only.

[帮助纠正😄]规范化 a 中压缩双精度（64 位）浮点元素的尾数，并将结果存储在 dst 中。 此内在函数本质上计算 ±(2^k)*|x.significand|，其中 k 取决于 interv 定义的区间范围，而符号取决于 sc 和源符号。 尾数归一化为 interv 指定的区间，可以取以下值： _MM_MANT_NORM_1_2 // 区间 [1, 2) _MM_MANT_NORM_p5_2 // 区间 [0.5, 2) _MM_MANT_NORM_p5_1 // 区间 [0.5, 1) _MM_MANT_NORM_p75_1p5 // 区间 [ 0.75, 1.5) 符号由 sc 确定，它可以采用以下值： _MM_MANT_SIGN_src // sign = sign(src) _MM_MANT_SIGN_zero // sign = 0 _MM_MANT_SIGN_nan // dst = NaN if sign(src) = 1 可以通过以下方式抑制异常 在 sae 参数中传递 _MM_FROUND_NO_EXC。

Normalize the mantissas of packed double-precision (64-bit) floating-point elements in a, and store the results in dst. This intrinsic essentially calculates ±(2^k)*|x.significand|, where k depends on the interval range defined by interv and the sign depends on sc and the source sign.
The mantissa is normalized to the interval specified by interv, which can take the following values:
_MM_MANT_NORM_1_2 // interval [1, 2)
_MM_MANT_NORM_p5_2 // interval [0.5, 2)
_MM_MANT_NORM_p5_1 // interval [0.5, 1)
_MM_MANT_NORM_p75_1p5 // interval [0.75, 1.5)
The sign is determined by sc which can take the following values:
_MM_MANT_SIGN_src // sign = sign(src)
_MM_MANT_SIGN_zero // sign = 0
_MM_MANT_SIGN_nan // dst = NaN if sign(src) = 1
Exceptions can be suppressed by passing _MM_FROUND_NO_EXC in the sae parameter.

英特尔的文档

Intel’s documentation