core::arch::x86
Function _mm512_maskz_roundscale_round_ph

pub unsafe fn _mm512_maskz_roundscale_round_ph(
    k: __mmask32,
    a: __m512h,
    const IMM8: i32,
    const SAE: i32,
) -> __m512h

🔬This is a nightly-only experimental API. (stdarch_x86_avx512_f16 #127213)

Available on (x86 or x86-64) and target feature avx512fp16 and x86 only.

Round packed half-precision (16-bit) floating-point elements in a to the number of fraction bits specified by imm8, and store the results in dst using zeromask k (elements are zeroed out when the corresponding mask bit is not set). Exceptions can be suppressed by passing _MM_FROUND_NO_EXC in the sae parameter

Rounding is done according to the imm8 parameter, which can be one of:

_MM_FROUND_TO_NEAREST_INT // round to nearest
_MM_FROUND_TO_NEG_INF     // round down
_MM_FROUND_TO_POS_INF     // round up
_MM_FROUND_TO_ZERO        // truncate
_MM_FROUND_CUR_DIRECTION  // use MXCSR.RC; see _MM_SET_ROUNDING_MODE

Intel’s documentation