Function core::arch::x86::_mm_loadl_pi

pub unsafe fn _mm_loadl_pi(a: __m128, p: *const __m64) -> __m128

🔬 This is a nightly-only experimental API. (stdsimd #48556)

This is supported on x86 and target feature sse only.

Load two floats from p into the lower half of a __m128. The upper half is copied from the upper half of a.

This corresponds to the MOVLPS / MOVLDP / VMOVLDP instructions.

#[cfg(target_arch = "x86")]
use std::arch::x86::*;
#[cfg(target_arch = "x86_64")]
use std::arch::x86_64::*;

let a = _mm_setr_ps(1.0, 2.0, 3.0, 4.0);
let data: [f32; 4] = [5.0, 6.0, 7.0, 8.0];
let r = _mm_loadh_pi(a, data[..].as_ptr() as *const _) ;
// assert_eq!(r, _mm_setr_ps(5.0, 6.0, 3.0, 4.0));Run