Function core::arch::x86::_mm_maskz_roundscale_sd

pub unsafe fn _mm_maskz_roundscale_sd(
    k: __mmask8,
    a: __m128d,
    b: __m128d,
    const IMM8: i32
) -> __m128d

🔬This is a nightly-only experimental API. (stdsimd #48556)

Available on (x86 or x86-64) and target feature avx512f and x86 only.

Round the lower double-precision (64-bit) floating-point element in b to the number of fraction bits specified by imm8, store the result in the lower element of dst using zeromask k (the element is zeroed out when mask bit 0 is not set), and copy the upper element from a to the upper element of dst.
Rounding is done according to the imm8[2:0] parameter, which can be one of:
_MM_FROUND_TO_NEAREST_INT // round to nearest
_MM_FROUND_TO_NEG_INF // round down
_MM_FROUND_TO_POS_INF // round up
_MM_FROUND_TO_ZERO // truncate
_MM_FROUND_CUR_DIRECTION // use MXCSR.RC; see _MM_SET_ROUNDING_MODE

Intel’s documentation

This documentation is an old archive. Please see https://rust.docs.kernel.org instead.