Function core::arch::x86::_mm_prefetch

pub unsafe fn _mm_prefetch<const STRATEGY: i32>(p: *const i8)

This is supported on x86 and target feature sse only.

[帮助纠正😄]使用给定的 STRATEGY 获取包含地址 p 的缓存行。

Fetch the cache line that contains address p using the given STRATEGY.

战略必须是以下之一：

The STRATEGY must be one of:

_MM_HINT_T0：获取缓存层次结构的所有级别。

_MM_HINT_T0: Fetch into all levels of the cache hierarchy.

_MM_HINT_T1：获取到 L2 和更高。

_MM_HINT_T1: Fetch into L2 and higher.

_MM_HINT_T2：获取 L3 和更高级别或特定于实现的选择（例如，如果没有 L3，则为 L2）。

_MM_HINT_T2: Fetch into L3 and higher or an implementation-specific choice (e.g., L2 if there is no L3).

_MM_HINT_NTA：使用非临时访问 (NTA) 提示获取数据。 它可能是一个比主存更近但在缓存层次结构之外的地方。 这用于在不污染缓存的情况下减少访问延迟。

_MM_HINT_NTA: Fetch data using the non-temporal access (NTA) hint. It may be a place closer than main memory but outside of the cache hierarchy. This is used to reduce access latency without polluting the cache.

_MM_HINT_ET0 和 _MM_HINT_ET1 与 _MM_HINT_T0 和 _MM_HINT_T1 类似，但表示预期写入地址。

_MM_HINT_ET0 and _MM_HINT_ET1 are similar to _MM_HINT_T0 and _MM_HINT_T1 but indicate an anticipation to write to the address.

实际实现取决于特定的 CPU。 该指令被认为是一个提示，因此 CPU 也可以简单地忽略该请求。

The actual implementation depends on the particular CPU. This instruction is considered a hint, so the CPU is also free to simply ignore the request.

预取数据量取决于特定 CPU 的高速缓存行大小，但至少为 32 字节。

The amount of prefetched data depends on the cache line size of the specific CPU, but it will be at least 32 bytes.

常见警告：

Common caveats:

大多数现代 CPU 已经根据预测的访问模式自动预取数据。

Most modern CPUs already automatically prefetch data based on predicted access patterns.

如果这会导致 TLB 未命中或页面错误，则通常不会获取数据。

Data is usually not fetched if this would cause a TLB miss or a page fault.

过多的预取可能会导致不必要的缓存驱逐。

Too much prefetching can cause unnecessary cache evictions.

如果没有足够的内存子系统资源（例如，请求缓冲区），预取也可能会失败。

Prefetching may also fail if there are not enough memory-subsystem resources (e.g., request buffers).

英特尔的文档

Intel’s documentation