Function core::arch::x86::_xrstor

pub unsafe fn _xrstor(mem_addr: *const u8, rs_mask: u64)

Available on (x86 or x86-64) and target feature xsave and x86 only.

Performs a full or partial restore of the enabled processor states using the state information stored in memory at mem_addr.

State is restored based on bits [62:0] in rs_mask, XCR0, and mem_addr.HEADER.XSTATE_BV. mem_addr must be aligned on a 64-byte boundary.

Intel’s documentation