Struct core::arch::x86::__m128i

#[repr(simd)]
pub struct __m128i(/* private fields */);

Available on x86 and (x86 or x86-64) only.

128-bit wide integer vector type, x86-specific

This type is the same as the __m128i type defined by Intel, representing a 128-bit SIMD register. Usage of this type typically corresponds to the sse and up target features for x86/x86_64.

Internally this type may be viewed as:

i8x16 - sixteen i8 variables packed together

i16x8 - eight i16 variables packed together

i32x4 - four i32 variables packed together

i64x2 - two i64 variables packed together

(as well as unsigned versions). Each intrinsic may interpret the internal bits differently, check the documentation of the intrinsic to see how it’s being used.

Note that this means that an instance of __m128i typically just means a “bag of bits” which is left up to interpretation at the point of use.

Most intrinsics using __m128i are prefixed with _mm_ and the integer types tend to correspond to suffixes like “epi8” or “epi32”.

Examples

#[cfg(target_arch = "x86")]
use std::arch::x86::*;
#[cfg(target_arch = "x86_64")]
use std::arch::x86_64::*;

let all_bytes_zero = _mm_setzero_si128();
let all_bytes_one = _mm_set1_epi8(1);
let four_i32 = _mm_set_epi32(1, 2, 3, 4);
Run