core::arch::x86_64

Function _mm256_sha512rnds2_epi64
·

pub fn _mm256_sha512rnds2_epi64(a: __m256i, b: __m256i, k: __m128i) -> __m256i

Available on x86-64 and target feature sha512 and target feature avx only.

This intrinsic performs two rounds of SHA512 operation using initial SHA512 state (C,D,G,H) from a, an initial SHA512 state (A,B,E,F) from b, and a pre-computed sum of the next two round message qwords and the corresponding round constants from c (only the two lower qwords of the third operand). The updated SHA512 state (A,B,E,F) is written to dst, and dst can be used as the updated state (C,D,G,H) in later rounds.

Intel’s documentation