core::arch::x86_64

Function _mm512_mask_min_epi16
·

pub fn _mm512_mask_min_epi16(
    src: __m512i,
    k: __mmask32,
    a: __m512i,
    b: __m512i,
) -> __m512i

Available on x86-64 and target feature avx512bw only.

Compare packed signed 16-bit integers in a and b, and store packed minimum values in dst using writemask k (elements are copied from src when the corresponding mask bit is not set).

Intel’s documentation