core::arch::x86_64::
Function _mm512_mask_sub_round_ph

pub unsafe fn _mm512_mask_sub_round_ph(
    src: __m512h,
    k: __mmask32,
    a: __m512h,
    b: __m512h,
    const ROUNDING: i32,
) -> __m512h

🔬This is a nightly-only experimental API. (stdarch_x86_avx512_f16 #127213)

Available on (x86 or x86-64) and target feature avx512fp16 and x86-64 only.

Subtract packed half-precision (16-bit) floating-point elements in b from a, and store the results in dst using writemask k (elements are copied from src when the corresponding mask bit is not set). Rounding is done according to the rounding parameter, which can be one of:

(_MM_FROUND_TO_NEAREST_INT |_MM_FROUND_NO_EXC) // round to nearest, and suppress exceptions
(_MM_FROUND_TO_NEG_INF |_MM_FROUND_NO_EXC)     // round down, and suppress exceptions
(_MM_FROUND_TO_POS_INF |_MM_FROUND_NO_EXC)     // round up, and suppress exceptions
(_MM_FROUND_TO_ZERO |_MM_FROUND_NO_EXC)        // truncate, and suppress exceptions
_MM_FROUND_CUR_DIRECTION

Intel’s documentation