Function core::arch::x86_64::_mm_extract_si64

pub unsafe fn _mm_extract_si64(x: __m128i, y: __m128i) -> __m128i

This is supported on x86-64 and target feature sse4a only.

[帮助纠正😄]从 x 的低 64 位中提取 y 指定的位范围。

Extracts the bit range specified by y from the lower 64 bits of x.

y 的 [13:8] 位指定要提取的位范围的索引。 y 的 [5:0] 位指定要提取的位范围的长度。 所有其他位都被忽略。

The [13:8] bits of y specify the index of the bit-range to extract. The [5:0] bits of y specify the length of the bit-range to extract. All other bits are ignored.

如果长度为零，则解释为 64。如果长度和索引为零，则提取 x 的低 64 位。

If the length is zero, it is interpreted as 64. If the length and index are zero, the lower 64 bits of x are extracted.

如果 length == 0 && index > 0 或 lenght + index > 64，则结果未定义。

If length == 0 && index > 0 or lenght + index > 64 the result is undefined.