Function core::arch::x86_64::_mm_maskz_scalef_round_sd

pub unsafe fn _mm_maskz_scalef_round_sd(
    k: __mmask8,
    a: __m128d,
    b: __m128d,
    const ROUNDING: i32
) -> __m128d

🔬 This is a nightly-only experimental API. (stdsimd #48556)

Available on x86-64 and target feature avx512f only.

Scale the packed double-precision (64-bit) floating-point elements in a using values from b, store the result in the lower element of dst using zeromask k (the element is zeroed out when mask bit 0 is not set), and copy the upper element from a to the upper element of dst.

Rounding is done according to the rounding[3:0] parameter, which can be one of:
(_MM_FROUND_TO_NEAREST_INT |_MM_FROUND_NO_EXC) // round to nearest, and suppress exceptions
(_MM_FROUND_TO_NEG_INF |_MM_FROUND_NO_EXC) // round down, and suppress exceptions
(_MM_FROUND_TO_POS_INF |_MM_FROUND_NO_EXC) // round up, and suppress exceptions
(_MM_FROUND_TO_ZERO |_MM_FROUND_NO_EXC) // truncate, and suppress exceptions
_MM_FROUND_CUR_DIRECTION // use MXCSR.RC; see _MM_SET_ROUNDING_MODE

Intel’s documentation