Function core::arch::x86_64::_mm_minpos_epu16

pub unsafe fn _mm_minpos_epu16(a: __m128i) -> __m128i

This is supported on x86-64 and target feature sse4.1 only.

[帮助纠正😄]在 128 位 __m128i Vector中查找最小的无符号 16 位元素，返回一个Vector，该Vector的第一个位置包含其值，第二个位置包含其索引； 所有其他元素都设置为零。

Finds the minimum unsigned 16-bit element in the 128-bit __m128i vector, returning a vector containing its value in its first position, and its index in its second position; all other elements are set to zero.

此内在函数对应于 VPHMINPOSUW / PHMINPOSUW 指令。

This intrinsic corresponds to the VPHMINPOSUW / PHMINPOSUW instruction.

论据：

Arguments:

a - __m128i 类型的 128 位Vector。

a - A 128-bit vector of type __m128i.

回报：

Returns:

一个 128 位的值，其中：

A 128-bit value where:

位 [15:0] - 包含在参数 a 中找到的最小值，

bits [15:0] - contain the minimum value found in parameter a,

bits [18:16] - 包含最小值的索引

bits [18:16] - contain the index of the minimum value

其余位设置为 0。

remaining bits are set to 0.

英特尔的文档

Intel’s documentation