Function core::arch::x86_64::_mm_round_ps

pub unsafe fn _mm_round_ps<const ROUNDING: i32>(a: __m128) -> __m128

This is supported on x86-64 and target feature sse4.1 only.

Round the packed single-precision (32-bit) floating-point elements in a using the ROUNDING parameter, and stores the results as packed single-precision floating-point elements. Rounding is done according to the rounding parameter, which can be one of:

#[cfg(target_arch = "x86")]
use std::arch::x86::*;
#[cfg(target_arch = "x86_64")]
use std::arch::x86_64::*;

// round to nearest, and suppress exceptions:
_MM_FROUND_TO_NEAREST_INT | _MM_FROUND_NO_EXC;
// round down, and suppress exceptions:
_MM_FROUND_TO_NEG_INF | _MM_FROUND_NO_EXC;
// round up, and suppress exceptions:
_MM_FROUND_TO_POS_INF | _MM_FROUND_NO_EXC;
// truncate, and suppress exceptions:
_MM_FROUND_TO_ZERO | _MM_FROUND_NO_EXC;
// use MXCSR.RC; see `_MM_SET_ROUNDING_MODE`:
_MM_FROUND_CUR_DIRECTION;
Run

Intel’s documentation