Function core::arch::x86_64::_mm_setcsr

pub unsafe fn _mm_setcsr(val: u32)

This is supported on x86-64 and target feature sse only.

[帮助纠正😄]使用 32 位无符号整数值设置 MXCSR 寄存器。

Sets the MXCSR register with the 32-bit unsigned integer value.

该寄存器控制 SIMD 指令如何处理浮点运算。 修改这个寄存器只会影响当前线程。

This register constrols how SIMD instructions handle floating point operations. Modifying this register only affects the current thread.

它包含几组标志：

It contains several groups of flags:

异常标志报告自上次重置以来发生了哪些异常。

Exception flags report which exceptions occurred since last they were reset.

屏蔽标志可用于屏蔽（忽略）某些异常。 默认情况下，这些标志都设置为 1，因此所有异常都被屏蔽。 当一个异常被屏蔽时，处理器简单地设置异常标志并继续操作。 如果该异常未被屏蔽，则该标志也会被设置，但还会调用一个异常处理程序。

Masking flags can be used to mask (ignore) certain exceptions. By default these flags are all set to 1, so all exceptions are masked. When an an exception is masked, the processor simply sets the exception flag and continues the operation. If the exception is unmasked, the flag is also set but additionally an exception handler is invoked.

舍入模式标志控制浮点指令的舍入模式。

Rounding mode flags control the rounding mode of floating point instructions.

denormals-are-zero 模式标志将所有将被非规格化（指数位全为零）的数字变为零。

The denormals-are-zero mode flag turns all numbers which would be denormalized (exponent bits are all zeros) into zeros.

Exception Flags

_MM_EXCEPT_INVALID：执行了无效操作（例如，将 Infinity 除以 Infinity）。

_MM_EXCEPT_INVALID: An invalid operation was performed (e.g., dividing Infinity by Infinity).

_MM_EXCEPT_DENORM：试图对非规范化数字进行操作。 主要是这会导致精度损失。

_MM_EXCEPT_DENORM: An operation attempted to operate on a denormalized number. Mainly this can cause loss of precision.

_MM_EXCEPT_DIV ZERO：发生被零除。

_MM_EXCEPT_DIV_ZERO: Division by zero occured.

_MM_EXCEPT_OVERFLOW：发生数值溢出异常，即结果太大而无法表示（例如，绝对值大于 2^128 的 f32）。

_MM_EXCEPT_OVERFLOW: A numeric overflow exception occured, i.e., a result was too large to be represented (e.g., an f32 with absolute value greater than 2^128).

_MM_EXCEPT_UNDERFLOW：发生数值下溢异常，即结果太小而无法以标准化方式表示（例如，绝对值小于 2^-126 的 f32。）

_MM_EXCEPT_UNDERFLOW: A numeric underflow exception occured, i.e., a result was too small to be represented in a normalized way (e.g., an f32 with absulte value smaller than 2^-126.)

_MM_EXCEPT_INEXACT：发生不精确结果异常（也称为精度异常）。 这意味着由于四舍五入而损失了一些精度。 例如，分数 1/3 不能以 32 或 64 位浮点数准确表示，计算它会导致引发此异常。 精确异常非常常见，因此它们通常被屏蔽。

_MM_EXCEPT_INEXACT: An inexact-result exception occured (a.k.a. precision exception). This means some precision was lost due to rounding. For example, the fraction 1/3 cannot be represented accurately in a 32 or 64 bit float and computing it would cause this exception to be raised. Precision exceptions are very common, so they are usually masked.

可以使用便利函数 _MM_GET_EXCEPTION_STATE 和 _MM_SET_EXCEPTION_STATE 读取和设置异常标志。 例如，检查某个操作是否导致溢出：

Exception flags can be read and set using the convenience functions _MM_GET_EXCEPTION_STATE and _MM_SET_EXCEPTION_STATE. For example, to check if an operation caused some overflow:

_MM_SET_EXCEPTION_STATE(0); // clear all exception flags
                            // perform calculations
if _MM_GET_EXCEPTION_STATE() & _MM_EXCEPT_OVERFLOW != 0 {
    // handle overflow
}
Run

Masking Flags

每个异常标志都有一个屏蔽标志：_MM_MASK_INVALID、_MM_MASK_DENORM、_MM_MASK_DIV_ZERO、_MM_MASK_OVERFLOW、_MM_MASK_UNDERFLOW、_MM_MASK_INEXACT。

There is one masking flag for each exception flag: _MM_MASK_INVALID, _MM_MASK_DENORM, _MM_MASK_DIV_ZERO, _MM_MASK_OVERFLOW, _MM_MASK_UNDERFLOW, _MM_MASK_INEXACT.

单个屏蔽位可以通过设置

A single masking bit can be set via

_MM_SET_EXCEPTION_MASK(_MM_MASK_UNDERFLOW);
Run

但是，由于默认情况下掩码位都设置为 1，因此更常见的是要禁用某些位。 例如，要取消屏蔽下溢异常，请使用：

However, since mask bits are by default all set to 1, it is more common to want to disable certain bits. For example, to unmask the underflow exception, use:

_mm_setcsr(_mm_getcsr() & !_MM_MASK_UNDERFLOW); // unmask underflow
exception
Run

警告：未屏蔽的异常将导致调用异常处理程序。 标准处理程序将简单地终止进程。 因此，在这种情况下，任何下溢异常都会终止当前进程，并使用类似信号：8，SIGFPE：错误算术运算。

Warning: an unmasked exception will cause an exception handler to be called. The standard handler will simply terminate the process. So, in this case any underflow exception would terminate the current process with something like signal: 8, SIGFPE: erroneous arithmetic operation.

Rounding Mode

使用两位描述舍入模式。 可以使用方便的包装器 _MM_GET_ROUNDING_MODE() 和 _MM_SET_ROUNDING_MODE(mode) 读取和设置它。

The rounding mode is describe using two bits. It can be read and set using the convenience wrappers _MM_GET_ROUNDING_MODE() and _MM_SET_ROUNDING_MODE(mode).

舍入模式有：

The rounding modes are:

_MM_ROUND_NEAREST：（默认）舍入到最接近无限精度值。 如果两个值同样接近，则舍入为偶数（即，最低有效位为零）。

_MM_ROUND_NEAREST: (default) Round to closest to the infinite precision value. If two values are equally close, round to even (i.e., least significant bit will be zero).

_MM_ROUND_DOWN：向负无穷大舍入。

_MM_ROUND_DOWN: Round toward negative Infinity.

_MM_ROUND_UP：向正无穷大舍入。

_MM_ROUND_UP: Round toward positive Infinity.

_MM_ROUND_TOWARD_ZERO：向零舍入（截断）。

_MM_ROUND_TOWARD_ZERO: Round towards zero (truncate).

例子：

Example:

_MM_SET_ROUNDING_MODE(_MM_ROUND_DOWN)
Run

Denormals-are-zero/Flush-to-zero Mode

如果设置了该位，则将被非规范化的值将设置为零。 这是默认关闭的。

If this bit is set, values that would be denormalized will be set to zero instead. This is turned off by default.

您可以通过辅助函数 _MM_GET_FLUSH_ZERO_MODE() 和 _MM_SET_FLUSH_ZERO_MODE() 读取和启用/禁用此模式：

You can read and enable/disable this mode via the helper functions _MM_GET_FLUSH_ZERO_MODE() and _MM_SET_FLUSH_ZERO_MODE():

_MM_SET_FLUSH_ZERO_MODE(_MM_FLUSH_ZERO_OFF); // turn off (default)
_MM_SET_FLUSH_ZERO_MODE(_MM_FLUSH_ZERO_ON); // turn on
Run

英特尔的文档

Intel’s documentation