Function core::arch::x86_64::_xrstors64

pub unsafe fn _xrstors64(mem_addr: *const u8, rs_mask: u64)

This is supported on x86-64 and target feature xsave,xsaves only.

[帮助纠正😄]使用存储在内存中 mem_addr 的状态信息执行启用的处理器状态的全部或部分恢复。

Performs a full or partial restore of the enabled processor states using the state information stored in memory at mem_addr.

xrstors 与 xrstor 的不同之处在于它可以恢复与 IA32_XSS MSR 中设置的位相对应的状态组件； xrstors 无法从扩展区域为标准格式的 xsave 区域恢复。 根据 rs_mask、XCR0 和 mem_addr.HEADER.XSTATE_BV 中的位 [62:0] 恢复状态。 mem_addr 必须在 64 字节边界上对齐。

xrstors differs from xrstor in that it can restore state components corresponding to bits set in the IA32_XSS MSR; xrstors cannot restore from an xsave area in which the extended region is in the standard form. State is restored based on bits [62:0] in rs_mask, XCR0, and mem_addr.HEADER.XSTATE_BV. mem_addr must be aligned on a 64-byte boundary.

英特尔的文档

Intel’s documentation