Function core::arch::x86_64::cmpxchg16b

pub unsafe fn cmpxchg16b(
    dst: *mut u128, 
    old: u128, 
    new: u128, 
    success: Ordering, 
    failure: Ordering
) -> u128

🔬 This is a nightly-only experimental API. (stdsimd #48556)

This is supported on x86-64 and target feature cmpxchg16b only.

[帮助纠正😄]以原子方式比较和交换 16 个字节（128 位）的数据。

Compares and exchange 16 bytes (128 bits) of data atomically.

此内在函数对应于 x86_64 处理器上的 cmpxchg16b 指令。 它执行原子比较和交换，如果内存中的当前值等于 old，则将 ptr 内存位置更新为 val。

This intrinsic corresponds to the cmpxchg16b instruction on x86_64 processors. It performs an atomic compare-and-swap, updating the ptr memory location to val if the current value in memory equals old.

Return value

此函数返回内存位置的先前值。 如果它等于旧，则内存已更新为新的。

This function returns the previous value at the memory location. If it is equal to old then the memory was updated to new.

Memory Orderings

这种原子操作与 AtomicUsize::compare_exchange 具有相同的内存排序语义，仅在 16 字节内存上操作，而不仅仅是指针。

This atomic operations has the same semantics of memory orderings as AtomicUsize::compare_exchange does, only operating on 16 bytes of memory instead of just a pointer.

有关内存排序的更多信息，请参阅标准库中其他 Atomic* 类型的 compare_exchange 文档。

For more information on memory orderings here see the compare_exchange documentation for other Atomic* types in the standard library.

Unsafety

此方法是不安全的，因为它采用原始指针并将尝试读取并可能写入指针处的内存。 指针还必须在 16 字节边界上对齐。

This method is unsafe because it takes a raw pointer and will attempt to read and possibly write the memory at the pointer. The pointer must also be aligned on a 16-byte boundary.

此方法还需要 cmpxchg16b CPU 功能在运行时可用才能正常工作。 如果运行二进制文件的 CPU 实际上不支持 cmpxchg16b 并且程序进入最终将到达此函数的执行路径，则行为未定义。

This method also requires the cmpxchg16b CPU feature to be available at runtime to work correctly. If the CPU running the binary does not actually support cmpxchg16b and the program enters an execution path that eventually would reach this function the behavior is undefined.

成功排序也必须更强或等于失败，否则此函数调用未定义。 有关详细信息，请参阅 Atomic* 文档的 compare_exchange 函数。 当 compare_exchange 发生panic时，这是未定义的行为。 当前，此函数使用未定义的指令中止进程。

The success ordering must also be stronger or equal to failure, or this function call is undefined. See the Atomic* documentation’s compare_exchange function for more information. When compare_exchange panics, this is undefined behavior. Currently this function aborts the process with an undefined instruction.