core::arch::x86_64

Struct __m128
ยท

pub struct __m128(/* private fields */);

Available on x86-64 only.

128-bit wide set of four f32 types, x86-specific

This type is the same as the __m128 type defined by Intel, representing a 128-bit SIMD register which internally is consisted of four packed f32 instances. Usage of this type typically corresponds to the sse and up target features for x86/x86_64.

Note that unlike __m128i, the integer version of the 128-bit registers, this __m128 type has one interpretation. Each instance of __m128 always corresponds to f32x4, or four f32 types packed together.

The in-memory representation of this type is the same as the one of an equivalent array (i.e. the in-memory order of elements is the same, and there is no padding); however, the alignment is different and equal to the size of the type. Note that the ABI for function calls may not be the same.

Most intrinsics using __m128 are prefixed with _mm_ and are suffixed with โpsโ (or otherwise contain โpsโ). Not to be confused with โpdโ which is used for __m128d.

Examples

#[cfg(target_arch = "x86")]
use std::arch::x86::*;
#[cfg(target_arch = "x86_64")]
use std::arch::x86_64::*;

let four_zeros = _mm_setzero_ps();
let four_ones = _mm_set1_ps(1.0);
let four_floats = _mm_set_ps(1.0, 2.0, 3.0, 4.0);