Module core::array

Unstable (fixed_size_array #27778)
: traits and impls are better expressed through generic integer constants

Implementations of things like Eq for fixed-length arrays up to a certain length. Eventually we should able to generalize to all lengths.

See also the array primitive type.

Traits

FixedSizeArray 	 [Unstable]
Utility trait implemented only on arrays of fixed size