Trait core::array::FixedSizeArray

pub trait FixedSizeArray<T> {
    fn as_slice(&self) -> &[T];
    fn as_mut_slice(&mut self) -> &mut [T];
}

Unstable (fixed_size_array #27778)
: traits and impls are better expressed through generic integer constants

Utility trait implemented only on arrays of fixed size

This trait can be used to implement other traits on fixed-size arrays without causing much metadata bloat.

Required Methods

fn as_slice(&self) -> &[T]

Unstable (fixed_size_array #27778)
: traits and impls are better expressed through generic integer constants

Converts the array to immutable slice

fn as_mut_slice(&mut self) -> &mut [T]

Unstable (fixed_size_array #27778)
: traits and impls are better expressed through generic integer constants

Converts the array to mutable slice