Module core::ascii

对 ASCII 字符串和字符的操作。

Operations on ASCII strings and characters.

Rust 中的大多数字符串操作都作用于 UTF-8 字符串。 但是，有时只考虑特定操作的 ASCII 字符集更有意义。

Most string operations in Rust act on UTF-8 strings. However, at times it makes more sense to only consider the ASCII character set for a specific operation.

escape_default 函数为给定字符的转义版本的字节提供了一个迭代器。

The escape_default function provides an iterator over the bytes of an escaped version of the character given.

Structs

EscapeDefault

An iterator over the escaped version of a byte.

Functions

escape_default

Returns an iterator that produces an escaped version of a u8.