Enum core::cell::BorrowState

pub enum BorrowState {
    Reading,
    Writing,
    Unused,
}

Unstable (borrow_state #27733)

An enumeration of values returned from the state method on a RefCell<T>.

Variants

Reading
Unstable (borrow_state #27733)

The cell is currently being read, there is at least one active borrow.

Writing
Unstable (borrow_state #27733)

The cell is currently being written to, there is an active borrow_mut.

Unused
Unstable (borrow_state #27733)

There are no outstanding borrows on this cell.