Function core::cell::clone_ref

pub fn clone_ref<'b, T: Clone>(orig: &Ref<'b, T>) -> Ref<'b, T>

Deprecated since 1.2.0
: moved to a Ref::clone associated function

Copies a Ref.

The RefCell is already immutably borrowed, so this cannot fail.

A Clone implementation would interfere with the widespread use of r.borrow().clone() to clone the contents of a RefCell.