Struct core::cell::RefCell

pub struct RefCell<T: ?Sized> {
    // some fields omitted
}

A mutable memory location with dynamically checked borrow rules

See the module-level documentation for more.

Methods

impl<T> RefCell<T>

const fn new(value: T) -> RefCell<T>

Creates a new RefCell containing value.

Examples
fn main() { use std::cell::RefCell; let c = RefCell::new(5); }

use std::cell::RefCell;

let c = RefCell::new(5);

fn into_inner(self) -> T

Consumes the RefCell, returning the wrapped value.

Examples
fn main() { use std::cell::RefCell; let c = RefCell::new(5); let five = c.into_inner(); }

use std::cell::RefCell;

let c = RefCell::new(5);

let five = c.into_inner();

impl<T: ?Sized> RefCell<T>

fn borrow_state(&self) -> BorrowState

Unstable (borrow_state #27733)

Query the current state of this RefCell

The returned value can be dispatched on to determine if a call to borrow or borrow_mut would succeed.

Examples
#![feature(borrow_state)] fn main() { use std::cell::{BorrowState, RefCell}; let c = RefCell::new(5); match c.borrow_state() { BorrowState::Writing => println!("Cannot be borrowed"), BorrowState::Reading => println!("Cannot be borrowed mutably"), BorrowState::Unused => println!("Can be borrowed (mutably as well)"), } }

#![feature(borrow_state)]

use std::cell::{BorrowState, RefCell};

let c = RefCell::new(5);

match c.borrow_state() {
    BorrowState::Writing => println!("Cannot be borrowed"),
    BorrowState::Reading => println!("Cannot be borrowed mutably"),
    BorrowState::Unused => println!("Can be borrowed (mutably as well)"),
}

fn borrow(&self) -> Ref<T>

Immutably borrows the wrapped value.

The borrow lasts until the returned Ref exits scope. Multiple immutable borrows can be taken out at the same time.

Panics

Panics if the value is currently mutably borrowed. For a non-panicking variant, use try_borrow.

Examples
fn main() { use std::cell::RefCell; let c = RefCell::new(5); let borrowed_five = c.borrow(); let borrowed_five2 = c.borrow(); }

use std::cell::RefCell;

let c = RefCell::new(5);

let borrowed_five = c.borrow();
let borrowed_five2 = c.borrow();

An example of panic:
fn main() { use std::cell::RefCell; use std::thread; let result = thread::spawn(move || { let c = RefCell::new(5); let m = c.borrow_mut(); let b = c.borrow(); // this causes a panic }).join(); assert!(result.is_err()); }

use std::cell::RefCell;
use std::thread;

let result = thread::spawn(move || {
   let c = RefCell::new(5);
   let m = c.borrow_mut();

   let b = c.borrow(); // this causes a panic
}).join();

assert!(result.is_err());

fn try_borrow(&self) -> Result<Ref<T>, BorrowError<T>>

Unstable (try_borrow #35070)

Immutably borrows the wrapped value, returning an error if the value is currently mutably borrowed.

The borrow lasts until the returned Ref exits scope. Multiple immutable borrows can be taken out at the same time.

This is the non-panicking variant of borrow.

Examples
#![feature(try_borrow)] fn main() { use std::cell::RefCell; let c = RefCell::new(5); { let m = c.borrow_mut(); assert!(c.try_borrow().is_err()); } { let m = c.borrow(); assert!(c.try_borrow().is_ok()); } }

#![feature(try_borrow)]

use std::cell::RefCell;

let c = RefCell::new(5);

{
    let m = c.borrow_mut();
    assert!(c.try_borrow().is_err());
}

{
    let m = c.borrow();
    assert!(c.try_borrow().is_ok());
}

fn borrow_mut(&self) -> RefMut<T>

Mutably borrows the wrapped value.

The borrow lasts until the returned RefMut exits scope. The value cannot be borrowed while this borrow is active.

Panics

Panics if the value is currently borrowed. For a non-panicking variant, use try_borrow_mut.

Examples
fn main() { use std::cell::RefCell; let c = RefCell::new(5); *c.borrow_mut() = 7; assert_eq!(*c.borrow(), 7); }

use std::cell::RefCell;

let c = RefCell::new(5);

*c.borrow_mut() = 7;

assert_eq!(*c.borrow(), 7);

An example of panic:
fn main() { use std::cell::RefCell; use std::thread; let result = thread::spawn(move || { let c = RefCell::new(5); let m = c.borrow(); let b = c.borrow_mut(); // this causes a panic }).join(); assert!(result.is_err()); }

use std::cell::RefCell;
use std::thread;

let result = thread::spawn(move || {
   let c = RefCell::new(5);
   let m = c.borrow();

   let b = c.borrow_mut(); // this causes a panic
}).join();

assert!(result.is_err());

fn try_borrow_mut(&self) -> Result<RefMut<T>, BorrowMutError<T>>

Unstable (try_borrow #35070)

Mutably borrows the wrapped value, returning an error if the value is currently borrowed.

The borrow lasts until the returned RefMut exits scope. The value cannot be borrowed while this borrow is active.

This is the non-panicking variant of borrow_mut.

Examples
#![feature(try_borrow)] fn main() { use std::cell::RefCell; let c = RefCell::new(5); { let m = c.borrow(); assert!(c.try_borrow_mut().is_err()); } assert!(c.try_borrow_mut().is_ok()); }

#![feature(try_borrow)]

use std::cell::RefCell;

let c = RefCell::new(5);

{
    let m = c.borrow();
    assert!(c.try_borrow_mut().is_err());
}

assert!(c.try_borrow_mut().is_ok());

unsafe fn as_unsafe_cell(&self) -> &UnsafeCell<T>

Deprecated since 1.12.0
: renamed to as_ptr

Unstable (as_unsafe_cell #27708)
: renamed to as_ptr

Returns a reference to the underlying UnsafeCell.

This can be used to circumvent RefCell's safety checks.

This function is unsafe because UnsafeCell's field is public.

Examples
#![feature(as_unsafe_cell)] fn main() { use std::cell::RefCell; let c = RefCell::new(5); let c = unsafe { c.as_unsafe_cell() }; }

#![feature(as_unsafe_cell)]

use std::cell::RefCell;

let c = RefCell::new(5);
let c = unsafe { c.as_unsafe_cell() };

fn as_ptr(&self) -> *mut T

Returns a raw pointer to the underlying data in this cell.

Examples
fn main() { use std::cell::RefCell; let c = RefCell::new(5); let ptr = c.as_ptr(); }

use std::cell::RefCell;

let c = RefCell::new(5);

let ptr = c.as_ptr();

fn get_mut(&mut self) -> &mut T

Returns a mutable reference to the underlying data.

This call borrows RefCell mutably (at compile-time) so there is no need for dynamic checks.

Examples
fn main() { use std::cell::RefCell; let mut c = RefCell::new(5); *c.get_mut() += 1; assert_eq!(c, RefCell::new(6)); }

use std::cell::RefCell;

let mut c = RefCell::new(5);
*c.get_mut() += 1;

assert_eq!(c, RefCell::new(6));