Struct core::cell::RefMut

pub struct RefMut<'b, T: ?Sized + 'b> {
    // some fields omitted
}

A wrapper type for a mutably borrowed value from a RefCell<T>.

See the module-level documentation for more.

Methods

impl<'b, T: ?Sized> RefMut<'b, T>

fn map<U: ?Sized, F>(orig: RefMut<'b, T>, f: F) -> RefMut<'b, U> where F: FnOnce(&mut T) -> &mut U

Unstable (cell_extras #27746)
: recently added

Make a new RefMut for a component of the borrowed data, e.g. an enum variant.

The RefCell is already mutably borrowed, so this cannot fail.

This is an associated function that needs to be used as RefMut::map(...). A method would interfere with methods of the same name on the contents of a RefCell used through Deref.

Example
#![feature(cell_extras)] fn main() { use std::cell::{RefCell, RefMut}; let c = RefCell::new((5, 'b')); { let b1: RefMut<(u32, char)> = c.borrow_mut(); let mut b2: RefMut<u32> = RefMut::map(b1, |t| &mut t.0); assert_eq!(*b2, 5); *b2 = 42; } assert_eq!(*c.borrow(), (42, 'b')); }

use std::cell::{RefCell, RefMut};

let c = RefCell::new((5, 'b'));
{
    let b1: RefMut<(u32, char)> = c.borrow_mut();
    let mut b2: RefMut<u32> = RefMut::map(b1, |t| &mut t.0);
    assert_eq!(*b2, 5);
    *b2 = 42;
}
assert_eq!(*c.borrow(), (42, 'b'));

fn filter_map<U: ?Sized, F>(orig: RefMut<'b, T>, f: F) -> Option<RefMut<'b, U>> where F: FnOnce(&mut T) -> Option<&mut U>

Unstable (cell_extras #27746)
: recently added

Make a new RefMut for a optional component of the borrowed data, e.g. an enum variant.

The RefCell is already mutably borrowed, so this cannot fail.

This is an associated function that needs to be used as RefMut::filter_map(...). A method would interfere with methods of the same name on the contents of a RefCell used through Deref.

Example
#![feature(cell_extras)] fn main() { use std::cell::{RefCell, RefMut}; let c = RefCell::new(Ok(5)); { let b1: RefMut<Result<u32, ()>> = c.borrow_mut(); let mut b2: RefMut<u32> = RefMut::filter_map(b1, |o| { o.as_mut().ok() }).unwrap(); assert_eq!(*b2, 5); *b2 = 42; } assert_eq!(*c.borrow(), Ok(42)); }

use std::cell::{RefCell, RefMut};

let c = RefCell::new(Ok(5));
{
    let b1: RefMut<Result<u32, ()>> = c.borrow_mut();
    let mut b2: RefMut<u32> = RefMut::filter_map(b1, |o| {
        o.as_mut().ok()
    }).unwrap();
    assert_eq!(*b2, 5);
    *b2 = 42;
}
assert_eq!(*c.borrow(), Ok(42));