Module core::char

Character manipulation.

For more details, see ::std_unicode::char (a.k.a. std::char)

Structs

EscapeDefault
An iterator that yields the literal escape code of a char.

EscapeUnicode
Returns an iterator that yields the hexadecimal Unicode escape of a character, as chars.

CharTryFromError 	 [
Experimental
]
The error type returned when a conversion from u32 to char fails.

DecodeUtf8 	 [
Experimental
]
An iterator over an iterator of bytes of the characters the bytes represent as UTF-8

EscapeDebug 	 [
Experimental
]
An iterator that yields the literal escape code of a char.

InvalidSequence 	 [
Experimental
]
<DecodeUtf8 as Iterator>::next returns this for an invalid input sequence.

Constants

MAX
The highest valid code point a char can have.

Functions

from_digit
Converts a digit in the given radix to a char.

from_u32
Converts a u32 to a char.

from_u32_unchecked⚠
Converts a u32 to a char, ignoring validity.

decode_utf8 	 [
Experimental
]
Decodes an Iterator of bytes as UTF-8.