Function core::char::from_digit

pub fn from_digit(num: u32, radix: u32) -> Option<char>

将给定基数中的数字转换为 `char`。

Converts a digit in the given radix to a char.

这里的“基数”有时也称为“基数”。 基数为 2 表示二进制数，基数为 10 表示十进制数，基数为 16 表示十六进制数，以给出一些常用值。 支持任意基数。

A ‘radix’ here is sometimes also called a ‘base’. A radix of two indicates a binary number, a radix of ten, decimal, and a radix of sixteen, hexadecimal, to give some common values. Arbitrary radices are supported.

如果输入不是给定基数中的数字，`from_digit()` 将返回 `None`。

from_digit() will return None if the input is not a digit in the given radix.

Panics

如果给定大于 36 的基数，则会出现panic。

Panics if given a radix larger than 36.

Examples

基本用法：

Basic usage:

use std::char;

let c = char::from_digit(4, 10);

assert_eq!(Some('4'), c);

// Decimal 11 is a single digit in base 16
let c = char::from_digit(11, 16);

assert_eq!(Some('b'), c);
Run

当输入不是数字时返回 `None`：

Returning None when the input is not a digit:

use std::char;

let c = char::from_digit(20, 10);

assert_eq!(None, c);
Run

传递一个大基数，引起panic：

Passing a large radix, causing a panic:

use std::char;

// this panics
let c = char::from_digit(1, 37);
Run