Function core::char::to_digitDeprecated

pub fn to_digit(c: char, radix: uint) -> Option<uint>

Converts a char to the corresponding digit

Return value

If c is between '0' and '9', the corresponding value between 0 and 9. If c is 'a' or 'A', 10. If c is 'b' or 'B', 11, etc. Returns none if the char does not refer to a digit in the given radix.

Panics

Panics if given a radix outside the range [0..36].