Function core::cmp::partial_minUnstable

pub fn partial_min<T: PartialOrd>(v1: T, v2: T) -> Option<T>

Compare and return the minimum of two values if there is one.

Returns the first argument if the comparison determines them to be equal.