Struct core::cmp::Reverse

pub struct Reverse<T>(pub T);

A helper struct for reverse ordering.

This struct is a helper to be used with functions like Vec::sort_by_key and can be used to reverse order a part of a key.

Example usage:

use std::cmp::Reverse;

let mut v = vec![1, 2, 3, 4, 5, 6];
v.sort_by_key(|&num| (num > 3, Reverse(num)));
assert_eq!(v, vec![3, 2, 1, 6, 5, 4]);Run