Module core::convert

Traits for conversions between types.

The traits in this module provide a general way to talk about conversions from one type to another. They follow the standard Rust conventions of as/into/from.

Like many traits, these are often used as bounds for generic functions, to support arguments of multiple types.

See each trait for usage examples.

Traits

AsMut
A cheap, mutable reference-to-mutable reference conversion.

AsRef
A cheap, reference-to-reference conversion.

From
Construct Self via a conversion.

Into
A conversion that consumes self, which may or may not be expensive.