Module core::core_arch::arm::neon

🔬This is a nightly-only experimental API. (stdsimd #48556)

Available on ARM only.

Functions

vbsl_s8_ 🔒 ⚠Experimental

vbslq_s8_ 🔒 ⚠Experimental

vcvtq_s32_f32⚠Experimentalneon and v7

Floating-point Convert to Signed fixed-point, rounding toward Zero (vector)

vcvtq_u32_f32⚠Experimentalneon and v7

Floating-point Convert to Unsigned fixed-point, rounding toward Zero (vector)

vld1_f32⚠Experimentalneon,v7

Load multiple single-element structures to one, two, three, or four registers.

vld1_p8⚠Experimentalneon,v7

Load multiple single-element structures to one, two, three, or four registers.

vld1_p16⚠Experimentalneon,v7

Load multiple single-element structures to one, two, three, or four registers.

vld1_p64⚠Experimentalneon,aes

Load multiple single-element structures to one, two, three, or four registers.

vld1_s8⚠Experimentalneon,v7

Load multiple single-element structures to one, two, three, or four registers.

vld1_s16⚠Experimentalneon,v7

Load multiple single-element structures to one, two, three, or four registers.

vld1_s32⚠Experimentalneon,v7

Load multiple single-element structures to one, two, three, or four registers.

vld1_s64⚠Experimentalneon,v7

Load multiple single-element structures to one, two, three, or four registers.

vld1_u8⚠Experimentalneon,v7

Load multiple single-element structures to one, two, three, or four registers.

vld1_u16⚠Experimentalneon,v7

Load multiple single-element structures to one, two, three, or four registers.

vld1_u32⚠Experimentalneon,v7

Load multiple single-element structures to one, two, three, or four registers.

vld1_u64⚠Experimentalneon,v7

Load multiple single-element structures to one, two, three, or four registers.

vld1_v1i64 🔒 ⚠Experimental

vld1_v2f32 🔒 ⚠Experimental

vld1_v2i32 🔒 ⚠Experimental

vld1_v4i16 🔒 ⚠Experimental

vld1_v8i8 🔒 ⚠Experimental

vld1q_f32⚠Experimentalneon,v7

Load multiple single-element structures to one, two, three, or four registers.

vld1q_p8⚠Experimentalneon,v7

Load multiple single-element structures to one, two, three, or four registers.

vld1q_p16⚠Experimentalneon,v7

Load multiple single-element structures to one, two, three, or four registers.

vld1q_p64⚠Experimentalneon,aes

Load multiple single-element structures to one, two, three, or four registers.

vld1q_s8⚠Experimentalneon,v7

Load multiple single-element structures to one, two, three, or four registers.

vld1q_s16⚠Experimentalneon,v7

Load multiple single-element structures to one, two, three, or four registers.

vld1q_s32⚠Experimentalneon,v7

Load multiple single-element structures to one, two, three, or four registers.

vld1q_s64⚠Experimentalneon,v7

Load multiple single-element structures to one, two, three, or four registers.

vld1q_u8⚠Experimentalneon,v7

Load multiple single-element structures to one, two, three, or four registers.

vld1q_u16⚠Experimentalneon,v7

Load multiple single-element structures to one, two, three, or four registers.

vld1q_u32⚠Experimentalneon,v7

Load multiple single-element structures to one, two, three, or four registers.

vld1q_u64⚠Experimentalneon,v7

Load multiple single-element structures to one, two, three, or four registers.

vld1q_v2i64 🔒 ⚠Experimental

vld1q_v4f32 🔒 ⚠Experimental

vld1q_v4i32 🔒 ⚠Experimental

vld1q_v8i16 🔒 ⚠Experimental

vld1q_v16i8 🔒 ⚠Experimental

vpadal_s8_ 🔒 ⚠Experimental

vpadal_s16_ 🔒 ⚠Experimental

vpadal_s32_ 🔒 ⚠Experimental

vpadal_u8_ 🔒 ⚠Experimental

vpadal_u16_ 🔒 ⚠Experimental

vpadal_u32_ 🔒 ⚠Experimental

vpadalq_s8_ 🔒 ⚠Experimental

vpadalq_s16_ 🔒 ⚠Experimental

vpadalq_s32_ 🔒 ⚠Experimental

vpadalq_u8_ 🔒 ⚠Experimental

vpadalq_u16_ 🔒 ⚠Experimental

vpadalq_u32_ 🔒 ⚠Experimental

vshiftins_v1i64 🔒 ⚠Experimental

vshiftins_v2i32 🔒 ⚠Experimental

vshiftins_v2i64 🔒 ⚠Experimental

vshiftins_v4i16 🔒 ⚠Experimental

vshiftins_v4i32 🔒 ⚠Experimental

vshiftins_v8i8 🔒 ⚠Experimental

vshiftins_v8i16 🔒 ⚠Experimental

vshiftins_v16i8 🔒 ⚠Experimental

vsli_n_p8⚠Experimentalneon,v7

Shift Left and Insert (immediate)

vsli_n_p16⚠Experimentalneon,v7

Shift Left and Insert (immediate)

vsli_n_p64⚠Experimentalneon,v7,aes

Shift Left and Insert (immediate)

vsli_n_s8⚠Experimentalneon,v7

Shift Left and Insert (immediate)

vsli_n_s16⚠Experimentalneon,v7

Shift Left and Insert (immediate)

vsli_n_s32⚠Experimentalneon,v7

Shift Left and Insert (immediate)

vsli_n_s64⚠Experimentalneon,v7

Shift Left and Insert (immediate)

vsli_n_u8⚠Experimentalneon,v7

Shift Left and Insert (immediate)

vsli_n_u16⚠Experimentalneon,v7

Shift Left and Insert (immediate)

vsli_n_u32⚠Experimentalneon,v7

Shift Left and Insert (immediate)

vsli_n_u64⚠Experimentalneon,v7

Shift Left and Insert (immediate)

vsliq_n_p8⚠Experimentalneon,v7

Shift Left and Insert (immediate)

vsliq_n_p16⚠Experimentalneon,v7

Shift Left and Insert (immediate)

vsliq_n_p64⚠Experimentalneon,v7,aes

Shift Left and Insert (immediate)

vsliq_n_s8⚠Experimentalneon,v7

Shift Left and Insert (immediate)

vsliq_n_s16⚠Experimentalneon,v7

Shift Left and Insert (immediate)

vsliq_n_s32⚠Experimentalneon,v7

Shift Left and Insert (immediate)

vsliq_n_s64⚠Experimentalneon,v7

Shift Left and Insert (immediate)

vsliq_n_u8⚠Experimentalneon,v7

Shift Left and Insert (immediate)

vsliq_n_u16⚠Experimentalneon,v7

Shift Left and Insert (immediate)

vsliq_n_u32⚠Experimentalneon,v7

Shift Left and Insert (immediate)

vsliq_n_u64⚠Experimentalneon,v7

Shift Left and Insert (immediate)

vsri_n_p8⚠Experimentalneon,v7

Shift Right and Insert (immediate)

vsri_n_p16⚠Experimentalneon,v7

Shift Right and Insert (immediate)

vsri_n_p64⚠Experimentalneon,v7,aes

Shift Right and Insert (immediate)

vsri_n_s8⚠Experimentalneon,v7

Shift Right and Insert (immediate)

vsri_n_s16⚠Experimentalneon,v7

Shift Right and Insert (immediate)

vsri_n_s32⚠Experimentalneon,v7

Shift Right and Insert (immediate)

vsri_n_s64⚠Experimentalneon,v7

Shift Right and Insert (immediate)

vsri_n_u8⚠Experimentalneon,v7

Shift Right and Insert (immediate)

vsri_n_u16⚠Experimentalneon,v7

Shift Right and Insert (immediate)

vsri_n_u32⚠Experimentalneon,v7

Shift Right and Insert (immediate)

vsri_n_u64⚠Experimentalneon,v7

Shift Right and Insert (immediate)

vsriq_n_p8⚠Experimentalneon,v7

Shift Right and Insert (immediate)

vsriq_n_p16⚠Experimentalneon,v7

Shift Right and Insert (immediate)

vsriq_n_p64⚠Experimentalneon,v7,aes

Shift Right and Insert (immediate)

vsriq_n_s8⚠Experimentalneon,v7

Shift Right and Insert (immediate)

vsriq_n_s16⚠Experimentalneon,v7

Shift Right and Insert (immediate)

vsriq_n_s32⚠Experimentalneon,v7

Shift Right and Insert (immediate)

vsriq_n_s64⚠Experimentalneon,v7

Shift Right and Insert (immediate)

vsriq_n_u8⚠Experimentalneon,v7

Shift Right and Insert (immediate)

vsriq_n_u16⚠Experimentalneon,v7

Shift Right and Insert (immediate)

vsriq_n_u32⚠Experimentalneon,v7

Shift Right and Insert (immediate)

vsriq_n_u64⚠Experimentalneon,v7

Shift Right and Insert (immediate)

vst1_f32⚠Experimentalneon,v7

vst1_p8⚠Experimentalneon,v7

Store multiple single-element structures from one, two, three, or four registers.

vst1_p16⚠Experimentalneon,v7

Store multiple single-element structures from one, two, three, or four registers.

vst1_p64⚠Experimentalneon,aes,v8

Store multiple single-element structures from one, two, three, or four registers.

vst1_s8⚠Experimentalneon,v7

Store multiple single-element structures from one, two, three, or four registers.

vst1_s16⚠Experimentalneon,v7

Store multiple single-element structures from one, two, three, or four registers.

vst1_s32⚠Experimentalneon,v7

Store multiple single-element structures from one, two, three, or four registers.

vst1_s64⚠Experimentalneon,v7

Store multiple single-element structures from one, two, three, or four registers.

vst1_u8⚠Experimentalneon,v7

Store multiple single-element structures from one, two, three, or four registers.

vst1_u16⚠Experimentalneon,v7

Store multiple single-element structures from one, two, three, or four registers.

vst1_u32⚠Experimentalneon,v7

Store multiple single-element structures from one, two, three, or four registers.

vst1_u64⚠Experimentalneon,v7

Store multiple single-element structures from one, two, three, or four registers.

vst1_v1i64 🔒 ⚠Experimental

vst1_v2f32 🔒 ⚠Experimental

vst1_v2i32 🔒 ⚠Experimental

vst1_v4i16 🔒 ⚠Experimental

vst1_v8i8 🔒 ⚠Experimental

vst1q_f32⚠Experimentalneon,v7

vst1q_p8⚠Experimentalneon,v7

Store multiple single-element structures from one, two, three, or four registers.

vst1q_p16⚠Experimentalneon,v7

Store multiple single-element structures from one, two, three, or four registers.

vst1q_p64⚠Experimentalneon,aes,v8

Store multiple single-element structures from one, two, three, or four registers.

vst1q_s8⚠Experimentalneon,v7

Store multiple single-element structures from one, two, three, or four registers.

vst1q_s16⚠Experimentalneon,v7

Store multiple single-element structures from one, two, three, or four registers.

vst1q_s32⚠Experimentalneon,v7

Store multiple single-element structures from one, two, three, or four registers.

vst1q_s64⚠Experimentalneon,v7

Store multiple single-element structures from one, two, three, or four registers.

vst1q_u8⚠Experimentalneon,v7

Store multiple single-element structures from one, two, three, or four registers.

vst1q_u16⚠Experimentalneon,v7

Store multiple single-element structures from one, two, three, or four registers.

vst1q_u32⚠Experimentalneon,v7

Store multiple single-element structures from one, two, three, or four registers.

vst1q_u64⚠Experimentalneon,v7

Store multiple single-element structures from one, two, three, or four registers.

vst1q_v2i64 🔒 ⚠Experimental

vst1q_v4f32 🔒 ⚠Experimental

vst1q_v4i32 🔒 ⚠Experimental

vst1q_v8i16 🔒 ⚠Experimental

vst1q_v16i8 🔒 ⚠Experimental

vtbl1 🔒 ⚠Experimental

vtbl1_p8⚠Experimentalneon,v7

Table look-up

vtbl1_s8⚠Experimentalneon,v7

Table look-up

vtbl1_u8⚠Experimentalneon,v7

Table look-up

vtbl2 🔒 ⚠Experimental

vtbl2_p8⚠Experimentalneon,v7

Table look-up

vtbl2_s8⚠Experimentalneon,v7

Table look-up

vtbl2_u8⚠Experimentalneon,v7

Table look-up

vtbl3 🔒 ⚠Experimental

vtbl3_p8⚠Experimentalneon,v7

Table look-up

vtbl3_s8⚠Experimentalneon,v7

Table look-up

vtbl3_u8⚠Experimentalneon,v7

Table look-up

vtbl4 🔒 ⚠Experimental

vtbl4_p8⚠Experimentalneon,v7

Table look-up

vtbl4_s8⚠Experimentalneon,v7

Table look-up

vtbl4_u8⚠Experimentalneon,v7

Table look-up

vtbx1 🔒 ⚠Experimental

vtbx1_p8⚠Experimentalneon,v7

Extended table look-up

vtbx1_s8⚠Experimentalneon,v7

Extended table look-up

vtbx1_u8⚠Experimentalneon,v7

Extended table look-up

vtbx2 🔒 ⚠Experimental

vtbx2_p8⚠Experimentalneon,v7

Extended table look-up

vtbx2_s8⚠Experimentalneon,v7

Extended table look-up

vtbx2_u8⚠Experimentalneon,v7

Extended table look-up

vtbx3 🔒 ⚠Experimental

vtbx3_p8⚠Experimentalneon,v7

Extended table look-up

vtbx3_s8⚠Experimentalneon,v7

Extended table look-up

vtbx3_u8⚠Experimentalneon,v7

Extended table look-up

vtbx4 🔒 ⚠Experimental

vtbx4_p8⚠Experimentalneon,v7

Extended table look-up

vtbx4_s8⚠Experimentalneon,v7

Extended table look-up

vtbx4_u8⚠Experimentalneon,v7

Extended table look-up

Type Definitions

p8 🔒 Experimental

p16 🔒 Experimental