Module core::core_arch::arm_shared::neon

🔬This is a nightly-only experimental API. (stdsimd #48556)

ARMv7 NEON intrinsics

Re-exports

pub use self::generated::*;

Experimental

Modules

generated 🔒 Experimental

Structs

float32x2_t

ARM-specific 64-bit wide vector of two packed f32.

float32x2x2_t

ARM-specific type containing two float32x2_t vectors.

float32x2x3_t

ARM-specific type containing three float32x2_t vectors.

float32x2x4_t

ARM-specific type containing four float32x2_t vectors.

float32x4_t

ARM-specific 128-bit wide vector of four packed f32.

float32x4x2_t

ARM-specific type containing two float32x4_t vectors.

float32x4x3_t

ARM-specific type containing three float32x4_t vectors.

float32x4x4_t

ARM-specific type containing four float32x4_t vectors.

int8x8_t

ARM-specific 64-bit wide vector of eight packed i8.

int8x8x2_t

ARM-specific type containing two int8x8_t vectors.

int8x8x3_t

ARM-specific type containing three int8x8_t vectors.

int8x8x4_t

ARM-specific type containing four int8x8_t vectors.

int8x16_t

ARM-specific 128-bit wide vector of sixteen packed i8.

int8x16x2_t

ARM-specific type containing two int8x16_t vectors.

int8x16x3_t

ARM-specific type containing three int8x16_t vectors.

int8x16x4_t

ARM-specific type containing four int8x16_t vectors.

int16x4_t

ARM-specific 64-bit wide vector of four packed i16.

int16x4x2_t

ARM-specific type containing two int16x4_t vectors.

int16x4x3_t

ARM-specific type containing three int16x4_t vectors.

int16x4x4_t

ARM-specific type containing four int16x4_t vectors.

int16x8_t

ARM-specific 128-bit wide vector of eight packed i16.

int16x8x2_t

ARM-specific type containing two int16x8_t vectors.

int16x8x3_t

ARM-specific type containing three int16x8_t vectors.

int16x8x4_t

ARM-specific type containing four int16x8_t vectors.

int32x2_t

ARM-specific 64-bit wide vector of two packed i32.

int32x2x2_t

ARM-specific type containing two int32x2_t vectors.

int32x2x3_t

ARM-specific type containing three int32x2_t vectors.

int32x2x4_t

ARM-specific type containing four int32x2_t vectors.

int32x4_t

ARM-specific 128-bit wide vector of four packed i32.

int32x4x2_t

ARM-specific type containing two int32x4_t vectors.

int32x4x3_t

ARM-specific type containing three int32x4_t vectors.

int32x4x4_t

ARM-specific type containing four int32x4_t vectors.

int64x1_t

ARM-specific 64-bit wide vector of one packed i64.

int64x1x2_t

ARM-specific type containing four int64x1_t vectors.

int64x1x3_t

ARM-specific type containing four int64x1_t vectors.

int64x1x4_t

ARM-specific type containing four int64x1_t vectors.

int64x2_t

ARM-specific 128-bit wide vector of two packed i64.

int64x2x2_t

ARM-specific type containing four int64x2_t vectors.

int64x2x3_t

ARM-specific type containing four int64x2_t vectors.

int64x2x4_t

ARM-specific type containing four int64x2_t vectors.

poly8x8_t

ARM-specific 64-bit wide polynomial vector of eight packed p8.

poly8x8x2_t

ARM-specific type containing two poly8x8_t vectors.

poly8x8x3_t

ARM-specific type containing three poly8x8_t vectors.

poly8x8x4_t

ARM-specific type containing four poly8x8_t vectors.

poly8x16_t

ARM-specific 128-bit wide vector of sixteen packed p8.

poly8x16x2_t

ARM-specific type containing two poly8x16_t vectors.

poly8x16x3_t

ARM-specific type containing three poly8x16_t vectors.

poly8x16x4_t

ARM-specific type containing four poly8x16_t vectors.

poly16x4_t

ARM-specific 64-bit wide vector of four packed p16.

poly16x4x2_t

ARM-specific type containing two poly16x4_t vectors.

poly16x4x3_t

ARM-specific type containing three poly16x4_t vectors.

poly16x4x4_t

ARM-specific type containing four poly16x4_t vectors.

poly16x8_t

ARM-specific 128-bit wide vector of eight packed p16.

poly16x8x2_t

ARM-specific type containing two poly16x8_t vectors.

poly16x8x3_t

ARM-specific type containing three poly16x8_t vectors.

poly16x8x4_t

ARM-specific type containing four poly16x8_t vectors.

poly64x1_t

ARM-specific 64-bit wide vector of one packed p64.

poly64x1x2_t

ARM-specific type containing four poly64x1_t vectors.

poly64x1x3_t

ARM-specific type containing four poly64x1_t vectors.

poly64x1x4_t

ARM-specific type containing four poly64x1_t vectors.

poly64x2_t

ARM-specific 128-bit wide vector of two packed p64.

poly64x2x2_t

ARM-specific type containing four poly64x2_t vectors.

poly64x2x3_t

ARM-specific type containing four poly64x2_t vectors.

poly64x2x4_t

ARM-specific type containing four poly64x2_t vectors.

uint8x8_t

ARM-specific 64-bit wide vector of eight packed u8.

uint8x8x2_t

ARM-specific type containing two uint8x8_t vectors.

uint8x8x3_t

ARM-specific type containing three uint8x8_t vectors.

uint8x8x4_t

ARM-specific type containing four uint8x8_t vectors.

uint8x16_t

ARM-specific 128-bit wide vector of sixteen packed u8.

uint8x16x2_t

ARM-specific type containing two uint8x16_t vectors.

uint8x16x3_t

ARM-specific type containing three uint8x16_t vectors.

uint8x16x4_t

ARM-specific type containing four uint8x16_t vectors.

uint16x4_t

ARM-specific 64-bit wide vector of four packed u16.

uint16x4x2_t

ARM-specific type containing two uint16x4_t vectors.

uint16x4x3_t

ARM-specific type containing three uint16x4_t vectors.

uint16x4x4_t

ARM-specific type containing four uint16x4_t vectors.

uint16x8_t

ARM-specific 128-bit wide vector of eight packed u16.

uint16x8x2_t

ARM-specific type containing two uint16x8_t vectors.

uint16x8x3_t

ARM-specific type containing three uint16x8_t vectors.

uint16x8x4_t

ARM-specific type containing four uint16x8_t vectors.

uint32x2_t

ARM-specific 64-bit wide vector of two packed u32.

uint32x2x2_t

ARM-specific type containing two uint32x2_t vectors.

uint32x2x3_t

ARM-specific type containing three uint32x2_t vectors.

uint32x2x4_t

ARM-specific type containing four uint32x2_t vectors.

uint32x4_t

ARM-specific 128-bit wide vector of four packed u32.

uint32x4x2_t

ARM-specific type containing two uint32x4_t vectors.

uint32x4x3_t

ARM-specific type containing three uint32x4_t vectors.

uint32x4x4_t

ARM-specific type containing four uint32x4_t vectors.

uint64x1_t

ARM-specific 64-bit wide vector of one packed u64.

uint64x1x2_t

ARM-specific type containing four uint64x1_t vectors.

uint64x1x3_t

ARM-specific type containing four uint64x1_t vectors.

uint64x1x4_t

ARM-specific type containing four uint64x1_t vectors.

uint64x2_t

ARM-specific 128-bit wide vector of two packed u64.

uint64x2x2_t

ARM-specific type containing four uint64x2_t vectors.

uint64x2x3_t

ARM-specific type containing four uint64x2_t vectors.

uint64x2x4_t

ARM-specific type containing four uint64x2_t vectors.

Functions

vabs_s8_ 🔒 ⚠Experimental

vabs_s16_ 🔒 ⚠Experimental

vabs_s32_ 🔒 ⚠Experimental

vabsq_s8_ 🔒 ⚠Experimental

vabsq_s16_ 🔒 ⚠Experimental

vabsq_s32_ 🔒 ⚠Experimental

vclz_s8_ 🔒 ⚠Experimental

vclz_s16_ 🔒 ⚠Experimental

vclz_s32_ 🔒 ⚠Experimental

vclzq_s8_ 🔒 ⚠Experimental

vclzq_s16_ 🔒 ⚠Experimental

vclzq_s32_ 🔒 ⚠Experimental

vcnt_s8_ 🔒 ⚠Experimental

vcntq_s8_ 🔒 ⚠Experimental

vcombine_p64⚠Experimentalneon

Vector combine

vcombine_s8⚠Experimentalneon

Vector combine

vcombine_s16⚠Experimentalneon

Vector combine

vcombine_s32⚠Experimentalneon

Vector combine

vcombine_s64⚠Experimentalneon

Vector combine

vcombine_u8⚠Experimentalneon

Vector combine

vcombine_u16⚠Experimentalneon

Vector combine

vcombine_u32⚠Experimentalneon

Vector combine

vcombine_u64⚠Experimentalneon

Vector combine

vmmlaq_s32⚠Experimentali8mm and neon

8-bit integer matrix multiply-accumulate

vmmlaq_u32⚠Experimentali8mm and neon

8-bit integer matrix multiply-accumulate

vpadd_s8_ 🔒 ⚠Experimental

vpadd_s16_ 🔒 ⚠Experimental

vpadd_s32_ 🔒 ⚠Experimental

vpaddl_s8_ 🔒 ⚠Experimental

vpaddl_s16_ 🔒 ⚠Experimental

vpaddl_s32_ 🔒 ⚠Experimental

vpaddl_u8_ 🔒 ⚠Experimental

vpaddl_u16_ 🔒 ⚠Experimental

vpaddl_u32_ 🔒 ⚠Experimental

vpaddlq_s8_ 🔒 ⚠Experimental

vpaddlq_s16_ 🔒 ⚠Experimental

vpaddlq_s32_ 🔒 ⚠Experimental

vpaddlq_u8_ 🔒 ⚠Experimental

vpaddlq_u16_ 🔒 ⚠Experimental

vpaddlq_u32_ 🔒 ⚠Experimental

vpmaxf_v2f32 🔒 ⚠Experimental

vpmaxs_v2i32 🔒 ⚠Experimental

vpmaxs_v4i16 🔒 ⚠Experimental

vpmaxs_v8i8 🔒 ⚠Experimental

vpmaxu_v2i32 🔒 ⚠Experimental

vpmaxu_v4i16 🔒 ⚠Experimental

vpmaxu_v8i8 🔒 ⚠Experimental

vpminf_v2f32 🔒 ⚠Experimental

vpmins_v2i32 🔒 ⚠Experimental

vpmins_v4i16 🔒 ⚠Experimental

vpmins_v8i8 🔒 ⚠Experimental

vpminu_v2i32 🔒 ⚠Experimental

vpminu_v4i16 🔒 ⚠Experimental

vpminu_v8i8 🔒 ⚠Experimental

vraddhn_s16_ 🔒 ⚠Experimental

vraddhn_s32_ 🔒 ⚠Experimental

vraddhn_s64_ 🔒 ⚠Experimental

vusmmlaq_s32⚠Experimentali8mm and neon

Unsigned and signed 8-bit integer matrix multiply-accumulate

vaba_s8⚠neon

vaba_s16⚠neon

vaba_s32⚠neon

vaba_u8⚠neon

vaba_u16⚠neon

vaba_u32⚠neon

vabaq_s8⚠neon

vabaq_s16⚠neon

vabaq_s32⚠neon

vabaq_u8⚠neon

vabaq_u16⚠neon

vabaq_u32⚠neon

vabs_s8⚠neon

Absolute value (wrapping).

vabs_s16⚠neon

Absolute value (wrapping).

vabs_s32⚠neon

Absolute value (wrapping).

vabsq_s8⚠neon

Absolute value (wrapping).

vabsq_s16⚠neon

Absolute value (wrapping).

vabsq_s32⚠neon

Absolute value (wrapping).

vadd_f32⚠neon

Vector add.

vadd_s8⚠neon

Vector add.

vadd_s16⚠neon

Vector add.

vadd_s32⚠neon

Vector add.

vadd_u8⚠neon

Vector add.

vadd_u16⚠neon

Vector add.

vadd_u32⚠neon

Vector add.

vaddhn_high_s16⚠neon

Add returning High Narrow (high half).

vaddhn_high_s32⚠neon

Add returning High Narrow (high half).

vaddhn_high_s64⚠neon

Add returning High Narrow (high half).

vaddhn_high_u16⚠neon

Add returning High Narrow (high half).

vaddhn_high_u32⚠neon

Add returning High Narrow (high half).

vaddhn_high_u64⚠neon

Add returning High Narrow (high half).

vaddhn_s16⚠neon

Add returning High Narrow.

vaddhn_s32⚠neon

Add returning High Narrow.

vaddhn_s64⚠neon

Add returning High Narrow.

vaddhn_u16⚠neon

Add returning High Narrow.

vaddhn_u32⚠neon

Add returning High Narrow.

vaddhn_u64⚠neon

Add returning High Narrow.

vaddl_high_s8⚠neon

Signed Add Long (vector, high half).

vaddl_high_s16⚠neon

Signed Add Long (vector, high half).

vaddl_high_s32⚠neon

Signed Add Long (vector, high half).

vaddl_high_u8⚠neon

Unsigned Add Long (vector, high half).

vaddl_high_u16⚠neon

Unsigned Add Long (vector, high half).

vaddl_high_u32⚠neon

Unsigned Add Long (vector, high half).

vaddl_s8⚠neon

Signed Add Long (vector).

vaddl_s16⚠neon

Signed Add Long (vector).

vaddl_s32⚠neon

Signed Add Long (vector).

vaddl_u8⚠neon

Unsigned Add Long (vector).

vaddl_u16⚠neon

Unsigned Add Long (vector).

vaddl_u32⚠neon

Unsigned Add Long (vector).

vaddq_f32⚠neon

Vector add.

vaddq_s8⚠neon

Vector add.

vaddq_s16⚠neon

Vector add.

vaddq_s32⚠neon

Vector add.

vaddq_s64⚠neon

Vector add.

vaddq_u8⚠neon

Vector add.

vaddq_u16⚠neon

Vector add.

vaddq_u32⚠neon

Vector add.

vaddq_u64⚠neon

Vector add.

vaddw_high_s8⚠neon

Signed Add Wide (high half).

vaddw_high_s16⚠neon

Signed Add Wide (high half).

vaddw_high_s32⚠neon

Signed Add Wide (high half).

vaddw_high_u8⚠neon

Unsigned Add Wide (high half).

vaddw_high_u16⚠neon

Unsigned Add Wide (high half).

vaddw_high_u32⚠neon

Unsigned Add Wide (high half).

vaddw_s8⚠neon

Signed Add Wide.

vaddw_s16⚠neon

Signed Add Wide.

vaddw_s32⚠neon

Signed Add Wide.

vaddw_u8⚠neon

Unsigned Add Wide.

vaddw_u16⚠neon

Unsigned Add Wide.

vaddw_u32⚠neon

Unsigned Add Wide.

vbic_s8⚠neon

Vector bitwise bit clear

vbic_s16⚠neon

Vector bitwise bit clear

vbic_s32⚠neon

Vector bitwise bit clear

vbic_s64⚠neon

Vector bitwise bit clear

vbic_u8⚠neon

Vector bitwise bit clear

vbic_u16⚠neon

Vector bitwise bit clear

vbic_u32⚠neon

Vector bitwise bit clear

vbic_u64⚠neon

Vector bitwise bit clear

vbicq_s8⚠neon

Vector bitwise bit clear

vbicq_s16⚠neon

Vector bitwise bit clear

vbicq_s32⚠neon

Vector bitwise bit clear

vbicq_s64⚠neon

Vector bitwise bit clear

vbicq_u8⚠neon

Vector bitwise bit clear

vbicq_u16⚠neon

Vector bitwise bit clear

vbicq_u32⚠neon

Vector bitwise bit clear

vbicq_u64⚠neon

Vector bitwise bit clear

vbsl_f32⚠neon

Bitwise Select.

vbsl_p8⚠neon

Bitwise Select.

vbsl_p16⚠neon

Bitwise Select.

vbsl_s8⚠neon

Bitwise Select instructions. This instruction sets each bit in the destination SIMD&FP register to the corresponding bit from the first source SIMD&FP register when the original destination bit was 1, otherwise from the second source SIMD&FP register. Bitwise Select.

vbsl_s16⚠neon

Bitwise Select.

vbsl_s32⚠neon

Bitwise Select.

vbsl_s64⚠neon

Bitwise Select.

vbsl_u8⚠neon

Bitwise Select.

vbsl_u16⚠neon

Bitwise Select.

vbsl_u32⚠neon

Bitwise Select.

vbsl_u64⚠neon

Bitwise Select.

vbslq_f32⚠neon

Bitwise Select. (128-bit)

vbslq_p8⚠neon

Bitwise Select. (128-bit)

vbslq_p16⚠neon

Bitwise Select. (128-bit)

vbslq_s8⚠neon

Bitwise Select. (128-bit)

vbslq_s16⚠neon

Bitwise Select. (128-bit)

vbslq_s32⚠neon

Bitwise Select. (128-bit)

vbslq_s64⚠neon

Bitwise Select. (128-bit)

vbslq_u8⚠neon

Bitwise Select. (128-bit)

vbslq_u16⚠neon

Bitwise Select. (128-bit)

vbslq_u32⚠neon

Bitwise Select. (128-bit)

vbslq_u64⚠neon

Bitwise Select. (128-bit)

vcnt_p8⚠neon

Population count per byte.

vcnt_s8⚠neon

Population count per byte.

vcnt_u8⚠neon

Population count per byte.

vcntq_p8⚠neon

Population count per byte.

vcntq_s8⚠neon

Population count per byte.

vcntq_u8⚠neon

Population count per byte.

vcombine_f32⚠neon

Vector combine

vcombine_p8⚠neon

Vector combine

vcombine_p16⚠neon

Vector combine

vdup_n_f32⚠neon

Duplicate vector element to vector or scalar

vdup_n_f32_vfp4 🔒 ⚠neon

Duplicate vector element to vector or scalar

vdup_n_p8⚠neon

Duplicate vector element to vector or scalar

vdup_n_p16⚠neon

Duplicate vector element to vector or scalar

vdup_n_s8⚠neon

Duplicate vector element to vector or scalar

vdup_n_s16⚠neon

Duplicate vector element to vector or scalar

vdup_n_s32⚠neon

Duplicate vector element to vector or scalar

vdup_n_s64⚠neon

Duplicate vector element to vector or scalar

vdup_n_u8⚠neon

Duplicate vector element to vector or scalar

vdup_n_u16⚠neon

Duplicate vector element to vector or scalar

vdup_n_u32⚠neon

Duplicate vector element to vector or scalar

vdup_n_u64⚠neon

Duplicate vector element to vector or scalar

vdupq_n_f32⚠neon

Duplicate vector element to vector or scalar

vdupq_n_f32_vfp4 🔒 ⚠neon

Duplicate vector element to vector or scalar

vdupq_n_p8⚠neon

Duplicate vector element to vector or scalar

vdupq_n_p16⚠neon

Duplicate vector element to vector or scalar

vdupq_n_s8⚠neon

Duplicate vector element to vector or scalar

vdupq_n_s16⚠neon

Duplicate vector element to vector or scalar

vdupq_n_s32⚠neon

Duplicate vector element to vector or scalar

vdupq_n_s64⚠neon

Duplicate vector element to vector or scalar

vdupq_n_u8⚠neon

Duplicate vector element to vector or scalar

vdupq_n_u16⚠neon

Duplicate vector element to vector or scalar

vdupq_n_u32⚠neon

Duplicate vector element to vector or scalar

vdupq_n_u64⚠neon

Duplicate vector element to vector or scalar

vext_s64⚠neon

Extract vector from pair of vectors

vext_u64⚠neon

Extract vector from pair of vectors

vget_high_f32⚠neon

Duplicate vector element to vector or scalar

vget_high_p8⚠neon

Duplicate vector element to vector or scalar

vget_high_p16⚠neon

Duplicate vector element to vector or scalar

vget_high_s8⚠neon

Duplicate vector element to vector or scalar

vget_high_s16⚠neon

Duplicate vector element to vector or scalar

vget_high_s32⚠neon

Duplicate vector element to vector or scalar

vget_high_s64⚠neon

Duplicate vector element to vector or scalar

vget_high_u8⚠neon

Duplicate vector element to vector or scalar

vget_high_u16⚠neon

Duplicate vector element to vector or scalar

vget_high_u32⚠neon

Duplicate vector element to vector or scalar

vget_high_u64⚠neon

Duplicate vector element to vector or scalar

vget_lane_f32⚠neon

Duplicate vector element to vector or scalar

vget_lane_p8⚠neon

Move vector element to general-purpose register

vget_lane_p16⚠neon

Move vector element to general-purpose register

vget_lane_p64⚠neon

Move vector element to general-purpose register

vget_lane_s8⚠neon

Move vector element to general-purpose register

vget_lane_s16⚠neon

Move vector element to general-purpose register

vget_lane_s32⚠neon

Move vector element to general-purpose register

vget_lane_s64⚠neon

Move vector element to general-purpose register

vget_lane_u8⚠neon

Move vector element to general-purpose register

vget_lane_u16⚠neon

Move vector element to general-purpose register

vget_lane_u32⚠neon

Move vector element to general-purpose register

vget_lane_u64⚠neon

Move vector element to general-purpose register

vget_low_f32⚠neon

Duplicate vector element to vector or scalar

vget_low_p8⚠neon

Duplicate vector element to vector or scalar

vget_low_p16⚠neon

Duplicate vector element to vector or scalar

vget_low_s8⚠neon

Duplicate vector element to vector or scalar

vget_low_s16⚠neon

Duplicate vector element to vector or scalar

vget_low_s32⚠neon

Duplicate vector element to vector or scalar

vget_low_s64⚠neon

Duplicate vector element to vector or scalar

vget_low_u8⚠neon

Duplicate vector element to vector or scalar

vget_low_u16⚠neon

Duplicate vector element to vector or scalar

vget_low_u32⚠neon

Duplicate vector element to vector or scalar

vget_low_u64⚠neon

Duplicate vector element to vector or scalar

vgetq_lane_f32⚠neon

Duplicate vector element to vector or scalar

vgetq_lane_p8⚠neon

Move vector element to general-purpose register

vgetq_lane_p16⚠neon

Move vector element to general-purpose register

vgetq_lane_p64⚠neon

Move vector element to general-purpose register

vgetq_lane_s8⚠neon

Move vector element to general-purpose register

vgetq_lane_s16⚠neon

Move vector element to general-purpose register

vgetq_lane_s32⚠neon

Move vector element to general-purpose register

vgetq_lane_s64⚠neon

Move vector element to general-purpose register

vgetq_lane_u8⚠neon

Move vector element to general-purpose register

vgetq_lane_u16⚠neon

Move vector element to general-purpose register

vgetq_lane_u32⚠neon

Move vector element to general-purpose register

vgetq_lane_u64⚠neon

Move vector element to general-purpose register

vld1_dup_f32⚠neon

Load one single-element structure and Replicate to all lanes (of one register).

vld1_dup_p8⚠neon

Load one single-element structure and Replicate to all lanes (of one register).

vld1_dup_p16⚠neon

Load one single-element structure and Replicate to all lanes (of one register).

vld1_dup_p64⚠neon,aes

Load one single-element structure and Replicate to all lanes (of one register).

vld1_dup_s8⚠neon

Load one single-element structure and Replicate to all lanes (of one register).

vld1_dup_s16⚠neon

Load one single-element structure and Replicate to all lanes (of one register).

vld1_dup_s32⚠neon

Load one single-element structure and Replicate to all lanes (of one register).

vld1_dup_s64⚠neon

Load one single-element structure and Replicate to all lanes (of one register).

vld1_dup_u8⚠neon

Load one single-element structure and Replicate to all lanes (of one register).

vld1_dup_u16⚠neon

Load one single-element structure and Replicate to all lanes (of one register).

vld1_dup_u32⚠neon

Load one single-element structure and Replicate to all lanes (of one register).

vld1_dup_u64⚠neon

Load one single-element structure and Replicate to all lanes (of one register).

vld1_lane_f32⚠neon

Load one single-element structure to one lane of one register.

vld1_lane_p8⚠neon

Load one single-element structure to one lane of one register.

vld1_lane_p16⚠neon

Load one single-element structure to one lane of one register.

vld1_lane_p64⚠neon,aes

Load one single-element structure to one lane of one register.

vld1_lane_s8⚠neon

Load one single-element structure to one lane of one register.

vld1_lane_s16⚠neon

Load one single-element structure to one lane of one register.

vld1_lane_s32⚠neon

Load one single-element structure to one lane of one register.

vld1_lane_s64⚠neon

Load one single-element structure to one lane of one register.

vld1_lane_u8⚠neon

Load one single-element structure to one lane of one register.

vld1_lane_u16⚠neon

Load one single-element structure to one lane of one register.

vld1_lane_u32⚠neon

Load one single-element structure to one lane of one register.

vld1_lane_u64⚠neon

Load one single-element structure to one lane of one register.

vld1q_dup_f32⚠neon

Load one single-element structure and Replicate to all lanes (of one register).

vld1q_dup_p8⚠neon

Load one single-element structure and Replicate to all lanes (of one register).

vld1q_dup_p16⚠neon

Load one single-element structure and Replicate to all lanes (of one register).

vld1q_dup_p64⚠neon,aes

Load one single-element structure and Replicate to all lanes (of one register).

vld1q_dup_s8⚠neon

Load one single-element structure and Replicate to all lanes (of one register).

vld1q_dup_s16⚠neon

Load one single-element structure and Replicate to all lanes (of one register).

vld1q_dup_s32⚠neon

Load one single-element structure and Replicate to all lanes (of one register).

vld1q_dup_s64⚠neon

Load one single-element structure and Replicate to all lanes (of one register).

vld1q_dup_u8⚠neon

Load one single-element structure and Replicate to all lanes (of one register).

vld1q_dup_u16⚠neon

Load one single-element structure and Replicate to all lanes (of one register).

vld1q_dup_u32⚠neon

Load one single-element structure and Replicate to all lanes (of one register).

vld1q_dup_u64⚠neon

Load one single-element structure and Replicate to all lanes (of one register).

vld1q_lane_f32⚠neon

Load one single-element structure to one lane of one register.

vld1q_lane_p8⚠neon

Load one single-element structure to one lane of one register.

vld1q_lane_p16⚠neon

Load one single-element structure to one lane of one register.

vld1q_lane_p64⚠neon,aes

Load one single-element structure to one lane of one register.

vld1q_lane_s8⚠neon

Load one single-element structure to one lane of one register.

vld1q_lane_s16⚠neon

Load one single-element structure to one lane of one register.

vld1q_lane_s32⚠neon

Load one single-element structure to one lane of one register.

vld1q_lane_s64⚠neon

Load one single-element structure to one lane of one register.

vld1q_lane_u8⚠neon

Load one single-element structure to one lane of one register.

vld1q_lane_u16⚠neon

Load one single-element structure to one lane of one register.

vld1q_lane_u32⚠neon

Load one single-element structure to one lane of one register.

vld1q_lane_u64⚠neon

Load one single-element structure to one lane of one register.

vldrq_p128⚠neon

Load SIMD&FP register (immediate offset)

vmov_n_f32⚠neon

Duplicate vector element to vector or scalar

vmov_n_p8⚠neon

Duplicate vector element to vector or scalar

vmov_n_p16⚠neon

Duplicate vector element to vector or scalar

vmov_n_s8⚠neon

Duplicate vector element to vector or scalar

vmov_n_s16⚠neon

Duplicate vector element to vector or scalar

vmov_n_s32⚠neon

Duplicate vector element to vector or scalar

vmov_n_s64⚠neon

Duplicate vector element to vector or scalar

vmov_n_u8⚠neon

Duplicate vector element to vector or scalar

vmov_n_u16⚠neon

Duplicate vector element to vector or scalar

vmov_n_u32⚠neon

Duplicate vector element to vector or scalar

vmov_n_u64⚠neon

Duplicate vector element to vector or scalar

vmovl_s8⚠neon

Vector long move.

vmovl_s16⚠neon

Vector long move.

vmovl_s32⚠neon

Vector long move.

vmovl_u8⚠neon

Vector long move.

vmovl_u16⚠neon

Vector long move.

vmovl_u32⚠neon

Vector long move.

vmovn_s16⚠neon

Vector narrow integer.

vmovn_s32⚠neon

Vector narrow integer.

vmovn_s64⚠neon

Vector narrow integer.

vmovn_u16⚠neon

Vector narrow integer.

vmovn_u32⚠neon

Vector narrow integer.

vmovn_u64⚠neon

Vector narrow integer.

vmovq_n_f32⚠neon

Duplicate vector element to vector or scalar

vmovq_n_p8⚠neon

Duplicate vector element to vector or scalar

vmovq_n_p16⚠neon

Duplicate vector element to vector or scalar

vmovq_n_s8⚠neon

Duplicate vector element to vector or scalar

vmovq_n_s16⚠neon

Duplicate vector element to vector or scalar

vmovq_n_s32⚠neon

Duplicate vector element to vector or scalar

vmovq_n_s64⚠neon

Duplicate vector element to vector or scalar

vmovq_n_u8⚠neon

Duplicate vector element to vector or scalar

vmovq_n_u16⚠neon

Duplicate vector element to vector or scalar

vmovq_n_u32⚠neon

Duplicate vector element to vector or scalar

vmovq_n_u64⚠neon

Duplicate vector element to vector or scalar

vmvn_p8⚠neon

Vector bitwise not.

vmvn_s8⚠neon

Vector bitwise not.

vmvn_s16⚠neon

Vector bitwise not.

vmvn_s32⚠neon

Vector bitwise not.

vmvn_u8⚠neon

Vector bitwise not.

vmvn_u16⚠neon

Vector bitwise not.

vmvn_u32⚠neon

Vector bitwise not.

vmvnq_p8⚠neon

Vector bitwise not.

vmvnq_s8⚠neon

Vector bitwise not.

vmvnq_s16⚠neon

Vector bitwise not.

vmvnq_s32⚠neon

Vector bitwise not.

vmvnq_u8⚠neon

Vector bitwise not.

vmvnq_u16⚠neon

Vector bitwise not.

vmvnq_u32⚠neon

Vector bitwise not.

vorn_s8⚠neon

Vector bitwise inclusive OR NOT

vorn_s16⚠neon

Vector bitwise inclusive OR NOT

vorn_s32⚠neon

Vector bitwise inclusive OR NOT

vorn_s64⚠neon

Vector bitwise inclusive OR NOT

vorn_u8⚠neon

Vector bitwise inclusive OR NOT

vorn_u16⚠neon

Vector bitwise inclusive OR NOT

vorn_u32⚠neon

Vector bitwise inclusive OR NOT

vorn_u64⚠neon

Vector bitwise inclusive OR NOT

vornq_s8⚠neon

Vector bitwise inclusive OR NOT

vornq_s16⚠neon

Vector bitwise inclusive OR NOT

vornq_s32⚠neon

Vector bitwise inclusive OR NOT

vornq_s64⚠neon

Vector bitwise inclusive OR NOT

vornq_u8⚠neon

Vector bitwise inclusive OR NOT

vornq_u16⚠neon

Vector bitwise inclusive OR NOT

vornq_u32⚠neon

Vector bitwise inclusive OR NOT

vornq_u64⚠neon

Vector bitwise inclusive OR NOT

vpadal_s8⚠neon

Signed Add and Accumulate Long Pairwise.

vpadal_s16⚠neon

Signed Add and Accumulate Long Pairwise.

vpadal_s32⚠neon

Signed Add and Accumulate Long Pairwise.

vpadal_u8⚠neon

Unsigned Add and Accumulate Long Pairwise.

vpadal_u16⚠neon

Unsigned Add and Accumulate Long Pairwise.

vpadal_u32⚠neon

Unsigned Add and Accumulate Long Pairwise.

vpadalq_s8⚠neon

Signed Add and Accumulate Long Pairwise.

vpadalq_s16⚠neon

Signed Add and Accumulate Long Pairwise.

vpadalq_s32⚠neon

Signed Add and Accumulate Long Pairwise.

vpadalq_u8⚠neon

Unsigned Add and Accumulate Long Pairwise.

vpadalq_u16⚠neon

Unsigned Add and Accumulate Long Pairwise.

vpadalq_u32⚠neon

Unsigned Add and Accumulate Long Pairwise.

vpadd_s8⚠neon

Add pairwise.

vpadd_s16⚠neon

Add pairwise.

vpadd_s32⚠neon

Add pairwise.

vpadd_u8⚠neon

Add pairwise.

vpadd_u16⚠neon

Add pairwise.

vpadd_u32⚠neon

Add pairwise.

vpaddl_s8⚠neon

Signed Add Long Pairwise.

vpaddl_s16⚠neon

Signed Add Long Pairwise.

vpaddl_s32⚠neon

Signed Add Long Pairwise.

vpaddl_u8⚠neon

Unsigned Add Long Pairwise.

vpaddl_u16⚠neon

Unsigned Add Long Pairwise.

vpaddl_u32⚠neon

Unsigned Add Long Pairwise.

vpaddlq_s8⚠neon

Signed Add Long Pairwise.

vpaddlq_s16⚠neon

Signed Add Long Pairwise.

vpaddlq_s32⚠neon

Signed Add Long Pairwise.

vpaddlq_u8⚠neon

Unsigned Add Long Pairwise.

vpaddlq_u16⚠neon

Unsigned Add Long Pairwise.

vpaddlq_u32⚠neon

Unsigned Add Long Pairwise.

vpmax_f32⚠neon

Folding maximum of adjacent pairs

vpmax_s8⚠neon

Folding maximum of adjacent pairs

vpmax_s16⚠neon

Folding maximum of adjacent pairs

vpmax_s32⚠neon

Folding maximum of adjacent pairs

vpmax_u8⚠neon

Folding maximum of adjacent pairs

vpmax_u16⚠neon

Folding maximum of adjacent pairs

vpmax_u32⚠neon

Folding maximum of adjacent pairs

vpmin_f32⚠neon

Folding minimum of adjacent pairs

vpmin_s8⚠neon

Folding minimum of adjacent pairs

vpmin_s16⚠neon

Folding minimum of adjacent pairs

vpmin_s32⚠neon

Folding minimum of adjacent pairs

vpmin_u8⚠neon

Folding minimum of adjacent pairs

vpmin_u16⚠neon

Folding minimum of adjacent pairs

vpmin_u32⚠neon

Folding minimum of adjacent pairs

vraddhn_high_s16⚠neon

Rounding Add returning High Narrow (high half).

vraddhn_high_s32⚠neon

Rounding Add returning High Narrow (high half).

vraddhn_high_s64⚠neon

Rounding Add returning High Narrow (high half).

vraddhn_high_u16⚠neon

Rounding Add returning High Narrow (high half).

vraddhn_high_u32⚠neon

Rounding Add returning High Narrow (high half).

vraddhn_high_u64⚠neon

Rounding Add returning High Narrow (high half).

vraddhn_s16⚠neon

Rounding Add returning High Narrow.

vraddhn_s32⚠neon

Rounding Add returning High Narrow.

vraddhn_s64⚠neon

Rounding Add returning High Narrow.

vraddhn_u16⚠neon

Rounding Add returning High Narrow.

vraddhn_u32⚠neon

Rounding Add returning High Narrow.

vraddhn_u64⚠neon

Rounding Add returning High Narrow.

vrev16_p8⚠neon

Reversing vector elements (swap endianness)

vrev16_s8⚠neon

Reversing vector elements (swap endianness)

vrev16_u8⚠neon

Reversing vector elements (swap endianness)

vrev16q_p8⚠neon

Reversing vector elements (swap endianness)

vrev16q_s8⚠neon

Reversing vector elements (swap endianness)

vrev16q_u8⚠neon

Reversing vector elements (swap endianness)

vrev32_p8⚠neon

Reversing vector elements (swap endianness)

vrev32_p16⚠neon

Reversing vector elements (swap endianness)

vrev32_s8⚠neon

Reversing vector elements (swap endianness)

vrev32_s16⚠neon

Reversing vector elements (swap endianness)

vrev32_u8⚠neon

Reversing vector elements (swap endianness)

vrev32_u16⚠neon

Reversing vector elements (swap endianness)

vrev32q_p8⚠neon

Reversing vector elements (swap endianness)

vrev32q_p16⚠neon

Reversing vector elements (swap endianness)

vrev32q_s8⚠neon

Reversing vector elements (swap endianness)

vrev32q_s16⚠neon

Reversing vector elements (swap endianness)

vrev32q_u8⚠neon

Reversing vector elements (swap endianness)

vrev32q_u16⚠neon

Reversing vector elements (swap endianness)

vrev64_f32⚠neon

Reversing vector elements (swap endianness)

vrev64_p8⚠neon

Reversing vector elements (swap endianness)

vrev64_p16⚠neon

Reversing vector elements (swap endianness)

vrev64_s8⚠neon

Reversing vector elements (swap endianness)

vrev64_s16⚠neon

Reversing vector elements (swap endianness)

vrev64_s32⚠neon

Reversing vector elements (swap endianness)

vrev64_u8⚠neon

Reversing vector elements (swap endianness)

vrev64_u16⚠neon

Reversing vector elements (swap endianness)

vrev64_u32⚠neon

Reversing vector elements (swap endianness)

vrev64q_f32⚠neon

Reversing vector elements (swap endianness)

vrev64q_p8⚠neon

Reversing vector elements (swap endianness)

vrev64q_p16⚠neon

Reversing vector elements (swap endianness)

vrev64q_s8⚠neon

Reversing vector elements (swap endianness)

vrev64q_s16⚠neon

Reversing vector elements (swap endianness)

vrev64q_s32⚠neon

Reversing vector elements (swap endianness)

vrev64q_u8⚠neon

Reversing vector elements (swap endianness)

vrev64q_u16⚠neon

Reversing vector elements (swap endianness)

vrev64q_u32⚠neon

Reversing vector elements (swap endianness)

vstrq_p128⚠neon

Store SIMD&FP register (immediate offset)

Type Definitions

p8 🔒 Experimental

p16 🔒 Experimental

p64 🔒 Experimental

p128 🔒 Experimental