Module core::core_arch::wasm32::relaxed_simd

🔬This is a nightly-only experimental API. (stdsimd #48556)

Available on target_family="wasm" only.

Functions

f32x4_relaxed_maddExperimentalrelaxed-simd

Computes a * b + c with either one rounding or two roundings.

f32x4_relaxed_maxExperimentalrelaxed-simd

A relaxed version of f32x4_max which is either f32x4_max or f32x4_pmax.

f32x4_relaxed_minExperimentalrelaxed-simd

A relaxed version of f32x4_min which is either f32x4_min or f32x4_pmin.

f32x4_relaxed_nmaddExperimentalrelaxed-simd

Computes -a * b + c with either one rounding or two roundings.

f64x2_relaxed_maddExperimentalrelaxed-simd

Computes a * b + c with either one rounding or two roundings.

f64x2_relaxed_maxExperimentalrelaxed-simd

A relaxed version of f64x2_max which is either f64x2_max or f64x2_pmax.

f64x2_relaxed_minExperimentalrelaxed-simd

A relaxed version of f64x2_min which is either f64x2_min or f64x2_pmin.

f64x2_relaxed_nmaddExperimentalrelaxed-simd

Computes -a * b + c with either one rounding or two roundings.

i8x16_relaxed_laneselectExperimentalrelaxed-simd

A relaxed version of v128_bitselect where this either behaves the same as v128_bitselect or the high bit of each lane m is inspected and the corresponding lane of a is chosen if the bit is 1 or the lane of b is chosen if it’s zero.

i8x16_relaxed_swizzleExperimentalrelaxed-simd

A relaxed version of i8x16_swizzle(a, s) which selects lanes from a using indices in s.

i16x8_relaxed_dot_i8x16_i7x16Experimentalrelaxed-simd

A relaxed dot-product instruction.

i16x8_relaxed_laneselectExperimentalrelaxed-simd

A relaxed version of v128_bitselect where this either behaves the same as v128_bitselect or the high bit of each lane m is inspected and the corresponding lane of a is chosen if the bit is 1 or the lane of b is chosen if it’s zero.

i16x8_relaxed_q15mulrExperimentalrelaxed-simd

A relaxed version of i16x8_relaxed_q15mulr where if both lanes are i16::MIN then the result is either i16::MIN or i16::MAX.

i32x4_relaxed_dot_i8x16_i7x16_addExperimentalrelaxed-simd

Similar to i16x8_relaxed_dot_i8x16_i7x16 except that the intermediate i16x8 result is fed into i32x4_extadd_pairwise_i16x8 followed by i32x4_add to add the value c to the result.

i32x4_relaxed_laneselectExperimentalrelaxed-simd

A relaxed version of v128_bitselect where this either behaves the same as v128_bitselect or the high bit of each lane m is inspected and the corresponding lane of a is chosen if the bit is 1 or the lane of b is chosen if it’s zero.

i32x4_relaxed_trunc_f32x4Experimentalrelaxed-simd

A relaxed version of i32x4_trunc_sat_f32x4(a) converts the f32 lanes of a to signed 32-bit integers.

i32x4_relaxed_trunc_f64x2_zeroExperimentalrelaxed-simd

A relaxed version of i32x4_trunc_sat_f64x2_zero(a) converts the f64 lanes of a to signed 32-bit integers and the upper two lanes are zero.

i64x2_relaxed_laneselectExperimentalrelaxed-simd

A relaxed version of v128_bitselect where this either behaves the same as v128_bitselect or the high bit of each lane m is inspected and the corresponding lane of a is chosen if the bit is 1 or the lane of b is chosen if it’s zero.

llvm_f32x4_fma 🔒 ⚠Experimental

llvm_f32x4_fms 🔒 ⚠Experimental

llvm_f32x4_relaxed_max 🔒 ⚠Experimental

llvm_f32x4_relaxed_min 🔒 ⚠Experimental

llvm_f64x2_fma 🔒 ⚠Experimental

llvm_f64x2_fms 🔒 ⚠Experimental

llvm_f64x2_relaxed_max 🔒 ⚠Experimental

llvm_f64x2_relaxed_min 🔒 ⚠Experimental

llvm_i8x16_laneselect 🔒 ⚠Experimental

llvm_i16x8_laneselect 🔒 ⚠Experimental

llvm_i16x8_relaxed_dot_i8x16_i7x16_s 🔒 ⚠Experimental

llvm_i32x4_laneselect 🔒 ⚠Experimental

llvm_i32x4_relaxed_dot_i8x16_i7x16_add_s 🔒 ⚠Experimental

llvm_i64x2_laneselect 🔒 ⚠Experimental

llvm_relaxed_q15mulr_signed 🔒 ⚠Experimental

llvm_relaxed_swizzle 🔒 ⚠Experimental

llvm_relaxed_trunc_signed 🔒 ⚠Experimental

llvm_relaxed_trunc_signed_zero 🔒 ⚠Experimental

llvm_relaxed_trunc_unsigned 🔒 ⚠Experimental

llvm_relaxed_trunc_unsigned_zero 🔒 ⚠Experimental

u32x4_relaxed_trunc_f32x4Experimentalrelaxed-simd

A relaxed version of u32x4_trunc_sat_f32x4(a) converts the f32 lanes of a to unsigned 32-bit integers.

u32x4_relaxed_trunc_f64x2_zeroExperimentalrelaxed-simd

A relaxed version of u32x4_trunc_sat_f64x2_zero(a) converts the f64 lanes of a to unsigned 32-bit integers and the upper two lanes are zero.