Module core::core_arch::x86::avx512cd

🔬This is a nightly-only experimental API. (stdsimd #48556)

Available on x86 or x86-64 only.

Functions

_mm256_broadcastmb_epi64⚠Experimentalavx512cd,avx512vl

Broadcast the low 8-bits from input mask k to all 64-bit elements of dst.

_mm256_broadcastmw_epi32⚠Experimentalavx512cd,avx512vl

Broadcast the low 16-bits from input mask k to all 32-bit elements of dst.

_mm256_conflict_epi32⚠Experimentalavx512cd,avx512vl

Test each 32-bit element of a for equality with all other elements in a closer to the least significant bit. Each element’s comparison forms a zero extended bit vector in dst.

_mm256_conflict_epi64⚠Experimentalavx512cd,avx512vl

Test each 64-bit element of a for equality with all other elements in a closer to the least significant bit. Each element’s comparison forms a zero extended bit vector in dst.

_mm256_lzcnt_epi32⚠Experimentalavx512cd,avx512vl

Counts the number of leading zero bits in each packed 32-bit integer in a, and store the results in dst.

_mm256_lzcnt_epi64⚠Experimentalavx512cd,avx512vl

Counts the number of leading zero bits in each packed 64-bit integer in a, and store the results in dst.

_mm256_mask_conflict_epi32⚠Experimentalavx512cd,avx512vl

Test each 32-bit element of a for equality with all other elements in a closer to the least significant bit using writemask k (elements are copied from src when the corresponding mask bit is not set). Each element’s comparison forms a zero extended bit vector in dst.

_mm256_mask_conflict_epi64⚠Experimentalavx512cd,avx512vl

Test each 64-bit element of a for equality with all other elements in a closer to the least significant bit using writemask k (elements are copied from src when the corresponding mask bit is not set). Each element’s comparison forms a zero extended bit vector in dst.

_mm256_mask_lzcnt_epi32⚠Experimentalavx512cd,avx512vl

Counts the number of leading zero bits in each packed 32-bit integer in a, and store the results in dst using writemask k (elements are copied from src when the corresponding mask bit is not set).

_mm256_mask_lzcnt_epi64⚠Experimentalavx512cd,avx512vl

Counts the number of leading zero bits in each packed 64-bit integer in a, and store the results in dst using writemask k (elements are copied from src when the corresponding mask bit is not set).

_mm256_maskz_conflict_epi32⚠Experimentalavx512cd,avx512vl

Test each 32-bit element of a for equality with all other elements in a closer to the least significant bit using zeromask k (elements are zeroed out when the corresponding mask bit is not set). Each element’s comparison forms a zero extended bit vector in dst.

_mm256_maskz_conflict_epi64⚠Experimentalavx512cd,avx512vl

Test each 64-bit element of a for equality with all other elements in a closer to the least significant bit using zeromask k (elements are zeroed out when the corresponding mask bit is not set). Each element’s comparison forms a zero extended bit vector in dst.

_mm256_maskz_lzcnt_epi32⚠Experimentalavx512cd,avx512vl

Counts the number of leading zero bits in each packed 32-bit integer in a, and store the results in dst using zeromask k (elements are zeroed out when the corresponding mask bit is not set).

_mm256_maskz_lzcnt_epi64⚠Experimentalavx512cd,avx512vl

Counts the number of leading zero bits in each packed 64-bit integer in a, and store the results in dst using zeromask k (elements are zeroed out when the corresponding mask bit is not set).

_mm512_broadcastmb_epi64⚠Experimentalavx512cd

Broadcast the low 8-bits from input mask k to all 64-bit elements of dst.

_mm512_broadcastmw_epi32⚠Experimentalavx512cd

Broadcast the low 16-bits from input mask k to all 32-bit elements of dst.

_mm512_conflict_epi32⚠Experimentalavx512cd

Test each 32-bit element of a for equality with all other elements in a closer to the least significant bit. Each element’s comparison forms a zero extended bit vector in dst.

_mm512_conflict_epi64⚠Experimentalavx512cd

Test each 64-bit element of a for equality with all other elements in a closer to the least significant bit. Each element’s comparison forms a zero extended bit vector in dst.

_mm512_lzcnt_epi32⚠Experimentalavx512cd

Counts the number of leading zero bits in each packed 32-bit integer in a, and store the results in dst.

_mm512_lzcnt_epi64⚠Experimentalavx512cd

Counts the number of leading zero bits in each packed 64-bit integer in a, and store the results in dst.

_mm512_mask_conflict_epi32⚠Experimentalavx512cd

Test each 32-bit element of a for equality with all other elements in a closer to the least significant bit using writemask k (elements are copied from src when the corresponding mask bit is not set). Each element’s comparison forms a zero extended bit vector in dst.

_mm512_mask_conflict_epi64⚠Experimentalavx512cd

Test each 64-bit element of a for equality with all other elements in a closer to the least significant bit using writemask k (elements are copied from src when the corresponding mask bit is not set). Each element’s comparison forms a zero extended bit vector in dst.

_mm512_mask_lzcnt_epi32⚠Experimentalavx512cd

Counts the number of leading zero bits in each packed 32-bit integer in a, and store the results in dst using writemask k (elements are copied from src when the corresponding mask bit is not set).

_mm512_mask_lzcnt_epi64⚠Experimentalavx512cd

Counts the number of leading zero bits in each packed 64-bit integer in a, and store the results in dst using writemask k (elements are copied from src when the corresponding mask bit is not set).

_mm512_maskz_conflict_epi32⚠Experimentalavx512cd

Test each 32-bit element of a for equality with all other elements in a closer to the least significant bit using zeromask k (elements are zeroed out when the corresponding mask bit is not set). Each element’s comparison forms a zero extended bit vector in dst.

_mm512_maskz_conflict_epi64⚠Experimentalavx512cd

Test each 64-bit element of a for equality with all other elements in a closer to the least significant bit using zeromask k (elements are zeroed out when the corresponding mask bit is not set). Each element’s comparison forms a zero extended bit vector in dst.

_mm512_maskz_lzcnt_epi32⚠Experimentalavx512cd

Counts the number of leading zero bits in each packed 32-bit integer in a, and store the results in dst using zeromask k (elements are zeroed out when the corresponding mask bit is not set).

_mm512_maskz_lzcnt_epi64⚠Experimentalavx512cd

Counts the number of leading zero bits in each packed 64-bit integer in a, and store the results in dst using zeromask k (elements are zeroed out when the corresponding mask bit is not set).

_mm_broadcastmb_epi64⚠Experimentalavx512cd,avx512vl

Broadcast the low 8-bits from input mask k to all 64-bit elements of dst.

_mm_broadcastmw_epi32⚠Experimentalavx512cd,avx512vl

Broadcast the low 16-bits from input mask k to all 32-bit elements of dst.

_mm_conflict_epi32⚠Experimentalavx512cd,avx512vl

Test each 32-bit element of a for equality with all other elements in a closer to the least significant bit. Each element’s comparison forms a zero extended bit vector in dst.

_mm_conflict_epi64⚠Experimentalavx512cd,avx512vl

Test each 64-bit element of a for equality with all other elements in a closer to the least significant bit. Each element’s comparison forms a zero extended bit vector in dst.

_mm_lzcnt_epi32⚠Experimentalavx512cd,avx512vl

Counts the number of leading zero bits in each packed 32-bit integer in a, and store the results in dst.

_mm_lzcnt_epi64⚠Experimentalavx512cd,avx512vl

Counts the number of leading zero bits in each packed 64-bit integer in a, and store the results in dst.

_mm_mask_conflict_epi32⚠Experimentalavx512cd,avx512vl

Test each 32-bit element of a for equality with all other elements in a closer to the least significant bit using writemask k (elements are copied from src when the corresponding mask bit is not set). Each element’s comparison forms a zero extended bit vector in dst.

_mm_mask_conflict_epi64⚠Experimentalavx512cd,avx512vl

Test each 64-bit element of a for equality with all other elements in a closer to the least significant bit using writemask k (elements are copied from src when the corresponding mask bit is not set). Each element’s comparison forms a zero extended bit vector in dst.

_mm_mask_lzcnt_epi32⚠Experimentalavx512cd,avx512vl

Counts the number of leading zero bits in each packed 32-bit integer in a, and store the results in dst using writemask k (elements are copied from src when the corresponding mask bit is not set).

_mm_mask_lzcnt_epi64⚠Experimentalavx512cd,avx512vl

Counts the number of leading zero bits in each packed 64-bit integer in a, and store the results in dst using writemask k (elements are copied from src when the corresponding mask bit is not set).

_mm_maskz_conflict_epi32⚠Experimentalavx512cd,avx512vl

Test each 32-bit element of a for equality with all other elements in a closer to the least significant bit using zeromask k (elements are zeroed out when the corresponding mask bit is not set). Each element’s comparison forms a zero extended bit vector in dst.

_mm_maskz_conflict_epi64⚠Experimentalavx512cd,avx512vl

Test each 64-bit element of a for equality with all other elements in a closer to the least significant bit using zeromask k (elements are zeroed out when the corresponding mask bit is not set). Each element’s comparison forms a zero extended bit vector in dst.

_mm_maskz_lzcnt_epi32⚠Experimentalavx512cd,avx512vl

Counts the number of leading zero bits in each packed 32-bit integer in a, and store the results in dst using zeromask k (elements are zeroed out when the corresponding mask bit is not set).

_mm_maskz_lzcnt_epi64⚠Experimentalavx512cd,avx512vl

Counts the number of leading zero bits in each packed 64-bit integer in a, and store the results in dst using zeromask k (elements are zeroed out when the corresponding mask bit is not set).

vpconflictd 🔒 ⚠Experimental

vpconflictd128 🔒 ⚠Experimental

vpconflictd256 🔒 ⚠Experimental

vpconflictq 🔒 ⚠Experimental

vpconflictq128 🔒 ⚠Experimental

vpconflictq256 🔒 ⚠Experimental

vplzcntd 🔒 ⚠Experimental

vplzcntd128 🔒 ⚠Experimental

vplzcntd256 🔒 ⚠Experimental

vplzcntq 🔒 ⚠Experimental

vplzcntq128 🔒 ⚠Experimental

vplzcntq256 🔒 ⚠Experimental