Module core::core_arch::x86::sha

🔬This is a nightly-only experimental API. (stdsimd #48556)

Available on x86 or x86-64 only.

Functions

sha1msg1 🔒 ⚠Experimental

sha1msg2 🔒 ⚠Experimental

sha1nexte 🔒 ⚠Experimental

sha1rnds4 🔒 ⚠Experimental

sha256msg1 🔒 ⚠Experimental

sha256msg2 🔒 ⚠Experimental

sha256rnds2 🔒 ⚠Experimental

_mm_sha1msg1_epu32⚠sha

Performs an intermediate calculation for the next four SHA1 message values (unsigned 32-bit integers) using previous message values from a and b, and returning the result.

_mm_sha1msg2_epu32⚠sha

Performs the final calculation for the next four SHA1 message values (unsigned 32-bit integers) using the intermediate result in a and the previous message values in b, and returns the result.

_mm_sha1nexte_epu32⚠sha

Calculate SHA1 state variable E after four rounds of operation from the current SHA1 state variable a, add that value to the scheduled values (unsigned 32-bit integers) in b, and returns the result.

_mm_sha1rnds4_epu32⚠sha

Performs four rounds of SHA1 operation using an initial SHA1 state (A,B,C,D) from a and some pre-computed sum of the next 4 round message values (unsigned 32-bit integers), and state variable E from b, and return the updated SHA1 state (A,B,C,D). FUNC contains the logic functions and round constants.

_mm_sha256msg1_epu32⚠sha

Performs an intermediate calculation for the next four SHA256 message values (unsigned 32-bit integers) using previous message values from a and b, and return the result.

_mm_sha256msg2_epu32⚠sha

Performs the final calculation for the next four SHA256 message values (unsigned 32-bit integers) using previous message values from a and b, and return the result.

_mm_sha256rnds2_epu32⚠sha

Performs 2 rounds of SHA256 operation using an initial SHA256 state (C,D,G,H) from a, an initial SHA256 state (A,B,E,F) from b, and a pre-computed sum of the next 2 round message values (unsigned 32-bit integers) and the corresponding round constants from k, and store the updated SHA256 state (A,B,E,F) in dst.