Module core::core_arch::x86::sse42

🔬This is a nightly-only experimental API. (stdsimd #48556)

Available on x86 or x86-64 only.

Streaming SIMD Extensions 4.2 (SSE4.2)

Extends SSE4.1 with STTNI (String and Text New Instructions).

Constants

_SIDD_BIT_MASK

Mask only: return the bit mask

_SIDD_CMP_EQUAL_ANY

For each character in a, find if it is in b (Default)

_SIDD_CMP_EQUAL_EACH

The strings defined by a and b are equal

_SIDD_CMP_EQUAL_ORDERED

Search for the defined substring in the target

_SIDD_CMP_RANGES

For each character in a, determine if b[0] <= c <= b[1] or b[1] <= c <= b[2]...

_SIDD_LEAST_SIGNIFICANT

Index only: return the least significant bit (Default)

_SIDD_MASKED_NEGATIVE_POLARITY

Negates results only before the end of the string

_SIDD_MASKED_POSITIVE_POLARITY

Do not negate results before the end of the string

_SIDD_MOST_SIGNIFICANT

Index only: return the most significant bit

_SIDD_NEGATIVE_POLARITY

Negates results

_SIDD_POSITIVE_POLARITY

Do not negate results (Default)

_SIDD_SBYTE_OPS

String contains signed 8-bit characters

_SIDD_SWORD_OPS

String contains unsigned 16-bit characters

_SIDD_UBYTE_OPS

String contains unsigned 8-bit characters (Default)

_SIDD_UNIT_MASK

Mask only: return the byte mask

_SIDD_UWORD_OPS

String contains unsigned 16-bit characters

Functions

crc32_32_8 🔒 ⚠Experimental

crc32_32_16 🔒 ⚠Experimental

crc32_32_32 🔒 ⚠Experimental

pcmpestri128 🔒 ⚠Experimental

pcmpestria128 🔒 ⚠Experimental

pcmpestric128 🔒 ⚠Experimental

pcmpestrio128 🔒 ⚠Experimental

pcmpestris128 🔒 ⚠Experimental

pcmpestriz128 🔒 ⚠Experimental

pcmpestrm128 🔒 ⚠Experimental

pcmpistri128 🔒 ⚠Experimental

pcmpistria128 🔒 ⚠Experimental

pcmpistric128 🔒 ⚠Experimental

pcmpistrio128 🔒 ⚠Experimental

pcmpistris128 🔒 ⚠Experimental

pcmpistriz128 🔒 ⚠Experimental

pcmpistrm128 🔒 ⚠Experimental

_mm_cmpestra⚠sse4.2

Compares packed strings in a and b with lengths la and lb using the control in IMM8, and return 1 if b did not contain a null character and the resulting mask was zero, and 0 otherwise.

_mm_cmpestrc⚠sse4.2

Compares packed strings in a and b with lengths la and lb using the control in IMM8, and return 1 if the resulting mask was non-zero, and 0 otherwise.

_mm_cmpestri⚠sse4.2

Compares packed strings a and b with lengths la and lb using the control in IMM8 and return the generated index. Similar to _mm_cmpistri with the exception that _mm_cmpistri implicitly determines the length of a and b.

_mm_cmpestrm⚠sse4.2

Compares packed strings in a and b with lengths la and lb using the control in IMM8, and return the generated mask.

_mm_cmpestro⚠sse4.2

Compares packed strings in a and b with lengths la and lb using the control in IMM8, and return bit 0 of the resulting bit mask.

_mm_cmpestrs⚠sse4.2

Compares packed strings in a and b with lengths la and lb using the control in IMM8, and return 1 if any character in a was null, and 0 otherwise.

_mm_cmpestrz⚠sse4.2

Compares packed strings in a and b with lengths la and lb using the control in IMM8, and return 1 if any character in b was null, and 0 otherwise.

_mm_cmpgt_epi64⚠sse4.2

Compares packed 64-bit integers in a and b for greater-than, return the results.

_mm_cmpistra⚠sse4.2

Compares packed strings with implicit lengths in a and b using the control in IMM8, and return 1 if b did not contain a null character and the resulting mask was zero, and 0 otherwise.

_mm_cmpistrc⚠sse4.2

Compares packed strings with implicit lengths in a and b using the control in IMM8, and return 1 if the resulting mask was non-zero, and 0 otherwise.

_mm_cmpistri⚠sse4.2

Compares packed strings with implicit lengths in a and b using the control in IMM8 and return the generated index. Similar to _mm_cmpestri with the exception that _mm_cmpestri requires the lengths of a and b to be explicitly specified.

_mm_cmpistrm⚠sse4.2

Compares packed strings with implicit lengths in a and b using the control in IMM8, and return the generated mask.

_mm_cmpistro⚠sse4.2

Compares packed strings with implicit lengths in a and b using the control in IMM8, and return bit 0 of the resulting bit mask.

_mm_cmpistrs⚠sse4.2

Compares packed strings with implicit lengths in a and b using the control in IMM8, and returns 1 if any character in a was null, and 0 otherwise.

_mm_cmpistrz⚠sse4.2

Compares packed strings with implicit lengths in a and b using the control in IMM8, and return 1 if any character in b was null. and 0 otherwise.

_mm_crc32_u8⚠sse4.2

Starting with the initial value in crc, return the accumulated CRC32-C value for unsigned 8-bit integer v.

_mm_crc32_u16⚠sse4.2

Starting with the initial value in crc, return the accumulated CRC32-C value for unsigned 16-bit integer v.

_mm_crc32_u32⚠sse4.2

Starting with the initial value in crc, return the accumulated CRC32-C value for unsigned 32-bit integer v.