Module core::core_arch::x86_64

🔬This is a nightly-only experimental API. (stdsimd #48556)

Available on x86-64 only.

x86_64 intrinsics

Re-exports

pub use self::fxsr::*;

Experimental

pub use self::sse::*;

Experimental

pub use self::sse2::*;

Experimental

pub use self::sse41::*;

Experimental

pub use self::sse42::*;

Experimental

pub use self::xsave::*;

Experimental

pub use self::abm::*;

Experimental

pub use self::avx::*;

Experimental

pub use self::bmi::*;

Experimental

pub use self::bmi2::*;

Experimental

pub use self::avx2::*;

Experimental

pub use self::avx512f::*;

Experimental

pub use self::bswap::*;

Experimental

pub use self::rdrand::*;

Experimental

pub use self::cmpxchg16b::*;

Experimental

pub use self::adx::*;

Experimental

pub use self::bt::*;

Experimental

Modules

abm 🔒 Experimental

Advanced Bit Manipulation (ABM) instructions

adx 🔒 Experimental

avx 🔒 Experimental

Advanced Vector Extensions (AVX)

avx2 🔒 Experimental

Advanced Vector Extensions 2 (AVX)

avx512f 🔒 Experimental

bmi 🔒 Experimental

Bit Manipulation Instruction (BMI) Set 1.0.

bmi2 🔒 Experimental

Bit Manipulation Instruction (BMI) Set 2.0.

bswap 🔒 Experimental

Byte swap intrinsics.

bt 🔒 Experimental

cmpxchg16b 🔒 Experimental

fxsr 🔒 Experimental

FXSR floating-point context fast save and restore.

macros 🔒 Experimental

Utility macros.

rdrand 🔒 Experimental

RDRAND and RDSEED instructions for returning random numbers from an Intel on-chip hardware random number generator which has been seeded by an on-chip entropy source.

sse 🔒 Experimental

x86_64 Streaming SIMD Extensions (SSE)

sse2 🔒 Experimental

x86_64’s Streaming SIMD Extensions 2 (SSE2)

sse41 🔒 Experimental

i686’s Streaming SIMD Extensions 4.1 (SSE4.1)

sse42 🔒 Experimental

x86_64’s Streaming SIMD Extensions 4.2 (SSE4.2)

xsave 🔒 Experimental

x86_64’s xsave and xsaveopt target feature intrinsics