Trait core::core_simd::lane_count::SupportedLaneCount

pub trait SupportedLaneCount: Sealed {
    type BitMask: Copy + Default + AsRef<[u8]> + AsMut<[u8]>;
}

🔬This is a nightly-only experimental API. (portable_simd #86656)

Statically guarantees that a lane count is marked as supported.

This trait is sealed: the list of implementors below is total. Users do not have the ability to mark additional LaneCount<N> values as supported. Only SIMD vectors with supported lane counts are constructable.

Required Associated Types

type BitMask: Copy + Default + AsRef<[u8]> + AsMut<[u8]>

🔬This is a nightly-only experimental API. (portable_simd #86656)