calimero_sdk
Enum ProposalAction

pub enum ProposalAction {
    ExternalFunctionCall {
        receiver_id: AccountId,
        method_name: String,
        args: String,
        deposit: u128,
    },
    Transfer {
        receiver_id: AccountId,
        amount: u128,
    },
    SetNumApprovals {
        num_approvals: u32,
    },
    SetActiveProposalsLimit {
        active_proposals_limit: u32,
    },
    SetContextValue {
        key: Box<[u8]>,
        value: Box<[u8]>,
    },
    DeleteProposal {
        proposal_id: ProposalId,
    },
}

A blockchain proposal action.

This enum represents the different actions that can be executed against a blockchain, and combined into a proposal.

Variants

ExternalFunctionCall

Call a method on a contract.

Fields

receiver_id: AccountId

The account ID of the contract to call.

method_name: String

The method name to call.

args: String

The arguments to pass to the method.

deposit: u128

The amount of tokens to attach to the call.

Transfer

Transfer tokens to an account.

Fields

receiver_id: AccountId

The account ID of the receiver.

amount: u128

The amount of tokens to transfer.

SetNumApprovals

Set the number of approvals required for a proposal to be executed.

Fields

num_approvals: u32

The number of approvals required.

SetActiveProposalsLimit

Set the number of active proposals allowed at once.

Fields

active_proposals_limit: u32

The number of active proposals allowed.

SetContextValue

Set a value in the contractโs context.

Fields

key: Box<[u8]>

The key to set.

value: Box<[u8]>

The value to set.

DeleteProposal

Delete a proposal.

Fields

proposal_id: ProposalId

The ID of the proposal to delete.