core::ffi::c_str

Struct Display

pub struct Display<'a> { /* private fields */ }

🔬This is a nightly-only experimental API. (cstr_display #139984)

Helper struct for safely printing a CStr with format! and {}.

A CStr might contain non-Unicode data. This struct implements the Display trait in a way that mitigates that. It is created by the display method on CStr. This may perform lossy conversion, depending on the platform. If you would like an implementation which escapes the CStr please use Debug instead.

Examples

#![feature(cstr_display)]

let s = c"Hello, world!";
println!("{}", s.display());