Type Alias core::ffi::NonZero_c_ushort

pub type NonZero_c_ushort = NonZeroU16;

🔬This is a nightly-only experimental API. (raw_os_nonzero #82363)

Type alias for NonZero version of c_ushort

Aliased Type

struct NonZero_c_ushort(/* private fields */);

Implementations

impl NonZeroU16

pub const unsafe fn new_unchecked(n: u16) -> Self

Creates a non-zero without checking whether the value is non-zero. This results in undefined behaviour if the value is zero.

Safety

The value must not be zero.

pub const fn new(n: u16) -> Option<Self>

Creates a non-zero if the given value is not zero.

pub const fn get(self) -> u16

Returns the value as a primitive type.

impl NonZeroU16

pub const fn leading_zeros(self) -> u32

Returns the number of leading zeros in the binary representation of self.

On many architectures, this function can perform better than leading_zeros() on the underlying integer type, as special handling of zero can be avoided.

Examples

Basic usage:

let n = std::num::NonZeroU16::new(u16::MAX).unwrap();

assert_eq!(n.leading_zeros(), 0);
Run

pub const fn trailing_zeros(self) -> u32

Returns the number of trailing zeros in the binary representation of self.

On many architectures, this function can perform better than trailing_zeros() on the underlying integer type, as special handling of zero can be avoided.

Examples

Basic usage:

let n = std::num::NonZeroU16::new(0b0101000).unwrap();

assert_eq!(n.trailing_zeros(), 3);
Run

impl NonZeroU16

pub const fn checked_add(self, other: u16) -> Option<NonZeroU16>

Adds an unsigned integer to a non-zero value. Checks for overflow and returns None on overflow. As a consequence, the result cannot wrap to zero.

Examples

let one = NonZeroU16::new(1)?;
let two = NonZeroU16::new(2)?;
let max = NonZeroU16::new(u16::MAX)?;

assert_eq!(Some(two), one.checked_add(1));
assert_eq!(None, max.checked_add(1));
Run

pub const fn saturating_add(self, other: u16) -> NonZeroU16

Adds an unsigned integer to a non-zero value. Return NonZeroU16::MAX on overflow.

Examples

let one = NonZeroU16::new(1)?;
let two = NonZeroU16::new(2)?;
let max = NonZeroU16::new(u16::MAX)?;

assert_eq!(two, one.saturating_add(1));
assert_eq!(max, max.saturating_add(1));
Run

pub const unsafe fn unchecked_add(self, other: u16) -> NonZeroU16

🔬This is a nightly-only experimental API. (nonzero_ops #84186)

Adds an unsigned integer to a non-zero value, assuming overflow cannot occur. Overflow is unchecked, and it is undefined behaviour to overflow even if the result would wrap to a non-zero value. The behaviour is undefined as soon as self + rhs > u16::MAX.

Examples

#![feature(nonzero_ops)]

let one = NonZeroU16::new(1)?;
let two = NonZeroU16::new(2)?;

assert_eq!(two, unsafe { one.unchecked_add(1) });
Run

pub const fn checked_next_power_of_two(self) -> Option<NonZeroU16>

Returns the smallest power of two greater than or equal to n. Checks for overflow and returns None if the next power of two is greater than the type’s maximum value. As a consequence, the result cannot wrap to zero.

Examples

let two = NonZeroU16::new(2)?;
let three = NonZeroU16::new(3)?;
let four = NonZeroU16::new(4)?;
let max = NonZeroU16::new(u16::MAX)?;

assert_eq!(Some(two), two.checked_next_power_of_two() );
assert_eq!(Some(four), three.checked_next_power_of_two() );
assert_eq!(None, max.checked_next_power_of_two() );
Run

pub const fn ilog2(self) -> u32

Returns the base 2 logarithm of the number, rounded down.

This is the same operation as u16::ilog2, except that it has no failure cases to worry about since this value can never be zero.

Examples

assert_eq!(NonZeroU16::new(7).unwrap().ilog2(), 2);
assert_eq!(NonZeroU16::new(8).unwrap().ilog2(), 3);
assert_eq!(NonZeroU16::new(9).unwrap().ilog2(), 3);
Run

pub const fn ilog10(self) -> u32

Returns the base 10 logarithm of the number, rounded down.

This is the same operation as u16::ilog10, except that it has no failure cases to worry about since this value can never be zero.

Examples

assert_eq!(NonZeroU16::new(99).unwrap().ilog10(), 1);
assert_eq!(NonZeroU16::new(100).unwrap().ilog10(), 2);
assert_eq!(NonZeroU16::new(101).unwrap().ilog10(), 2);
Run

pub fn midpoint(self, rhs: Self) -> Self

🔬This is a nightly-only experimental API. (num_midpoint #110840)

Calculates the middle point of self and rhs.

midpoint(a, b) is (a + b) >> 1 as if it were performed in a sufficiently-large signed integral type. This implies that the result is always rounded towards negative infinity and that no overflow will ever occur.

Examples

#![feature(num_midpoint)]

let one = NonZeroU16::new(1)?;
let two = NonZeroU16::new(2)?;
let four = NonZeroU16::new(4)?;

assert_eq!(one.midpoint(four), two);
assert_eq!(four.midpoint(one), two);
Run

impl NonZeroU16

pub const fn checked_mul(self, other: NonZeroU16) -> Option<NonZeroU16>

Multiplies two non-zero integers together. Checks for overflow and returns None on overflow. As a consequence, the result cannot wrap to zero.

Examples

let two = NonZeroU16::new(2)?;
let four = NonZeroU16::new(4)?;
let max = NonZeroU16::new(u16::MAX)?;

assert_eq!(Some(four), two.checked_mul(two));
assert_eq!(None, max.checked_mul(two));
Run

pub const fn saturating_mul(self, other: NonZeroU16) -> NonZeroU16

Multiplies two non-zero integers together. Return NonZeroU16::MAX on overflow.

Examples

let two = NonZeroU16::new(2)?;
let four = NonZeroU16::new(4)?;
let max = NonZeroU16::new(u16::MAX)?;

assert_eq!(four, two.saturating_mul(two));
assert_eq!(max, four.saturating_mul(max));
Run

pub const unsafe fn unchecked_mul(self, other: NonZeroU16) -> NonZeroU16

🔬This is a nightly-only experimental API. (nonzero_ops #84186)

Multiplies two non-zero integers together, assuming overflow cannot occur. Overflow is unchecked, and it is undefined behaviour to overflow even if the result would wrap to a non-zero value. The behaviour is undefined as soon as self * rhs > u16::MAX.

Examples

#![feature(nonzero_ops)]

let two = NonZeroU16::new(2)?;
let four = NonZeroU16::new(4)?;

assert_eq!(four, unsafe { two.unchecked_mul(two) });
Run

pub const fn checked_pow(self, other: u32) -> Option<NonZeroU16>

Raises non-zero value to an integer power. Checks for overflow and returns None on overflow. As a consequence, the result cannot wrap to zero.

Examples

let three = NonZeroU16::new(3)?;
let twenty_seven = NonZeroU16::new(27)?;
let half_max = NonZeroU16::new(u16::MAX / 2)?;

assert_eq!(Some(twenty_seven), three.checked_pow(3));
assert_eq!(None, half_max.checked_pow(3));
Run

pub const fn saturating_pow(self, other: u32) -> NonZeroU16

Raise non-zero value to an integer power. Return NonZeroU16::MAX on overflow.

Examples

let three = NonZeroU16::new(3)?;
let twenty_seven = NonZeroU16::new(27)?;
let max = NonZeroU16::new(u16::MAX)?;

assert_eq!(twenty_seven, three.saturating_pow(3));
assert_eq!(max, max.saturating_pow(3));
Run

impl NonZeroU16

pub const fn is_power_of_two(self) -> bool

Returns true if and only if self == (1 << k) for some k.

On many architectures, this function can perform better than is_power_of_two() on the underlying integer type, as special handling of zero can be avoided.

Examples

Basic usage:

let eight = std::num::NonZeroU16::new(8).unwrap();
assert!(eight.is_power_of_two());
let ten = std::num::NonZeroU16::new(10).unwrap();
assert!(!ten.is_power_of_two());
Run

impl NonZeroU16

pub const MIN: Self = _

The smallest value that can be represented by this non-zero integer type, 1.

Examples

assert_eq!(NonZeroU16::MIN.get(), 1u16);
Run

pub const MAX: Self = _

The largest value that can be represented by this non-zero integer type, equal to u16::MAX.

Examples

assert_eq!(NonZeroU16::MAX.get(), u16::MAX);
Run

impl NonZeroU16

pub const BITS: u32 = 16u32

The size of this non-zero integer type in bits.

This value is equal to u16::BITS.

Examples

assert_eq!(NonZeroU16::BITS, u16::BITS);
Run