Type Alias core::ffi::c_double

pub type c_double = f64;

Equivalent to C’s double type.

This type will almost always be f64, which is guaranteed to be an IEEE 754 double-precision float in Rust. That said, the standard technically only guarantees that it be a floating-point number with at least the precision of a float, and it may be f32 or something entirely different from the IEEE-754 standard.

Implementations

impl f64

pub const RADIX: u32 = 2u32

The radix or base of the internal representation of f64.

pub const MANTISSA_DIGITS: u32 = 53u32

Number of significant digits in base 2.

pub const DIGITS: u32 = 15u32

Approximate number of significant digits in base 10.

pub const EPSILON: f64 = 2.2204460492503131E-16f64

Machine epsilon value for f64.

This is the difference between 1.0 and the next larger representable number.

pub const MIN: f64 = -1.7976931348623157E+308f64

Smallest finite f64 value.

pub const MIN_POSITIVE: f64 = 2.2250738585072014E-308f64

Smallest positive normal f64 value.

pub const MAX: f64 = 1.7976931348623157E+308f64

Largest finite f64 value.

pub const MIN_EXP: i32 = -1_021i32

One greater than the minimum possible normal power of 2 exponent.

pub const MAX_EXP: i32 = 1_024i32

Maximum possible power of 2 exponent.

pub const MIN_10_EXP: i32 = -307i32

Minimum possible normal power of 10 exponent.

pub const MAX_10_EXP: i32 = 308i32

Maximum possible power of 10 exponent.

pub const NAN: f64 = NaNf64

Not a Number (NaN).

Note that IEEE 754 doesn’t define just a single NaN value; a plethora of bit patterns are considered to be NaN. Furthermore, the standard makes a difference between a “signaling” and a “quiet” NaN, and allows inspecting its “payload” (the unspecified bits in the bit pattern). This constant isn’t guaranteed to equal to any specific NaN bitpattern, and the stability of its representation over Rust versions and target platforms isn’t guaranteed.

pub const INFINITY: f64 = +Inff64

Infinity (∞).

pub const NEG_INFINITY: f64 = -Inff64

Negative infinity (−∞).

pub fn is_nan(self) -> bool

Returns true if this value is NaN.

let nan = f64::NAN;
let f = 7.0_f64;

assert!(nan.is_nan());
assert!(!f.is_nan());
Run

pub fn is_infinite(self) -> bool

Returns true if this value is positive infinity or negative infinity, and false otherwise.

let f = 7.0f64;
let inf = f64::INFINITY;
let neg_inf = f64::NEG_INFINITY;
let nan = f64::NAN;

assert!(!f.is_infinite());
assert!(!nan.is_infinite());

assert!(inf.is_infinite());
assert!(neg_inf.is_infinite());
Run

pub fn is_finite(self) -> bool

Returns true if this number is neither infinite nor NaN.

let f = 7.0f64;
let inf: f64 = f64::INFINITY;
let neg_inf: f64 = f64::NEG_INFINITY;
let nan: f64 = f64::NAN;

assert!(f.is_finite());

assert!(!nan.is_finite());
assert!(!inf.is_finite());
assert!(!neg_inf.is_finite());
Run

pub fn is_subnormal(self) -> bool

Returns true if the number is subnormal.

let min = f64::MIN_POSITIVE; // 2.2250738585072014e-308_f64
let max = f64::MAX;
let lower_than_min = 1.0e-308_f64;
let zero = 0.0_f64;

assert!(!min.is_subnormal());
assert!(!max.is_subnormal());

assert!(!zero.is_subnormal());
assert!(!f64::NAN.is_subnormal());
assert!(!f64::INFINITY.is_subnormal());
// Values between `0` and `min` are Subnormal.
assert!(lower_than_min.is_subnormal());
Run

pub fn is_normal(self) -> bool

Returns true if the number is neither zero, infinite, subnormal, or NaN.

let min = f64::MIN_POSITIVE; // 2.2250738585072014e-308f64
let max = f64::MAX;
let lower_than_min = 1.0e-308_f64;
let zero = 0.0f64;

assert!(min.is_normal());
assert!(max.is_normal());

assert!(!zero.is_normal());
assert!(!f64::NAN.is_normal());
assert!(!f64::INFINITY.is_normal());
// Values between `0` and `min` are Subnormal.
assert!(!lower_than_min.is_normal());
Run

pub fn classify(self) -> FpCategory

Returns the floating point category of the number. If only one property is going to be tested, it is generally faster to use the specific predicate instead.

use std::num::FpCategory;

let num = 12.4_f64;
let inf = f64::INFINITY;

assert_eq!(num.classify(), FpCategory::Normal);
assert_eq!(inf.classify(), FpCategory::Infinite);
Run

pub fn is_sign_positive(self) -> bool

Returns true if self has a positive sign, including +0.0, NaNs with positive sign bit and positive infinity. Note that IEEE 754 doesn’t assign any meaning to the sign bit in case of a NaN, and as Rust doesn’t guarantee that the bit pattern of NaNs are conserved over arithmetic operations, the result of is_sign_positive on a NaN might produce an unexpected result in some cases. See explanation of NaN as a special value for more info.

let f = 7.0_f64;
let g = -7.0_f64;

assert!(f.is_sign_positive());
assert!(!g.is_sign_positive());
Run

pub fn is_sign_negative(self) -> bool

Returns true if self has a negative sign, including -0.0, NaNs with negative sign bit and negative infinity. Note that IEEE 754 doesn’t assign any meaning to the sign bit in case of a NaN, and as Rust doesn’t guarantee that the bit pattern of NaNs are conserved over arithmetic operations, the result of is_sign_negative on a NaN might produce an unexpected result in some cases. See explanation of NaN as a special value for more info.

let f = 7.0_f64;
let g = -7.0_f64;

assert!(!f.is_sign_negative());
assert!(g.is_sign_negative());
Run

pub fn next_up(self) -> Self

🔬This is a nightly-only experimental API. (float_next_up_down #91399)

Returns the least number greater than self.

Let TINY be the smallest representable positive f64. Then,

if self.is_nan(), this returns self;

if self is NEG_INFINITY, this returns MIN;

if self is -TINY, this returns -0.0;

if self is -0.0 or +0.0, this returns TINY;

if self is MAX or INFINITY, this returns INFINITY;

otherwise the unique least value greater than self is returned.

The identity x.next_up() == -(-x).next_down() holds for all non-NaN x. When x is finite x == x.next_up().next_down() also holds.

#![feature(float_next_up_down)]
// f64::EPSILON is the difference between 1.0 and the next number up.
assert_eq!(1.0f64.next_up(), 1.0 + f64::EPSILON);
// But not for most numbers.
assert!(0.1f64.next_up() < 0.1 + f64::EPSILON);
assert_eq!(9007199254740992f64.next_up(), 9007199254740994.0);
Run

pub fn next_down(self) -> Self

🔬This is a nightly-only experimental API. (float_next_up_down #91399)

Returns the greatest number less than self.

Let TINY be the smallest representable positive f64. Then,

if self.is_nan(), this returns self;

if self is INFINITY, this returns MAX;

if self is TINY, this returns 0.0;

if self is -0.0 or +0.0, this returns -TINY;

if self is MIN or NEG_INFINITY, this returns NEG_INFINITY;

otherwise the unique greatest value less than self is returned.

The identity x.next_down() == -(-x).next_up() holds for all non-NaN x. When x is finite x == x.next_down().next_up() also holds.

#![feature(float_next_up_down)]
let x = 1.0f64;
// Clamp value into range [0, 1).
let clamped = x.clamp(0.0, 1.0f64.next_down());
assert!(clamped < 1.0);
assert_eq!(clamped.next_up(), 1.0);
Run

pub fn recip(self) -> f64

Takes the reciprocal (inverse) of a number, 1/x.

let x = 2.0_f64;
let abs_difference = (x.recip() - (1.0 / x)).abs();

assert!(abs_difference < 1e-10);
Run

pub fn to_degrees(self) -> f64

Converts radians to degrees.

let angle = std::f64::consts::PI;

let abs_difference = (angle.to_degrees() - 180.0).abs();

assert!(abs_difference < 1e-10);
Run

pub fn to_radians(self) -> f64

Converts degrees to radians.

let angle = 180.0_f64;

let abs_difference = (angle.to_radians() - std::f64::consts::PI).abs();

assert!(abs_difference < 1e-10);
Run

pub fn max(self, other: f64) -> f64

Returns the maximum of the two numbers, ignoring NaN.

If one of the arguments is NaN, then the other argument is returned. This follows the IEEE 754-2008 semantics for maxNum, except for handling of signaling NaNs; this function handles all NaNs the same way and avoids maxNum’s problems with associativity. This also matches the behavior of libm’s fmax.

let x = 1.0_f64;
let y = 2.0_f64;

assert_eq!(x.max(y), y);
Run

pub fn min(self, other: f64) -> f64

Returns the minimum of the two numbers, ignoring NaN.

If one of the arguments is NaN, then the other argument is returned. This follows the IEEE 754-2008 semantics for minNum, except for handling of signaling NaNs; this function handles all NaNs the same way and avoids minNum’s problems with associativity. This also matches the behavior of libm’s fmin.

let x = 1.0_f64;
let y = 2.0_f64;

assert_eq!(x.min(y), x);
Run

pub fn maximum(self, other: f64) -> f64

🔬This is a nightly-only experimental API. (float_minimum_maximum #91079)

Returns the maximum of the two numbers, propagating NaN.

This returns NaN when either argument is NaN, as opposed to f64::max which only returns NaN when both arguments are NaN.

#![feature(float_minimum_maximum)]
let x = 1.0_f64;
let y = 2.0_f64;

assert_eq!(x.maximum(y), y);
assert!(x.maximum(f64::NAN).is_nan());
Run

If one of the arguments is NaN, then NaN is returned. Otherwise this returns the greater of the two numbers. For this operation, -0.0 is considered to be less than +0.0. Note that this follows the semantics specified in IEEE 754-2019.

Also note that “propagation” of NaNs here doesn’t necessarily mean that the bitpattern of a NaN operand is conserved; see explanation of NaN as a special value for more info.

pub fn minimum(self, other: f64) -> f64

🔬This is a nightly-only experimental API. (float_minimum_maximum #91079)

Returns the minimum of the two numbers, propagating NaN.

This returns NaN when either argument is NaN, as opposed to f64::min which only returns NaN when both arguments are NaN.

#![feature(float_minimum_maximum)]
let x = 1.0_f64;
let y = 2.0_f64;

assert_eq!(x.minimum(y), x);
assert!(x.minimum(f64::NAN).is_nan());
Run

If one of the arguments is NaN, then NaN is returned. Otherwise this returns the lesser of the two numbers. For this operation, -0.0 is considered to be less than +0.0. Note that this follows the semantics specified in IEEE 754-2019.

Also note that “propagation” of NaNs here doesn’t necessarily mean that the bitpattern of a NaN operand is conserved; see explanation of NaN as a special value for more info.

pub fn midpoint(self, other: f64) -> f64

🔬This is a nightly-only experimental API. (num_midpoint #110840)

Calculates the middle point of self and rhs.

This returns NaN when either argument is NaN or if a combination of +inf and -inf is provided as arguments.

Examples

#![feature(num_midpoint)]
assert_eq!(1f64.midpoint(4.0), 2.5);
assert_eq!((-5.5f64).midpoint(8.0), 1.25);
Run

pub unsafe fn to_int_unchecked<Int>(self) -> Intwhere Self: FloatToInt<Int>,

Rounds toward zero and converts to any primitive integer type, assuming that the value is finite and fits in that type.

let value = 4.6_f64;
let rounded = unsafe { value.to_int_unchecked::<u16>() };
assert_eq!(rounded, 4);

let value = -128.9_f64;
let rounded = unsafe { value.to_int_unchecked::<i8>() };
assert_eq!(rounded, i8::MIN);
Run

Safety

The value must:

Not be NaN

Not be infinite

Be representable in the return type Int, after truncating off its fractional part

pub fn to_bits(self) -> u64

Raw transmutation to u64.

This is currently identical to transmute::<f64, u64>(self) on all platforms.

See from_bits for some discussion of the portability of this operation (there are almost no issues).

Note that this function is distinct from as casting, which attempts to preserve the numeric value, and not the bitwise value.

Examples

assert!((1f64).to_bits() != 1f64 as u64); // to_bits() is not casting!
assert_eq!((12.5f64).to_bits(), 0x4029000000000000);

Run

pub fn from_bits(v: u64) -> Self

Raw transmutation from u64.

This is currently identical to transmute::<u64, f64>(v) on all platforms. It turns out this is incredibly portable, for two reasons:

Floats and Ints have the same endianness on all supported platforms.

IEEE 754 very precisely specifies the bit layout of floats.

However there is one caveat: prior to the 2008 version of IEEE 754, how to interpret the NaN signaling bit wasn’t actually specified. Most platforms (notably x86 and ARM) picked the interpretation that was ultimately standardized in 2008, but some didn’t (notably MIPS). As a result, all signaling NaNs on MIPS are quiet NaNs on x86, and vice-versa.

Rather than trying to preserve signaling-ness cross-platform, this implementation favors preserving the exact bits. This means that any payloads encoded in NaNs will be preserved even if the result of this method is sent over the network from an x86 machine to a MIPS one.

If the results of this method are only manipulated by the same architecture that produced them, then there is no portability concern.

If the input isn’t NaN, then there is no portability concern.

If you don’t care about signaling-ness (very likely), then there is no portability concern.

Note that this function is distinct from as casting, which attempts to preserve the numeric value, and not the bitwise value.

Examples

let v = f64::from_bits(0x4029000000000000);
assert_eq!(v, 12.5);
Run

pub fn to_be_bytes(self) -> [u8; 8]

Return the memory representation of this floating point number as a byte array in big-endian (network) byte order.

See from_bits for some discussion of the portability of this operation (there are almost no issues).

Examples

let bytes = 12.5f64.to_be_bytes();
assert_eq!(bytes, [0x40, 0x29, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00]);
Run

pub fn to_le_bytes(self) -> [u8; 8]

Return the memory representation of this floating point number as a byte array in little-endian byte order.

See from_bits for some discussion of the portability of this operation (there are almost no issues).

Examples

let bytes = 12.5f64.to_le_bytes();
assert_eq!(bytes, [0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x29, 0x40]);
Run

pub fn to_ne_bytes(self) -> [u8; 8]

Return the memory representation of this floating point number as a byte array in native byte order.

As the target platform’s native endianness is used, portable code should use to_be_bytes or to_le_bytes, as appropriate, instead.

See from_bits for some discussion of the portability of this operation (there are almost no issues).

Examples

let bytes = 12.5f64.to_ne_bytes();
assert_eq!(
    bytes,
    if cfg!(target_endian = "big") {
        [0x40, 0x29, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00]
    } else {
        [0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x29, 0x40]
    }
);
Run

pub fn from_be_bytes(bytes: [u8; 8]) -> Self

Create a floating point value from its representation as a byte array in big endian.

See from_bits for some discussion of the portability of this operation (there are almost no issues).

Examples

let value = f64::from_be_bytes([0x40, 0x29, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00]);
assert_eq!(value, 12.5);
Run

pub fn from_le_bytes(bytes: [u8; 8]) -> Self

Create a floating point value from its representation as a byte array in little endian.

See from_bits for some discussion of the portability of this operation (there are almost no issues).

Examples

let value = f64::from_le_bytes([0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x29, 0x40]);
assert_eq!(value, 12.5);
Run

pub fn from_ne_bytes(bytes: [u8; 8]) -> Self

Create a floating point value from its representation as a byte array in native endian.

As the target platform’s native endianness is used, portable code likely wants to use from_be_bytes or from_le_bytes, as appropriate instead.

See from_bits for some discussion of the portability of this operation (there are almost no issues).

Examples

let value = f64::from_ne_bytes(if cfg!(target_endian = "big") {
    [0x40, 0x29, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00]
} else {
    [0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x29, 0x40]
});
assert_eq!(value, 12.5);
Run

pub fn total_cmp(&self, other: &Self) -> Ordering

Return the ordering between self and other.

Unlike the standard partial comparison between floating point numbers, this comparison always produces an ordering in accordance to the totalOrder predicate as defined in the IEEE 754 (2008 revision) floating point standard. The values are ordered in the following sequence:

negative quiet NaN

negative signaling NaN

negative infinity

negative numbers

negative subnormal numbers

negative zero

positive zero

positive subnormal numbers

positive numbers

positive infinity

positive signaling NaN

positive quiet NaN.

The ordering established by this function does not always agree with the PartialOrd and PartialEq implementations of f64. For example, they consider negative and positive zero equal, while total_cmp doesn’t.

The interpretation of the signaling NaN bit follows the definition in the IEEE 754 standard, which may not match the interpretation by some of the older, non-conformant (e.g. MIPS) hardware implementations.

Example

struct GoodBoy {
    name: String,
    weight: f64,
}

let mut bois = vec![
    GoodBoy { name: "Pucci".to_owned(), weight: 0.1 },
    GoodBoy { name: "Woofer".to_owned(), weight: 99.0 },
    GoodBoy { name: "Yapper".to_owned(), weight: 10.0 },
    GoodBoy { name: "Chonk".to_owned(), weight: f64::INFINITY },
    GoodBoy { name: "Abs. Unit".to_owned(), weight: f64::NAN },
    GoodBoy { name: "Floaty".to_owned(), weight: -5.0 },
];

bois.sort_by(|a, b| a.weight.total_cmp(&b.weight));
Run

pub fn clamp(self, min: f64, max: f64) -> f64

Restrict a value to a certain interval unless it is NaN.

Returns max if self is greater than max, and min if self is less than min. Otherwise this returns self.

Note that this function returns NaN if the initial value was NaN as well.

Panics

Panics if min > max, min is NaN, or max is NaN.

Examples

assert!((-3.0f64).clamp(-2.0, 1.0) == -2.0);
assert!((0.0f64).clamp(-2.0, 1.0) == 0.0);
assert!((2.0f64).clamp(-2.0, 1.0) == 1.0);
assert!((f64::NAN).clamp(-2.0, 1.0).is_nan());
Run