Module core::fmt::num

Integer and floating-point number formatting

Modules

imp ğŸ”’

Macros

debug ğŸ”’

impl_Display ğŸ”’

impl_Exp ğŸ”’

impl_int ğŸ”’

impl_uint ğŸ”’

int_base ğŸ”’

integer ğŸ”’

radix ğŸ”’

Structs

Binary ğŸ”’

A binary (base 2) radix

LowerHex ğŸ”’

A hexadecimal (base 16) radix, formatted with lower-case characters

Octal ğŸ”’

An octal (base 8) radix

UpperHex ğŸ”’

A hexadecimal (base 16) radix, formatted with upper-case characters

Statics

DEC_DIGITS_LUT ğŸ”’

Traits

DisplayInt ğŸ”’

GenericRadix ğŸ”’

A type that represents a specific radix

Functions

exp_u128 ğŸ”’

fmt_u128 ğŸ”’

Specialized optimization for u128. Instead of taking two items at a time, it splits into at most 2 u64s, and then chunks by 10e16, 10e8, 10e4, 10e2, and then 10e1. It also has to handle 1 last item, as 10^40 > 2^128 > 10^39, whereas 10^20 > 2^64 > 10^19.

parse_u64_into ğŸ”’

Helper function for writing a u64 into buf going from last to first, with curr.

u128_mulhi ğŸ”’

Multiply unsigned 128 bit integers, return upper 128 bits of the result

udiv_1e19 ğŸ”’

Partition of n into n > 1e19 and rem <= 1e19