Trait core::fmt::UpperHex

pub trait UpperHex {
    fn fmt(&self, &mut Formatter) -> Result;
}

Format trait for the X character.

The UpperHex trait should format its output as a number in hexidecimal, with A through F in upper case.

The alternate flag, #, adds a 0x in front of the output.

For more information on formatters, see the module-level documentation.

Examples

Basic usage with i32:
fn main() { let x = 42; // 42 is '2A' in hex assert_eq!(format!("{:X}", x), "2A"); assert_eq!(format!("{:#X}", x), "0x2A"); }

let x = 42; // 42 is '2A' in hex

assert_eq!(format!("{:X}", x), "2A");
assert_eq!(format!("{:#X}", x), "0x2A");

Implementing UpperHex on a type:
fn main() { use std::fmt; struct Length(i32); impl fmt::UpperHex for Length { fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result { let val = self.0; write!(f, "{:X}", val) // delegate to i32's implementation } } let l = Length(9); println!("l as hex is: {:X}", l); }

use std::fmt;

struct Length(i32);

impl fmt::UpperHex for Length {
    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
        let val = self.0;

        write!(f, "{:X}", val) // delegate to i32's implementation
    }
}

let l = Length(9);

println!("l as hex is: {:X}", l);

Required Methods

fn fmt(&self, &mut Formatter) -> Result

Formats the value using the given formatter.