Function core::future::poll_fn

pub fn poll_fn<T, F>(f: F) -> PollFn<F>where
    F: FnMut(&mut Context<'_>) -> Poll<T>,

Creates a future that wraps a function returning Poll.

Polling the future delegates to the wrapped function. If the returned future is pinned, then the captured environment of the wrapped function is also pinned in-place, so as long as the closure does not move out of its captures it can soundly create pinned references to them.

Examples

use core::future::poll_fn;
use std::task::{Context, Poll};

fn read_line(_cx: &mut Context<'_>) -> Poll<String> {
    Poll::Ready("Hello, World!".into())
}

let read_future = poll_fn(read_line);
assert_eq!(read_future.await, "Hello, World!".to_owned());
Run

This documentation is an old archive. Please see https://rust.docs.kernel.org instead.