Function core::hash::hash

pub fn hash<T: Hash, H: Hasher + Default>(value: &T) -> u64

Unstable
: module was recently redesigned

Hash a value with the default SipHasher algorithm (two initial keys of 0).

The specified value will be hashed with this hasher and then the resulting hash will be returned.