Struct core::hash::BuildHasherDefault

pub struct BuildHasherDefault<H>(_);

The BuildHasherDefault structure is used in scenarios where one has a type that implements Hasher and Default, but needs that type to implement BuildHasher.

This structure is zero-sized and does not need construction.

Examples

Using BuildHasherDefault to specify a custom BuildHasher for HashMap:

use std::collections::HashMap;
use std::hash::{BuildHasherDefault, Hasher};

#[derive(Default)]
struct MyHasher;

impl Hasher for MyHasher {
    fn write(&mut self, bytes: &[u8]) {
        // Your hashing algorithm goes here!
       unimplemented!()
    }

    fn finish(&self) -> u64 {
        // Your hashing algorithm goes here!
        unimplemented!()
    }
}

type MyBuildHasher = BuildHasherDefault<MyHasher>;

let hash_map = HashMap::<u32, u32, MyBuildHasher>::default();Run