Struct core::hash::SipHasher

pub struct SipHasher {
    // some fields omitted
}

An implementation of SipHash 2-4.

See: http://131002.net/siphash/

Consider this as a main "general-purpose" hash for all hashtables: it runs at good speed (competitive with spooky and city) and permits strong keyed hashing. Key your hashtables from a strong RNG, such as rand::Rng.

Although the SipHash algorithm is considered to be cryptographically strong, this implementation has not been reviewed for such purposes. As such, all cryptographic uses of this implementation are strongly discouraged.

Methods

impl SipHasher

fn new() -> SipHasher

Creates a new SipHasher with the two initial keys set to 0.

fn new_with_keys(key0: u64, key1: u64) -> SipHasher

Creates a SipHasher that is keyed off the provided keys.