Function core::hint::black_box

pub fn black_box<T>(dummy: T) -> T

🔬 This is a nightly-only experimental API. (bench_black_box #64102)

一个恒等函数，提示编译器对 black_box 可以做的事情极度悲观。

An identity function that hints to the compiler to be maximally pessimistic about what black_box could do.

与 std::convert::identity 不同，Rust 编译器被鼓励假设 black_box 可以以允许 Rust 代码的任何可能有效方式使用 dummy ，而不会在调用代码中引入未定义的行为。 此属性使 black_box 可用于编写不需要某些优化的代码，例如基准测试。

Unlike std::convert::identity, a Rust compiler is encouraged to assume that black_box can use dummy in any possible valid way that Rust code is allowed to without introducing undefined behavior in the calling code. This property makes black_box useful for writing code in which certain optimizations are not desired, such as benchmarks.

但是请注意，black_box 仅（并且只能）在“尽力而为”的基础上提供。 它可以阻止优化的程度可能会因平台和使用的代码生成后端而异。 程序不能以任何方式依赖 black_box 的正确性。

Note however, that black_box is only (and can only be) provided on a “best-effort” basis. The extent to which it can block optimisations may vary depending upon the platform and code-gen backend used. Programs cannot rely on black_box for correctness in any way.