Function core::hint::unreachable_unchecked

pub unsafe fn unreachable_unchecked() -> !

通知编译器代码中的这一点不可到达，从而实现进一步优化。

Informs the compiler that this point in the code is not reachable, enabling further optimizations.

Safety

达到此功能是完全未定义的行为 (UB)。 特别是，编译器假定所有 UB 都不能发生，因此将消除所有到达 unreachable_unchecked() 调用的分支。

Reaching this function is completely undefined behavior (UB). In particular, the compiler assumes that all UB must never happen, and therefore will eliminate all branches that reach to a call to unreachable_unchecked().

像所有 UB 实例一样，如果这个假设被证明是错误的，即 unreachable_unchecked() 调用实际上在所有可能的控制流中都可达，编译器将应用错误的优化策略，有时甚至可能损坏看似无关的代码，导致 难以调试的问题。

Like all instances of UB, if this assumption turns out to be wrong, i.e., the unreachable_unchecked() call is actually reachable among all possible control flow, the compiler will apply the wrong optimization strategy, and may sometimes even corrupt seemingly unrelated code, causing difficult-to-debug problems.

仅当您可以证明代码永远不会调用它时才使用此函数。 否则，请考虑使用不可访问！ 宏，不允许优化，但在执行时会panic。

Use this function only when you can prove that the code will never call it. Otherwise, consider using the unreachable! macro, which does not allow optimizations but will panic when executed.

Example

fn div_1(a: u32, b: u32) -> u32 {
    use std::hint::unreachable_unchecked;

    // `b.saturating_add(1)` is always positive (not zero),
    // hence `checked_div` will never return `None`.
    // Therefore, the else branch is unreachable.
    a.checked_div(b.saturating_add(1))
        .unwrap_or_else(|| unsafe { unreachable_unchecked() })
}

assert_eq!(div_1(7, 0), 7);
assert_eq!(div_1(9, 1), 4);
assert_eq!(div_1(11, u32::MAX), 0);
Run