Module core::i128

👎Deprecating in a future Rust version: all constants in this module replaced by associated constants on i128

Constants for the 128-bit signed integer type.

See also the i128 primitive type.

New code should use the associated constants directly on the primitive type.

Constants

MAXDeprecation planned

The largest value that can be represented by this integer type. Use i128::MAX instead.

MINDeprecation planned

The smallest value that can be represented by this integer type. Use i128::MIN instead.

This documentation is an old archive. Please see https://rust.docs.kernel.org instead.