Module core::i64

ğŸ‘ Deprecating in a future Rust version:
all constants in this module replaced by associated constants on i64

64 ä½æœ‰ç¬¦å·æ•´æ•°ç±»å‹çš„å¸¸é‡ã€‚

See also the i64 primitive type.

æ–°ä»£ç åº”ç›´æ¥åœ¨åŸå‹ç±»å‹ä¸Šä½¿ç”¨å…³è”çš„å¸¸é‡ã€‚

Constants

MAXDeprecation planned

The largest value that can be represented by this integer type. Use i64::MAX instead.

MINDeprecation planned

The smallest value that can be represented by this integer type. Use i64::MIN instead.