core::intrinsics::fallback

Trait FunnelShift

pub trait FunnelShift: Copy + 'static {
    // Required methods
    unsafe fn unchecked_funnel_shl(self, rhs: Self, shift: u32) -> Self;
    unsafe fn unchecked_funnel_shr(self, rhs: Self, shift: u32) -> Self;
}

🔬This is a nightly-only experimental API. (core_intrinsics_fallbacks)

Required Methods

unsafe fn unchecked_funnel_shl(self, rhs: Self, shift: u32) -> Self

🔬This is a nightly-only experimental API. (core_intrinsics_fallbacks)

See super::unchecked_funnel_shl; we just need the trait indirection to handle different types since calling intrinsics with generics doesn’t work.

unsafe fn unchecked_funnel_shr(self, rhs: Self, shift: u32) -> Self

🔬This is a nightly-only experimental API. (core_intrinsics_fallbacks)

See super::unchecked_funnel_shr; we just need the trait indirection to handle different types since calling intrinsics with generics doesn’t work.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.