Function core::intrinsics::arith_offset

pub unsafe extern "rust-intrinsic" fn arith_offset<T>(
    dst: *const T, 
    offset: isize
) -> *const T

ğŸ”¬ This is a nightly-only experimental API. (core_intrinsics)

intrinsics are unlikely to ever be stabilized, instead they should be used through stabilized interfaces in the rest of the standard library

è®¡ç®—ä¸æŒ‡é’ˆçš„åç»é‡ (å¯èƒ½ä¼šè‡ªåŠ¨æ¢è¡Œ)ã€‚

è¿™è¢«å®ç°ä¸ºå†…éƒ¨å‡½æ•°ï¼Œä»¥é¿å…ä¸æ•´æ•°è¿›è¡Œç›¸äº’è½¬æ¢ï¼Œå› ä¸ºè¯¥è½¬æ¢ä¼šç¦æ­¢æŸäº›ä¼˜åŒ–ã€‚

Safety

ä¸ offset å†…éƒ¨å‡½æ•°ä¸åŒï¼Œæ­¤å†…éƒ¨å‡½æ•°ä¸ä¼šé™åˆ¶ç»“æœæŒ‡é’ˆæŒ‡å‘å·²åˆ†é…å¯¹è±¡çš„æœ«å°¾æˆ–æŒ‡å‘è¯¥å¯¹è±¡æœ«å°¾ä¸€ä¸ªå­—èŠ‚ï¼Œå¹¶ä¸”ä½¿ç”¨äºŒè¿›åˆ¶è¡¥ç ç®—æ³•è¿›è¡Œæ¢è¡Œã€‚ ç»“æœå€¼ä¸ä¸€å®šæœ‰æ•ˆåœ°ç”¨äºå®é™…è®¿é—®å†…å­˜ã€‚

æ­¤å†…éƒ¨å‡½æ•°çš„ç¨³å®šç‰ˆæœ¬ä¸º pointer::wrapping_offsetã€‚