core::intrinsics

Function atomic_umin

pub unsafe fn atomic_umin<T: Copy, const ORD: AtomicOrdering>(
    dst: *mut T,
    src: T,
) -> T

🔬This is a nightly-only experimental API. (core_intrinsics)

Minimum with the current value using an unsigned comparison. T must be an unsigned integer type.

The stabilized version of this intrinsic is available on the atomic unsigned integer types via the fetch_min method. For example, AtomicU32::fetch_min.