Function core::intrinsics::atomic_xchg_release

pub unsafe extern "rust-intrinsic" fn atomic_xchg_release<T: Copy>(
    dst: *mut T,
    src: T
) -> T

🔬This is a nightly-only experimental API. (core_intrinsics)

Stores the value at the specified memory location, returning the old value.

The stabilized version of this intrinsic is available on the atomic types via the swap method by passing Ordering::Release as the order. For example, AtomicBool::swap.