Function core::intrinsics::const_eval_select

pub unsafe fn const_eval_select<ARG, F, G, RET>(
    arg: ARG, 
    _called_in_const: F, 
    called_at_rt: G
) -> RET where
    F: ~constFnOnce<ARG, Output = RET>,
    G: FnOnce<ARG, Output = RET> + ~constDrop, 

🔬 This is a nightly-only experimental API. (const_eval_select)

根据上下文选择要调用的函数。

Selects which function to call depending on the context.

如果这个函数是在编译时计算的，那么对这个内部函数的调用将被替换为对 `called_in_const` 的调用。 否则，它会被调用 `called_at_rt` 替换。

If this function is evaluated at compile-time, then a call to this intrinsic will be replaced with a call to called_in_const. It gets replaced with a call to called_at_rt otherwise.

Type Requirements

这两个功能必须都是功能项。 它们不能是函数指针或闭包。

The two functions must be both function items. They cannot be function pointers or closures.

`arg` 将是传递给两个函数之一的参数，因此，两个函数必须接受相同类型的参数。 这两个函数都必须返回 RET。

arg will be the arguments that will be passed to either one of the two functions, therefore, both functions must accept the same type of arguments. Both functions must return RET.

Safety

此内在函数允许破坏 `const fn` 中的“引用透明性”，因此是“不安全的”。

This intrinsic allows breaking referential transparency in const fn and is therefore unsafe.

使用此内在函数的代码必须非常小心，以确保 `const fn` 保持引用透明，与评估时无关。

Code that uses this intrinsic must be extremely careful to ensure that const fns remain referentially-transparent independently of when they are evaluated.

Rust 编译器假定用编译时评估它产生的结果替换对 `const fn` 的调用是合理的。 如果在运行时评估函数会产生不同的结果，或者有任何其他可观察到的副作用，则行为是未定义的。

The Rust compiler assumes that it is sound to replace a call to a const fn with the result produced by evaluating it at compile-time. If evaluating the function at run-time were to produce a different result, or have any other observable side-effects, the behavior is undefined.