Function core::intrinsics::ctlz_nonzero

pub unsafe extern "rust-intrinsic" fn ctlz_nonzero<T>(x: T) -> T

🔬 This is a nightly-only experimental API. (core_intrinsics)

intrinsics are unlikely to ever be stabilized, instead they should be used through stabilized interfaces in the rest of the standard library

Like ctlz, but extra-unsafe as it returns undef when given an x with value 0.

Examples

#![feature(core_intrinsics)]

use std::intrinsics::ctlz_nonzero;

let x = 0b0001_1100_u8;
let num_leading = unsafe { ctlz_nonzero(x) };
assert_eq!(num_leading, 3);Run