core::intrinsics::cttz16

pub unsafe extern "rust-intrinsic" fn cttz16(x: u16) -> u16

Unstable (core_intrinsics #0)
: intrinsics are unlikely to ever be stabilized, instead they should be used through stabilized interfaces in the rest of the standard library

Returns the number of trailing bits unset in a u16.