Function core::intrinsics::cttz_nonzero

pub unsafe extern "rust-intrinsic" fn cttz_nonzero<T: Copy>(x: T) -> T

🔬 This is a nightly-only experimental API. (core_intrinsics)

intrinsics are unlikely to ever be stabilized, instead they should be used through stabilized interfaces in the rest of the standard library

Like cttz, but extra-unsafe as it returns undef when given an x with value 0.

This intrinsic does not have a stable counterpart.

Examples

#![feature(core_intrinsics)]

use std::intrinsics::cttz_nonzero;

let x = 0b0011_1000_u8;
let num_trailing = unsafe { cttz_nonzero(x) };
assert_eq!(num_trailing, 3);
Run