Function core::intrinsics::needs_drop

pub extern "rust-intrinsic" fn needs_drop<T>() -> bool

🔬 This is a nightly-only experimental API. (core_intrinsics)

如果以 `T` 给出的实际类型需要滴胶，则返回 `true`； 如果为 `T` 提供的实际类型实现了 `Copy`，则返回 `false`。

Returns true if the actual type given as T requires drop glue; returns false if the actual type provided for T implements Copy.

如果实际类型既不需要滴胶也没有实现`Copy`，那么这个函数的返回值是未指定的。

If the actual type neither requires drop glue nor implements Copy, then the return value of this function is unspecified.

请注意，与大多数内在函数不同，这是可以安全调用的； 它不需要“不安全”块。 因此，实现不得要求用户维护任何安全不变量。

Note that, unlike most intrinsics, this is safe to call; it does not require an unsafe block. Therefore, implementations must not require the user to uphold any safety invariants.

这个内在函数的稳定版本是“mem::needs_drop”。

The stabilized version of this intrinsic is mem::needs_drop.