Function core::intrinsics::prefetch_read_instruction

pub unsafe extern "rust-intrinsic" fn prefetch_read_instruction<T>(
    data: *const T, 
    locality: i32
)

ğŸ”¬ This is a nightly-only experimental API. (core_intrinsics)

intrinsics are unlikely to ever be stabilized, instead they should be used through stabilized interfaces in the rest of the standard library

prefetch å†…è”å‡½æ•°æ˜¯å¯¹ä»£ç ç”Ÿæˆå™¨çš„æç¤ºï¼Œå¦‚æœæ”¯æŒçš„è¯ï¼Œå®ƒä¼šæ’å…¥ä¸€ä¸ªé¢„å–æŒ‡ä»¤ã€‚å¦åˆ™ï¼Œå®ƒæ˜¯æ— æ“ä½œçš„ã€‚ é¢„å–å¯¹ç¨‹åºçš„è¡Œä¸ºæ²¡æœ‰å½±å“ï¼Œä½†å¯ä»¥æ›´æ”¹å…¶æ€èƒ½ç‰¹å¾ã€‚

locality å‚æ•°å¿…é¡»æ˜¯ä¸€ä¸ªå¸¸é‡æ•´æ•°ï¼Œå¹¶ä¸”æ˜¯æ—¶é—´å±€éƒ¨æ€è¯´æ˜ç¬¦ï¼ŒèŒƒå›´ä» (0) (æ— å±€éƒ¨æ€) åˆ° (3) (åœ¨ç¼“å­˜ä¸­æå…¶å±€éƒ¨åŒ–)ã€‚

æ­¤å†…è”å‡½æ•°æ²¡æœ‰ç¨³å®šçš„å¯¹åº”å¯¹è±¡ã€‚