core

Module io

🔬This is a nightly-only experimental API. (core_io #154046)

Traits, helpers, and type definitions for core I/O functionality.

Structs

BorrowedBufExperimental

A borrowed buffer of initially uninitialized elements, which is incrementally filled.

BorrowedCursorExperimental

A writeable view of the unfilled portion of a BorrowedBuf.

ChainExperimental

Adapter to chain together two readers.

CursorExperimental

A Cursor wraps an in-memory buffer and provides it with a Seek implementation.

EmptyExperimental

Empty ignores any data written via Write, and will always be empty (returning zero bytes) when read via Read.

IoSliceExperimental

A buffer type used with Write::write_vectored.

IoSliceMutExperimental

A buffer type used with Read::read_vectored.

RepeatExperimental

A reader which yields one byte over and over and over and over and over and…

SinkExperimental

A writer which will move data into the void.

TakeExperimental

Reader adapter which limits the bytes read from an underlying reader.

Enums

ErrorKindExperimental

A list specifying general categories of I/O error.

Functions

emptyExperimental

Creates a value that is always at EOF for reads, and ignores all data written.

repeatExperimental

Creates an instance of a reader that infinitely repeats one byte.

sinkExperimental

Creates an instance of a writer which will successfully consume all data.

Type Aliases

RawOsErrorExperimental

The type of raw OS error codes.