core::io

Module prelude

🔬This is a nightly-only experimental API. (core_io #154046)

The I/O Prelude.

The purpose of this module is to alleviate imports of many common I/O traits by adding a glob import to the top of I/O heavy modules:

use core::io::prelude::*;

Re-exports

pub use crate::io::Seek;

pub use crate::io::Write;