core::io

Struct BorrowedCursor

Write

pub struct BorrowedCursor<'a, T> { /* private fields */ }

🔬This is a nightly-only experimental API. (core_io_borrowed_buf #117693)

A writeable view of the unfilled portion of a BorrowedBuf.

The unfilled portion may be uninitialized; see BorrowedBuf for details.

Data can be written directly to the cursor by using append or indirectly by getting a slice of part or all of the cursor and writing into the slice. In the indirect case, the caller must call advance after writing to inform the cursor how many elements have been written.

Once elements are written to the cursor, they become part of the filled portion of the underlying BorrowedBuf and can no longer be accessed or re-written by the cursor. In other words, the cursor tracks the unfilled part of the underlying BorrowedBuf.

The lifetime 'a is a bound on the lifetime of the underlying buffer (which means it is a bound on the elements in that buffer by transitivity).

Implementations

impl<'a, T: Copy> BorrowedCursor<'a, T>

pub fn reborrow<'this>(&'this mut self) -> BorrowedCursor<'this, T>

🔬This is a nightly-only experimental API. (core_io_borrowed_buf #117693)

Reborrows this cursor by cloning it with a smaller lifetime.

Since a cursor maintains unique access to its underlying buffer, the borrowed cursor is not accessible while the new cursor exists.

pub fn capacity(&self) -> usize

🔬This is a nightly-only experimental API. (core_io_borrowed_buf #117693)

Returns the available space in the cursor.

pub fn written(&self) -> usize

🔬This is a nightly-only experimental API. (core_io_borrowed_buf #117693)

Returns the number of elements written to the BorrowedBuf this cursor was created from.

In particular, the count returned is shared by all reborrows of the cursor.

pub fn is_init(&self) -> bool

🔬This is a nightly-only experimental API. (borrowed_buf_init #160476)

Returns true if the buffer is initialized.

pub unsafe fn set_init(&mut self)

🔬This is a nightly-only experimental API. (borrowed_buf_init #160476)

Set the buffer as fully initialized.

Safety

All the elements of the cursor must be initialized.

pub unsafe fn as_mut(&mut self) -> &mut [MaybeUninit<T>]

🔬This is a nightly-only experimental API. (core_io_borrowed_buf #117693)

Returns a mutable reference to the whole cursor.

Safety

The caller must not uninitialize any elements of the cursor if it is initialized.

pub fn advance_checked(&mut self, n: usize) -> &mut Self

🔬This is a nightly-only experimental API. (borrowed_buf_init #160476)

Advances the cursor by asserting that n elements have been filled.

After advancing, the n elements are no longer accessible via the cursor and can only be accessed via the underlying buffer. I.e., the buffer’s filled portion grows by n elements and its unfilled portion (and the capacity of this cursor) shrinks by n elements.

If less than n elements initialized (by the cursor’s point of view), set_init should be called first.

Panics

Panics if there are less than n elements initialized.

pub unsafe fn advance(&mut self, n: usize) -> &mut Self

🔬This is a nightly-only experimental API. (core_io_borrowed_buf #117693)

Advances the cursor by asserting that n elements have been filled.

After advancing, the n elements are no longer accessible via the cursor and can only be accessed via the underlying buffer. I.e., the buffer’s filled portion grows by n elements and its unfilled portion (and the capacity of this cursor) shrinks by n elements.

Safety

The caller must ensure that the first n elements of the cursor have been initialized.

pub fn append(&mut self, buf: &[T])

🔬This is a nightly-only experimental API. (core_io_borrowed_buf #117693)

Append elements to the cursor, advancing position within its buffer.

Panics

Panics if self.capacity() is less than buf.len().

pub fn with_unfilled_buf<R>( &mut self, f: impl FnOnce(&mut BorrowedBuf<'_, T>) -> R, ) -> R

🔬This is a nightly-only experimental API. (core_io_borrowed_buf #117693)

Runs the given closure with a BorrowedBuf containing the unfilled part of the cursor.

This enables inspecting what was written to the cursor.

Panics

Panics if the BorrowedBuf given to the closure is replaced by another one.

impl<'a, T: Default + Copy> BorrowedCursor<'a, T>

pub fn ensure_init(&mut self) -> &mut [T]

🔬This is a nightly-only experimental API. (borrowed_buf_init #160476)

Initializes all elements in the cursor with their default value and returns them.