Module core::iter::order

Deprecated since 1.4.0
: use the equivalent methods on Iterator instead

Functions for lexicographical ordering of sequences.

Lexicographical ordering through <, <=, >=, > requires that the elements implement both PartialEq and PartialOrd.

If two sequences are equal up until the point where one ends, the shorter sequence compares less.

Functions

cmp 	 [Deprecated]
Order a and b lexicographically using Ord

eq 	 [Deprecated]
Compare a and b for equality (Using partial equality, PartialEq)

equals 	 [Deprecated]
Compare a and b for equality using Eq

ge 	 [Deprecated]
Returns a >= b lexicographically (Using partial order, PartialOrd)

gt 	 [Deprecated]
Returns a > b lexicographically (Using partial order, PartialOrd)

le 	 [Deprecated]
Returns a <= b lexicographically (Using partial order, PartialOrd)

lt 	 [Deprecated]
Returns a < b lexicographically (Using partial order, PartialOrd)

ne 	 [Deprecated]
Compares a and b for nonequality (Using partial equality, PartialEq)

partial_cmp 	 [Deprecated]
Order a and b lexicographically using PartialOrd