Trait core::iter::AdditiveIteratorUnstable

pub trait AdditiveIterator<A> {
    fn sum(self) -> A;
}

A trait for iterators over elements which can be added together

Required Methods

fn sum(self) -> A

Iterates over the entire iterator, summing up all the elements

Examples
fn main() { use std::iter::AdditiveIterator; let a = [1, 2, 3, 4, 5]; let mut it = a.iter().cloned(); assert!(it.sum() == 15); }

use std::iter::AdditiveIterator;

let a = [1, 2, 3, 4, 5];
let mut it = a.iter().cloned();
assert!(it.sum() == 15);