Trait core::iter::IntoIterator

pub trait IntoIterator {
    type Item;
    type IntoIter: Iterator<Item=Self::Item>;
    fn into_iter(self) -> Self::IntoIter;
}

Conversion into an Iterator

Implementing this trait allows you to use your type with Rust's for loop. See the module level documentation for more details.

Associated Types

type Item

The type of the elements being iterated

type IntoIter: Iterator<Item=Self::Item>

A container for iterating over elements of type Item

Required Methods

fn into_iter(self) -> Self::IntoIter

Consumes Self and returns an iterator over it