Trait core::iter::Sum

pub trait Sum<A = Self>: Sized {
    fn sum<I: Iterator<Item = A>>(iter: I) -> Self;
}

Trait to represent types that can be created by summing up an iterator.

This trait is used to implement the sum method on iterators. Types which implement the trait can be generated by the sum method. Like FromIterator this trait should rarely be called directly and instead interacted with through Iterator::sum.

Required Methods

fn sum<I: Iterator<Item = A>>(iter: I) -> Self

Method which takes an iterator and generates Self from the elements by "summing up" the items.