Macro core::assert

macro_rules! assert {
    ($cond : expr $(,) ?) => { ... };
    ($cond : expr, $($arg : tt) +) => { ... };
}

断言布尔表达式在运行时为真。

Asserts that a boolean expression is true at runtime.

这将引发panic！ 如果提供的表达式在运行时无法评估为 true，则为宏。

This will invoke the panic! macro if the provided expression cannot be evaluated to true at runtime.

Uses

断言总是在调试和发布版本中检查，并且不能被禁用。 请参阅 debug_assert！ 对于默认情况下在发布版本中未启用的断言。

Assertions are always checked in both debug and release builds, and cannot be disabled. See debug_assert! for assertions that are not enabled in release builds by default.

不安全的代码可能依赖于断言！ 强制执行运行时不变量，如果违反可能导致不安全。

Unsafe code may rely on assert! to enforce run-time invariants that, if violated could lead to unsafety.

断言的其他用例！ 包括在安全代码中测试和强制执行运行时不变量（其违规不会导致不安全）。

Other use-cases of assert! include testing and enforcing run-time invariants in safe code (whose violation cannot result in unsafety).

Custom Messages

这个宏有第二种形式，可以提供自定义紧急消息，带或不带参数用于格式化。 请参阅 std::fmt 了解此表单的语法。 仅当断言失败时才会评估用作格式参数的表达式。

This macro has a second form, where a custom panic message can be provided with or without arguments for formatting. See std::fmt for syntax for this form. Expressions used as format arguments will only be evaluated if the assertion fails.

Examples

// the panic message for these assertions is the stringified value of the
// expression given.
assert!(true);

fn some_computation() -> bool { true } // a very simple function

assert!(some_computation());

// assert with a custom message
let x = true;
assert!(x, "x wasn't true!");

let a = 3; let b = 27;
assert!(a + b == 30, "a = {}, b = {}", a, b);
Run