Macro core::cfg

macro_rules! cfg {
    ($($cfg:tt)*) => { ... };
}

Evaluates boolean combinations of configuration flags at compile-time.

In addition to the #[cfg] attribute, this macro is provided to allow boolean expression evaluation of configuration flags. This frequently leads to less duplicated code.

The syntax given to this macro is the same syntax as the cfg attribute.

Examples

let my_directory = if cfg!(windows) {
    "windows-specific-directory"
} else {
    "unix-directory"
};Run