Macro core::debug_assert_eq

macro_rules! debug_assert_eq {
    ($($arg : tt) *) => { ... };
}

断言两个表达式彼此相等。

Asserts that two expressions are equal to each other.

panic时，此宏将打印表达式的值及其调试表示。

On panic, this macro will print the values of the expressions with their debug representations.

与 "assert_eq!" 不同，`debug_assert_eq!` 语句仅在非优化构建中默认启用。 除非将 `-C debug-assertions` 传递给编译器，否则优化的构建将不会执行 `debug_assert_eq!` 语句。 这使得 `debug_assert_eq!` 对于成本太高而无法出现在发布版本中的检查很有用，但在开发过程中可能会有所帮助。 扩展 `debug_assert_eq!` 的结果总是经过类型检查。

Unlike assert_eq!, debug_assert_eq! statements are only enabled in non optimized builds by default. An optimized build will not execute debug_assert_eq! statements unless -C debug-assertions is passed to the compiler. This makes debug_assert_eq! useful for checks that are too expensive to be present in a release build but may be helpful during development. The result of expanding debug_assert_eq! is always type checked.

Examples

let a = 3;
let b = 1 + 2;
debug_assert_eq!(a, b);
Run