Macro core::include

macro_rules! include {
    ($file : expr $(,) ?) => { ... };
}

根据上下文将文件解析为表达式或项。

Parses a file as an expression or an item according to the context.

该文件相对于当前文件定位（类似于如何找到模块）。 提供的路径在编译时以特定于平台的方式进行解释。 因此，例如，使用包含反斜杠 `\` 的 Windows 路径的调用将无法在 Unix 上正确编译。

The file is located relative to the current file (similarly to how modules are found). The provided path is interpreted in a platform-specific way at compile time. So, for instance, an invocation with a Windows path containing backslashes \ would not compile correctly on Unix.

使用这个宏通常是个坏主意，因为如果文件被解析为表达式，它将不卫生地放置在周围的代码中。 如果当前文件中存在同名的变量或函数，这可能会导致变量或函数与文件预期的不同。

Using this macro is often a bad idea, because if the file is parsed as an expression, it is going to be placed in the surrounding code unhygienically. This could result in variables or functions being different from what the file expected if there are variables or functions that have the same name in the current file.

Examples

假设同一目录下有两个文件，内容如下：

Assume there are two files in the same directory with the following contents:

文件‘monkeys.in’：

File ‘monkeys.in’:

['🙈', '🙊', '🙉']
    .iter()
    .cycle()
    .take(6)
    .collect::<String>()
Run

文件“main.rs”：

File ‘main.rs’:

fn main() {
    let my_string = include!("monkeys.in");
    assert_eq!("🙈🙊🙉🙈🙊🙉", my_string);
    println!("{}", my_string);
}
Run

编译“main.rs”并运行生成的二进制文件将打印“🙈🙊🙉🙈🙊🙉”。

Compiling ‘main.rs’ and running the resulting binary will print “🙈🙊🙉🙈🙊🙉”.