Macro core::module_path

macro_rules! module_path {
    () => { ... };
}

Expands to a string that represents the current module path.

The current module path can be thought of as the hierarchy of modules leading back up to the crate root. The first component of the path returned is the name of the crate currently being compiled.

Examples

mod test {
    pub fn foo() {
        assert!(module_path!().ends_with("test"));
    }
}

test::foo();
Run

This documentation is an old archive. Please see https://rust.docs.kernel.org instead.