Macro core::panic

macro_rules! panic {
    ($($arg : tt) *) => { ... };
}

panic当前线程。

Panics the current thread.

这允许程序立即终止并向程序的调用者提供反馈。 panic！ 应在程序达到不可恢复状态时使用。

This allows a program to terminate immediately and provide feedback to the caller of the program. panic! should be used when a program reaches an unrecoverable state.

这个宏是在示例代码和测试中断言条件的完美方式。 panic！ 与 Option 和 Result 枚举的 unwrap 方法密切相关。 两种实现都称为panic！ 当它们设置为 None 或 Err 变体时。

This macro is the perfect way to assert conditions in example code and in tests. panic! is closely tied with the unwrap method of both Option and Result enums. Both implementations call panic! when they are set to None or Err variants.

使用 panic!() 时，您可以指定一个字符串有效负载，它是使用格式构建的！ 句法。 在将panic注入调用 Rust 线程时使用该有效负载，导致线程完全panic。

When using panic!() you can specify a string payload, that is built using the format! syntax. That payload is used when injecting the panic into the calling Rust thread, causing the thread to panic entirely.

默认 std 钩子的行为，即在调用 panic 后直接运行的代码，是将消息有效负载连同 panic!() 调用的文件/行/列信息一起打印到 stderr。 您可以使用 std::panic::set_hook() 覆盖panic钩子。 在钩子内部，panic 可以作为 &dyn Any + Send 访问，其中包含用于常规 panic!() 调用的 &str 或字符串。 要对另一个其他类型的值进行panic，可以使用 panic_any。

The behavior of the default std hook, i.e. the code that runs directly after the panic is invoked, is to print the message payload to stderr along with the file/line/column information of the panic!() call. You can override the panic hook using std::panic::set_hook(). Inside the hook a panic can be accessed as a &dyn Any + Send, which contains either a &str or String for regular panic!() invocations. To panic with a value of another other type, panic_any can be used.

结果枚举通常是从错误中恢复的更好解决方案，而不是使用panic！ 宏。 应该使用这个宏来避免使用不正确的值，例如来自外部源的值。 有关错误处理的详细信息可在书中找到。

Result enum is often a better solution for recovering from errors than using the panic! macro. This macro should be used to avoid proceeding using incorrect values, such as from external sources. Detailed information about error handling is found in the book.

另请参阅宏 compile_error!，以了解在编译期间引发错误。

See also the macro compile_error!, for raising errors during compilation.

Current implementation

如果主线程发生panic，它将终止所有线程并以代码 101 结束程序。

If the main thread panics it will terminate all your threads and end your program with code 101.

Examples

panic!();
panic!("this is a terrible mistake!");
panic!("this is a {} {message}", "fancy", message = "message");
std::panic::panic_any(4); // panic with the value of 4 to be collected elsewhere
Run