Macro core::todo

macro_rules! todo {
    () => { ... };
    ($($arg : tt) +) => { ... };
}

表示未完成的代码。

Indicates unfinished code.

如果您正在制作原型并且只是希望对代码进行类型检查，这可能会很有用。

This can be useful if you are prototyping and are just looking to have your code typecheck.

“未实现！”的区别！ 和 `todo!` 是，虽然 `todo!` 传达了稍后实现功能的意图并且消息“尚未实现”，但 `unimplemented!` 没有这样的声明。 它的信息是“未实施”。 还有一些 IDE 会标记 `todo!`s。

The difference between unimplemented! and todo! is that while todo! conveys an intent of implementing the functionality later and the message is “not yet implemented”, unimplemented! makes no such claims. Its message is “not implemented”. Also some IDEs will mark todo!s.

Panics

这将永远“panic！”。

This will always panic!.

Examples

这是一些正在进行的代码的示例。 我们有一个Trait `Foo`：

Here’s an example of some in-progress code. We have a trait Foo:

trait Foo {
    fn bar(&self);
    fn baz(&self);
}
Run

我们想在我们的一种类型上实现“Foo”，但我们也想首先处理“bar()”。 为了让我们的代码能够编译，我们需要实现 `baz()`，所以我们可以使用 `todo!`：

We want to implement Foo on one of our types, but we also want to work on just bar() first. In order for our code to compile, we need to implement baz(), so we can use todo!:

struct MyStruct;

impl Foo for MyStruct {
    fn bar(&self) {
        // implementation goes here
    }

    fn baz(&self) {
        // let's not worry about implementing baz() for now
        todo!();
    }
}

fn main() {
    let s = MyStruct;
    s.bar();

    // we aren't even using baz(), so this is fine.
}
Run