Struct core::marker::NoCopy

pub struct NoCopy;

Unstable
: likely to change with new variance strategy

A type which is considered "not POD", meaning that it is not implicitly copyable. This is typically embedded in other types to ensure that they are never copied, even if they lack a destructor.