Trait core::marker::DiscriminantKind

pub trait DiscriminantKind {
    type Discriminant: Clone + Copy + Debug + Eq + PartialEq + Hash + Send + Sync + Unpin;
}

🔬 This is a nightly-only experimental API. (discriminant_kind)

this trait is unlikely to ever be stabilized, use mem::discriminant instead

用于指示枚举判别式类型的编译器内部Trait。

Compiler-internal trait used to indicate the type of enum discriminants.

这个特性会自动为每种类型实现，并且不会为 mem::Discriminant 添加任何保证。 在 DiscriminantKind::Discriminant 和 mem::Discriminant 之间转换是未定义的行为。

This trait is automatically implemented for every type and does not add any guarantees to mem::Discriminant. It is undefined behavior to transmute between DiscriminantKind::Discriminant and mem::Discriminant.

Associated Types

type Discriminant: Clone + Copy + Debug + Eq + PartialEq + Hash + Send + Sync + Unpin

🔬 This is a nightly-only experimental API. (discriminant_kind)

this trait is unlikely to ever be stabilized, use mem::discriminant instead

The type of the discriminant, which must satisfy the trait bounds required by mem::Discriminant.