core::marker

Trait UnsafeUnpin

pub unsafe auto trait UnsafeUnpin { }

🔬This is a nightly-only experimental API. (unsafe_unpin #125735)

Used to determine whether a type contains any UnsafePinned (or PhantomPinned) internally, but not through an indirection. This affects, for example, whether we emit noalias metadata for &mut T or not.

This is part of RFC 3467, and is tracked by #125735.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".