Function core::mem::min_align_of_val

pub fn min_align_of_val<T: ?Sized>(val: &T) -> usize

👎Deprecated since 1.2.0: use align_of_val instead

Returns the ABI-required minimum alignment of the type of the value that val points to in bytes.

Every reference to a value of the type T must be a multiple of this number.

Examples

use std::mem;

assert_eq!(4, mem::min_align_of_val(&5i32));
Run

This documentation is an old archive. Please see https://rust.docs.kernel.org instead.