Function core::mem::zeroed

pub unsafe fn zeroed<T>() -> T

Creates a value initialized to zero.

This function is similar to allocating space for a local variable and zeroing it out (an unsafe operation).

Care must be taken when using this function, if the type T has a destructor and the value falls out of scope (due to unwinding or returning) before being initialized, then the destructor will run on zeroed data, likely leading to crashes.

This is useful for FFI functions sometimes, but should generally be avoided.

Examples
fn main() { use std::mem; let x: i32 = unsafe { mem::zeroed() }; }

use std::mem;

let x: i32 = unsafe { mem::zeroed() };