core::mem

Module type_info

🔬This is a nightly-only experimental API. (type_info #146922)

MVP for exposing compile-time information about types in a runtime or const-eval processable way.

Structs

ArrayExperimental

Compile-time type information about arrays.

BoolExperimental

Compile-time type information about bool.

CharExperimental

Compile-time type information about char.

FieldExperimental

Compile-time type information about fields of tuples, structs and enum variants.

FloatExperimental

Compile-time type information about floating-point types.

IntExperimental

Compile-time type information about signed and unsigned integer types.

PointerExperimental

Compile-time type information about pointers.

ReferenceExperimental

Compile-time type information about references.

StrExperimental

Compile-time type information about string slice types.

TupleExperimental

Compile-time type information about tuples.

TypeExperimental

Compile-time type information.

Enums

TypeKindExperimental

Compile-time type information.