Module core::nonzero

Unstable (nonzero #27730)
: needs an RFC to flesh out the design

Exposes the NonZero lang item which provides optimization hints.

Structs

NonZero 	 [Unstable]
A wrapper type for raw pointers and integers that will never be NULL or 0 that might allow certain optimizations.

Traits

Zeroable 	 [Unstable]
Unsafe trait to indicate what types are usable with the NonZero struct