Module core::num::dec2flt::num

Unstable (dec2flt)
: internal routines only exposed for testing

Utility functions for bignums that don't make too much sense to turn into methods.

Reexports

use prelude::v1::*;

use cmp::Ordering::{self, Less, Equal, Greater};

pub use num::bignum::Big32x40 as Big;

Functions

compare_with_half_ulp 	 [Unstable]
Test whether truncating all bits less significant than ones_place introduces a relative error less, equal, or greater than 0.5 ULP.

digits_to_big 	 [Unstable]
Convert a string of ASCII digits into a bignum.

from_str_unchecked 	 [Unstable]
Convert an ASCII string containing only decimal digits to a u64.

get_bits 	 [Unstable]
Extract a range of bits. Index 0 is the least significant bit and the range is half-open as usual. Panics if asked to extract more bits than fit into the return type.

to_u64 	 [Unstable]
Unwraps a bignum into a 64 bit integer. Panics if the number is too large.