Enum core::num::IntErrorKind

#[non_exhaustive]
pub enum IntErrorKind {
    Empty,
    InvalidDigit,
    PosOverflow,
    NegOverflow,
    Zero,
}

Enum to store the various types of errors that can cause parsing an integer to fail.

Example

if let Err(e) = i32::from_str_radix("a12", 10) {
    println!("Failed conversion to i32: {:?}", e.kind());
}
Run

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Empty

Value being parsed is empty.

This variant will be constructed when parsing an empty string.

InvalidDigit

Contains an invalid digit in its context.

Among other causes, this variant will be constructed when parsing a string that contains a non-ASCII char.

This variant is also constructed when a + or - is misplaced within a string either on its own or in the middle of a number.

PosOverflow

Integer is too large to store in target integer type.

NegOverflow

Integer is too small to store in target integer type.

Zero

Value was Zero

This variant will be emitted when the parsing string has a value of zero, which would be illegal for non-zero types.