Module core::num::int_log10

Functions

i8

i16

i32

i64

i128

less_than_5 🔒

panic_for_nonpositive_argument

Instantiate this panic logic once, rather than for all the ilog methods on every single primitive type.

u8

These functions compute the integer logarithm of their type, assuming that someone has already checked that the value is strictly positive.

u16

u32

u64

u128

usize