Module core::num::nonzero

Definitions of integer that is known not to equal zero.

Macros

from_str_radix_nzint_impl 🔒

impl_nonzero_fmt 🔒

nonzero_bits 🔒

nonzero_integers 🔒

nonzero_integers_div 🔒

nonzero_leading_trailing_zeros 🔒

nonzero_min_max_signed 🔒

nonzero_min_max_unsigned 🔒

nonzero_signed_operations 🔒

nonzero_unsigned_is_power_of_two 🔒

nonzero_unsigned_operations 🔒

nonzero_unsigned_signed_operations 🔒

sign_dependent_expr 🔒

Structs

NonZeroI8

An integer that is known not to equal zero.

NonZeroI16

An integer that is known not to equal zero.

NonZeroI32

An integer that is known not to equal zero.

NonZeroI64

An integer that is known not to equal zero.

NonZeroI128

An integer that is known not to equal zero.

NonZeroIsize

An integer that is known not to equal zero.

NonZeroU8

An integer that is known not to equal zero.

NonZeroU16

An integer that is known not to equal zero.

NonZeroU32

An integer that is known not to equal zero.

NonZeroU64

An integer that is known not to equal zero.

NonZeroU128

An integer that is known not to equal zero.

NonZeroUsize

An integer that is known not to equal zero.