core::num

Struct NonZero

pub struct NonZero<T: ZeroablePrimitive>(/* private fields */);

A value that is known not to equal zero.

This enables some memory layout optimization. For example, Option<NonZero<u32>> is the same size as u32:

use core::{num::NonZero};

assert_eq!(size_of::<Option<NonZero<u32>>>(), size_of::<u32>());

Layout

NonZero<T> is guaranteed to have the same layout and bit validity as T with the exception that the all-zero bit pattern is invalid. Option<NonZero<T>> is guaranteed to be compatible with T, including in FFI.

Thanks to the null pointer optimization, NonZero<T> and Option<NonZero<T>> are guaranteed to have the same size and alignment:

use std::num::NonZero;

assert_eq!(size_of::<NonZero<u32>>(), size_of::<Option<NonZero<u32>>>());
assert_eq!(align_of::<NonZero<u32>>(), align_of::<Option<NonZero<u32>>>());

Note on generic usage

NonZero<T> can only be used with some standard library primitive types (such as u8, i32, and etc.). The type parameter T must implement the internal trait ZeroablePrimitive, which is currently permanently unstable and cannot be implemented by users. Therefore, you cannot use NonZero<T> with your own types, nor can you implement traits for all NonZero<T>, only for concrete types.

Implementations

impl<T> NonZero<T>
where T: ZeroablePrimitive,

pub const fn new(n: T) -> Option<Self>

Creates a non-zero if the given value is not zero.

pub const unsafe fn new_unchecked(n: T) -> Self

Creates a non-zero without checking whether the value is non-zero. This results in undefined behavior if the value is zero.

Safety

The value must not be zero.

pub fn from_mut(n: &mut T) -> Option<&mut Self>

This is a nightly-only experimental API. (nonzero_from_mut #106290)

Converts a reference to a non-zero mutable reference if the referenced value is not zero.

pub unsafe fn from_mut_unchecked(n: &mut T) -> &mut Self

This is a nightly-only experimental API. (nonzero_from_mut #106290)

Converts a mutable reference to a non-zero mutable reference without checking whether the referenced value is non-zero. This results in undefined behavior if the referenced value is zero.

Safety

The referenced value must not be zero.

pub const fn get(self) -> T

Returns the contained value as a primitive type.

impl NonZero<u8>

pub const BITS: u32 = 8u32

The size of this non-zero integer type in bits.

This value is equal to u8::BITS.

Examples

assert_eq!(NonZero::<u8>::BITS, u8::BITS);

pub const MIN: Self

The smallest value that can be represented by this non-zero integer type, 1.

Examples

assert_eq!(NonZero::<u8>::MIN.get(), 1u8);

pub const MAX: Self

The largest value that can be represented by this non-zero integer type, equal to u8::MAX.

Examples

assert_eq!(NonZero::<u8>::MAX.get(), u8::MAX);

pub const fn leading_zeros(self) -> u32

Returns the number of leading zeros in the binary representation of self.

On many architectures, this function can perform better than leading_zeros() on the underlying integer type, as special handling of zero can be avoided.

Examples

let n = NonZero::<u8>::new(u8::MAX)?;

assert_eq!(n.leading_zeros(), 0);

pub const fn trailing_zeros(self) -> u32

Returns the number of trailing zeros in the binary representation of self.

On many architectures, this function can perform better than trailing_zeros() on the underlying integer type, as special handling of zero can be avoided.

Examples

let n = NonZero::<u8>::new(0b0101000)?;

assert_eq!(n.trailing_zeros(), 3);

pub const fn isolate_highest_one(self) -> Self

This is a nightly-only experimental API. (isolate_most_least_significant_one #136909)

Returns self with only the most significant bit set.

Example

#![feature(isolate_most_least_significant_one)]

let a = NonZero::<u8>::new(0b_01100100)?;
let b = NonZero::<u8>::new(0b_01000000)?;

assert_eq!(a.isolate_highest_one(), b);

pub const fn isolate_lowest_one(self) -> Self

This is a nightly-only experimental API. (isolate_most_least_significant_one #136909)

Returns self with only the least significant bit set.

Example

#![feature(isolate_most_least_significant_one)]

let a = NonZero::<u8>::new(0b_01100100)?;
let b = NonZero::<u8>::new(0b_00000100)?;

assert_eq!(a.isolate_lowest_one(), b);

pub const fn highest_one(self) -> u32

This is a nightly-only experimental API. (int_lowest_highest_one #145203)

Returns the index of the highest bit set to one in self.

Examples

#![feature(int_lowest_highest_one)]

assert_eq!(NonZero::<u8>::new(0x1)?.highest_one(), 0);
assert_eq!(NonZero::<u8>::new(0x10)?.highest_one(), 4);
assert_eq!(NonZero::<u8>::new(0x1f)?.highest_one(), 4);

pub const fn lowest_one(self) -> u32

This is a nightly-only experimental API. (int_lowest_highest_one #145203)

Returns the index of the lowest bit set to one in self.

Examples

#![feature(int_lowest_highest_one)]

assert_eq!(NonZero::<u8>::new(0x1)?.lowest_one(), 0);
assert_eq!(NonZero::<u8>::new(0x10)?.lowest_one(), 4);
assert_eq!(NonZero::<u8>::new(0x1f)?.lowest_one(), 0);

pub const fn count_ones(self) -> NonZero<u32>

Returns the number of ones in the binary representation of self.

Examples

let a = NonZero::<u8>::new(0b100_0000)?;
let b = NonZero::<u8>::new(0b100_0011)?;

assert_eq!(a.count_ones(), NonZero::new(1)?);
assert_eq!(b.count_ones(), NonZero::new(3)?);

pub const fn rotate_left(self, n: u32) -> Self

This is a nightly-only experimental API. (nonzero_bitwise #128281)

Shifts the bits to the left by a specified amount, n, wrapping the truncated bits to the end of the resulting integer.

Please note this isnt the same operation as the << shifting operator!

Examples

#![feature(nonzero_bitwise)]
let n = NonZero::new(0x82u8)?;
let m = NonZero::new(0xa)?;

assert_eq!(n.rotate_left(2), m);

pub const fn rotate_right(self, n: u32) -> Self

This is a nightly-only experimental API. (nonzero_bitwise #128281)

Shifts the bits to the right by a specified amount, n, wrapping the truncated bits to the beginning of the resulting integer.

Please note this isnt the same operation as the >> shifting operator!

Examples

#![feature(nonzero_bitwise)]
let n = NonZero::new(0xau8)?;
let m = NonZero::new(0x82)?;

assert_eq!(n.rotate_right(2), m);

pub const fn swap_bytes(self) -> Self

This is a nightly-only experimental API. (nonzero_bitwise #128281)

Reverses the byte order of the integer.

Examples

#![feature(nonzero_bitwise)]
let n = NonZero::new(0x12u8)?;
let m = n.swap_bytes();

assert_eq!(m, NonZero::new(0x12)?);

pub const fn reverse_bits(self) -> Self

This is a nightly-only experimental API. (nonzero_bitwise #128281)

Reverses the order of bits in the integer. The least significant bit becomes the most significant bit, second least-significant bit becomes second most-significant bit, etc.

Examples

#![feature(nonzero_bitwise)]
let n = NonZero::new(0x12u8)?;
let m = n.reverse_bits();

assert_eq!(m, NonZero::new(0x48)?);

pub const fn from_be(x: Self) -> Self

This is a nightly-only experimental API. (nonzero_bitwise #128281)

Converts an integer from big endian to the targets endianness.

On big endian this is a no-op. On little endian the bytes are swapped.

Examples

#![feature(nonzero_bitwise)]
use std::num::NonZeroU8;
let n = NonZero::new(0x1Au8)?;

if cfg!(target_endian = "big") {
    assert_eq!(NonZeroU8::from_be(n), n)
} else {
    assert_eq!(NonZeroU8::from_be(n), n.swap_bytes())
}

pub const fn from_le(x: Self) -> Self

This is a nightly-only experimental API. (nonzero_bitwise #128281)

Converts an integer from little endian to the targets endianness.

On little endian this is a no-op. On big endian the bytes are swapped.

Examples

#![feature(nonzero_bitwise)]
use std::num::NonZeroU8;
let n = NonZero::new(0x1Au8)?;

if cfg!(target_endian = "little") {
    assert_eq!(NonZeroU8::from_le(n), n)
} else {
    assert_eq!(NonZeroU8::from_le(n), n.swap_bytes())
}

pub const fn to_be(self) -> Self

This is a nightly-only experimental API. (nonzero_bitwise #128281)

Converts self to big endian from the targets endianness.

On big endian this is a no-op. On little endian the bytes are swapped.

Examples

#![feature(nonzero_bitwise)]
let n = NonZero::new(0x1Au8)?;

if cfg!(target_endian = "big") {
    assert_eq!(n.to_be(), n)
} else {
    assert_eq!(n.to_be(), n.swap_bytes())
}

pub const fn to_le(self) -> Self

This is a nightly-only experimental API. (nonzero_bitwise #128281)

Converts self to little endian from the targets endianness.

On little endian this is a no-op. On big endian the bytes are swapped.

Examples

#![feature(nonzero_bitwise)]
let n = NonZero::new(0x1Au8)?;

if cfg!(target_endian = "little") {
    assert_eq!(n.to_le(), n)
} else {
    assert_eq!(n.to_le(), n.swap_bytes())
}

pub const fn checked_add(self, other: u8) -> Option<Self>

Adds an unsigned integer to a non-zero value. Checks for overflow and returns None on overflow. As a consequence, the result cannot wrap to zero.

Examples

let one = NonZero::new(1u8)?;
let two = NonZero::new(2u8)?;
let max = NonZero::new(u8::MAX)?;

assert_eq!(Some(two), one.checked_add(1));
assert_eq!(None, max.checked_add(1));

pub const fn saturating_add(self, other: u8) -> Self

Adds an unsigned integer to a non-zero value. Return NonZero::<u8>::MAX on overflow.

Examples

let one = NonZero::new(1u8)?;
let two = NonZero::new(2u8)?;
let max = NonZero::new(u8::MAX)?;

assert_eq!(two, one.saturating_add(1));
assert_eq!(max, max.saturating_add(1));

pub const unsafe fn unchecked_add(self, other: u8) -> Self

This is a nightly-only experimental API. (nonzero_ops #84186)

Adds an unsigned integer to a non-zero value, assuming overflow cannot occur. Overflow is unchecked, and it is undefined behavior to overflow even if the result would wrap to a non-zero value. The behavior is undefined as soon as self + rhs > u8::MAX.

Examples

#![feature(nonzero_ops)]

let one = NonZero::new(1u8)?;
let two = NonZero::new(2u8)?;

assert_eq!(two, unsafe { one.unchecked_add(1) });

pub const fn checked_next_power_of_two(self) -> Option<Self>

Returns the smallest power of two greater than or equal to self. Checks for overflow and returns None if the next power of two is greater than the types maximum value. As a consequence, the result cannot wrap to zero.

Examples

let two = NonZero::new(2u8)?;
let three = NonZero::new(3u8)?;
let four = NonZero::new(4u8)?;
let max = NonZero::new(u8::MAX)?;

assert_eq!(Some(two), two.checked_next_power_of_two() );
assert_eq!(Some(four), three.checked_next_power_of_two() );
assert_eq!(None, max.checked_next_power_of_two() );

pub const fn ilog2(self) -> u32

Returns the base 2 logarithm of the number, rounded down.

This is the same operation as u8::ilog2, except that it has no failure cases to worry about since this value can never be zero.

Examples

assert_eq!(NonZero::new(7u8)?.ilog2(), 2);
assert_eq!(NonZero::new(8u8)?.ilog2(), 3);
assert_eq!(NonZero::new(9u8)?.ilog2(), 3);

pub const fn ilog10(self) -> u32

Returns the base 10 logarithm of the number, rounded down.

This is the same operation as u8::ilog10, except that it has no failure cases to worry about since this value can never be zero.

Examples

assert_eq!(NonZero::new(99u8)?.ilog10(), 1);
assert_eq!(NonZero::new(100u8)?.ilog10(), 2);
assert_eq!(NonZero::new(101u8)?.ilog10(), 2);

pub const fn midpoint(self, rhs: Self) -> Self

Calculates the midpoint (average) between self and rhs.

midpoint(a, b) is (a + b) >> 1 as if it were performed in a sufficiently-large signed integral type. This implies that the result is always rounded towards negative infinity and that no overflow will ever occur.

Examples

let one = NonZero::new(1u8)?;
let two = NonZero::new(2u8)?;
let four = NonZero::new(4u8)?;

assert_eq!(one.midpoint(four), two);
assert_eq!(four.midpoint(one), two);

pub const fn is_power_of_two(self) -> bool

Returns true if and only if self == (1 << k) for some k.

On many architectures, this function can perform better than is_power_of_two() on the underlying integer type, as special handling of zero can be avoided.

Examples

let eight = NonZero::new(8u8)?;
assert!(eight.is_power_of_two());
let ten = NonZero::new(10u8)?;
assert!(!ten.is_power_of_two());

pub const fn isqrt(self) -> Self

Returns the square root of the number, rounded down.

Examples

let ten = NonZero::new(10u8)?;
let three = NonZero::new(3u8)?;

assert_eq!(ten.isqrt(), three);

pub const fn cast_signed(self) -> NonZero<i8>

Returns the bit pattern of self reinterpreted as a signed integer of the same size.

Examples

let n = NonZero::<u8>::MAX;

assert_eq!(n.cast_signed(), NonZero::new(-1i8).unwrap());

pub const fn checked_mul(self, other: Self) -> Option<Self>

Multiplies two non-zero integers together. Checks for overflow and returns None on overflow. As a consequence, the result cannot wrap to zero.

Examples

let two = NonZero::new(2u8)?;
let four = NonZero::new(4u8)?;
let max = NonZero::new(u8::MAX)?;

assert_eq!(Some(four), two.checked_mul(two));
assert_eq!(None, max.checked_mul(two));

pub const fn saturating_mul(self, other: Self) -> Self

Multiplies two non-zero integers together. Return NonZero::<u8>::MAX on overflow.

Examples

let two = NonZero::new(2u8)?;
let four = NonZero::new(4u8)?;
let max = NonZero::new(u8::MAX)?;

assert_eq!(four, two.saturating_mul(two));
assert_eq!(max, four.saturating_mul(max));

pub const unsafe fn unchecked_mul(self, other: Self) -> Self

This is a nightly-only experimental API. (nonzero_ops #84186)

Multiplies two non-zero integers together, assuming overflow cannot occur. Overflow is unchecked, and it is undefined behavior to overflow even if the result would wrap to a non-zero value. The behavior is undefined as soon as self * rhs > u8::MAX.

Examples

#![feature(nonzero_ops)]

let two = NonZero::new(2u8)?;
let four = NonZero::new(4u8)?;

assert_eq!(four, unsafe { two.unchecked_mul(two) });

pub const fn checked_pow(self, other: u32) -> Option<Self>

Raises non-zero value to an integer power. Checks for overflow and returns None on overflow. As a consequence, the result cannot wrap to zero.

Examples

let three = NonZero::new(3u8)?;
let twenty_seven = NonZero::new(27u8)?;
let half_max = NonZero::new(u8::MAX / 2)?;

assert_eq!(Some(twenty_seven), three.checked_pow(3));
assert_eq!(None, half_max.checked_pow(3));

pub const fn saturating_pow(self, other: u32) -> Self

Raise non-zero value to an integer power. Return NonZero::<u8>::MAX on overflow.

Examples

let three = NonZero::new(3u8)?;
let twenty_seven = NonZero::new(27u8)?;
let max = NonZero::new(u8::MAX)?;

assert_eq!(twenty_seven, three.saturating_pow(3));
assert_eq!(max, max.saturating_pow(3));

impl NonZero<u8>

pub const fn div_ceil(self, rhs: Self) -> Self

Calculates the quotient of self and rhs, rounding the result towards positive infinity.

The result is guaranteed to be non-zero.

Examples

let one = NonZero::new(1u8).unwrap();
let max = NonZero::new(u8::MAX).unwrap();
assert_eq!(one.div_ceil(max), one);

let two = NonZero::new(2u8).unwrap();
let three = NonZero::new(3u8).unwrap();
assert_eq!(three.div_ceil(two), two);

impl NonZero<u16>

pub const BITS: u32 = 16u32

The size of this non-zero integer type in bits.

This value is equal to u16::BITS.

Examples

assert_eq!(NonZero::<u16>::BITS, u16::BITS);

pub const MIN: Self

The smallest value that can be represented by this non-zero integer type, 1.

Examples

assert_eq!(NonZero::<u16>::MIN.get(), 1u16);

pub const MAX: Self

The largest value that can be represented by this non-zero integer type, equal to u16::MAX.

Examples

assert_eq!(NonZero::<u16>::MAX.get(), u16::MAX);

pub const fn leading_zeros(self) -> u32

Returns the number of leading zeros in the binary representation of self.

On many architectures, this function can perform better than leading_zeros() on the underlying integer type, as special handling of zero can be avoided.

Examples

let n = NonZero::<u16>::new(u16::MAX)?;

assert_eq!(n.leading_zeros(), 0);

pub const fn trailing_zeros(self) -> u32

Returns the number of trailing zeros in the binary representation of self.

On many architectures, this function can perform better than trailing_zeros() on the underlying integer type, as special handling of zero can be avoided.

Examples

let n = NonZero::<u16>::new(0b0101000)?;

assert_eq!(n.trailing_zeros(), 3);

pub const fn isolate_highest_one(self) -> Self

This is a nightly-only experimental API. (isolate_most_least_significant_one #136909)

Returns self with only the most significant bit set.

Example

#![feature(isolate_most_least_significant_one)]

let a = NonZero::<u16>::new(0b_01100100)?;
let b = NonZero::<u16>::new(0b_01000000)?;

assert_eq!(a.isolate_highest_one(), b);

pub const fn isolate_lowest_one(self) -> Self

This is a nightly-only experimental API. (isolate_most_least_significant_one #136909)

Returns self with only the least significant bit set.

Example

#![feature(isolate_most_least_significant_one)]

let a = NonZero::<u16>::new(0b_01100100)?;
let b = NonZero::<u16>::new(0b_00000100)?;

assert_eq!(a.isolate_lowest_one(), b);

pub const fn highest_one(self) -> u32

This is a nightly-only experimental API. (int_lowest_highest_one #145203)

Returns the index of the highest bit set to one in self.

Examples

#![feature(int_lowest_highest_one)]

assert_eq!(NonZero::<u16>::new(0x1)?.highest_one(), 0);
assert_eq!(NonZero::<u16>::new(0x10)?.highest_one(), 4);
assert_eq!(NonZero::<u16>::new(0x1f)?.highest_one(), 4);

pub const fn lowest_one(self) -> u32

This is a nightly-only experimental API. (int_lowest_highest_one #145203)

Returns the index of the lowest bit set to one in self.

Examples

#![feature(int_lowest_highest_one)]

assert_eq!(NonZero::<u16>::new(0x1)?.lowest_one(), 0);
assert_eq!(NonZero::<u16>::new(0x10)?.lowest_one(), 4);
assert_eq!(NonZero::<u16>::new(0x1f)?.lowest_one(), 0);

pub const fn count_ones(self) -> NonZero<u32>

Returns the number of ones in the binary representation of self.

Examples

let a = NonZero::<u16>::new(0b100_0000)?;
let b = NonZero::<u16>::new(0b100_0011)?;

assert_eq!(a.count_ones(), NonZero::new(1)?);
assert_eq!(b.count_ones(), NonZero::new(3)?);

pub const fn rotate_left(self, n: u32) -> Self

This is a nightly-only experimental API. (nonzero_bitwise #128281)

Shifts the bits to the left by a specified amount, n, wrapping the truncated bits to the end of the resulting integer.

Please note this isnt the same operation as the << shifting operator!

Examples

#![feature(nonzero_bitwise)]
let n = NonZero::new(0xa003u16)?;
let m = NonZero::new(0x3a)?;

assert_eq!(n.rotate_left(4), m);

pub const fn rotate_right(self, n: u32) -> Self

This is a nightly-only experimental API. (nonzero_bitwise #128281)

Shifts the bits to the right by a specified amount, n, wrapping the truncated bits to the beginning of the resulting integer.

Please note this isnt the same operation as the >> shifting operator!

Examples

#![feature(nonzero_bitwise)]
let n = NonZero::new(0x3au16)?;
let m = NonZero::new(0xa003)?;

assert_eq!(n.rotate_right(4), m);

pub const fn swap_bytes(self) -> Self

This is a nightly-only experimental API. (nonzero_bitwise #128281)

Reverses the byte order of the integer.

Examples

#![feature(nonzero_bitwise)]
let n = NonZero::new(0x1234u16)?;
let m = n.swap_bytes();

assert_eq!(m, NonZero::new(0x3412)?);

pub const fn reverse_bits(self) -> Self

This is a nightly-only experimental API. (nonzero_bitwise #128281)

Reverses the order of bits in the integer. The least significant bit becomes the most significant bit, second least-significant bit becomes second most-significant bit, etc.

Examples

#![feature(nonzero_bitwise)]
let n = NonZero::new(0x1234u16)?;
let m = n.reverse_bits();

assert_eq!(m, NonZero::new(0x2c48)?);

pub const fn from_be(x: Self) -> Self

This is a nightly-only experimental API. (nonzero_bitwise #128281)

Converts an integer from big endian to the targets endianness.

On big endian this is a no-op. On little endian the bytes are swapped.

Examples

#![feature(nonzero_bitwise)]
use std::num::NonZeroU16;
let n = NonZero::new(0x1Au16)?;

if cfg!(target_endian = "big") {
    assert_eq!(NonZeroU16::from_be(n), n)
} else {
    assert_eq!(NonZeroU16::from_be(n), n.swap_bytes())
}

pub const fn from_le(x: Self) -> Self

This is a nightly-only experimental API. (nonzero_bitwise #128281)

Converts an integer from little endian to the targets endianness.

On little endian this is a no-op. On big endian the bytes are swapped.

Examples

#![feature(nonzero_bitwise)]
use std::num::NonZeroU16;
let n = NonZero::new(0x1Au16)?;

if cfg!(target_endian = "little") {
    assert_eq!(NonZeroU16::from_le(n), n)
} else {
    assert_eq!(NonZeroU16::from_le(n), n.swap_bytes())
}

pub const fn to_be(self) -> Self

This is a nightly-only experimental API. (nonzero_bitwise #128281)

Converts self to big endian from the targets endianness.

On big endian this is a no-op. On little endian the bytes are swapped.

Examples

#![feature(nonzero_bitwise)]
let n = NonZero::new(0x1Au16)?;

if cfg!(target_endian = "big") {
    assert_eq!(n.to_be(), n)
} else {
    assert_eq!(n.to_be(), n.swap_bytes())
}

pub const fn to_le(self) -> Self

This is a nightly-only experimental API. (nonzero_bitwise #128281)

Converts self to little endian from the targets endianness.

On little endian this is a no-op. On big endian the bytes are swapped.

Examples

#![feature(nonzero_bitwise)]
let n = NonZero::new(0x1Au16)?;

if cfg!(target_endian = "little") {
    assert_eq!(n.to_le(), n)
} else {
    assert_eq!(n.to_le(), n.swap_bytes())
}

pub const fn checked_add(self, other: u16) -> Option<Self>

Adds an unsigned integer to a non-zero value. Checks for overflow and returns None on overflow. As a consequence, the result cannot wrap to zero.

Examples

let one = NonZero::new(1u16)?;
let two = NonZero::new(2u16)?;
let max = NonZero::new(u16::MAX)?;

assert_eq!(Some(two), one.checked_add(1));
assert_eq!(None, max.checked_add(1));

pub const fn saturating_add(self, other: u16) -> Self

Adds an unsigned integer to a non-zero value. Return NonZero::<u16>::MAX on overflow.

Examples

let one = NonZero::new(1u16)?;
let two = NonZero::new(2u16)?;
let max = NonZero::new(u16::MAX)?;

assert_eq!(two, one.saturating_add(1));
assert_eq!(max, max.saturating_add(1));

pub const unsafe fn unchecked_add(self, other: u16) -> Self

This is a nightly-only experimental API. (nonzero_ops #84186)

Adds an unsigned integer to a non-zero value, assuming overflow cannot occur. Overflow is unchecked, and it is undefined behavior to overflow even if the result would wrap to a non-zero value. The behavior is undefined as soon as self + rhs > u16::MAX.

Examples

#![feature(nonzero_ops)]

let one = NonZero::new(1u16)?;
let two = NonZero::new(2u16)?;

assert_eq!(two, unsafe { one.unchecked_add(1) });

pub const fn checked_next_power_of_two(self) -> Option<Self>

Returns the smallest power of two greater than or equal to self. Checks for overflow and returns None if the next power of two is greater than the types maximum value. As a consequence, the result cannot wrap to zero.

Examples

let two = NonZero::new(2u16)?;
let three = NonZero::new(3u16)?;
let four = NonZero::new(4u16)?;
let max = NonZero::new(u16::MAX)?;

assert_eq!(Some(two), two.checked_next_power_of_two() );
assert_eq!(Some(four), three.checked_next_power_of_two() );
assert_eq!(None, max.checked_next_power_of_two() );

pub const fn ilog2(self) -> u32

Returns the base 2 logarithm of the number, rounded down.

This is the same operation as u16::ilog2, except that it has no failure cases to worry about since this value can never be zero.

Examples

assert_eq!(NonZero::new(7u16)?.ilog2(), 2);
assert_eq!(NonZero::new(8u16)?.ilog2(), 3);
assert_eq!(NonZero::new(9u16)?.ilog2(), 3);

pub const fn ilog10(self) -> u32

Returns the base 10 logarithm of the number, rounded down.

This is the same operation as u16::ilog10, except that it has no failure cases to worry about since this value can never be zero.

Examples

assert_eq!(NonZero::new(99u16)?.ilog10(), 1);
assert_eq!(NonZero::new(100u16)?.ilog10(), 2);
assert_eq!(NonZero::new(101u16)?.ilog10(), 2);

pub const fn midpoint(self, rhs: Self) -> Self

Calculates the midpoint (average) between self and rhs.

midpoint(a, b) is (a + b) >> 1 as if it were performed in a sufficiently-large signed integral type. This implies that the result is always rounded towards negative infinity and that no overflow will ever occur.

Examples

let one = NonZero::new(1u16)?;
let two = NonZero::new(2u16)?;
let four = NonZero::new(4u16)?;

assert_eq!(one.midpoint(four), two);
assert_eq!(four.midpoint(one), two);

pub const fn is_power_of_two(self) -> bool

Returns true if and only if self == (1 << k) for some k.

On many architectures, this function can perform better than is_power_of_two() on the underlying integer type, as special handling of zero can be avoided.

Examples

let eight = NonZero::new(8u16)?;
assert!(eight.is_power_of_two());
let ten = NonZero::new(10u16)?;
assert!(!ten.is_power_of_two());

pub const fn isqrt(self) -> Self

Returns the square root of the number, rounded down.

Examples

let ten = NonZero::new(10u16)?;
let three = NonZero::new(3u16)?;

assert_eq!(ten.isqrt(), three);

pub const fn cast_signed(self) -> NonZero<i16>

Returns the bit pattern of self reinterpreted as a signed integer of the same size.

Examples

let n = NonZero::<u16>::MAX;

assert_eq!(n.cast_signed(), NonZero::new(-1i16).unwrap());

pub const fn checked_mul(self, other: Self) -> Option<Self>

Multiplies two non-zero integers together. Checks for overflow and returns None on overflow. As a consequence, the result cannot wrap to zero.

Examples

let two = NonZero::new(2u16)?;
let four = NonZero::new(4u16)?;
let max = NonZero::new(u16::MAX)?;

assert_eq!(Some(four), two.checked_mul(two));
assert_eq!(None, max.checked_mul(two));

pub const fn saturating_mul(self, other: Self) -> Self

Multiplies two non-zero integers together. Return NonZero::<u16>::MAX on overflow.

Examples

let two = NonZero::new(2u16)?;
let four = NonZero::new(4u16)?;
let max = NonZero::new(u16::MAX)?;

assert_eq!(four, two.saturating_mul(two));
assert_eq!(max, four.saturating_mul(max));

pub const unsafe fn unchecked_mul(self, other: Self) -> Self

This is a nightly-only experimental API. (nonzero_ops #84186)

Multiplies two non-zero integers together, assuming overflow cannot occur. Overflow is unchecked, and it is undefined behavior to overflow even if the result would wrap to a non-zero value. The behavior is undefined as soon as self * rhs > u16::MAX.

Examples

#![feature(nonzero_ops)]

let two = NonZero::new(2u16)?;
let four = NonZero::new(4u16)?;

assert_eq!(four, unsafe { two.unchecked_mul(two) });

pub const fn checked_pow(self, other: u32) -> Option<Self>

Raises non-zero value to an integer power. Checks for overflow and returns None on overflow. As a consequence, the result cannot wrap to zero.

Examples

let three = NonZero::new(3u16)?;
let twenty_seven = NonZero::new(27u16)?;
let half_max = NonZero::new(u16::MAX / 2)?;

assert_eq!(Some(twenty_seven), three.checked_pow(3));
assert_eq!(None, half_max.checked_pow(3));

pub const fn saturating_pow(self, other: u32) -> Self

Raise non-zero value to an integer power. Return NonZero::<u16>::MAX on overflow.

Examples

let three = NonZero::new(3u16)?;
let twenty_seven = NonZero::new(27u16)?;
let max = NonZero::new(u16::MAX)?;

assert_eq!(twenty_seven, three.saturating_pow(3));
assert_eq!(max, max.saturating_pow(3));

impl NonZero<u16>

pub const fn div_ceil(self, rhs: Self) -> Self

Calculates the quotient of self and rhs, rounding the result towards positive infinity.

The result is guaranteed to be non-zero.

Examples

let one = NonZero::new(1u16).unwrap();
let max = NonZero::new(u16::MAX).unwrap();
assert_eq!(one.div_ceil(max), one);

let two = NonZero::new(2u16).unwrap();
let three = NonZero::new(3u16).unwrap();
assert_eq!(three.div_ceil(two), two);

impl NonZero<u32>

pub const BITS: u32 = 32u32

The size of this non-zero integer type in bits.

This value is equal to u32::BITS.

Examples

assert_eq!(NonZero::<u32>::BITS, u32::BITS);

pub const MIN: Self

The smallest value that can be represented by this non-zero integer type, 1.

Examples

assert_eq!(NonZero::<u32>::MIN.get(), 1u32);

pub const MAX: Self

The largest value that can be represented by this non-zero integer type, equal to u32::MAX.

Examples

assert_eq!(NonZero::<u32>::MAX.get(), u32::MAX);

pub const fn leading_zeros(self) -> u32

Returns the number of leading zeros in the binary representation of self.

On many architectures, this function can perform better than leading_zeros() on the underlying integer type, as special handling of zero can be avoided.

Examples

let n = NonZero::<u32>::new(u32::MAX)?;

assert_eq!(n.leading_zeros(), 0);

pub const fn trailing_zeros(self) -> u32

Returns the number of trailing zeros in the binary representation of self.

On many architectures, this function can perform better than trailing_zeros() on the underlying integer type, as special handling of zero can be avoided.

Examples

let n = NonZero::<u32>::new(0b0101000)?;

assert_eq!(n.trailing_zeros(), 3);

pub const fn isolate_highest_one(self) -> Self

This is a nightly-only experimental API. (isolate_most_least_significant_one #136909)

Returns self with only the most significant bit set.

Example

#![feature(isolate_most_least_significant_one)]

let a = NonZero::<u32>::new(0b_01100100)?;
let b = NonZero::<u32>::new(0b_01000000)?;

assert_eq!(a.isolate_highest_one(), b);

pub const fn isolate_lowest_one(self) -> Self

This is a nightly-only experimental API. (isolate_most_least_significant_one #136909)

Returns self with only the least significant bit set.

Example

#![feature(isolate_most_least_significant_one)]

let a = NonZero::<u32>::new(0b_01100100)?;
let b = NonZero::<u32>::new(0b_00000100)?;

assert_eq!(a.isolate_lowest_one(), b);

pub const fn highest_one(self) -> u32

This is a nightly-only experimental API. (int_lowest_highest_one #145203)

Returns the index of the highest bit set to one in self.

Examples

#![feature(int_lowest_highest_one)]

assert_eq!(NonZero::<u32>::new(0x1)?.highest_one(), 0);
assert_eq!(NonZero::<u32>::new(0x10)?.highest_one(), 4);
assert_eq!(NonZero::<u32>::new(0x1f)?.highest_one(), 4);

pub const fn lowest_one(self) -> u32

This is a nightly-only experimental API. (int_lowest_highest_one #145203)

Returns the index of the lowest bit set to one in self.

Examples

#![feature(int_lowest_highest_one)]

assert_eq!(NonZero::<u32>::new(0x1)?.lowest_one(), 0);
assert_eq!(NonZero::<u32>::new(0x10)?.lowest_one(), 4);
assert_eq!(NonZero::<u32>::new(0x1f)?.lowest_one(), 0);

pub const fn count_ones(self) -> NonZero<u32>

Returns the number of ones in the binary representation of self.

Examples

let a = NonZero::<u32>::new(0b100_0000)?;
let b = NonZero::<u32>::new(0b100_0011)?;

assert_eq!(a.count_ones(), NonZero::new(1)?);
assert_eq!(b.count_ones(), NonZero::new(3)?);

pub const fn rotate_left(self, n: u32) -> Self

This is a nightly-only experimental API. (nonzero_bitwise #128281)

Shifts the bits to the left by a specified amount, n, wrapping the truncated bits to the end of the resulting integer.

Please note this isnt the same operation as the << shifting operator!

Examples

#![feature(nonzero_bitwise)]
let n = NonZero::new(0x10000b3u32)?;
let m = NonZero::new(0xb301)?;

assert_eq!(n.rotate_left(8), m);

pub const fn rotate_right(self, n: u32) -> Self

This is a nightly-only experimental API. (nonzero_bitwise #128281)

Shifts the bits to the right by a specified amount, n, wrapping the truncated bits to the beginning of the resulting integer.

Please note this isnt the same operation as the >> shifting operator!

Examples

#![feature(nonzero_bitwise)]
let n = NonZero::new(0xb301u32)?;
let m = NonZero::new(0x10000b3)?;

assert_eq!(n.rotate_right(8), m);

pub const fn swap_bytes(self) -> Self

This is a nightly-only experimental API. (nonzero_bitwise #128281)

Reverses the byte order of the integer.

Examples

#![feature(nonzero_bitwise)]
let n = NonZero::new(0x12345678u32)?;
let m = n.swap_bytes();

assert_eq!(m, NonZero::new(0x78563412)?);

pub const fn reverse_bits(self) -> Self

This is a nightly-only experimental API. (nonzero_bitwise #128281)

Reverses the order of bits in the integer. The least significant bit becomes the most significant bit, second least-significant bit becomes second most-significant bit, etc.

Examples

#![feature(nonzero_bitwise)]
let n = NonZero::new(0x12345678u32)?;
let m = n.reverse_bits();

assert_eq!(m, NonZero::new(0x1e6a2c48)?);

pub const fn from_be(x: Self) -> Self

This is a nightly-only experimental API. (nonzero_bitwise #128281)

Converts an integer from big endian to the targets endianness.

On big endian this is a no-op. On little endian the bytes are swapped.

Examples

#![feature(nonzero_bitwise)]
use std::num::NonZeroU32;
let n = NonZero::new(0x1Au32)?;

if cfg!(target_endian = "big") {
    assert_eq!(NonZeroU32::from_be(n), n)
} else {
    assert_eq!(NonZeroU32::from_be(n), n.swap_bytes())
}

pub const fn from_le(x: Self) -> Self

This is a nightly-only experimental API. (nonzero_bitwise #128281)

Converts an integer from little endian to the targets endianness.

On little endian this is a no-op. On big endian the bytes are swapped.

Examples

#![feature(nonzero_bitwise)]
use std::num::NonZeroU32;
let n = NonZero::new(0x1Au32)?;

if cfg!(target_endian = "little") {
    assert_eq!(NonZeroU32::from_le(n), n)
} else {
    assert_eq!(NonZeroU32::from_le(n), n.swap_bytes())
}

pub const fn to_be(self) -> Self

This is a nightly-only experimental API. (nonzero_bitwise #128281)

Converts self to big endian from the targets endianness.

On big endian this is a no-op. On little endian the bytes are swapped.

Examples

#![feature(nonzero_bitwise)]
let n = NonZero::new(0x1Au32)?;

if cfg!(target_endian = "big") {
    assert_eq!(n.to_be(), n)
} else {
    assert_eq!(n.to_be(), n.swap_bytes())
}

pub const fn to_le(self) -> Self

This is a nightly-only experimental API. (nonzero_bitwise #128281)

Converts self to little endian from the targets endianness.

On little endian this is a no-op. On big endian the bytes are swapped.

Examples

#![feature(nonzero_bitwise)]
let n = NonZero::new(0x1Au32)?;

if cfg!(target_endian = "little") {
    assert_eq!(n.to_le(), n)
} else {
    assert_eq!(n.to_le(), n.swap_bytes())
}

pub const fn checked_add(self, other: u32) -> Option<Self>

Adds an unsigned integer to a non-zero value. Checks for overflow and returns None on overflow. As a consequence, the result cannot wrap to zero.

Examples

let one = NonZero::new(1u32)?;
let two = NonZero::new(2u32)?;
let max = NonZero::new(u32::MAX)?;

assert_eq!(Some(two), one.checked_add(1));
assert_eq!(None, max.checked_add(1));

pub const fn saturating_add(self, other: u32) -> Self

Adds an unsigned integer to a non-zero value. Return NonZero::<u32>::MAX on overflow.

Examples

let one = NonZero::new(1u32)?;
let two = NonZero::new(2u32)?;
let max = NonZero::new(u32::MAX)?;

assert_eq!(two, one.saturating_add(1));
assert_eq!(max, max.saturating_add(1));

pub const unsafe fn unchecked_add(self, other: u32) -> Self

This is a nightly-only experimental API. (nonzero_ops #84186)

Adds an unsigned integer to a non-zero value, assuming overflow cannot occur. Overflow is unchecked, and it is undefined behavior to overflow even if the result would wrap to a non-zero value. The behavior is undefined as soon as self + rhs > u32::MAX.

Examples

#![feature(nonzero_ops)]

let one = NonZero::new(1u32)?;
let two = NonZero::new(2u32)?;

assert_eq!(two, unsafe { one.unchecked_add(1) });

pub const fn checked_next_power_of_two(self) -> Option<Self>

Returns the smallest power of two greater than or equal to self. Checks for overflow and returns None if the next power of two is greater than the types maximum value. As a consequence, the result cannot wrap to zero.

Examples

let two = NonZero::new(2u32)?;
let three = NonZero::new(3u32)?;
let four = NonZero::new(4u32)?;
let max = NonZero::new(u32::MAX)?;

assert_eq!(Some(two), two.checked_next_power_of_two() );
assert_eq!(Some(four), three.checked_next_power_of_two() );
assert_eq!(None, max.checked_next_power_of_two() );

pub const fn ilog2(self) -> u32

Returns the base 2 logarithm of the number, rounded down.

This is the same operation as u32::ilog2, except that it has no failure cases to worry about since this value can never be zero.

Examples

assert_eq!(NonZero::new(7u32)?.ilog2(), 2);
assert_eq!(NonZero::new(8u32)?.ilog2(), 3);
assert_eq!(NonZero::new(9u32)?.ilog2(), 3);

pub const fn ilog10(self) -> u32

Returns the base 10 logarithm of the number, rounded down.

This is the same operation as u32::ilog10, except that it has no failure cases to worry about since this value can never be zero.

Examples

assert_eq!(NonZero::new(99u32)?.ilog10(), 1);
assert_eq!(NonZero::new(100u32)?.ilog10(), 2);
assert_eq!(NonZero::new(101u32)?.ilog10(), 2);

pub const fn midpoint(self, rhs: Self) -> Self

Calculates the midpoint (average) between self and rhs.

midpoint(a, b) is (a + b) >> 1 as if it were performed in a sufficiently-large signed integral type. This implies that the result is always rounded towards negative infinity and that no overflow will ever occur.

Examples

let one = NonZero::new(1u32)?;
let two = NonZero::new(2u32)?;
let four = NonZero::new(4u32)?;

assert_eq!(one.midpoint(four), two);
assert_eq!(four.midpoint(one), two);

pub const fn is_power_of_two(self) -> bool

Returns true if and only if self == (1 << k) for some k.

On many architectures, this function can perform better than is_power_of_two() on the underlying integer type, as special handling of zero can be avoided.

Examples

let eight = NonZero::new(8u32)?;
assert!(eight.is_power_of_two());
let ten = NonZero::new(10u32)?;
assert!(!ten.is_power_of_two());

pub const fn isqrt(self) -> Self

Returns the square root of the number, rounded down.

Examples

let ten = NonZero::new(10u32)?;
let three = NonZero::new(3u32)?;

assert_eq!(ten.isqrt(), three);

pub const fn cast_signed(self) -> NonZero<i32>

Returns the bit pattern of self reinterpreted as a signed integer of the same size.

Examples

let n = NonZero::<u32>::MAX;

assert_eq!(n.cast_signed(), NonZero::new(-1i32).unwrap());

pub const fn checked_mul(self, other: Self) -> Option<Self>

Multiplies two non-zero integers together. Checks for overflow and returns None on overflow. As a consequence, the result cannot wrap to zero.

Examples

let two = NonZero::new(2u32)?;
let four = NonZero::new(4u32)?;
let max = NonZero::new(u32::MAX)?;

assert_eq!(Some(four), two.checked_mul(two));
assert_eq!(None, max.checked_mul(two));

pub const fn saturating_mul(self, other: Self) -> Self

Multiplies two non-zero integers together. Return NonZero::<u32>::MAX on overflow.

Examples

let two = NonZero::new(2u32)?;
let four = NonZero::new(4u32)?;
let max = NonZero::new(u32::MAX)?;

assert_eq!(four, two.saturating_mul(two));
assert_eq!(max, four.saturating_mul(max));

pub const unsafe fn unchecked_mul(self, other: Self) -> Self

This is a nightly-only experimental API. (nonzero_ops #84186)

Multiplies two non-zero integers together, assuming overflow cannot occur. Overflow is unchecked, and it is undefined behavior to overflow even if the result would wrap to a non-zero value. The behavior is undefined as soon as self * rhs > u32::MAX.

Examples

#![feature(nonzero_ops)]

let two = NonZero::new(2u32)?;
let four = NonZero::new(4u32)?;

assert_eq!(four, unsafe { two.unchecked_mul(two) });

pub const fn checked_pow(self, other: u32) -> Option<Self>

Raises non-zero value to an integer power. Checks for overflow and returns None on overflow. As a consequence, the result cannot wrap to zero.

Examples

let three = NonZero::new(3u32)?;
let twenty_seven = NonZero::new(27u32)?;
let half_max = NonZero::new(u32::MAX / 2)?;

assert_eq!(Some(twenty_seven), three.checked_pow(3));
assert_eq!(None, half_max.checked_pow(3));

pub const fn saturating_pow(self, other: u32) -> Self

Raise non-zero value to an integer power. Return NonZero::<u32>::MAX on overflow.

Examples

let three = NonZero::new(3u32)?;
let twenty_seven = NonZero::new(27u32)?;
let max = NonZero::new(u32::MAX)?;

assert_eq!(twenty_seven, three.saturating_pow(3));
assert_eq!(max, max.saturating_pow(3));