Struct core::num::NonZeroI32

#[repr(transparent)]
pub struct NonZeroI32(_);

An integer that is known not to equal zero.

This enables some memory layout optimization. For example, Option<NonZeroI32> is the same size as i32:

use std::mem::size_of;
assert_eq!(size_of::<Option<core::num::NonZeroI32>>(), size_of::<i32>());
Run

Implementations

impl NonZeroI32

pub const unsafe fn new_unchecked(n: i32) -> Self

Creates a non-zero without checking the value.

Safety

The value must not be zero.

pub const fn new(n: i32) -> Option<Self>

Creates a non-zero if the given value is not zero.

pub const fn get(self) -> i32

Returns the value as a primitive type.

impl NonZeroI32

pub const fn leading_zeros(self) -> u32

Returns the number of leading zeros in the binary representation of self.

On many architectures, this function can perform better than leading_zeros() on the underlying integer type, as special handling of zero can be avoided.

Examples

Basic usage:

let n = std::num::NonZeroI32::new(-1i32).unwrap();

assert_eq!(n.leading_zeros(), 0);
Run

pub const fn trailing_zeros(self) -> u32

Returns the number of trailing zeros in the binary representation of self.

On many architectures, this function can perform better than trailing_zeros() on the underlying integer type, as special handling of zero can be avoided.

Examples

Basic usage:

let n = std::num::NonZeroI32::new(0b0101000).unwrap();

assert_eq!(n.trailing_zeros(), 3);
Run