core::num

Struct Saturating

#[repr(transparent)]
pub struct Saturating<T>(pub T);

Provides intentionally-saturating arithmetic on T.

Operations like + on u32 values are intended to never overflow, and in some debug configurations overflow is detected and results in a panic. While most arithmetic falls into this category, some code explicitly expects and relies upon saturating arithmetic.

Saturating arithmetic can be achieved either through methods like saturating_add, or through the Saturating<T> type, which says that all standard arithmetic operations on the underlying value are intended to have saturating semantics.

The underlying value can be retrieved through the .0 index of the Saturating tuple.

Examples

use std::num::Saturating;

let max = Saturating(u32::MAX);
let one = Saturating(1u32);

assert_eq!(u32::MAX, (max + one).0);

Tuple Fields
0: T
Implementations

impl Saturating<usize>

pub const MIN: Self

Returns the smallest value that can be represented by this integer type.

Examples

use std::num::Saturating;

assert_eq!(<Saturating<usize>>::MIN, Saturating(usize::MIN));

pub const MAX: Self

Returns the largest value that can be represented by this integer type.

Examples

use std::num::Saturating;

assert_eq!(<Saturating<usize>>::MAX, Saturating(usize::MAX));

pub const BITS: u32 = 64u32

Returns the size of this integer type in bits.

Examples

use std::num::Saturating;

assert_eq!(<Saturating<usize>>::BITS, usize::BITS);

pub const fn count_ones(self) -> u32

Returns the number of ones in the binary representation of self.

Examples

use std::num::Saturating;

let n = Saturating(0b01001100usize);

assert_eq!(n.count_ones(), 3);

pub const fn count_zeros(self) -> u32

Returns the number of zeros in the binary representation of self.

Examples

use std::num::Saturating;

assert_eq!(Saturating(!0usize).count_zeros(), 0);

pub const fn trailing_zeros(self) -> u32

Returns the number of trailing zeros in the binary representation of self.

Examples

use std::num::Saturating;

let n = Saturating(0b0101000usize);

assert_eq!(n.trailing_zeros(), 3);

pub const fn rotate_left(self, n: u32) -> Self

Shifts the bits to the left by a specified amount, n, saturating the truncated bits to the end of the resulting integer.

Please note this isnt the same operation as the << shifting operator!

Examples

use std::num::Saturating;

let n: Saturating<i64> = Saturating(0x0123456789ABCDEF);
let m: Saturating<i64> = Saturating(-0x76543210FEDCBA99);

assert_eq!(n.rotate_left(32), m);

pub const fn rotate_right(self, n: u32) -> Self

Shifts the bits to the right by a specified amount, n, saturating the truncated bits to the beginning of the resulting integer.

Please note this isnt the same operation as the >> shifting operator!

Examples

use std::num::Saturating;

let n: Saturating<i64> = Saturating(0x0123456789ABCDEF);
let m: Saturating<i64> = Saturating(-0xFEDCBA987654322);

assert_eq!(n.rotate_right(4), m);

pub const fn swap_bytes(self) -> Self

Reverses the byte order of the integer.

Examples

use std::num::Saturating;

let n: Saturating<i16> = Saturating(0b0000000_01010101);
assert_eq!(n, Saturating(85));

let m = n.swap_bytes();

assert_eq!(m, Saturating(0b01010101_00000000));
assert_eq!(m, Saturating(21760));

pub const fn reverse_bits(self) -> Self

Reverses the bit pattern of the integer.

Examples

Please note that this example is shared among integer types, which is why i16 is used.

use std::num::Saturating;

let n = Saturating(0b0000000_01010101i16);
assert_eq!(n, Saturating(85));

let m = n.reverse_bits();

assert_eq!(m.0 as u16, 0b10101010_00000000);
assert_eq!(m, Saturating(-22016));

pub const fn from_be(x: Self) -> Self

Converts an integer from big endian to the targets endianness.

On big endian this is a no-op. On little endian the bytes are swapped.

Examples

use std::num::Saturating;

let n = Saturating(0x1Ausize);

if cfg!(target_endian = "big") {
    assert_eq!(<Saturating<usize>>::from_be(n), n)
} else {
    assert_eq!(<Saturating<usize>>::from_be(n), n.swap_bytes())
}

pub const fn from_le(x: Self) -> Self

Converts an integer from little endian to the targets endianness.

On little endian this is a no-op. On big endian the bytes are swapped.

Examples

use std::num::Saturating;

let n = Saturating(0x1Ausize);

if cfg!(target_endian = "little") {
    assert_eq!(<Saturating<usize>>::from_le(n), n)
} else {
    assert_eq!(<Saturating<usize>>::from_le(n), n.swap_bytes())
}

pub const fn to_be(self) -> Self

Converts self to big endian from the targets endianness.

On big endian this is a no-op. On little endian the bytes are swapped.

Examples

use std::num::Saturating;

let n = Saturating(0x1Ausize);

if cfg!(target_endian = "big") {
    assert_eq!(n.to_be(), n)
} else {
    assert_eq!(n.to_be(), n.swap_bytes())
}

pub const fn to_le(self) -> Self

Converts self to little endian from the targets endianness.

On little endian this is a no-op. On big endian the bytes are swapped.

Examples

use std::num::Saturating;

let n = Saturating(0x1Ausize);

if cfg!(target_endian = "little") {
    assert_eq!(n.to_le(), n)
} else {
    assert_eq!(n.to_le(), n.swap_bytes())
}

pub const fn pow(self, exp: u32) -> Self

Raises self to the power of exp, using exponentiation by squaring.

Examples

use std::num::Saturating;

assert_eq!(Saturating(3usize).pow(4), Saturating(81));

Results that are too large are saturated:

use std::num::Saturating;

assert_eq!(Saturating(3i8).pow(5), Saturating(127));
assert_eq!(Saturating(3i8).pow(6), Saturating(127));

impl Saturating<u8>

pub const MIN: Self

Returns the smallest value that can be represented by this integer type.

Examples

use std::num::Saturating;

assert_eq!(<Saturating<u8>>::MIN, Saturating(u8::MIN));

pub const MAX: Self

Returns the largest value that can be represented by this integer type.

Examples

use std::num::Saturating;

assert_eq!(<Saturating<u8>>::MAX, Saturating(u8::MAX));

pub const BITS: u32 = 8u32

Returns the size of this integer type in bits.

Examples

use std::num::Saturating;

assert_eq!(<Saturating<u8>>::BITS, u8::BITS);

pub const fn count_ones(self) -> u32

Returns the number of ones in the binary representation of self.

Examples

use std::num::Saturating;

let n = Saturating(0b01001100u8);

assert_eq!(n.count_ones(), 3);

pub const fn count_zeros(self) -> u32

Returns the number of zeros in the binary representation of self.

Examples

use std::num::Saturating;

assert_eq!(Saturating(!0u8).count_zeros(), 0);

pub const fn trailing_zeros(self) -> u32

Returns the number of trailing zeros in the binary representation of self.

Examples

use std::num::Saturating;

let n = Saturating(0b0101000u8);

assert_eq!(n.trailing_zeros(), 3);

pub const fn rotate_left(self, n: u32) -> Self

Shifts the bits to the left by a specified amount, n, saturating the truncated bits to the end of the resulting integer.

Please note this isnt the same operation as the << shifting operator!

Examples

use std::num::Saturating;

let n: Saturating<i64> = Saturating(0x0123456789ABCDEF);
let m: Saturating<i64> = Saturating(-0x76543210FEDCBA99);

assert_eq!(n.rotate_left(32), m);

pub const fn rotate_right(self, n: u32) -> Self

Shifts the bits to the right by a specified amount, n, saturating the truncated bits to the beginning of the resulting integer.

Please note this isnt the same operation as the >> shifting operator!

Examples

use std::num::Saturating;

let n: Saturating<i64> = Saturating(0x0123456789ABCDEF);
let m: Saturating<i64> = Saturating(-0xFEDCBA987654322);

assert_eq!(n.rotate_right(4), m);

pub const fn swap_bytes(self) -> Self

Reverses the byte order of the integer.

Examples

use std::num::Saturating;

let n: Saturating<i16> = Saturating(0b0000000_01010101);
assert_eq!(n, Saturating(85));

let m = n.swap_bytes();

assert_eq!(m, Saturating(0b01010101_00000000));
assert_eq!(m, Saturating(21760));

pub const fn reverse_bits(self) -> Self

Reverses the bit pattern of the integer.

Examples

Please note that this example is shared among integer types, which is why i16 is used.

use std::num::Saturating;

let n = Saturating(0b0000000_01010101i16);
assert_eq!(n, Saturating(85));

let m = n.reverse_bits();

assert_eq!(m.0 as u16, 0b10101010_00000000);
assert_eq!(m, Saturating(-22016));

pub const fn from_be(x: Self) -> Self

Converts an integer from big endian to the targets endianness.

On big endian this is a no-op. On little endian the bytes are swapped.

Examples

use std::num::Saturating;

let n = Saturating(0x1Au8);

if cfg!(target_endian = "big") {
    assert_eq!(<Saturating<u8>>::from_be(n), n)
} else {
    assert_eq!(<Saturating<u8>>::from_be(n), n.swap_bytes())
}

pub const fn from_le(x: Self) -> Self

Converts an integer from little endian to the targets endianness.

On little endian this is a no-op. On big endian the bytes are swapped.

Examples

use std::num::Saturating;

let n = Saturating(0x1Au8);

if cfg!(target_endian = "little") {
    assert_eq!(<Saturating<u8>>::from_le(n), n)
} else {
    assert_eq!(<Saturating<u8>>::from_le(n), n.swap_bytes())
}

pub const fn to_be(self) -> Self

Converts self to big endian from the targets endianness.

On big endian this is a no-op. On little endian the bytes are swapped.

Examples

use std::num::Saturating;

let n = Saturating(0x1Au8);

if cfg!(target_endian = "big") {
    assert_eq!(n.to_be(), n)
} else {
    assert_eq!(n.to_be(), n.swap_bytes())
}

pub const fn to_le(self) -> Self

Converts self to little endian from the targets endianness.

On little endian this is a no-op. On big endian the bytes are swapped.

Examples

use std::num::Saturating;

let n = Saturating(0x1Au8);

if cfg!(target_endian = "little") {
    assert_eq!(n.to_le(), n)
} else {
    assert_eq!(n.to_le(), n.swap_bytes())
}

pub const fn pow(self, exp: u32) -> Self

Raises self to the power of exp, using exponentiation by squaring.

Examples

use std::num::Saturating;

assert_eq!(Saturating(3u8).pow(4), Saturating(81));

Results that are too large are saturated:

use std::num::Saturating;

assert_eq!(Saturating(3i8).pow(5), Saturating(127));
assert_eq!(Saturating(3i8).pow(6), Saturating(127));

impl Saturating<u16>

pub const MIN: Self

Returns the smallest value that can be represented by this integer type.

Examples

use std::num::Saturating;

assert_eq!(<Saturating<u16>>::MIN, Saturating(u16::MIN));

pub const MAX: Self

Returns the largest value that can be represented by this integer type.

Examples

use std::num::Saturating;

assert_eq!(<Saturating<u16>>::MAX, Saturating(u16::MAX));

pub const BITS: u32 = 16u32

Returns the size of this integer type in bits.

Examples

use std::num::Saturating;

assert_eq!(<Saturating<u16>>::BITS, u16::BITS);

pub const fn count_ones(self) -> u32

Returns the number of ones in the binary representation of self.

Examples

use std::num::Saturating;

let n = Saturating(0b01001100u16);

assert_eq!(n.count_ones(), 3);

pub const fn count_zeros(self) -> u32

Returns the number of zeros in the binary representation of self.

Examples

use std::num::Saturating;

assert_eq!(Saturating(!0u16).count_zeros(), 0);

pub const fn trailing_zeros(self) -> u32

Returns the number of trailing zeros in the binary representation of self.

Examples

use std::num::Saturating;

let n = Saturating(0b0101000u16);

assert_eq!(n.trailing_zeros(), 3);

pub const fn rotate_left(self, n: u32) -> Self

Shifts the bits to the left by a specified amount, n, saturating the truncated bits to the end of the resulting integer.

Please note this isnt the same operation as the << shifting operator!

Examples

use std::num::Saturating;

let n: Saturating<i64> = Saturating(0x0123456789ABCDEF);
let m: Saturating<i64> = Saturating(-0x76543210FEDCBA99);

assert_eq!(n.rotate_left(32), m);

pub const fn rotate_right(self, n: u32) -> Self

Shifts the bits to the right by a specified amount, n, saturating the truncated bits to the beginning of the resulting integer.

Please note this isnt the same operation as the >> shifting operator!

Examples

use std::num::Saturating;

let n: Saturating<i64> = Saturating(0x0123456789ABCDEF);
let m: Saturating<i64> = Saturating(-0xFEDCBA987654322);

assert_eq!(n.rotate_right(4), m);

pub const fn swap_bytes(self) -> Self

Reverses the byte order of the integer.

Examples

use std::num::Saturating;

let n: Saturating<i16> = Saturating(0b0000000_01010101);
assert_eq!(n, Saturating(85));

let m = n.swap_bytes();

assert_eq!(m, Saturating(0b01010101_00000000));
assert_eq!(m, Saturating(21760));

pub const fn reverse_bits(self) -> Self

Reverses the bit pattern of the integer.

Examples

Please note that this example is shared among integer types, which is why i16 is used.

use std::num::Saturating;

let n = Saturating(0b0000000_01010101i16);
assert_eq!(n, Saturating(85));

let m = n.reverse_bits();

assert_eq!(m.0 as u16, 0b10101010_00000000);
assert_eq!(m, Saturating(-22016));

pub const fn from_be(x: Self) -> Self

Converts an integer from big endian to the targets endianness.

On big endian this is a no-op. On little endian the bytes are swapped.

Examples

use std::num::Saturating;

let n = Saturating(0x1Au16);

if cfg!(target_endian = "big") {
    assert_eq!(<Saturating<u16>>::from_be(n), n)
} else {
    assert_eq!(<Saturating<u16>>::from_be(n), n.swap_bytes())
}

pub const fn from_le(x: Self) -> Self

Converts an integer from little endian to the targets endianness.

On little endian this is a no-op. On big endian the bytes are swapped.

Examples

use std::num::Saturating;

let n = Saturating(0x1Au16);

if cfg!(target_endian = "little") {
    assert_eq!(<Saturating<u16>>::from_le(n), n)
} else {
    assert_eq!(<Saturating<u16>>::from_le(n), n.swap_bytes())
}

pub const fn to_be(self) -> Self

Converts self to big endian from the targets endianness.

On big endian this is a no-op. On little endian the bytes are swapped.

Examples

use std::num::Saturating;

let n = Saturating(0x1Au16);

if cfg!(target_endian = "big") {
    assert_eq!(n.to_be(), n)
} else {
    assert_eq!(n.to_be(), n.swap_bytes())
}

pub const fn to_le(self) -> Self

Converts self to little endian from the targets endianness.

On little endian this is a no-op. On big endian the bytes are swapped.

Examples

use std::num::Saturating;

let n = Saturating(0x1Au16);

if cfg!(target_endian = "little") {
    assert_eq!(n.to_le(), n)
} else {
    assert_eq!(n.to_le(), n.swap_bytes())
}

pub const fn pow(self, exp: u32) -> Self

Raises self to the power of exp, using exponentiation by squaring.

Examples

use std::num::Saturating;

assert_eq!(Saturating(3u16).pow(4), Saturating(81));

Results that are too large are saturated:

use std::num::Saturating;

assert_eq!(Saturating(3i8).pow(5), Saturating(127));
assert_eq!(Saturating(3i8).pow(6), Saturating(127));

impl Saturating<u32>

pub const MIN: Self

Returns the smallest value that can be represented by this integer type.

Examples

use std::num::Saturating;

assert_eq!(<Saturating<u32>>::MIN, Saturating(u32::MIN));

pub const MAX: Self

Returns the largest value that can be represented by this integer type.

Examples

use std::num::Saturating;

assert_eq!(<Saturating<u32>>::MAX, Saturating(u32::MAX));

pub const BITS: u32 = 32u32

Returns the size of this integer type in bits.

Examples

use std::num::Saturating;

assert_eq!(<Saturating<u32>>::BITS, u32::BITS);

pub const fn count_ones(self) -> u32

Returns the number of ones in the binary representation of self.

Examples

use std::num::Saturating;

let n = Saturating(0b01001100u32);

assert_eq!(n.count_ones(), 3);

pub const fn count_zeros(self) -> u32

Returns the number of zeros in the binary representation of self.

Examples

use std::num::Saturating;

assert_eq!(Saturating(!0u32).count_zeros(), 0);

pub const fn trailing_zeros(self) -> u32

Returns the number of trailing zeros in the binary representation of self.

Examples

use std::num::Saturating;

let n = Saturating(0b0101000u32);

assert_eq!(n.trailing_zeros(), 3);

pub const fn rotate_left(self, n: u32) -> Self

Shifts the bits to the left by a specified amount, n, saturating the truncated bits to the end of the resulting integer.

Please note this isnt the same operation as the << shifting operator!

Examples

use std::num::Saturating;

let n: Saturating<i64> = Saturating(0x0123456789ABCDEF);
let m: Saturating<i64> = Saturating(-0x76543210FEDCBA99);

assert_eq!(n.rotate_left(32), m);

pub const fn rotate_right(self, n: u32) -> Self

Shifts the bits to the right by a specified amount, n, saturating the truncated bits to the beginning of the resulting integer.

Please note this isnt the same operation as the >> shifting operator!

Examples

use std::num::Saturating;

let n: Saturating<i64> = Saturating(0x0123456789ABCDEF);
let m: Saturating<i64> = Saturating(-0xFEDCBA987654322);

assert_eq!(n.rotate_right(4), m);

pub const fn swap_bytes(self) -> Self

Reverses the byte order of the integer.

Examples

use std::num::Saturating;

let n: Saturating<i16> = Saturating(0b0000000_01010101);
assert_eq!(n, Saturating(85));

let m = n.swap_bytes();

assert_eq!(m, Saturating(0b01010101_00000000));
assert_eq!(m, Saturating(21760));

pub const fn reverse_bits(self) -> Self

Reverses the bit pattern of the integer.

Examples

Please note that this example is shared among integer types, which is why i16 is used.

use std::num::Saturating;

let n = Saturating(0b0000000_01010101i16);
assert_eq!(n, Saturating(85));

let m = n.reverse_bits();

assert_eq!(m.0 as u16, 0b10101010_00000000);
assert_eq!(m, Saturating(-22016));

pub const fn from_be(x: Self) -> Self

Converts an integer from big endian to the targets endianness.

On big endian this is a no-op. On little endian the bytes are swapped.

Examples

use std::num::Saturating;

let n = Saturating(0x1Au32);

if cfg!(target_endian = "big") {
    assert_eq!(<Saturating<u32>>::from_be(n), n)
} else {
    assert_eq!(<Saturating<u32>>::from_be(n), n.swap_bytes())
}

pub const fn from_le(x: Self) -> Self

Converts an integer from little endian to the targets endianness.

On little endian this is a no-op. On big endian the bytes are swapped.

Examples

use std::num::Saturating;

let n = Saturating(0x1Au32);

if cfg!(target_endian = "little") {
    assert_eq!(<Saturating<u32>>::from_le(n), n)
} else {
    assert_eq!(<Saturating<u32>>::from_le(n), n.swap_bytes())
}

pub const fn to_be(self) -> Self

Converts self to big endian from the targets endianness.

On big endian this is a no-op. On little endian the bytes are swapped.

Examples

use std::num::Saturating;

let n = Saturating(0x1Au32);

if cfg!(target_endian = "big") {
    assert_eq!(n.to_be(), n)
} else {
    assert_eq!(n.to_be(), n.swap_bytes())
}

pub const fn to_le(self) -> Self

Converts self to little endian from the targets endianness.

On little endian this is a no-op. On big endian the bytes are swapped.

Examples

use std::num::Saturating;

let n = Saturating(0x1Au32);

if cfg!(target_endian = "little") {
    assert_eq!(n.to_le(), n)
} else {
    assert_eq!(n.to_le(), n.swap_bytes())
}

pub const fn pow(self, exp: u32) -> Self

Raises self to the power of exp, using exponentiation by squaring.

Examples

use std::num::Saturating;

assert_eq!(Saturating(3u32).pow(4), Saturating(81));

Results that are too large are saturated:

use std::num::Saturating;

assert_eq!(Saturating(3i8).pow(5), Saturating(127));
assert_eq!(Saturating(3i8).pow(6), Saturating(127));

impl Saturating<u64>

pub const MIN: Self

Returns the smallest value that can be represented by this integer type.

Examples

use std::num::Saturating;

assert_eq!(<Saturating<u64>>::MIN, Saturating(u64::MIN));

pub const MAX: Self

Returns the largest value that can be represented by this integer type.

Examples

use std::num::Saturating;

assert_eq!(<Saturating<u64>>::MAX, Saturating(u64::MAX));

pub const BITS: u32 = 64u32

Returns the size of this integer type in bits.

Examples

use std::num::Saturating;

assert_eq!(<Saturating<u64>>::BITS, u64::BITS);

pub const fn count_ones(self) -> u32

Returns the number of ones in the binary representation of self.

Examples

use std::num::Saturating;

let n = Saturating(0b01001100u64);

assert_eq!(n.count_ones(), 3);

pub const fn count_zeros(self) -> u32

Returns the number of zeros in the binary representation of self.

Examples

use std::num::Saturating;

assert_eq!(Saturating(!0u64).count_zeros(), 0);

pub const fn trailing_zeros(self) -> u32

Returns the number of trailing zeros in the binary representation of self.

Examples

use std::num::Saturating;

let n = Saturating(0b0101000u64);

assert_eq!(n.trailing_zeros(), 3);

pub const fn rotate_left(self, n: u32) -> Self

Shifts the bits to the left by a specified amount, n, saturating the truncated bits to the end of the resulting integer.

Please note this isnt the same operation as the << shifting operator!

Examples

use std::num::Saturating;

let n: Saturating<i64> = Saturating(0x0123456789ABCDEF);
let m: Saturating<i64> = Saturating(-0x76543210FEDCBA99);

assert_eq!(n.rotate_left(32), m);

pub const fn rotate_right(self, n: u32) -> Self

Shifts the bits to the right by a specified amount, n, saturating the truncated bits to the beginning of the resulting integer.

Please note this isnt the same operation as the >> shifting operator!

Examples

use std::num::Saturating;

let n: Saturating<i64> = Saturating(0x0123456789ABCDEF);
let m: Saturating<i64> = Saturating(-0xFEDCBA987654322);

assert_eq!(n.rotate_right(4), m);

pub const fn swap_bytes(self) -> Self

Reverses the byte order of the integer.

Examples

use std::num::Saturating;

let n: Saturating<i16> = Saturating(0b0000000_01010101);
assert_eq!(n, Saturating(85));

let m = n.swap_bytes();

assert_eq!(m, Saturating(0b01010101_00000000));
assert_eq!(m, Saturating(21760));

pub const fn reverse_bits(self) -> Self

Reverses the bit pattern of the integer.

Examples

Please note that this example is shared among integer types, which is why i16 is used.

use std::num::Saturating;

let n = Saturating(0b0000000_01010101i16);
assert_eq!(n, Saturating(85));

let m = n.reverse_bits();

assert_eq!(m.0 as u16, 0b10101010_00000000);
assert_eq!(m, Saturating(-22016));

pub const fn from_be(x: Self) -> Self

Converts an integer from big endian to the targets endianness.

On big endian this is a no-op. On little endian the bytes are swapped.

Examples

use std::num::Saturating;

let n = Saturating(0x1Au64);

if cfg!(target_endian = "big") {
    assert_eq!(<Saturating<u64>>::from_be(n), n)
} else {
    assert_eq!(<Saturating<u64>>::from_be(n), n.swap_bytes())
}

pub const fn from_le(x: Self) -> Self

Converts an integer from little endian to the targets endianness.

On little endian this is a no-op. On big endian the bytes are swapped.

Examples

use std::num::Saturating;

let n = Saturating(0x1Au64);

if cfg!(target_endian = "little") {
    assert_eq!(<Saturating<u64>>::from_le(n), n)
} else {
    assert_eq!(<Saturating<u64>>::from_le(n), n.swap_bytes())
}

pub const fn to_be(self) -> Self

Converts self to big endian from the targets endianness.

On big endian this is a no-op. On little endian the bytes are swapped.

Examples

use std::num::Saturating;

let n = Saturating(0x1Au64);

if cfg!(target_endian = "big") {
    assert_eq!(n.to_be(), n)
} else {
    assert_eq!(n.to_be(), n.swap_bytes())
}

pub const fn to_le(self) -> Self

Converts self to little endian from the targets endianness.

On little endian this is a no-op. On big endian the bytes are swapped.

Examples

use std::num::Saturating;

let n = Saturating(0x1Au64);

if cfg!(target_endian = "little") {
    assert_eq!(n.to_le(), n)
} else {
    assert_eq!(n.to_le(), n.swap_bytes())
}

pub const fn pow(self, exp: u32) -> Self

Raises self to the power of exp, using exponentiation by squaring.

Examples

use std::num::Saturating;

assert_eq!(Saturating(3u64).pow(4), Saturating(81));

Results that are too large are saturated:

use std::num::Saturating;

assert_eq!(Saturating(3i8).pow(5), Saturating(127));
assert_eq!(Saturating(3i8).pow(6), Saturating(127));

impl Saturating<u128>

pub const MIN: Self

Returns the smallest value that can be represented by this integer type.

Examples

use std::num::Saturating;

assert_eq!(<Saturating<u128>>::MIN, Saturating(u128::MIN));

pub const MAX: Self

R