Type Alias core::num::NonZeroI128

pub type NonZeroI128 = NonZero<i128>;

An integer that is known not to equal zero.

This enables some memory layout optimization. For example, Option<NonZeroI128> is the same size as i128:

use std::mem::size_of;
assert_eq!(size_of::<Option<core::num::NonZeroI128>>(), size_of::<i128>());
Run

Layout

NonZeroI128 is guaranteed to have the same layout and bit validity as i128 with the exception that 0 is not a valid instance. Option<NonZeroI128> is guaranteed to be compatible with i128, including in FFI.

Thanks to the null pointer optimization, NonZeroI128 and Option<NonZeroI128> are guaranteed to have the same size and alignment:

use std::num::NonZeroI128;

assert_eq!(size_of::<NonZeroI128>(), size_of::<Option<NonZeroI128>>());
assert_eq!(align_of::<NonZeroI128>(), align_of::<Option<NonZeroI128>>());
Run

Aliased Type

struct NonZeroI128(/* private fields */);

Implementations

impl NonZeroI128

pub const BITS: u32 = 128u32

The size of this non-zero integer type in bits.

This value is equal to i128::BITS.

Examples

assert_eq!(NonZeroI128::BITS, i128::BITS);
Run

pub const fn leading_zeros(self) -> u32

Returns the number of leading zeros in the binary representation of self.

On many architectures, this function can perform better than leading_zeros() on the underlying integer type, as special handling of zero can be avoided.

Examples

Basic usage:

let n = std::num::NonZeroI128::new(-1i128).unwrap();

assert_eq!(n.leading_zeros(), 0);
Run

pub const fn trailing_zeros(self) -> u32

Returns the number of trailing zeros in the binary representation of self.

On many architectures, this function can perform better than trailing_zeros() on the underlying integer type, as special handling of zero can be avoided.

Examples

Basic usage:

let n = std::num::NonZeroI128::new(0b0101000).unwrap();

assert_eq!(n.trailing_zeros(), 3);
Run

pub fn count_ones(self) -> NonZero<u32>

🔬This is a nightly-only experimental API. (non_zero_count_ones #120287)

Returns the number of ones in the binary representation of self.

Examples

Basic usage:

#![feature(generic_nonzero, non_zero_count_ones)]
let a = NonZero::<i128>::new(0b100_0000)?;
let b = NonZero::<i128>::new(0b100_0011)?;

assert_eq!(a.count_ones(), NonZero::new(1)?);
assert_eq!(b.count_ones(), NonZero::new(3)?);
Run

pub const MIN: Self = _

The smallest value that can be represented by this non-zero integer type, equal to i128::MIN.

Note: While most integer types are defined for every whole number between MIN and MAX, signed non-zero integers are a special case. They have a “gap” at 0.

Examples

assert_eq!(NonZeroI128::MIN.get(), i128::MIN);
Run

pub const MAX: Self = _

The largest value that can be represented by this non-zero integer type, equal to i128::MAX.

Note: While most integer types are defined for every whole number between MIN and MAX, signed non-zero integers are a special case. They have a “gap” at 0.

Examples

assert_eq!(NonZeroI128::MAX.get(), i128::MAX);
Run

pub const fn abs(self) -> Self

Computes the absolute value of self. See i128::abs for documentation on overflow behaviour.

Example

let pos = NonZeroI128::new(1)?;
let neg = NonZeroI128::new(-1)?;

assert_eq!(pos, pos.abs());
assert_eq!(pos, neg.abs());
Run

pub const fn checked_abs(self) -> Option<Self>

Checked absolute value. Checks for overflow and returns None if self == NonZeroI128::MIN. The result cannot be zero.

Example

let pos = NonZeroI128::new(1)?;
let neg = NonZeroI128::new(-1)?;
let min = NonZeroI128::new(i128::MIN)?;

assert_eq!(Some(pos), neg.checked_abs());
assert_eq!(None, min.checked_abs());
Run

pub const fn overflowing_abs(self) -> (Self, bool)

Computes the absolute value of self, with overflow information, see i128::overflowing_abs.

Example

let pos = NonZeroI128::new(1)?;
let neg = NonZeroI128::new(-1)?;
let min = NonZeroI128::new(i128::MIN)?;

assert_eq!((pos, false), pos.overflowing_abs());
assert_eq!((pos, false), neg.overflowing_abs());
assert_eq!((min, true), min.overflowing_abs());
Run

pub const fn saturating_abs(self) -> Self

Saturating absolute value, see i128::saturating_abs.

Example

let pos = NonZeroI128::new(1)?;
let neg = NonZeroI128::new(-1)?;
let min = NonZeroI128::new(i128::MIN)?;
let min_plus = NonZeroI128::new(i128::MIN + 1)?;
let max = NonZeroI128::new(i128::MAX)?;

assert_eq!(pos, pos.saturating_abs());
assert_eq!(pos, neg.saturating_abs());
assert_eq!(max, min.saturating_abs());
assert_eq!(max, min_plus.saturating_abs());
Run

pub const fn wrapping_abs(self) -> Self

Wrapping absolute value, see i128::wrapping_abs.

Example

let pos = NonZeroI128::new(1)?;
let neg = NonZeroI128::new(-1)?;
let min = NonZeroI128::new(i128::MIN)?;

assert_eq!(pos, pos.wrapping_abs());
assert_eq!(pos, neg.wrapping_abs());
assert_eq!(min, min.wrapping_abs());
assert_eq!(max, (-max).wrapping_abs());
Run

pub const fn unsigned_abs(self) -> NonZeroU128

Computes the absolute value of self without any wrapping or panicking.

Example

let u_pos = NonZeroU128::new(1)?;
let i_pos = NonZeroI128::new(1)?;
let i_neg = NonZeroI128::new(-1)?;
let i_min = NonZeroI128::new(i128::MIN)?;
let u_max = NonZeroU128::new(u128::MAX / 2 + 1)?;

assert_eq!(u_pos, i_pos.unsigned_abs());
assert_eq!(u_pos, i_neg.unsigned_abs());
assert_eq!(u_max, i_min.unsigned_abs());
Run

pub const fn is_positive(self) -> bool

Returns true if self is positive and false if the number is negative.

Example

let pos_five = NonZeroI128::new(5)?;
let neg_five = NonZeroI128::new(-5)?;

assert!(pos_five.is_positive());
assert!(!neg_five.is_positive());
Run

pub const fn is_negative(self) -> bool

Returns true if self is negative and false if the number is positive.

Example

let pos_five = NonZeroI128::new(5)?;
let neg_five = NonZeroI128::new(-5)?;

assert!(neg_five.is_negative());
assert!(!pos_five.is_negative());
Run

pub const fn checked_neg(self) -> Option<Self>

Checked negation. Computes -self, returning None if self == NonZeroI128::MIN.

Example

let pos_five = NonZeroI128::new(5)?;
let neg_five = NonZeroI128::new(-5)?;
let min = NonZeroI128::new(i128::MIN)?;

assert_eq!(pos_five.checked_neg(), Some(neg_five));
assert_eq!(min.checked_neg(), None);
Run

pub const fn overflowing_neg(self) -> (Self, bool)

Negates self, overflowing if this is equal to the minimum value.

See i128::overflowing_neg for documentation on overflow behaviour.

Example

let pos_five = NonZeroI128::new(5)?;
let neg_five = NonZeroI128::new(-5)?;
let min = NonZeroI128::new(i128::MIN)?;

assert_eq!(pos_five.overflowing_neg(), (neg_five, false));
assert_eq!(min.overflowing_neg(), (min, true));
Run

pub const fn saturating_neg(self) -> Self

Saturating negation. Computes -self, returning NonZeroI128::MAX if self == NonZeroI128::MIN instead of overflowing.

Example

let pos_five = NonZeroI128::new(5)?;
let neg_five = NonZeroI128::new(-5)?;
let min = NonZeroI128::new(i128::MIN)?;
let min_plus_one = NonZeroI128::new(i128::MIN + 1)?;
let max = NonZeroI128::new(i128::MAX)?;

assert_eq!(pos_five.saturating_neg(), neg_five);
assert_eq!(min.saturating_neg(), max);
assert_eq!(max.saturating_neg(), min_plus_one);
Run

pub const fn wrapping_neg(self) -> Self

Wrapping (modular) negation. Computes -self, wrapping around at the boundary of the type.

See i128::wrapping_neg for documentation on overflow behaviour.

Example

let pos_five = NonZeroI128::new(5)?;
let neg_five = NonZeroI128::new(-5)?;
let min = NonZeroI128::new(i128::MIN)?;

assert_eq!(pos_five.wrapping_neg(), neg_five);
assert_eq!(min.wrapping_neg(), min);
Run

pub const fn checked_mul(self, other: Self) -> Option<Self>

Multiplies two non-zero integers together. Checks for overflow and returns None on overflow. As a consequence, the result cannot wrap to zero.

Examples

let two = NonZeroI128::new(2)?;
let four = NonZeroI128::new(4)?;
let max = NonZeroI128::new(i128::MAX)?;

assert_eq!(Some(four), two.checked_mul(two));
assert_eq!(None, max.checked_mul(two));
Run

pub const fn saturating_mul(self, other: Self) -> Self

Multiplies two non-zero integers together. Return NonZeroI128::MAX on overflow.

Examples

let two = NonZeroI128::new(2)?;
let four = NonZeroI128::new(4)?;
let max = NonZeroI128::new(i128::MAX)?;

assert_eq!(four, two.saturating_mul(two));
assert_eq!(max, four.saturating_mul(max));
Run

pub const unsafe fn unchecked_mul(self, other: Self) -> Self

🔬This is a nightly-only experimental API. (nonzero_ops #84186)

Multiplies two non-zero integers together, assuming overflow cannot occur. Overflow is unchecked, and it is undefined behaviour to overflow even if the result would wrap to a non-zero value. The behaviour is undefined as soon as self * rhs > i128::MAX, or self * rhs < i128::MIN.

Examples

#![feature(nonzero_ops)]

let two = NonZeroI128::new(2)?;
let four = NonZeroI128::new(4)?;

assert_eq!(four, unsafe { two.unchecked_mul(two) });
Run

pub const fn checked_pow(self, other: u32) -> Option<Self>

Raises non-zero value to an integer power. Checks for overflow and returns None on overflow. As a consequence, the result cannot wrap to zero.

Examples

let three = NonZeroI128::new(3)?;
let twenty_seven = NonZeroI128::new(27)?;
let half_max = NonZeroI128::new(i128::MAX / 2)?;

assert_eq!(Some(twenty_seven), three.checked_pow(3));
assert_eq!(None, half_max.checked_pow(3));
Run

pub const fn saturating_pow(self, other: u32) -> Self

Raise non-zero value to an integer power. Return NonZeroI128::MIN or NonZeroI128::MAX on overflow.

Examples

let three = NonZeroI128::new(3)?;
let twenty_seven = NonZeroI128::new(27)?;
let max = NonZeroI128::new(i128::MAX)?;

assert_eq!(twenty_seven, three.saturating_pow(3));
assert_eq!(max, max.saturating_pow(3));
Run