Type Alias core::num::NonZeroU16

pub type NonZeroU16 = NonZero<u16>;

An integer that is known not to equal zero.

This enables some memory layout optimization. For example, Option<NonZeroU16> is the same size as u16:

use std::mem::size_of;
assert_eq!(size_of::<Option<core::num::NonZeroU16>>(), size_of::<u16>());
Run

Layout

NonZeroU16 is guaranteed to have the same layout and bit validity as u16 with the exception that 0 is not a valid instance. Option<NonZeroU16> is guaranteed to be compatible with u16, including in FFI.

Thanks to the null pointer optimization, NonZeroU16 and Option<NonZeroU16> are guaranteed to have the same size and alignment:

use std::num::NonZeroU16;

assert_eq!(size_of::<NonZeroU16>(), size_of::<Option<NonZeroU16>>());
assert_eq!(align_of::<NonZeroU16>(), align_of::<Option<NonZeroU16>>());
Run

Aliased Type

struct NonZeroU16(/* private fields */);

Implementations

impl NonZeroU16

pub const unsafe fn new_unchecked(n: u16) -> Self

Creates a non-zero without checking whether the value is non-zero. This results in undefined behaviour if the value is zero.

Safety

The value must not be zero.

pub const fn new(n: u16) -> Option<Self>

Creates a non-zero if the given value is not zero.

pub unsafe fn from_mut_unchecked(n: &mut u16) -> &mut Self

🔬This is a nightly-only experimental API. (nonzero_from_mut #106290)

Converts a primitive mutable reference to a non-zero mutable reference without checking whether the referenced value is non-zero. This results in undefined behavior if *n is zero.

Safety

The referenced value must not be currently zero.

pub fn from_mut(n: &mut u16) -> Option<&mut Self>

🔬This is a nightly-only experimental API. (nonzero_from_mut #106290)

Converts a primitive mutable reference to a non-zero mutable reference if the referenced integer is not zero.

pub const fn get(self) -> u16

Returns the value as a primitive type.

pub const BITS: u32 = 16u32

The size of this non-zero integer type in bits.

This value is equal to u16::BITS.

Examples

assert_eq!(NonZeroU16::BITS, u16::BITS);
Run

pub const fn leading_zeros(self) -> u32

Returns the number of leading zeros in the binary representation of self.

On many architectures, this function can perform better than leading_zeros() on the underlying integer type, as special handling of zero can be avoided.

Examples

Basic usage:

let n = std::num::NonZeroU16::new(u16::MAX).unwrap();

assert_eq!(n.leading_zeros(), 0);
Run

pub const fn trailing_zeros(self) -> u32

Returns the number of trailing zeros in the binary representation of self.

On many architectures, this function can perform better than trailing_zeros() on the underlying integer type, as special handling of zero can be avoided.

Examples

Basic usage:

let n = std::num::NonZeroU16::new(0b0101000).unwrap();

assert_eq!(n.trailing_zeros(), 3);
Run

pub fn count_ones(self) -> NonZeroU32

🔬This is a nightly-only experimental API. (non_zero_count_ones #120287)

Returns the number of ones in the binary representation of self.

Examples

Basic usage:

#![feature(non_zero_count_ones)]

let one = num::NonZeroU32::new(1)?;
let three = num::NonZeroU32::new(3)?;
let a = NonZeroU16::new(0b100_0000)?;
let b = NonZeroU16::new(0b100_0011)?;

assert_eq!(a.count_ones(), one);
assert_eq!(b.count_ones(), three);
Run

pub const MIN: Self = _

The smallest value that can be represented by this non-zero integer type, 1.

Examples

assert_eq!(NonZeroU16::MIN.get(), 1u16);
Run

pub const MAX: Self = _

The largest value that can be represented by this non-zero integer type, equal to u16::MAX.

Examples

assert_eq!(NonZeroU16::MAX.get(), u16::MAX);
Run

pub const fn checked_add(self, other: u16) -> Option<Self>

Adds an unsigned integer to a non-zero value. Checks for overflow and returns None on overflow. As a consequence, the result cannot wrap to zero.

Examples

let one = NonZeroU16::new(1)?;
let two = NonZeroU16::new(2)?;
let max = NonZeroU16::new(u16::MAX)?;

assert_eq!(Some(two), one.checked_add(1));
assert_eq!(None, max.checked_add(1));
Run

pub const fn saturating_add(self, other: u16) -> Self

Adds an unsigned integer to a non-zero value. Return NonZeroU16::MAX on overflow.

Examples

let one = NonZeroU16::new(1)?;
let two = NonZeroU16::new(2)?;
let max = NonZeroU16::new(u16::MAX)?;

assert_eq!(two, one.saturating_add(1));
assert_eq!(max, max.saturating_add(1));
Run

pub const unsafe fn unchecked_add(self, other: u16) -> Self

🔬This is a nightly-only experimental API. (nonzero_ops #84186)

Adds an unsigned integer to a non-zero value, assuming overflow cannot occur. Overflow is unchecked, and it is undefined behaviour to overflow even if the result would wrap to a non-zero value. The behaviour is undefined as soon as self + rhs > u16::MAX.

Examples

#![feature(nonzero_ops)]

let one = NonZeroU16::new(1)?;
let two = NonZeroU16::new(2)?;

assert_eq!(two, unsafe { one.unchecked_add(1) });
Run

pub const fn checked_next_power_of_two(self) -> Option<Self>

Returns the smallest power of two greater than or equal to n. Checks for overflow and returns None if the next power of two is greater than the type’s maximum value. As a consequence, the result cannot wrap to zero.

Examples

let two = NonZeroU16::new(2)?;
let three = NonZeroU16::new(3)?;
let four = NonZeroU16::new(4)?;
let max = NonZeroU16::new(u16::MAX)?;

assert_eq!(Some(two), two.checked_next_power_of_two() );
assert_eq!(Some(four), three.checked_next_power_of_two() );
assert_eq!(None, max.checked_next_power_of_two() );
Run

pub const fn ilog2(self) -> u32

Returns the base 2 logarithm of the number, rounded down.

This is the same operation as u16::ilog2, except that it has no failure cases to worry about since this value can never be zero.

Examples

assert_eq!(NonZeroU16::new(7).unwrap().ilog2(), 2);
assert_eq!(NonZeroU16::new(8).unwrap().ilog2(), 3);
assert_eq!(NonZeroU16::new(9).unwrap().ilog2(), 3);
Run

pub const fn ilog10(self) -> u32

Returns the base 10 logarithm of the number, rounded down.

This is the same operation as u16::ilog10, except that it has no failure cases to worry about since this value can never be zero.

Examples

assert_eq!(NonZeroU16::new(99).unwrap().ilog10(), 1);
assert_eq!(NonZeroU16::new(100).unwrap().ilog10(), 2);
assert_eq!(NonZeroU16::new(101).unwrap().ilog10(), 2);
Run

pub fn midpoint(self, rhs: Self) -> Self

🔬This is a nightly-only experimental API. (num_midpoint #110840)

Calculates the middle point of self and rhs.

midpoint(a, b) is (a + b) >> 1 as if it were performed in a sufficiently-large signed integral type. This implies that the result is always rounded towards negative infinity and that no overflow will ever occur.

Examples

#![feature(num_midpoint)]

let one = NonZeroU16::new(1)?;
let two = NonZeroU16::new(2)?;
let four = NonZeroU16::new(4)?;

assert_eq!(one.midpoint(four), two);
assert_eq!(four.midpoint(one), two);
Run

pub const fn is_power_of_two(self) -> bool

Returns true if and only if self == (1 << k) for some k.

On many architectures, this function can perform better than is_power_of_two() on the underlying integer type, as special handling of zero can be avoided.

Examples

Basic usage:

let eight = std::num::NonZeroU16::new(8).unwrap();
assert!(eight.is_power_of_two());
let ten = std::num::NonZeroU16::new(10).unwrap();
assert!(!ten.is_power_of_two());
Run

pub const fn checked_mul(self, other: Self) -> Option<Self>

Multiplies two non-zero integers together. Checks for overflow and returns None on overflow. As a consequence, the result cannot wrap to zero.

Examples

let two = NonZeroU16::new(2)?;
let four = NonZeroU16::new(4)?;
let max = NonZeroU16::new(u16::MAX)?;

assert_eq!(Some(four), two.checked_mul(two));
assert_eq!(None, max.checked_mul(two));
Run

pub const fn saturating_mul(self, other: Self) -> Self

Multiplies two non-zero integers together. Return NonZeroU16::MAX on overflow.

Examples

let two = NonZeroU16::new(2)?;
let four = NonZeroU16::new(4)?;
let max = NonZeroU16::new(u16::MAX)?;

assert_eq!(four, two.saturating_mul(two));
assert_eq!(max, four.saturating_mul(max));
Run

pub const unsafe fn unchecked_mul(self, other: Self) -> Self

🔬This is a nightly-only experimental API. (nonzero_ops #84186)

Multiplies two non-zero integers together, assuming overflow cannot occur. Overflow is unchecked, and it is undefined behaviour to overflow even if the result would wrap to a non-zero value. The behaviour is undefined as soon as self * rhs > u16::MAX.

Examples

#![feature(nonzero_ops)]

let two = NonZeroU16::new(2)?;
let four = NonZeroU16::new(4)?;

assert_eq!(four, unsafe { two.unchecked_mul(two) });
Run

pub const fn checked_pow(self, other: u32) -> Option<Self>

Raises non-zero value to an integer power. Checks for overflow and returns None on overflow. As a consequence, the result cannot wrap to zero.

Examples

let three = NonZeroU16::new(3)?;
let twenty_seven = NonZeroU16::new(27)?;
let half_max = NonZeroU16::new(u16::MAX / 2)?;

assert_eq!(Some(twenty_seven), three.checked_pow(3));
assert_eq!(None, half_max.checked_pow(3));
Run

pub const fn saturating_pow(self, other: u32) -> Self

Raise non-zero value to an integer power. Return NonZeroU16::MAX on overflow.

Examples

let three = NonZeroU16::new(3)?;
let twenty_seven = NonZeroU16::new(27)?;
let max = NonZeroU16::new(u16::MAX)?;

assert_eq!(twenty_seven, three.saturating_pow(3));
assert_eq!(max, max.saturating_pow(3));
Run