Enum core::ops::GeneratorState

#[lang = "generator_state"]

pub enum GeneratorState<Y, R> {
    Yielded(Y),
    Complete(R),
}

🔬 This is a nightly-only experimental API. (generator_trait #43122)

The result of a generator resumption.

This enum is returned from the Generator::resume method and indicates the possible return values of a generator. Currently this corresponds to either a suspension point (Yielded) or a termination point (Complete).

Variants
Yielded(Y)

🔬 This is a nightly-only experimental API. (generator_trait #43122)

The generator suspended with a value.

This state indicates that a generator has been suspended, and typically corresponds to a yield statement. The value provided in this variant corresponds to the expression passed to yield and allows generators to provide a value each time they yield.

Complete(R)

🔬 This is a nightly-only experimental API. (generator_trait #43122)

The generator completed with a return value.

This state indicates that a generator has finished execution with the provided value. Once a generator has returned Complete it is considered a programmer error to call resume again.