Struct core::ops::Range

pub struct Range<Idx> {
    pub start: Idx,
    pub end: Idx,
}

A (half-open) range bounded inclusively below and exclusively above (start..end).

The Range start..end contains all values with x >= start and x < end. It is empty unless start < end.

Examples

assert_eq!((3..5), std::ops::Range { start: 3, end: 5 });
assert_eq!(3 + 4 + 5, (3..6).sum());

let arr = [0, 1, 2, 3, 4];
assert_eq!(arr[ ..  ], [0,1,2,3,4]);
assert_eq!(arr[ .. 3], [0,1,2    ]);
assert_eq!(arr[ ..=3], [0,1,2,3  ]);
assert_eq!(arr[1..  ], [  1,2,3,4]);
assert_eq!(arr[1.. 3], [  1,2    ]);  // Range
assert_eq!(arr[1..=3], [  1,2,3  ]);Run

Fields
start: Idx

The lower bound of the range (inclusive).

end: Idx

The upper bound of the range (exclusive).

Methods

impl<Idx: PartialOrd<Idx>> Range<Idx>

pub fn contains<U: ?Sized>(&self, item: &U) -> bool where
Idx: PartialOrd<U>,
U: PartialOrd<Idx>,

Returns true if item is contained in the range.

Examples

use std::f32;

assert!(!(3..5).contains(&2));
assert!( (3..5).contains(&3));
assert!( (3..5).contains(&4));
assert!(!(3..5).contains(&5));

assert!(!(3..3).contains(&3));
assert!(!(3..2).contains(&3));

assert!( (0.0..1.0).contains(&0.5));
assert!(!(0.0..1.0).contains(&f32::NAN));
assert!(!(0.0..f32::NAN).contains(&0.5));
assert!(!(f32::NAN..1.0).contains(&0.5));Run

pub fn is_empty(&self) -> bool

ðŸ”¬ This is a nightly-only experimental API. (range_is_empty #48111)

recently added

Returns true if the range contains no items.

Examples

#![feature(range_is_empty)]

assert!(!(3..5).is_empty());
assert!( (3..3).is_empty());
assert!( (3..2).is_empty());Run

The range is empty if either side is incomparable:

#![feature(range_is_empty)]

use std::f32::NAN;
assert!(!(3.0..5.0).is_empty());
assert!( (3.0..NAN).is_empty());
assert!( (NAN..5.0).is_empty());Run