Struct core::ops::RangeFrom

pub struct RangeFrom<Idx> {
    pub start: Idx,
}

A range only bounded inclusively below (start..).

The RangeFrom start.. contains all values with x >= start.

Note: Currently, no overflow checking is done for the Iterator implementation; if you use an integer range and the integer overflows, it might panic in debug mode or create an endless loop in release mode. This overflow behavior might change in the future.

Examples

assert_eq!((2..), std::ops::RangeFrom { start: 2 });
assert_eq!(2 + 3 + 4, (2..).take(3).sum());

let arr = [0, 1, 2, 3];
assert_eq!(arr[ .. ], [0,1,2,3]);
assert_eq!(arr[ ..3], [0,1,2  ]);
assert_eq!(arr[1.. ], [  1,2,3]);  // RangeFrom
assert_eq!(arr[1..3], [  1,2  ]);Run

Fields
start: Idx

The lower bound of the range (inclusive).

Methods

impl<Idx: PartialOrd<Idx>> RangeFrom<Idx>

pub fn contains(&self, item: Idx) -> bool

🔬 This is a nightly-only experimental API. (range_contains #32311)

recently added as per RFC

Returns true if item is contained in the range.

Examples

#![feature(range_contains)]

assert!(!(3..).contains(2));
assert!( (3..).contains(3));
assert!( (3..).contains(1_000_000_000));Run