Struct core::ops::RangeToInclusive

pub struct RangeToInclusive<Idx> {
    pub end: Idx,
}

A range only bounded inclusively above (..=end).

The RangeToInclusive ..=end contains all values with x <= end. It cannot serve as an Iterator because it doesn't have a starting point.

Examples

The ..=end syntax is a RangeToInclusive:

assert_eq!((..=5), std::ops::RangeToInclusive{ end: 5 });Run

It does not have an IntoIterator implementation, so you can't use it in a for loop directly. This won't compile:

// error[E0277]: the trait bound `std::ops::RangeToInclusive<{integer}>:
// std::iter::Iterator` is not satisfied
for i in ..=5 {
    // ...
}Run

When used as a slicing index, RangeToInclusive produces a slice of all array elements up to and including the index indicated by end.

let arr = [0, 1, 2, 3, 4];
assert_eq!(arr[ ..  ], [0,1,2,3,4]);
assert_eq!(arr[ .. 3], [0,1,2    ]);
assert_eq!(arr[ ..=3], [0,1,2,3  ]);  // RangeToInclusive
assert_eq!(arr[1..  ], [  1,2,3,4]);
assert_eq!(arr[1.. 3], [  1,2    ]);
assert_eq!(arr[1..=3], [  1,2,3  ]);Run

Fields
end: Idx

The upper bound of the range (inclusive)

Methods

impl<Idx: PartialOrd<Idx>> RangeToInclusive<Idx>

pub fn contains<U: ?Sized>(&self, item: &U) -> bool where
Idx: PartialOrd<U>,
U: PartialOrd<Idx>,

Returns true if item is contained in the range.

Examples

use std::f32;

assert!( (..=5).contains(&-1_000_000_000));
assert!( (..=5).contains(&5));
assert!(!(..=5).contains(&6));

assert!( (..=1.0).contains(&1.0));
assert!(!(..=1.0).contains(&f32::NAN));
assert!(!(..=f32::NAN).contains(&0.5));Run