Trait core::ops::AsyncFnOnce

pub trait AsyncFnOnce<Args: Tuple> {
    type CallOnceFuture: Future<Output = Self::Output>;
    type Output;

    // Required method
    extern "rust-call" fn async_call_once(
        self,
        args: Args,
    ) -> Self::CallOnceFuture;
}

🔬This is a nightly-only experimental API. (async_fn_traits)

An async-aware version of the FnOnce trait.

All async fn and functions returning futures implement this trait.

Required Associated Types

type CallOnceFuture: Future<Output = Self::Output>

🔬This is a nightly-only experimental API. (async_fn_traits)

Future returned by AsyncFnOnce::async_call_once.

type Output

🔬This is a nightly-only experimental API. (async_fn_traits)

Output type of the called closure’s future.

Required Methods

extern "rust-call" fn async_call_once( self, args: Args, ) -> Self::CallOnceFuture

🔬This is a nightly-only experimental API. (async_fn_traits)

Call the AsyncFnOnce, returning a future which may move out of the called closure.