core::ops

Trait CoerceShared

pub trait CoerceShared: Reborrow {
    type Target: Copy;
}

🔬This is a nightly-only experimental API. (reborrow #145612)

Allows reborrowable value to be reborrowed as shared, creating a copy that disables the source for writes for the lifetime of the copy.

Required Associated Types

type Target: Copy

🔬This is a nightly-only experimental API. (reborrow #145612)

The type of this value when reborrowed as shared.