Trait core::ops::DerefMutExperimental

pub trait DerefMut<Sized? Result> for Sized? : Deref<Result> {
    fn deref_mut<'a>(&'a mut self) -> &'a mut Result;
}

The DerefMut trait is used to specify the functionality of dereferencing mutably like *v = 1;

Example

A struct with a single field which is modifiable via dereferencing the struct.
struct DerefMutExample<T> { value: T } impl<T> Deref<T> for DerefMutExample<T> { fn deref<'a>(&'a self) -> &'a T { &self.value } } impl<T> DerefMut<T> for DerefMutExample<T> { fn deref_mut<'a>(&'a mut self) -> &'a mut T { &mut self.value } } fn main() { let mut x = DerefMutExample { value: 'a' }; *x = 'b'; assert_eq!('b', *x); }

struct DerefMutExample<T> {
    value: T
}

impl<T> Deref<T> for DerefMutExample<T> {
    fn deref<'a>(&'a self) -> &'a T {
        &self.value
    }
}

impl<T> DerefMut<T> for DerefMutExample<T> {
    fn deref_mut<'a>(&'a mut self) -> &'a mut T {
        &mut self.value
    }
}

fn main() {
    let mut x = DerefMutExample { value: 'a' };
    *x = 'b';
    assert_eq!('b', *x);
}

Required Methods

fn deref_mut<'a>(&'a mut self) -> &'a mut Result

The method called to mutably dereference a value