Trait core::ops::FnUnstable

pub trait Fn<Args> {
    type Output;

    fn call(&self, args: Args) -> Output;
}

A version of the call operator that takes an immutable receiver.

Associated Types

type Output

Required Methods

fn call(&self, args: Args) -> Output

This is called when the call operator is used.