Module core::pin

将数据固定到其在内存中的位置的类型。

Types that pin data to its location in memory.

有时保证对象不会移动很有用，因为它们在内存中的位置不会改变，因此可以依赖。 这种情况的一个主要示例是构建自引用结构，因为移动带有指向自身的指针的对象将使它们无效，这可能导致未定义的行为。

It is sometimes useful to have objects that are guaranteed not to move, in the sense that their placement in memory does not change, and can thus be relied upon. A prime example of such a scenario would be building self-referential structs, as moving an object with pointers to itself will invalidate them, which could cause undefined behavior.

在高层次上， Pin
确保任何指针类型 P 的指针在内存中都有一个稳定的位置，这意味着它不能移动到其他地方，并且它的内存在它被丢弃之前不能被释放。 我们说指针是“固定的”。 当讨论结合固定数据和非固定数据的类型时，事情会变得更加微妙； 请参阅下面的更多细节。

At a high level, a Pin<P> ensures that the pointee of any pointer type P has a stable location in memory, meaning it cannot be moved elsewhere and its memory cannot be deallocated until it gets dropped. We say that the pointee is “pinned”. Things get more subtle when discussing types that combine pinned with non-pinned data; see below for more details.

默认情况下，Rust 中的所有类型都是可移动的。 Rust 允许按值传递所有类型，并且常见的智能指针类型（例如 Box 和 &mut T）允许替换和移动它们包含的值：您可以移出 Box，也可以使用 mem： ：交换。 Pin
包装了一个指针类型 P，因此 Pin<>> 的功能很像普通的 Box：当 Pin<>> 被删除时，它的内容也被删除，并且内存被释放 . 类似地，Pin<&mut T> 很像 &mut T。但是，Pin
不会让客户端实际获得 Box 或 &mut T 来固定数据，这意味着您不能使用诸如 mem:: 之类的操作。 交换：

By default, all types in Rust are movable. Rust allows passing all types by-value, and common smart-pointer types such as Box<T> and &mut T allow replacing and moving the values they contain: you can move out of a Box<T>, or you can use mem::swap. Pin<P> wraps a pointer type P, so Pin<Box<T>> functions much like a regular Box<T>: when a Pin<Box<T>> gets dropped, so do its contents, and the memory gets deallocated. Similarly, Pin<&mut T> is a lot like &mut T. However, Pin<P> does not let clients actually obtain a Box<T> or &mut T to pinned data, which implies that you cannot use operations such as mem::swap:

use std::pin::Pin;
fn swap_pins<T>(x: Pin<&mut T>, y: Pin<&mut T>) {
    // `mem::swap` needs `&mut T`, but we cannot get it.
    // We are stuck, we cannot swap the contents of these references.
    // We could use `Pin::get_unchecked_mut`, but that is unsafe for a reason:
    // we are not allowed to use it for moving things out of the `Pin`.
}
Run

值得重申的是 Pin
并没有改变 Rust 编译器认为所有类型都可移动的事实。 mem::swap 对任何 T 都是可调用的。相反，Pin
防止某些值（由封装在 Pin
中的指针指向）被移动，因为它无法调用需要 &mut T 的方法（如 mem ：：交换）。

It is worth reiterating that Pin<P> does not change the fact that a Rust compiler considers all types movable. mem::swap remains callable for any T. Instead, Pin<P> prevents certain values (pointed to by pointers wrapped in Pin<P>) from being moved by making it impossible to call methods that require &mut T on them (like mem::swap).

Pin
可用于包装任何指针类型 P，因此它与 Deref 和 DerefMut 交互。 Pin
其中 P: Deref 应被视为指向固定 P::Target 的“P 样式指针”——因此，Pin<>> 是指向固定 T 的拥有指针，而 Pin <>> 是一个指向固定 T 的引用计数指针。为了正确起见，Pin
依赖于 Deref 和 DerefMut 的实现，不会移出它们的 self 参数，而只会返回一个指向 pinned 的指针 在固定指针上调用它们时的数据。

Pin<P> can be used to wrap any pointer type P, and as such it interacts with Deref and DerefMut. A Pin<P> where P: Deref should be considered as a “P-style pointer” to a pinned P::Target – so, a Pin<Box<T>> is an owned pointer to a pinned T, and a Pin<Rc<T>> is a reference-counted pointer to a pinned T. For correctness, Pin<P> relies on the implementations of Deref and DerefMut not to move out of their self parameter, and only ever to return a pointer to pinned data when they are called on a pinned pointer.

Unpin

许多类型总是可以自由移动，即使在固定时也是如此，因为它们不依赖于稳定的地址。 这包括所有基本类型（如 bool、i32 和引用）以及仅由这些类型组成的类型。 不关心固定的类型实现了 Unpin 自动Trait，它取消了 Pin
的效果。 对于 T：Unpin、Pin<>> 和 Box 功能相同，Pin<&mut T> 和 &mut T 也是如此。

Many types are always freely movable, even when pinned, because they do not rely on having a stable address. This includes all the basic types (like bool, i32, and references) as well as types consisting solely of these types. Types that do not care about pinning implement the Unpin auto-trait, which cancels the effect of Pin<P>. For T: Unpin, Pin<Box<T>> and Box<T> function identically, as do Pin<&mut T> and &mut T.

请注意，pinning 和 Unpin 只影响指向的类型 P::Target，而不影响封装在 Pin
中的指针类型 P 本身。 例如， Box 是否为 Unpin 对 Pin<>> 的行为没有影响（这里， T 是指向的类型）。

Note that pinning and Unpin only affect the pointed-to type P::Target, not the pointer type P itself that got wrapped in Pin<P>. For example, whether or not Box<T> is Unpin has no effect on the behavior of Pin<Box<T>> (here, T is the pointed-to type).

Example: self-referential struct

在我们更详细地解释与 Pin
相关的保证和选择之前，我们将讨论一些如何使用它的示例。 随意跳到理论讨论继续的地方。

Before we go into more details to explain the guarantees and choices associated with Pin<P>, we discuss some examples for how it might be used. Feel free to skip to where the theoretical discussion continues.

use std::pin::Pin;
use std::marker::PhantomPinned;
use std::ptr::NonNull;

// This is a self-referential struct because the slice field points to the data field.
// We cannot inform the compiler about that with a normal reference,
// as this pattern cannot be described with the usual borrowing rules.
// Instead we use a raw pointer, though one which is known not to be null,
// as we know it's pointing at the string.
struct Unmovable {
    data: String,
    slice: NonNull<String>,
    _pin: PhantomPinned,
}

impl Unmovable {
    // To ensure the data doesn't move when the function returns,
    // we place it in the heap where it will stay for the lifetime of the object,
    // and the only way to access it would be through a pointer to it.
    fn new(data: String) -> Pin<Box<Self>> {
        let res = Unmovable {
            data,
            // we only create the pointer once the data is in place
            // otherwise it will have already moved before we even started
            slice: NonNull::dangling(),
            _pin: PhantomPinned,
        };
        let mut boxed = Box::pin(res);

        let slice = NonNull::from(&boxed.data);
        // we know this is safe because modifying a field doesn't move the whole struct
        unsafe {
            let mut_ref: Pin<&mut Self> = Pin::as_mut(&mut boxed);
            Pin::get_unchecked_mut(mut_ref).slice = slice;
        }
        boxed
    }
}

let unmoved = Unmovable::new("hello".to_string());
// The pointer should point to the correct location,
// so long as the struct hasn't moved.
// Meanwhile, we are free to move the pointer around.
let mut still_unmoved = unmoved;
assert_eq!(still_unmoved.slice, NonNull::from(&still_unmoved.data));

// Since our type doesn't implement Unpin, this will fail to compile:
// let mut new_unmoved = Unmovable::new("world".to_string());
// std::mem::swap(&mut *still_unmoved, &mut *new_unmoved);
Run

Example: intrusive doubly-linked list

在侵入式双向链表中，集合实际上并不为元素本身分配内存。 分配由客户端控制，元素可以存在于比集合更短的堆栈帧上。

In an intrusive doubly-linked list, the collection does not actually allocate the memory for the elements itself. Allocation is controlled by the clients, and elements can live on a stack frame that lives shorter than the collection does.

为了完成这项工作，每个元素都有指向列表中其前任和后继的指针。 元素只能在固定时添加，因为移动元素会使指针无效。 此外，链表元素的 Drop 实现将修补其前任和后继的指针，以将其从链表中删除。

To make this work, every element has pointers to its predecessor and successor in the list. Elements can only be added when they are pinned, because moving the elements around would invalidate the pointers. Moreover, the Drop implementation of a linked list element will patch the pointers of its predecessor and successor to remove itself from the list.

至关重要的是，我们必须能够依赖 drop 被调用。 如果一个元素可以在不调用 drop 的情况下被释放或以其他方式无效，则从其相邻元素指向它的指针将变为无效，这将破坏数据结构。

Crucially, we have to be able to rely on drop being called. If an element could be deallocated or otherwise invalidated without calling drop, the pointers into it from its neighboring elements would become invalid, which would break the data structure.

因此，固定还带有与丢弃相关的保证。

Therefore, pinning also comes with a drop-related guarantee.

Drop guarantee

固定的目的是为了能够依赖一些数据在内存中的放置。 为了完成这项工作，不仅要限制移动数据； 释放、重新利用或以其他方式使用于存储数据的内存无效也受到限制。 具体来说，对于固定数据，您必须保持其内存从固定到调用 drop 的那一刻不会失效或重新利用的不变量。 只有一旦 drop 返回或panic，内存可能会被重用。

The purpose of pinning is to be able to rely on the placement of some data in memory. To make this work, not just moving the data is restricted; deallocating, repurposing, or otherwise invalidating the memory used to store the data is restricted, too. Concretely, for pinned data you have to maintain the invariant that its memory will not get invalidated or repurposed from the moment it gets pinned until when drop is called. Only once drop returns or panics, the memory may be reused.

内存可以通过释放来“失效”，也可以通过将 Some(v) 替换为 None 或调用 Vec::set_len 来“杀死”Vector中的一些元素。 它可以通过使用 ptr::write 来重新利用它来覆盖它，而无需先调用析构函数。 在不调用 drop 的情况下，这些都不允许用于固定数据。

Memory can be “invalidated” by deallocation, but also by replacing a Some(v) by None, or calling Vec::set_len to “kill” some elements off of a vector. It can be repurposed by using ptr::write to overwrite it without calling the destructor first. None of this is allowed for pinned data without calling drop.

这正是上一节中的侵入式链表需要正确运行的保证。

This is exactly the kind of guarantee that the intrusive linked list from the previous section needs to function correctly.

请注意，此保证并不意味着内存不会泄漏！ 永远不要在固定元素上调用 drop 仍然完全可以（例如，您仍然可以在 Pin<>> 上调用 mem::forget）。 在双向链表的示例中，该元素将仅保留在列表中。 但是，您不能在不调用 drop 的情况下释放或重用存储。

Notice that this guarantee does not mean that memory does not leak! It is still completely okay to not ever call drop on a pinned element (e.g., you can still call mem::forget on a Pin<Box<T>>). In the example of the doubly-linked list, that element would just stay in the list. However you must not free or reuse the storage without calling drop.

Drop implementation

如果你的类型使用 pinning（比如上面的两个例子），你在实现 Drop 时要小心。 drop 函数采用 &mut self，但即使您的类型之前已固定，也会调用它！ 就好像编译器自动调用了 Pin::get_unchecked_mut。

If your type uses pinning (such as the two examples above), you have to be careful when implementing Drop. The drop function takes &mut self, but this is called even if your type was previously pinned! It is as if the compiler automatically called Pin::get_unchecked_mut.

这永远不会导致安全代码出现问题，因为实现依赖于固定的类型需要不安全的代码，但请注意决定在您的类型中使用固定（例如，通过在 Pin<&Self> 或 Pin<&mut 上实现一些操作 Self>) 也会对 Dropimplementation 产生影响：如果您的类型的元素可能已固定，则必须将 Drop 视为隐式采用 Pin<&mut Self>。

This can never cause a problem in safe code because implementing a type that relies on pinning requires unsafe code, but be aware that deciding to make use of pinning in your type (for example by implementing some operation on Pin<&Self> or Pin<&mut Self>) has consequences for your Dropimplementation as well: if an element of your type could have been pinned, you must treat Drop as implicitly taking Pin<&mut Self>.

例如，您可以按如下方式实现 Drop：

For example, you could implement Drop as follows:

impl Drop for Type {
    fn drop(&mut self) {
        // `new_unchecked` is okay because we know this value is never used
        // again after being dropped.
        inner_drop(unsafe { Pin::new_unchecked(self)});
        fn inner_drop(this: Pin<&mut Type>) {
            // Actual drop code goes here.
        }
    }
}
Run

函数 inner_drop 具有 drop 应该具有的类型，因此可以确保您不会以与 pinning 冲突的方式意外使用 self/this。

The function inner_drop has the type that drop should have, so this makes sure that you do not accidentally use self/this in a way that is in conflict with pinning.

此外，如果您的类型是#[repr(packed)]，编译器将自动移动字段以便能够删除它们。 它甚至可以对恰好对齐的字段执行此操作。 因此，您不能将 pinning 与 #[repr(packed)] 类型一起使用。

Moreover, if your type is #[repr(packed)], the compiler will automatically move fields around to be able to drop them. It might even do that for fields that happen to be sufficiently aligned. As a consequence, you cannot use pinning with a #[repr(packed)] type.

Projections and Structural Pinning

在使用固定结构时，问题出现了如何以仅采用 Pin<&mut Struct> 的方法访问该结构的字段。 通常的方法是编写将 Pin<&mut Struct> 转换为对字段的引用的辅助方法（所谓的投影），但是该引用应该具有什么类型？ 是 Pin<&mut Field> 还是 &mut 字段？ 枚举的字段也会出现同样的问题，在考虑容器/包装器类型（例如 Vec、Box 或 RefCell 时）也会出现同样的问题。 （这个问题既适用于可变引用，也适用于共享引用，我们在这里只使用更常见的可变引用来说明。）

When working with pinned structs, the question arises how one can access the fields of that struct in a method that takes just Pin<&mut Struct>. The usual approach is to write helper methods (so called projections) that turn Pin<&mut Struct> into a reference to the field, but what type should that reference have? Is it Pin<&mut Field> or &mut Field? The same question arises with the fields of an enum, and also when considering container/wrapper types such as Vec<T>, Box<T>, or RefCell<T>. (This question applies to both mutable and shared references, we just use the more common case of mutable references here for illustration.)

事实证明，实际上是由数据结构的作者来决定特定字段的固定投影是将 Pin<&mut Struct> 变成 Pin<&mut Field> 还是 &mut Field。 但是有一些限制，最重要的限制是一致性：每个字段都可以投影到固定参考，或者将固定作为投影的一部分移除。 如果两者都是针对同一个领域进行的，那可能是不合理的！

It turns out that it is actually up to the author of the data structure to decide whether the pinned projection for a particular field turns Pin<&mut Struct> into Pin<&mut Field> or &mut Field. There are some constraints though, and the most important constraint is consistency: every field can be either projected to a pinned reference, or have pinning removed as part of the projection. If both are done for the same field, that will likely be unsound!

作为数据结构的作者，您可以决定每个字段是否固定“传播”到该字段。 传播的固定也称为“结构”，因为它遵循类型的结构。 在下面的小节中，我们描述了对任何一种选择都必须考虑的因素。

As the author of a data structure you get to decide for each field whether pinning “propagates” to this field or not. Pinning that propagates is also called “structural”, because it follows the structure of the type. In the following subsections, we describe the considerations that have to be made for either choice.

Pinning is not structural for field

固定结构的字段可能未固定，这似乎违反直觉，但这实际上是最简单的选择：如果从未创建 Pin<&mut Field>，则不会出错！ 因此，如果您决定某个字段没有结构固定，您只需确保永远不要创建对该字段的固定引用。

It may seem counter-intuitive that the field of a pinned struct might not be pinned, but that is actually the easiest choice: if a Pin<&mut Field> is never created, nothing can go wrong! So, if you decide that some field does not have structural pinning, all you have to ensure is that you never create a pinned reference to that field.

没有结构钉扎的字段可能有一个将 Pin<&mut Struct> 转换为 &mut 字段的投影方法：

Fields without structural pinning may have a projection method that turns Pin<&mut Struct> into &mut Field:

impl Struct {
    fn pin_get_field(self: Pin<&mut Self>) -> &mut Field {
        // This is okay because `field` is never considered pinned.
        unsafe { &mut self.get_unchecked_mut().field }
    }
}
Run

即使字段类型不是 Unpin，您也可以为 Struct 实现 Unpin。 当没有创建 Pin<&mut Field> 时，该类型对固定的看法是不相关的。

You may also impl Unpin for Struct even if the type of field is not Unpin. What that type thinks about pinning is not relevant when no Pin<&mut Field> is ever created.

Pinning is structural for field

另一种选择是确定字段的固定是“结构的”，这意味着如果结构被固定，那么字段也是如此。

The other option is to decide that pinning is “structural” for field, meaning that if the struct is pinned then so is the field.

这允许编写一个创建 Pin<&mut Field> 的投影，从而见证该字段被固定：

This allows writing a projection that creates a Pin<&mut Field>, thus witnessing that the field is pinned:

impl Struct {
    fn pin_get_field(self: Pin<&mut Self>) -> Pin<&mut Field> {
        // This is okay because `field` is pinned when `self` is.
        unsafe { self.map_unchecked_mut(|s| &mut s.field) }
    }
}
Run

但是，结构钉扎有一些额外的要求：

However, structural pinning comes with a few extra requirements:

如果所有结构字段都是 Unpin，则结构必须是 Unpin。 这是默认设置，但 Unpin 是一个安全特性，因此作为结构的作者，您有责任不要为 Struct 添加类似 impl Unpin 之类的内容。 （请注意，添加投影操作需要不安全的代码，因此 Unpin 是一个安全Trait这一事实并没有违反这样一个原则，即您只需要在使用 unsafe 时担心这些。）

The struct must only be Unpin if all the structural fields are Unpin. This is the default, but Unpin is a safe trait, so as the author of the struct it is your responsibility not to add something like impl<T> Unpin for Struct<T>. (Notice that adding a projection operation requires unsafe code, so the fact that Unpin is a safe trait does not break the principle that you only have to worry about any of this if you use unsafe.)

结构的析构函数不得将结构字段移出其参数。 这就是上一节中提出的确切观点：drop 接受 &mut self，但结构（以及它的字段）之前可能已被固定。 您必须保证不会在 Drop 实现中移动字段。 特别是，如前所述，这意味着您的结构不能是#[repr(packed)]。 请参阅该部分以了解如何以编译器可以帮助您不会意外破坏固定的方式编写 drop。

The destructor of the struct must not move structural fields out of its argument. This is the exact point that was raised in the previous section: drop takes &mut self, but the struct (and hence its fields) might have been pinned before. You have to guarantee that you do not move a field inside your Drop implementation. In particular, as explained previously, this means that your struct must not be #[repr(packed)]. See that section for how to write drop in a way that the compiler can help you not accidentally break pinning.

您必须确保遵守 Drop 保证：一旦您的结构被固定，包含内容的内存不会在不调用内容的析构函数的情况下被覆盖或释放。 这可能很棘手，正如 VecDeque 所见证的那样：如果析构函数之一发生panic，则 VecDeque 的析构函数可能无法对所有元素调用 drop。 这违反了 Drop 保证，因为它可能导致元素在没有调用其析构函数的情况下被释放。 （VecDeque 没有固定投影，所以这不会导致不健全。）

You must make sure that you uphold the Drop guarantee: once your struct is pinned, the memory that contains the content is not overwritten or deallocated without calling the content’s destructors. This can be tricky, as witnessed by VecDeque<T>: the destructor of VecDeque<T> can fail to call drop on all elements if one of the destructors panics. This violates the Drop guarantee, because it can lead to elements being deallocated without their destructor being called. (VecDeque<T> has no pinning projections, so this does not cause unsoundness.)

当您的类型被固定时，您不得提供任何其他可能导致数据移出结构字段的操作。例如，如果结构包含 Option 并且存在类型为 fn(Pin<&mut Struct>) -> Option 的 take-like 操作，则该操作可用于将 T 移出固定的 Struct - 这意味着对于保存此数据的字段，固定不能是结构性的。对于将数据移出固定类型的更复杂示例，想象一下如果 RefCell 有一个方法 fn get_pin_mut(self: Pin<&mut Self>) -> Pin<&mut T>。然后我们可以执行以下操作： ⓘfn exploit_ref_cell(rc: Pin<&mut RefCell>) { { let p = rc.as_mut().get_pin_mut(); } // 在这里，我们获得了对 `T` 的固定访问权限。让 rc_shr: &RefCell = rc.into_ref().get_ref();让 b = rc_shr.borrow_mut();让内容 = &mut *b; // 这里我们有 `&mut T` 到​​相同的数据。 }Run 这是灾难性的，这意味着我们可以首先固定 RefCell 的内容（使用 RefCell::get_pin_mut），然后使用我们稍后获得的可变引用移动该内容。

You must not offer any other operations that could lead to data being moved out of the structural fields when your type is pinned. For example, if the struct contains an Option<T> and there is a take-like operation with type fn(Pin<&mut Struct<T>>) -> Option<T>, that operation can be used to move a T out of a pinned Struct<T> – which means pinning cannot be structural for the field holding this data.

For a more complex example of moving data out of a pinned type, imagine if RefCell<T> had a method fn get_pin_mut(self: Pin<&mut Self>) -> Pin<&mut T>. Then we could do the following:

fn exploit_ref_cell<T>(rc: Pin<&mut RefCell<T>>) {
    { let p = rc.as_mut().get_pin_mut(); } // Here we get pinned access to the `T`.
    let rc_shr: &RefCell<T> = rc.into_ref().get_ref();
    let b = rc_shr.borrow_mut();
    let content = &mut *b; // And here we have `&mut T` to the same data.
}
Run

This is catastrophic, it means we can first pin the content of the RefCell<T> (using RefCell::get_pin_mut) and then move that content using the mutable reference we got later.

Examples

对于像 Vec 这样的类型，两种可能性（结构钉扎与否）都是有意义的。 具有结构固定的 Vec 可以具有 get_pin/get_pin_mut 方法来获取对元素的固定引用。 但是，它不能允许在固定的 Vec 上调用 pop，因为这会移动（结构固定的）内容！ 它也不允许推送，这可能会重新分配并因此也移动内容。

For a type like Vec<T>, both possibilities (structural pinning or not) make sense. A Vec<T> with structural pinning could have get_pin/get_pin_mut methods to get pinned references to elements. However, it could not allow calling pop on a pinned Vec<T> because that would move the (structurally pinned) contents! Nor could it allow push, which might reallocate and thus also move the contents.

没有结构固定的 Vec 可以 impl Unpin for Vec，因为内容永远不会固定，而且 Vec 本身也可以移动。 在这一点上，钉扎根本对Vector没有影响。

A Vec<T> without structural pinning could impl<T> Unpin for Vec<T>, because the contents are never pinned and the Vec<T> itself is fine with being moved as well. At that point pinning just has no effect on the vector at all.

在标准库中，指针类型通常没有结构固定，因此它们不提供固定投影。 这就是为什么 Box: Unpin 适用于所有 T。对指针类型这样做是有意义的，因为移动 Box 实际上不会移动 T：Box 可以自由移动（也称为 Unpin ) 即使 T 不是。 事实上，即使 Pin<>> 和 Pin<&mut T> 本身也总是 Unpin，原因相同：它们的内容（T）是固定的，但指针本身可以在不移动固定数据的情况下移动。 对于 Box 和 Pin<>>，内容是否被钉住完全与指针是否被钉住无关，意味着钉住不是结构性的。

In the standard library, pointer types generally do not have structural pinning, and thus they do not offer pinning projections. This is why Box<T>: Unpin holds for all T. It makes sense to do this for pointer types, because moving the Box<T> does not actually move the T: the Box<T> can be freely movable (aka Unpin) even if the T is not. In fact, even Pin<Box<T>> and Pin<&mut T> are always Unpin themselves, for the same reason: their contents (the T) are pinned, but the pointers themselves can be moved without moving the pinned data. For both Box<T> and Pin<Box<T>>, whether the content is pinned is entirely independent of whether the pointer is pinned, meaning pinning is not structural.

在实现 Future 组合器时，您通常需要对嵌套的 futures 进行结构固定，因为您需要获取对它们的固定引用才能调用 poll。 但是，如果您的组合器包含不需要固定的任何其他数据，您可以使这些字段不是结构化的，因此即使您只有 Pin<&mut Self> （例如在您自己的 poll 实现中），也可以使用可变引用自由访问它们 ）。

When implementing a Future combinator, you will usually need structural pinning for the nested futures, as you need to get pinned references to them to call poll. But if your combinator contains any other data that does not need to be pinned, you can make those fields not structural and hence freely access them with a mutable reference even when you just have Pin<&mut Self> (such as in your own poll implementation).

Structs

Pin

A pinned pointer.