Module core::preludeExperimental

The core prelude

This module is intended for users of libcore which do not link to libstd as well. This module is not imported by default, but using the entire contents of this module will provide all of the useful traits and types in libcore that one would expect from the standard library as well.

There is no method to automatically inject this prelude, and this prelude is a subset of the standard library's prelude.

Example
extern crate core; fn main() { #![feature(globs)] use core::prelude::*; }

#![feature(globs)]

use core::prelude::*;

Reexports

pub use kinds::{Copy, Send, Sized, Sync};

pub use ops::{Add, Sub, Mul, Div, Rem, Neg, Not};

pub use ops::{BitAnd, BitOr, BitXor};

pub use ops::{Drop, Deref, DerefMut};

pub use ops::{Shl, Shr};

pub use ops::{Index, IndexMut};

pub use ops::{Slice, SliceMut};

pub use ops::{Fn, FnMut, FnOnce};

pub use iter::range;

pub use mem::drop;

pub use str::from_str;

pub use char::Char;

pub use clone::Clone;

pub use cmp::{PartialEq, PartialOrd, Eq, Ord};

pub use cmp::{Ordering, Equiv};

pub use cmp::Ordering::{Less, Equal, Greater};

pub use iter::{FromIterator, Extend, IteratorExt};

pub use iter::{Iterator, DoubleEndedIterator, DoubleEndedIteratorExt, RandomAccessIterator};

pub use iter::{IteratorCloneExt, CloneIteratorExt};

pub use iter::{IteratorOrdExt, MutableDoubleEndedIterator, ExactSizeIterator};

pub use num::{ToPrimitive, FromPrimitive};

pub use option::Option;

pub use option::Option::{Some, None};

pub use ptr::RawPtr;

pub use result::Result;

pub use result::Result::{Ok, Err};

pub use str::{Str, StrExt};

pub use slice::{PartialEqSliceExt, OrdSliceExt};

pub use slice::{AsSlice, SliceExt};

Traits

Tuple1

Tuple10

Tuple11

Tuple12

Tuple2

Tuple3

Tuple4

Tuple5

Tuple6

Tuple7

Tuple8

Tuple9