Module core::prelude::rust_2024

🔬This is a nightly-only experimental API. (prelude_2024 #121042)

The 2024 version of the core prelude.

See the module-level documentation for more.

Re-exports

pub use crate::iter::FromIterator;

pub use crate::convert::TryFrom;

pub use crate::convert::TryInto;

pub use crate::marker::Copy;

pub use crate::marker::Copy;

pub use crate::marker::Send;

pub use crate::marker::Sized;

pub use crate::marker::Sync;

pub use crate::marker::Unpin;

pub use crate::ops::Drop;

pub use crate::ops::Fn;

pub use crate::ops::FnMut;

pub use crate::ops::FnOnce;

pub use crate::mem::drop;

pub use crate::mem::align_of;

pub use crate::mem::align_of_val;

pub use crate::mem::size_of;

pub use crate::mem::size_of_val;

pub use crate::clone::Clone;

pub use crate::clone::Clone;

pub use crate::cmp::Eq;

pub use crate::cmp::Eq;

pub use crate::cmp::Ord;

pub use crate::cmp::Ord;

pub use crate::cmp::PartialEq;

pub use crate::cmp::PartialEq;

pub use crate::cmp::PartialOrd;

pub use crate::cmp::PartialOrd;

pub use crate::convert::AsMut;

pub use crate::convert::AsRef;

pub use crate::convert::From;

pub use crate::convert::Into;

pub use crate::default::Default;

pub use crate::default::Default;

pub use crate::iter::DoubleEndedIterator;

pub use crate::iter::ExactSizeIterator;

pub use crate::iter::Extend;

pub use crate::iter::IntoIterator;

pub use crate::iter::Iterator;

pub use crate::option::Option;

pub use crate::option::Option::None;

pub use crate::option::Option::Some;

pub use crate::result::Result;

pub use crate::result::Result::Err;

pub use crate::result::Result::Ok;

pub use crate::fmt::macros::Debug;

pub use crate::hash::macros::Hash;

pub use crate::assert;

pub use crate::cfg;

pub use crate::column;

pub use crate::compile_error;

pub use crate::concat;

pub use crate::concat_idents;

Experimental

pub use crate::env;

pub use crate::file;

pub use crate::format_args;

pub use crate::format_args_nl;

Experimental

pub use crate::include;

pub use crate::include_bytes;

pub use crate::include_str;

pub use crate::line;

pub use crate::log_syntax;

Experimental

pub use crate::module_path;

pub use crate::option_env;

pub use crate::stringify;

pub use crate::trace_macros;

Experimental

pub use crate::future::Future;

pub use crate::future::IntoFuture;

pub use crate::concat_bytes;

Experimental

Macros

derefExperimental

Unstable placeholder for deref patterns.

type_ascribeExperimental

Unstable placeholder for type ascription.

Attribute Macros

derive

Attribute macro used to apply derive macros.

global_allocator

Attribute macro applied to a static to register it as a global allocator.

test

Attribute macro applied to a function to turn it into a unit test.

alloc_error_handlerExperimental

Attribute macro applied to a function to register it as a handler for allocation failure.

benchExperimental

Attribute macro applied to a function to turn it into a benchmark test.

cfg_accessibleExperimental

Keeps the item it’s applied to if the passed path is accessible, and removes it otherwise.

cfg_evalExperimental

Expands all #[cfg] and #[cfg_attr] attributes in the code fragment it’s applied to.

derive_constExperimental

Attribute macro used to apply derive macros for implementing traits in a const context.

test_caseExperimental

An implementation detail of the #[test] and #[bench] macros.