Module core::prelude::v1

The core prelude

This module is intended for users of libcore which do not link to libstd as well. This module is imported by default when #![no_std] is used in the same manner as the standard library's prelude.

Reexports

pub use marker::{Copy, Send, Sized, Sync};

pub use ops::{Drop, Fn, FnMut, FnOnce};

pub use mem::drop;

pub use clone::Clone;

pub use cmp::{PartialEq, PartialOrd, Eq, Ord};

pub use convert::{AsRef, AsMut, Into, From};

pub use default::Default;

pub use iter::{Iterator, Extend, IntoIterator};

pub use iter::{DoubleEndedIterator, ExactSizeIterator};

pub use option::Option::{self, Some, None};

pub use result::Result::{self, Ok, Err};

pub use slice::SliceExt;

pub use str::StrExt;

pub use char::CharExt;