Function core::ptr::read_and_zero

pub unsafe fn read_and_zero<T>(dest: *mut T) -> T

Unstable
: may play a larger role in std::ptr future extensions

Reads the value from src and nulls it out without dropping it.

Safety

This is unsafe for the same reasons that read is unsafe.