Function core::ptr::write_volatile

pub unsafe fn write_volatile<T>(dst: *mut T, src: T)

使用给定值执行内存位置的易失性写入，而不读取或删除旧值。

Performs a volatile write of a memory location with the given value without reading or dropping the old value.

易失性操作旨在作用于 I/O 内存，并保证编译器不会在其他易失性操作中忽略或重新排序。

Volatile operations are intended to act on I/O memory, and are guaranteed to not be elided or reordered by the compiler across other volatile operations.

write_volatile 不会删除 dst 的内容。 这是安全的，但它可能会泄漏分配或资源，因此应注意不要覆盖应删除的对象。

write_volatile does not drop the contents of dst. This is safe, but it could leak allocations or resources, so care should be taken not to overwrite an object that should be dropped.

此外，它不会删除 src。 从语义上讲，src 被移动到 dst 指向的位置。

Additionally, it does not drop src. Semantically, src is moved into the location pointed to by dst.

Notes

Rust 目前没有严格和正式定义的内存模型，因此“易失性”在这里的精确语义会随着时间而改变。 话虽如此，语义最终几乎总是与 C11 对 volatile 的定义非常相似。

Rust does not currently have a rigorously and formally defined memory model, so the precise semantics of what “volatile” means here is subject to change over time. That being said, the semantics will almost always end up pretty similar to C11’s definition of volatile.

编译器不应更改易失性内存操作的相对顺序或数量。 但是，对零大小类型的易失性内存操作（例如，如果将零大小类型传递给 write_volatile）是 noops 并且可以被忽略。

The compiler shouldn’t change the relative order or number of volatile memory operations. However, volatile memory operations on zero-sized types (e.g., if a zero-sized type is passed to write_volatile) are noops and may be ignored.

Safety

如果违反以下任何条件，则行为未定义：

Behavior is undefined if any of the following conditions are violated:

dst 必须对写入有效。

dst must be valid for writes.

dst 必须正确对齐。

dst must be properly aligned.

请注意，即使 T 的大小为 0，指针也必须非空且正确对齐。

Note that even if T has size 0, the pointer must be non-null and properly aligned.

就像在 C 中一样，一个操作是否是 volatile 与涉及来自多个线程的并发访问的问题没有任何关系。 在这方面，易失性访问的行为与非原子访问完全相同。 特别是， write_volatile 与同一位置上的任何其他操作（读取或写入）之间的竞争是未定义的行为。

Just like in C, whether an operation is volatile has no bearing whatsoever on questions involving concurrent access from multiple threads. Volatile accesses behave exactly like non-atomic accesses in that regard. In particular, a race between a write_volatile and any other operation (reading or writing) on the same location is undefined behavior.

Examples

基本用法：

Basic usage:

let mut x = 0;
let y = &mut x as *mut i32;
let z = 12;

unsafe {
    std::ptr::write_volatile(y, z);
    assert_eq!(std::ptr::read_volatile(y), 12);
}
Run