Struct core::ptr::Shared

pub struct Shared<T: ?Sized> {
    // some fields omitted
}

Unstable (shared #27730)
: needs an RFC to flesh out design

A wrapper around a raw *mut T that indicates that the possessor of this wrapper has shared ownership of the referent. Useful for building abstractions like Rc<T> or Arc<T>, which internally use raw pointers to manage the memory that they own.

Methods

impl<T: ?Sized> Shared<T>

unsafe fn new(ptr: *mut T) -> Self

Unstable (shared #27730)

Creates a new Shared.