Struct core::ptr::Unique

pub struct Unique<T: ?Sized> {
    // some fields omitted
}

Unstable (unique #27730)
: needs an RFC to flesh out design

A wrapper around a raw *mut T that indicates that the possessor of this wrapper owns the referent. This in turn implies that the Unique<T> is Send/Sync if T is Send/Sync, unlike a raw *mut T (which conveys no particular ownership semantics). It also implies that the referent of the pointer should not be modified without a unique path to the Unique reference. Useful for building abstractions like Vec<T> or Box<T>, which internally use raw pointers to manage the memory that they own.

Methods

impl<T: ?Sized> Unique<T>

unsafe fn new(ptr: *mut T) -> Unique<T>

Unstable (unique #27730)

Creates a new Unique.

unsafe fn get(&self) -> &T

Unstable (unique #27730)

Dereferences the content.

unsafe fn get_mut(&mut self) -> &mut T

Unstable (unique #27730)

Mutably dereferences the content.