Module core::simd

Unstable
: needs an RFC to flesh out the design

SIMD vectors.

These types can be used for accessing basic SIMD operations. Each of them implements the standard arithmetic operator traits (Add, Sub, Mul, Div, Rem, Shl, Shr) through compiler magic, rather than explicitly. Currently comparison operators are not implemented. To use SSE3+, you must enable the features, like -C target-feature=sse3,sse4.1,sse4.2, or a more specific target-cpu. No other SIMD intrinsics or high-level wrappers are provided beyond this module.
#![feature(core_simd)] extern crate core; fn main() { use std::simd::f32x4; let a = f32x4(40.0, 41.0, 42.0, 43.0); let b = f32x4(1.0, 1.1, 3.4, 9.8); println!("{:?}", a + b); }

fn main() {
    use std::simd::f32x4;
    let a = f32x4(40.0, 41.0, 42.0, 43.0);
    let b = f32x4(1.0, 1.1, 3.4, 9.8);
    println!("{:?}", a + b);
}

Stability Note

These are all experimental. The interface may change entirely, without warning.

Structs

f32x4 	 [Unstable]

f64x2 	 [Unstable]

i16x8 	 [Unstable]

i32x4 	 [Unstable]

i64x2 	 [Unstable]

i8x16 	 [Unstable]

u16x8 	 [Unstable]

u32x4 	 [Unstable]

u64x2 	 [Unstable]

u8x16 	 [Unstable]