Struct core::simd::i16x16

#[repr(simd)]

pub struct i16x16(_, _, _, _, _, _, _, _, _, _, _, _, _, _, _, _);

🔬 This is a nightly-only experimental API. (stdsimd #48556)

A 256-bit vector with 16 i16 lanes.

Methods

impl i16x16

pub const fn new(
x0: i16,
x1: i16,
x2: i16,
x3: i16,
x4: i16,
x5: i16,
x6: i16,
x7: i16,
x8: i16,
x9: i16,
x10: i16,
x11: i16,
x12: i16,
x13: i16,
x14: i16,
x15: i16
) -> Self

🔬 This is a nightly-only experimental API. (stdsimd #48556)

Creates a new instance with each vector elements initialized with the provided values.

pub const fn lanes() -> usize

🔬 This is a nightly-only experimental API. (stdsimd #48556)

Returns the number of vector lanes.

pub const fn splat(value: i16) -> Self

🔬 This is a nightly-only experimental API. (stdsimd #48556)

Constructs a new instance with each element initialized to value.

pub fn extract(self, index: usize) -> i16

🔬 This is a nightly-only experimental API. (stdsimd #48556)

Extracts the value at index.

Panics

If index >= Self::lanes().

pub unsafe fn extract_unchecked(self, index: usize) -> i16

🔬 This is a nightly-only experimental API. (stdsimd #48556)

Extracts the value at index.

If index >= Self::lanes() the behavior is undefined.

pub fn replace(self, index: usize, new_value: i16) -> Self

🔬 This is a nightly-only experimental API. (stdsimd #48556)

Returns a new vector where the value at index is replaced by new_value.

Panics

If index >= Self::lanes().

pub unsafe fn replace_unchecked(self, index: usize, new_value: i16) -> Self

🔬 This is a nightly-only experimental API. (stdsimd #48556)

Returns a new vector where the value at index is replaced by new_value.

Panics

If index >= Self::lanes().

impl i16x16

pub fn store_aligned(self, slice: &mut [i16])

🔬 This is a nightly-only experimental API. (stdsimd #48556)

Writes the values of the vector to the slice.

Panics

If slice.len() < Self::lanes() or &slice[0] is not aligned to an align_of::<Self>() boundary.

pub fn store_unaligned(self, slice: &mut [i16])

🔬 This is a nightly-only experimental API. (stdsimd #48556)

Writes the values of the vector to the slice.

Panics

If slice.len() < Self::lanes().

pub unsafe fn store_aligned_unchecked(self, slice: &mut [i16])

🔬 This is a nightly-only experimental API. (stdsimd #48556)

Writes the values of the vector to the slice.

Precondition

If slice.len() < Self::lanes() or &slice[0] is not aligned to an align_of::<Self>() boundary, the behavior is undefined.

pub unsafe fn store_unaligned_unchecked(self, slice: &mut [i16])

🔬 This is a nightly-only experimental API. (stdsimd #48556)

Writes the values of the vector to the slice.

Precondition

If slice.len() < Self::lanes() the behavior is undefined.

pub fn load_aligned(slice: &[i16]) -> Self

🔬 This is a nightly-only experimental API. (stdsimd #48556)

Instantiates a new vector with the values of the slice.

Panics

If slice.len() < Self::lanes() or &slice[0] is not aligned to an align_of::<Self>() boundary.

pub fn load_unaligned(slice: &[i16]) -> Self

🔬 This is a nightly-only experimental API. (stdsimd #48556)

Instantiates a new vector with the values of the slice.

Panics

If slice.len() < Self::lanes().

pub unsafe fn load_aligned_unchecked(slice: &[i16]) -> Self

🔬 This is a nightly-only experimental API. (stdsimd #48556)

Instantiates a new vector with the values of the slice.

Precondition

If slice.len() < Self::lanes() or &slice[0] is not aligned to an align_of::<Self>() boundary, the behavior is undefined.

pub unsafe fn load_unaligned_unchecked(slice: &[i16]) -> Self

🔬 This is a nightly-only experimental API. (stdsimd #48556)

Instantiates a new vector with the values of the slice.

Precondition

If slice.len() < Self::lanes() the behavior is undefined.

impl i16x16

pub fn eq(self, other: i16x16) -> b8x16

🔬 This is a nightly-only experimental API. (stdsimd #48556)

Lane-wise equality comparison.

pub fn ne(self, other: i16x16) -> b8x16

🔬 This is a nightly-only experimental API. (stdsimd #48556)

Lane-wise inequality comparison.

pub fn lt(self, other: i16x16) -> b8x16

🔬 This is a nightly-only experimental API. (stdsimd #48556)

Lane-wise less-than comparison.

pub fn le(self, other: i16x16) -> b8x16

🔬 This is a nightly-only experimental API. (stdsimd #48556)

Lane-wise less-than-or-equals comparison.

pub fn gt(self, other: i16x16) -> b8x16

🔬 This is a nightly-only experimental API. (stdsimd #48556)

Lane-wise greater-than comparison.

pub fn ge(self, other: i16x16) -> b8x16

🔬 This is a nightly-only experimental API. (stdsimd #48556)

Lane-wise greater-than-or-equals comparison.

impl i16x16

pub fn sum(self) -> i16

🔬 This is a nightly-only experimental API. (stdsimd #48556)

Lane-wise addition of the vector elements.

pub fn product(self) -> i16

🔬 This is a nightly-only experimental API. (stdsimd #48556)

Lane-wise multiplication of the vector elements.

impl i16x16

pub fn max(self) -> i16

🔬 This is a nightly-only experimental API. (stdsimd #48556)

Largest vector value.

pub fn min(self) -> i16

🔬 This is a nightly-only experimental API. (stdsimd #48556)

Smallest vector value.

impl i16x16

pub fn and(self) -> i16

🔬 This is a nightly-only experimental API. (stdsimd #48556)

Lane-wise bitwise and of the vector elements.

pub fn or(self) -> i16

🔬 This is a nightly-only experimental API. (stdsimd #48556)

Lane-wise bitwise or of the vector elements.

pub fn xor(self) -> i16

🔬 This is a nightly-only experimental API. (stdsimd #48556)

Lane-wise bitwise xor of the vector elements.