Module core::sliceStable

Slice management and manipulation

For more details std::slice.

Modules

bytes
Operations on [u8].

raw
Unsafe operations

Structs

Chunks
An iterator over a slice in (non-overlapping) chunks (size elements at a time).

Iter
Immutable slice iterator

IterMut
Mutable slice iterator.

MutChunks
An iterator over a slice in (non-overlapping) mutable chunks (size elements at a time). When the slice len is not evenly divided by the chunk size, the last slice of the iteration will be the remainder.

MutSplits
An iterator over the subslices of the vector which are separated by elements that match pred.

Splits
An iterator over subslices separated by elements that match a predicate function.

SplitsN
An iterator over subslices separated by elements that match a predicate function, splitting at most a fixed number of times.

Windows
An iterator over overlapping subslices of length size.

Enums

BinarySearchResult
The result of calling binary_search.

Traits

AsSlice
Data that is viewable as a slice.

CloneSliceExt
Extension methods for slices on Clone elements

ImmutableIntSlice
Extension methods for immutable slices containing integers.

MutableIntSlice
Extension methods for mutable slices containing integers.

OrdSliceExt
Extension methods for slices containing Ord elements.

PartialEqSliceExt
Extension methods for slices containing PartialEq elements.

SliceExt
Extension methods for slices.

Functions

from_raw_buf
Forms a slice from a pointer and a length.

from_raw_mut_buf
Performs the same functionality as from_raw_buf, except that a mutable slice is returned.

mut_ref_slice
Converts a pointer to A into a slice of length 1 (without copying).

ref_slice
Converts a pointer to A into a slice of length 1 (without copying).