Function core::slice::range

pub fn range<R>(range: R, bounds: RangeTo<usize>) -> Range<usize> where
    R: RangeBounds<usize>, 

🔬 This is a nightly-only experimental API. (slice_range #76393)

执行范围的边界检查。

Performs bounds-checking of a range.

此方法类似于切片的 Index::index，但它返回与 range 等效的 Range。 您可以使用此方法将任何范围转换为开始值和结束值。

This method is similar to Index::index for slices, but it returns a Range equivalent to range. You can use this method to turn any range into start and end values.

bounds 是用于边界检查的切片范围。 它应该是一个以切片长度结束的 RangeTo 范围。

bounds is the range of the slice to use for bounds-checking. It should be a RangeTo range that ends at the length of the slice.

对于具有给定范围的切片，返回的 Range 可以安全地传递给 slice::get_unchecked 和 slice::get_unchecked_mut。

The returned Range is safe to pass to slice::get_unchecked and slice::get_unchecked_mut for slices with the given range.

Panics

如果范围超出范围，会感到panic。

Panics if range would be out of bounds.

Examples

#![feature(slice_range)]

use std::slice;

let v = [10, 40, 30];
assert_eq!(1..2, slice::range(1..2, ..v.len()));
assert_eq!(0..2, slice::range(..2, ..v.len()));
assert_eq!(1..3, slice::range(1.., ..v.len()));
Run

当 Index::index 发生panic时发生panic：

Panics when Index::index would panic:

#![feature(slice_range)]

use std::slice;

slice::range(2..1, ..3);
Run

#![feature(slice_range)]

use std::slice;

slice::range(1..4, ..3);
Run

#![feature(slice_range)]

use std::slice;

slice::range(1..=usize::MAX, ..3);
Run