Module core::slice::iter

Definitions of a bunch of iterators for [T].

Modules

macros 🔒

Macros used by iterators of slice.

Macros

next_back_unchecked 🔒

next_back_unchecked 🔒

next_unchecked 🔒

next_unchecked 🔒

Structs

ArrayChunksExperimental

An iterator over a slice in (non-overlapping) chunks (N elements at a time), starting at the beginning of the slice.

ArrayChunksMutExperimental

An iterator over a slice in (non-overlapping) mutable chunks (N elements at a time), starting at the beginning of the slice.

ArrayWindowsExperimental

A windowed iterator over a slice in overlapping chunks (N elements at a time), starting at the beginning of the slice

GroupByExperimental

An iterator over slice in (non-overlapping) chunks separated by a predicate.

GroupByMutExperimental

An iterator over slice in (non-overlapping) mutable chunks separated by a predicate.

Chunks

An iterator over a slice in (non-overlapping) chunks (chunk_size elements at a time), starting at the beginning of the slice.

ChunksExact

An iterator over a slice in (non-overlapping) chunks (chunk_size elements at a time), starting at the beginning of the slice.

ChunksExactMut

An iterator over a slice in (non-overlapping) mutable chunks (chunk_size elements at a time), starting at the beginning of the slice.

ChunksMut

An iterator over a slice in (non-overlapping) mutable chunks (chunk_size elements at a time), starting at the beginning of the slice.

GenericSplitN 🔒

An private iterator over subslices separated by elements that match a predicate function, splitting at most a fixed number of times.

Iter

Immutable slice iterator

IterMut

Mutable slice iterator.

RChunks

An iterator over a slice in (non-overlapping) chunks (chunk_size elements at a time), starting at the end of the slice.

RChunksExact

An iterator over a slice in (non-overlapping) chunks (chunk_size elements at a time), starting at the end of the slice.

RChunksExactMut

An iterator over a slice in (non-overlapping) mutable chunks (chunk_size elements at a time), starting at the end of the slice.

RChunksMut

An iterator over a slice in (non-overlapping) mutable chunks (chunk_size elements at a time), starting at the end of the slice.

RSplit

An iterator over subslices separated by elements that match a predicate function, starting from the end of the slice.

RSplitMut

An iterator over the subslices of the vector which are separated by elements that match pred, starting from the end of the slice.

RSplitN

An iterator over subslices separated by elements that match a predicate function, limited to a given number of splits, starting from the end of the slice.

RSplitNMut

An iterator over subslices separated by elements that match a predicate function, limited to a given number of splits, starting from the end of the slice.

Split

An iterator over subslices separated by elements that match a predicate function.

SplitInclusive

An iterator over subslices separated by elements that match a predicate function. Unlike Split, it contains the matched part as a terminator of the subslice.

SplitInclusiveMut

An iterator over the mutable subslices of the vector which are separated by elements that match pred. Unlike SplitMut, it contains the matched parts in the ends of the subslices.

SplitMut

An iterator over the mutable subslices of the vector which are separated by elements that match pred.

SplitN

An iterator over subslices separated by elements that match a predicate function, limited to a given number of splits.

SplitNMut

An iterator over subslices separated by elements that match a predicate function, limited to a given number of splits.

Windows

An iterator over overlapping subslices of length size.

Traits

SplitIter 🔒

An internal abstraction over the splitting iterators, so that splitn, splitn_mut etc can be implemented once.