Trait core::slice::ImmutableSliceUnstable

pub trait ImmutableSlice<'a, T> {
    fn slice(&self, start: uint, end: uint) -> &'a [T];
    fn slice_from(&self, start: uint) -> &'a [T];
    fn slice_to(&self, end: uint) -> &'a [T];
    fn split_at(&self, mid: uint) -> (&'a [T], &'a [T]);
    fn iter(self) -> Items<'a, T>;
    fn split(self, pred: |&T| -> bool) -> Splits<'a, T>;
    fn splitn(self, n: uint, pred: |&T| -> bool) -> SplitsN<'a, T>;
    fn rsplitn(self, n: uint, pred: |&T| -> bool) -> SplitsN<'a, T>;
    fn windows(self, size: uint) -> Windows<'a, T>;
    fn chunks(self, size: uint) -> Chunks<'a, T>;
    fn get(&self, index: uint) -> Option<&'a T>;
    fn head(&self) -> Option<&'a T>;
    fn tail(&self) -> &'a [T];
    fn tailn(&self, n: uint) -> &'a [T];
    fn init(&self) -> &'a [T];
    fn initn(&self, n: uint) -> &'a [T];
    fn last(&self) -> Option<&'a T>;
    unsafe fn unsafe_ref(self, index: uint) -> &'a T;
    unsafe fn unsafe_get(self, index: uint) -> &'a T;
    fn as_ptr(&self) -> *const T;
    fn bsearch(&self, f: |&T| -> Ordering) -> Option<uint>;
    fn binary_search(&self, f: |&T| -> Ordering) -> BinarySearchResult;
    fn shift_ref(&mut self) -> Option<&'a T>;
    fn pop_ref(&mut self) -> Option<&'a T>;
}

Extension methods for immutable slices.

Required Methods

fn slice(&self, start: uint, end: uint) -> &'a [T]

Returns a subslice spanning the interval [start, end).

Fails when the end of the new slice lies beyond the end of the original slice (i.e. when end > self.len()) or when start > end.

Slicing with start equal to end yields an empty slice.

fn slice_from(&self, start: uint) -> &'a [T]

Returns a subslice from start to the end of the slice.

Fails when start is strictly greater than the length of the original slice.

Slicing from self.len() yields an empty slice.

fn slice_to(&self, end: uint) -> &'a [T]

Returns a subslice from the start of the slice to end.

Fails when end is strictly greater than the length of the original slice.

Slicing to 0 yields an empty slice.

fn split_at(&self, mid: uint) -> (&'a [T], &'a [T])

Divides one slice into two at an index.

The first will contain all indices from [0, mid) (excluding the index mid itself) and the second will contain all indices from [mid, len) (excluding the index len itself).

Fails if mid > len.

fn iter(self) -> Items<'a, T>

Returns an iterator over the vector

fn split(self, pred: |&T| -> bool) -> Splits<'a, T>

Returns an iterator over the subslices of the vector which are separated by elements that match pred. The matched element is not contained in the subslices.

fn splitn(self, n: uint, pred: |&T| -> bool) -> SplitsN<'a, T>

Returns an iterator over the subslices of the vector which are separated by elements that match pred, limited to splitting at most n times. The matched element is not contained in the subslices.

fn rsplitn(self, n: uint, pred: |&T| -> bool) -> SplitsN<'a, T>

Returns an iterator over the subslices of the vector which are separated by elements that match pred limited to splitting at most n times. This starts at the end of the vector and works backwards. The matched element is not contained in the subslices.

fn windows(self, size: uint) -> Windows<'a, T>

Returns an iterator over all contiguous windows of length size. The windows overlap. If the vector is shorter than size, the iterator returns no values.

Failure

Fails if size is 0.

Example

Print the adjacent pairs of a vector (i.e. [1,2], [2,3], [3,4]):
fn main() { let v = &[1i, 2, 3, 4]; for win in v.windows(2) { println!("{}", win); } }

let v = &[1i, 2, 3, 4];
for win in v.windows(2) {
    println!("{}", win);
}

fn chunks(self, size: uint) -> Chunks<'a, T>

Returns an iterator over size elements of the vector at a time. The chunks do not overlap. If size does not divide the length of the vector, then the last chunk will not have length size.

Failure

Fails if size is 0.

Example

Print the vector two elements at a time (i.e. [1,2], [3,4], [5]):
fn main() { let v = &[1i, 2, 3, 4, 5]; for win in v.chunks(2) { println!("{}", win); } }

let v = &[1i, 2, 3, 4, 5];
for win in v.chunks(2) {
    println!("{}", win);
}

fn get(&self, index: uint) -> Option<&'a T>

Returns the element of a vector at the given index, or None if the index is out of bounds

fn head(&self) -> Option<&'a T>

Returns the first element of a vector, or None if it is empty

fn tail(&self) -> &'a [T]

Returns all but the first element of a vector

fn tailn(&self, n: uint) -> &'a [T]

Returns all but the first `n' elements of a vector

fn init(&self) -> &'a [T]

Returns all but the last element of a vector

fn initn(&self, n: uint) -> &'a [T]

Returns all but the last `n' elements of a vector

fn last(&self) -> Option<&'a T>

Returns the last element of a vector, or None if it is empty.

unsafe fn unsafe_ref(self, index: uint) -> &'a T

Returns a pointer to the element at the given index, without doing bounds checking.

unsafe fn unsafe_get(self, index: uint) -> &'a T

Returns a pointer to the element at the given index, without doing bounds checking.

fn as_ptr(&self) -> *const T

Returns an unsafe pointer to the vector's buffer

The caller must ensure that the vector outlives the pointer this function returns, or else it will end up pointing to garbage.

Modifying the vector may cause its buffer to be reallocated, which would also make any pointers to it invalid.

fn bsearch(&self, f: |&T| -> Ordering) -> Option<uint>

Binary search a sorted vector with a comparator function.

The comparator function should implement an order consistent with the sort order of the underlying vector, returning an order code that indicates whether its argument is Less, Equal or Greater the desired target.

Returns the index where the comparator returned Equal, or None if not found.

fn binary_search(&self, f: |&T| -> Ordering) -> BinarySearchResult

Binary search a sorted vector with a comparator function.

The comparator function should implement an order consistent with the sort order of the underlying vector, returning an order code that indicates whether its argument is Less, Equal or Greater the desired target.

If the value is found then Found is returned, containing the index of the matching element; if the value is not found then NotFound is returned, containing the index where a matching element could be inserted while maintaining sorted order.

fn shift_ref(&mut self) -> Option<&'a T>

Returns an immutable reference to the first element in this slice and adjusts the slice in place so that it no longer contains that element. O(1).

Equivalent to:
fn main() { if self.len() == 0 { return None } let head = &self[0]; *self = self.slice_from(1); Some(head) }

    if self.len() == 0 { return None }
    let head = &self[0];
    *self = self.slice_from(1);
    Some(head)

Returns None if vector is empty

fn pop_ref(&mut self) -> Option<&'a T>

Returns an immutable reference to the last element in this slice and adjusts the slice in place so that it no longer contains that element. O(1).

Equivalent to:
fn main() { if self.len() == 0 { return None; } let tail = &self[self.len() - 1]; *self = self.slice_to(self.len() - 1); Some(tail) }

    if self.len() == 0 { return None; }
    let tail = &self[self.len() - 1];
    *self = self.slice_to(self.len() - 1);
    Some(tail)

Returns None if slice is empty.