Trait core::slice::MutableCloneableSliceUnstable

pub trait MutableCloneableSlice<T> {
    fn clone_from_slice(self, &[T]) -> uint;

    fn copy_from(self, s: &[T]) -> uint { ... }
}

Trait for &[T] where T is Cloneable

Required Methods

fn clone_from_slice(self, &[T]) -> uint

Copies as many elements from src as it can into self (the shorter of self.len() and src.len()). Returns the number of elements copied.

Example
fn main() { use std::slice::MutableCloneableSlice; let mut dst = [0i, 0, 0]; let src = [1i, 2]; assert!(dst.copy_from(src) == 2); assert!(dst == [1, 2, 0]); let src2 = [3i, 4, 5, 6]; assert!(dst.copy_from(src2) == 3); assert!(dst == [3i, 4, 5]); }

use std::slice::MutableCloneableSlice;

let mut dst = [0i, 0, 0];
let src = [1i, 2];

assert!(dst.copy_from(src) == 2);
assert!(dst == [1, 2, 0]);

let src2 = [3i, 4, 5, 6];
assert!(dst.copy_from(src2) == 3);
assert!(dst == [3i, 4, 5]);

Provided Methods

fn copy_from(self, s: &[T]) -> uint

Copies as many elements from src as it can into self (the shorter of self.len() and src.len()). Returns the number of elements copied.