Module core::str::iter

Iterators for str methods.

Macros

derive_pattern_clone 🔒

This macro generates a Clone impl for string pattern API wrapper types of the form X<’a, P>

escape_types_impls 🔒

generate_pattern_iterators 🔒

This macro generates two public iterator structs wrapping a private internal one that makes use of the Pattern API.

Structs

Bytes

An iterator over the bytes of a string slice.

CharIndices

An iterator over the chars of a string slice, and their positions.

Chars

An iterator over the chars of a string slice.

EncodeUtf16

An iterator of u16 over the string encoded as UTF-16.

EscapeDebug

The return type of str::escape_debug.

EscapeDefault

The return type of str::escape_default.

EscapeUnicode

The return type of str::escape_unicode.

Lines

An iterator over the lines of a string, as string slices.

LinesAnyDeprecated

Created with the method lines_any.

MatchIndices

Created with the method match_indices.

MatchIndicesInternal 🔒

Matches

Created with the method matches.

MatchesInternal 🔒

RMatchIndices

Created with the method rmatch_indices.

RMatches

Created with the method rmatches.

RSplit

Created with the method rsplit.

RSplitN

Created with the method rsplitn.

RSplitTerminator

Created with the method rsplit_terminator.

Split

Created with the method split.

SplitAsciiWhitespace

An iterator over the non-ASCII-whitespace substrings of a string, separated by any amount of ASCII whitespace.

SplitInclusive

An iterator over the substrings of a string, terminated by a substring matching to a predicate function Unlike Split, it contains the matched part as a terminator of the subslice.

SplitInternal 🔒

SplitN

Created with the method splitn.

SplitNInternal 🔒

SplitTerminator

Created with the method split_terminator.

SplitWhitespace

An iterator over the non-whitespace substrings of a string, separated by any amount of whitespace.