Enum core::str::pattern::SearchStep

pub enum SearchStep {
    Match(usize, usize),
    Reject(usize, usize),
    Done,
}

Unstable (pattern #27721)
: API not fully fleshed out and ready to be stabilized

Result of calling Searcher::next() or ReverseSearcher::next_back().

Variants

Match
Unstable (pattern #27721)
: API not fully fleshed out and ready to be stabilized

Expresses that a match of the pattern has been found at haystack[a..b].

Reject
Unstable (pattern #27721)
: API not fully fleshed out and ready to be stabilized

Expresses that haystack[a..b] has been rejected as a possible match of the pattern.

Note that there might be more than one Reject between two Matches, there is no requirement for them to be combined into one.

Done
Unstable (pattern #27721)
: API not fully fleshed out and ready to be stabilized

Expresses that every byte of the haystack has been visted, ending the iteration.