Trait core::str::pattern::Searcher

pub unsafe trait Searcher<'a> {
    fn haystack(&self) -> &'a str;

    fn next(&mut self) -> SearchStep;

    fn next_match(&mut self) -> Option<(usize, usize)> { ... }

    fn next_reject(&mut self) -> Option<(usize, usize)> { ... }
}

ğŸ”¬ This is a nightly-only experimental API. (pattern #27721)

API not fully fleshed out and ready to be stabilized

å­—ç¬¦ä¸²æ¨¡å¼çš„æœç´¢è€…ã€‚

trait æä¾›äº†ä»å­—ç¬¦ä¸²çš„å¼€å¤´ (left) å¼€å‹æœç´¢æ¨¡å¼çš„éé‡å åŒ¹é…çš„æ–¹æ³•ã€‚

å°†é€šè¿‡ Pattern trait çš„å…³è” Searcher ç±»å‹å®ç°ã€‚

trait è¢«æ ‡è®°ä¸ºä¸å®‰å…¨ï¼Œå› ä¸º next() æ–¹æ³•è¿”å›çš„ç´¢å¼•å¿…é¡»ä½äº haystack ä¸­çš„æœ‰æ•ˆ utf8 è¾¹ç•Œä¸Šã€‚ è¿™ä½¿ trait çš„ä½¿ç”¨è€…å¯ä»¥å¯¹ haystack è¿›è¡Œåˆ‡ç‰‡ï¼Œè€Œæ— éœ€è¿›è¡Œå…¶ä»–è¿è¡Œæ—¶æ£€æŸ¥ã€‚

Required methods

fn haystack(&self) -> &'a str

ğŸ”¬ This is a nightly-only experimental API. (pattern #27721)

API not fully fleshed out and ready to be stabilized

Getter æŸ¥æ‰¾è¦åœ¨å…¶ä¸­æœç´¢çš„åŸºç¡€å­—ç¬¦ä¸²

æ€»æ˜¯è¿”å›ç›¸åŒçš„ &strã€‚

fn next(&mut self) -> SearchStep

ğŸ”¬ This is a nightly-only experimental API. (pattern #27721)

API not fully fleshed out and ready to be stabilized

ä»å¤´å¼€å‹æ‰è¡Œä¸‹ä¸€ä¸ªæœç´¢æ­¥éª¤ã€‚

å¦‚æœ haystack[a..b] ä¸æ¨¡å¼åŒ¹é…ï¼Œåˆ™è¿”å› Match(a, b)ã€‚

å¦‚æœ haystack[a..b] ç”šè‡³éƒ¨åˆ†ä¸åŒ¹é…ï¼Œåˆ™è¿”å› Reject(a, b)ã€‚

å¦‚æœå·²è®¿é—® haystack çš„æ¯ä¸ªå­—èŠ‚ï¼Œåˆ™è¿”å› Doneã€‚

ç›´åˆ° Done çš„ Match å’Œ Reject å€¼æµå°†åŒ…å«ç›¸é‚»ï¼Œä¸é‡å ï¼Œè¦†ç›–æ•´ä¸ª haystack å¹¶ä½äº utf8 è¾¹ç•Œä¸Šçš„ç´¢å¼•èŒƒå›´ã€‚

Match ç»“æœéœ€è¦åŒ…å«æ•´ä¸ªåŒ¹é…çš„æ¨¡å¼ï¼Œä½†æ˜¯ Reject ç»“æœå¯ä»¥åˆ†ä¸ºä»»æ„å¤šä¸ªç›¸é‚»çš„ç‰‡æ®µã€‚ä¸¤ä¸ªèŒƒå›´çš„é•¿åº¦éƒ½å¯ä»¥ä¸ºé›¶ã€‚

ä¾‹å¦‚ï¼Œæ¨¡å¼ "aaa" å’Œ haystack "cbaaaaab" å¯èƒ½ä¼šäºç”Ÿæµ [Reject(0, 1), Reject(1, 2), Match(2, 5), Reject(5, 8)]

Provided methods

fn next_match(&mut self) -> Option<(usize, usize)>

ğŸ”¬ This is a nightly-only experimental API. (pattern #27721)

API not fully fleshed out and ready to be stabilized

æŸ¥æ‰¾ä¸‹ä¸€ä¸ª Match ç»“æœã€‚è¯·å‚è next()ã€‚

ä¸ next() ä¸åŒï¼Œä¸èƒ½ä¿è¯æ­¤å’Œ next_reject çš„è¿”å›èŒƒå›´ä¼šé‡å ã€‚ è¿™å°†è¿”å› (start_match, end_match)ï¼Œå…¶ä¸­ start_match æ˜¯åŒ¹é…å¼€å‹çš„ç´¢å¼•ï¼Œend_match æ˜¯åŒ¹é…ç»“æŸåçš„ç´¢å¼•ã€‚

fn next_reject(&mut self) -> Option<(usize, usize)>

ğŸ”¬ This is a nightly-only experimental API. (pattern #27721)

API not fully fleshed out and ready to be stabilized

æŸ¥æ‰¾ä¸‹ä¸€ä¸ª Reject ç»“æœã€‚è¯·å‚è next() å’Œ next_match()ã€‚

ä¸ next() ä¸åŒï¼Œä¸èƒ½ä¿è¯æ­¤å’Œ next_match çš„è¿”å›èŒƒå›´ä¼šé‡å ã€‚