Struct core::str::CharRange

pub struct CharRange {
    pub ch: char,
    pub next: usize,
}

Unstable (str_char #27754)
: existence of this struct is uncertain as it is frequently able to be replaced with char.len_utf8() and/or char/char_indices iterators

Struct that contains a char and the index of the first byte of the next char in a string. This can be used as a data structure for iterating over the UTF-8 bytes of a string.

Fields

ch
Unstable (str_char #27754)
: existence of this struct is uncertain as it is frequently able to be replaced with char.len_utf8() and/or char/char_indices iterators

Current char

next
Unstable (str_char #27754)
: existence of this struct is uncertain as it is frequently able to be replaced with char.len_utf8() and/or char/char_indices iterators

Index of the first byte of the next char