Module core::stream

🔬 This is a nightly-only experimental API. (async_stream #79024)

可组合的异步迭代。

Composable asynchronous iteration.

如果期货是异步值，那么流就是异步迭代器。 如果您发现自己有某种异步集合，并且需要对所述集合的元素执行操作，那么您很快就会遇到“流”。 流在惯用的异步 Rust 代码中大量使用，因此值得熟悉它们。

If futures are asynchronous values, then streams are asynchronous iterators. If you’ve found yourself with an asynchronous collection of some kind, and needed to perform an operation on the elements of said collection, you’ll quickly run into ‘streams’. Streams are heavily used in idiomatic asynchronous Rust code, so it’s worth becoming familiar with them.

在解释更多之前，让我们先谈谈这个模块的结构：

Before explaining more, let’s talk about how this module is structured:

Organization

该模块主要按类型组织：

This module is largely organized by type:

Trait是核心部分：这些Trait定义了存在的流类型以及您可以使用它们做什么。 这些Trait的方法值得花一些额外的研究时间。

Traits are the core portion: these traits define what kind of streams exist and what you can do with them. The methods of these traits are worth putting some extra study time into.

函数提供了一些有用的方法来创建一些基本的流。

Functions provide some helpful ways to create some basic streams.

结构通常是此模块Trait上各种方法的返回类型。 您通常希望查看创建结构的方法，而不是结构本身。 有关原因的更多详细信息，请参阅“实现流”。

Structs are often the return types of the various methods on this module’s traits. You’ll usually want to look at the method that creates the struct, rather than the struct itself. For more detail about why, see ‘Implementing Stream’.

而已！ 让我们深入研究流。

That’s it! Let’s dig into streams.

Stream

这个模块的核心和灵魂是 Stream 特性。 Stream 的核心是这样的：

The heart and soul of this module is the Stream trait. The core of Stream looks like this:

trait Stream {
    type Item;
    fn poll_next(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>>;
}
Run

与 Iterator 不同，Stream 区分了实现 Stream 时使用的 poll_next 方法和使用流时使用的（待实现的）next 方法。 Stream 的消费者只需要考虑 next，当调用它时，它会返回一个产生 Option 的未来。

Unlike Iterator, Stream makes a distinction between the poll_next method which is used when implementing a Stream, and a (to-be-implemented) next method which is used when consuming a stream. Consumers of Stream only need to consider next, which when called, returns a future which yields Option<Stream::Item>.

只要有元素，next 返回的 future 将产生 Some(Item) ，一旦它们都被耗尽，将产生 None 表示迭代已完成。 如果我们正在等待异步解决，future 将等到流准备好再次屈服。

The future returned by next will yield Some(Item) as long as there are elements, and once they’ve all been exhausted, will yield None to indicate that iteration is finished. If we’re waiting on something asynchronous to resolve, the future will wait until the stream is ready to yield again.

个别流可能会选择恢复迭代，因此再次调用 next 最终可能会或可能不会最终在某个时候再次产生 Some(Item) 。

Individual streams may choose to resume iteration, and so calling next again may or may not eventually yield Some(Item) again at some point.

Stream 的完整定义还包括许多其他方法，但它们是默认方法，构建在 poll_next 之上，因此您可以免费获得它们。

Stream’s full definition includes a number of other methods as well, but they are default methods, built on top of poll_next, and so you get them for free.

Implementing Stream

创建自己的流包括两个步骤：创建一个结构来保存流的状态，然后为该结构实现 Stream。

Creating a stream of your own involves two steps: creating a struct to hold the stream’s state, and then implementing Stream for that struct.

让我们创建一个名为 Counter 的流，从 1 计数到 5：

Let’s make a stream named Counter which counts from 1 to 5:

#![feature(async_stream)]

// First, the struct:

/// A stream which counts from one to five
struct Counter {
    count: usize,
}

// we want our count to start at one, so let's add a new() method to help.
// This isn't strictly necessary, but is convenient. Note that we start
// `count` at zero, we'll see why in `poll_next()`'s implementation below.
impl Counter {
    fn new() -> Counter {
        Counter { count: 0 }
    }
}

// Then, we implement `Stream` for our `Counter`:

impl Stream for Counter {
    // we will be counting with usize
    type Item = usize;

    // poll_next() is the only required method
    fn poll_next(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
        // Increment our count. This is why we started at zero.
        self.count += 1;

        // Check to see if we've finished counting or not.
        if self.count < 6 {
            Poll::Ready(Some(self.count))
        } else {
            Poll::Ready(None)
        }
    }
}
Run

Laziness

流是懒惰的。 这意味着仅仅创建一个流并不能做很多事情。 在你打电话之前，什么都没有发生。 在仅为其副作用创建流时，这有时会造成混淆。 编译器会警告我们这种行为：

Streams are lazy. This means that just creating a stream doesn’t do a whole lot. Nothing really happens until you call next. This is sometimes a source of confusion when creating a stream solely for its side effects. The compiler will warn us about this kind of behavior:

warning: unused result that must be used: streams do nothing unless polled

Structs

FromIterExperimental

A stream that was created from iterator.

Traits

StreamExperimental

An interface for dealing with asynchronous iterators.

Functions

from_iterExperimental

Converts an iterator into a stream.