Trait core::stream::Stream

pub trait Stream {
    type Item;
    fn poll_next(
        self: Pin<&mut Self>, 
        cx: &mut Context<'_>
    ) -> Poll<Option<Self::Item>>;

    fn size_hint(&self) -> (usize, Option<usize>) { ... }
}

🔬 This is a nightly-only experimental API. (async_stream #79024)

处理异步迭代器的接口。

An interface for dealing with asynchronous iterators.

这是主流Trait。 有关流概念的更多信息，请参阅模块级文档。 特别是，您可能想知道如何实现 Stream。

This is the main stream trait. For more about the concept of streams generally, please see the module-level documentation. In particular, you may want to know how to implement Stream.

Associated Types

type Item

🔬 This is a nightly-only experimental API. (async_stream #79024)

The type of items yielded by the stream.

Required methods

fn poll_next(
self: Pin<&mut Self>,
cx: &mut Context<'_>
) -> Poll<Option<Self::Item>>

🔬 This is a nightly-only experimental API. (async_stream #79024)

尝试提取此流的下一个值，如果该值尚不可用，则注册当前任务以进行唤醒，如果流耗尽则返回 None。

Attempt to pull out the next value of this stream, registering the current task for wakeup if the value is not yet available, and returning None if the stream is exhausted.

Return value

有几个可能的返回值，每个表示不同的流状态：

There are several possible return values, each indicating a distinct stream state:

Poll::Pending 意味着这个流的下一个值还没有准备好。 实现将确保在下一个值可能准备好时通知当前任务。

Poll::Pending means that this stream’s next value is not ready yet. Implementations will ensure that the current task will be notified when the next value may be ready.

Poll::Ready(Some(val)) 表示流已成功生成值 val，并且可能会在后续 poll_next 调用中生成更多值。

Poll::Ready(Some(val)) means that the stream has successfully produced a value, val, and may produce further values on subsequent poll_next calls.

Poll::Ready(None) 表示流已终止，不应再次调用 poll_next。

Poll::Ready(None) means that the stream has terminated, and poll_next should not be invoked again.

Panics

一旦流完成（从 poll_next 返回 Ready(None)），再次调用其 poll_next 方法可能会出现panic、永远阻塞或导致其他类型的问题； Stream trait 对这种调用的效果没有任何要求。 然而，由于 poll_next 方法没有被标记为不安全，Rust 的通常规则适用：无论流的状态如何，调用都不能导致未定义的行为（内存损坏、不安全函数的错误使用等）。

Once a stream has finished (returned Ready(None) from poll_next), calling its poll_next method again may panic, block forever, or cause other kinds of problems; the Stream trait places no requirements on the effects of such a call. However, as the poll_next method is not marked unsafe, Rust’s usual rules apply: calls must never cause undefined behavior (memory corruption, incorrect use of unsafe functions, or the like), regardless of the stream’s state.

Provided methods

fn size_hint(&self) -> (usize, Option<usize>)

🔬 This is a nightly-only experimental API. (async_stream #79024)

返回流剩余长度的界限。

Returns the bounds on the remaining length of the stream.

具体来说， size_hint() 返回一个元组，其中第一个元素是下限，第二个元素是上限。

Specifically, size_hint() returns a tuple where the first element is the lower bound, and the second element is the upper bound.

返回的元组的后半部分是 Option。 这里的 None 意味着要么没有已知的上限，要么上限大于 usize。

The second half of the tuple that is returned is an Option<usize>. A None here means that either there is no known upper bound, or the upper bound is larger than usize.

Implementation notes

不强制流实现产生声明的元素数量。 有问题的流可能会产生小于元素的下限或大于上限的结果。

It is not enforced that a stream implementation yields the declared number of elements. A buggy stream may yield less than the lower bound or more than the upper bound of elements.

size_hint() 主要用于优化，例如为流的元素保留空间，但不能信任，例如，在不安全的代码中省略边界检查。 size_hint() 的错误实现不应导致内存安全违规。

size_hint() is primarily intended to be used for optimizations such as reserving space for the elements of the stream, but must not be trusted to e.g., omit bounds checks in unsafe code. An incorrect implementation of size_hint() should not lead to memory safety violations.

也就是说，实现应该提供正确的估计，否则将违反 trait 的协议。

That said, the implementation should provide a correct estimation, because otherwise it would be a violation of the trait’s protocol.

默认实现返回 (0, None) 这对于任何流都是正确的。

The default implementation returns (0, None) which is correct for any stream.