Struct core::NodeAPI

pub struct NodeAPI { /* private fields */ }

Object that allows interaction with a TAPLE node.

It has methods to perform all available read and write operations, as well as an additional action to stop a running node. he interaction is performed thanks to the implementation of a trait known as ApiModuleInterface. Consequently, it is necessary to import the trait in order to properly use the object.