Struct core::sync::atomic::AtomicI32

pub struct AtomicI32 { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (integer_atomics #32976)

An integer type which can be safely shared between threads.

This type has the same in-memory representation as the underlying integer type, i32. For more about the differences between atomic types and non-atomic types, please see the module-level documentation.

Please note that examples are shared between atomic variants of primitive integer types, so it's normal that they are all demonstrating AtomicIsize.

Methods

impl AtomicI32

pub const fn new(v: i32) -> Self

🔬 This is a nightly-only experimental API. (integer_atomics #32976)

Creates a new atomic integer.

Examples

use std::sync::atomic::AtomicIsize;

let atomic_forty_two  = AtomicIsize::new(42);Run

pub fn get_mut(&mut self) -> &mut i32

🔬 This is a nightly-only experimental API. (integer_atomics #32976)

Returns a mutable reference to the underlying integer.

This is safe because the mutable reference guarantees that no other threads are concurrently accessing the atomic data.

Examples

use std::sync::atomic::{AtomicIsize, Ordering};

let mut some_isize = AtomicIsize::new(10);
assert_eq!(*some_isize.get_mut(), 10);
*some_isize.get_mut() = 5;
assert_eq!(some_isize.load(Ordering::SeqCst), 5);Run

pub fn into_inner(self) -> i32

🔬 This is a nightly-only experimental API. (integer_atomics #32976)

Consumes the atomic and returns the contained value.

This is safe because passing self by value guarantees that no other threads are concurrently accessing the atomic data.

Examples

use std::sync::atomic::AtomicIsize;

let some_isize = AtomicIsize::new(5);
assert_eq!(some_isize.into_inner(), 5);Run

pub fn load(&self, order: Ordering) -> i32

🔬 This is a nightly-only experimental API. (integer_atomics #32976)

Loads a value from the atomic integer.

load takes an Ordering argument which describes the memory ordering of this operation.

Panics

Panics if order is Release or AcqRel.

Examples

use std::sync::atomic::{AtomicIsize, Ordering};

let some_isize = AtomicIsize::new(5);

assert_eq!(some_isize.load(Ordering::Relaxed), 5);Run

pub fn store(&self, val: i32, order: Ordering)

🔬 This is a nightly-only experimental API. (integer_atomics #32976)

Stores a value into the atomic integer.

store takes an Ordering argument which describes the memory ordering of this operation.

Examples

use std::sync::atomic::{AtomicIsize, Ordering};

let some_isize = AtomicIsize::new(5);

some_isize.store(10, Ordering::Relaxed);
assert_eq!(some_isize.load(Ordering::Relaxed), 10);Run

Panics

Panics if order is Acquire or AcqRel.

pub fn swap(&self, val: i32, order: Ordering) -> i32

🔬 This is a nightly-only experimental API. (integer_atomics #32976)

Stores a value into the atomic integer, returning the previous value.

swap takes an Ordering argument which describes the memory ordering of this operation.

Examples

use std::sync::atomic::{AtomicIsize, Ordering};

let some_isize = AtomicIsize::new(5);

assert_eq!(some_isize.swap(10, Ordering::Relaxed), 5);Run

pub fn compare_and_swap(&self, current: i32, new: i32, order: Ordering) -> i32

🔬 This is a nightly-only experimental API. (integer_atomics #32976)

Stores a value into the atomic integer if the current value is the same as the current value.

The return value is always the previous value. If it is equal to current, then the value was updated.

compare_and_swap also takes an Ordering argument which describes the memory ordering of this operation.

Examples

use std::sync::atomic::{AtomicIsize, Ordering};

let some_isize = AtomicIsize::new(5);

assert_eq!(some_isize.compare_and_swap(5, 10, Ordering::Relaxed), 5);
assert_eq!(some_isize.load(Ordering::Relaxed), 10);

assert_eq!(some_isize.compare_and_swap(6, 12, Ordering::Relaxed), 10);
assert_eq!(some_isize.load(Ordering::Relaxed), 10);Run

pub fn compare_exchange(
&self,
current: i32,
new: i32,
success: Ordering,
failure: Ordering
) -> Result<i32, i32>

🔬 This is a nightly-only experimental API. (integer_atomics #32976)

Stores a value into the atomic integer if the current value is the same as the current value.

The return value is a result indicating whether the new value was written and containing the previous value. On success this value is guaranteed to be equal to current.

compare_exchange takes two Ordering arguments to describe the memory ordering of this operation. The first describes the required ordering if the operation succeeds while the second describes the required ordering when the operation fails. The failure ordering can't be Release or AcqRel and must be equivalent or weaker than the success ordering.

Examples

use std::sync::atomic::{AtomicIsize, Ordering};

let some_isize = AtomicIsize::new(5);

assert_eq!(some_isize.compare_exchange(5, 10,
                                       Ordering::Acquire,
                                       Ordering::Relaxed),
           Ok(5));
assert_eq!(some_isize.load(Ordering::Relaxed), 10);

assert_eq!(some_isize.compare_exchange(6, 12,
                                       Ordering::SeqCst,
                                       Ordering::Acquire),
           Err(10));
assert_eq!(some_isize.load(Ordering::Relaxed), 10);Run

pub fn compare_exchange_weak(
&self,
current: i32,
new: i32,
success: Ordering,
failure: Ordering
) -> Result<i32, i32>

🔬 This is a nightly-only experimental API. (integer_atomics #32976)

Stores a value into the atomic integer if the current value is the same as the current value.

Unlike compare_exchange, this function is allowed to spuriously fail even when the comparison succeeds, which can result in more efficient code on some platforms. The return value is a result indicating whether the new value was written and containing the previous value.

compare_exchange_weak takes two Ordering arguments to describe the memory ordering of this operation. The first describes the required ordering if the operation succeeds while the second describes the required ordering when the operation fails. The failure ordering can't be Release or AcqRel and must be equivalent or weaker than the success ordering.

Examples

use std::sync::atomic::{AtomicIsize, Ordering};

let val = AtomicIsize::new(4);

let mut old = val.load(Ordering::Relaxed);
loop {
    let new = old * 2;
    match val.compare_exchange_weak(old, new, Ordering::SeqCst, Ordering::Relaxed) {
        Ok(_) => break,
        Err(x) => old = x,
    }
}Run

pub fn fetch_add(&self, val: i32, order: Ordering) -> i32

🔬 This is a nightly-only experimental API. (integer_atomics #32976)

Adds to the current value, returning the previous value.

This operation wraps around on overflow.

Examples

use std::sync::atomic::{AtomicIsize, Ordering};

let foo = AtomicIsize::new(0);
assert_eq!(foo.fetch_add(10, Ordering::SeqCst), 0);
assert_eq!(foo.load(Ordering::SeqCst), 10);Run

pub fn fetch_sub(&self, val: i32, order: Ordering) -> i32

🔬 This is a nightly-only experimental API. (integer_atomics #32976)

Subtracts from the current value, returning the previous value.

This operation wraps around on overflow.

Examples

use std::sync::atomic::{AtomicIsize, Ordering};

let foo = AtomicIsize::new(0);
assert_eq!(foo.fetch_sub(10, Ordering::SeqCst), 0);
assert_eq!(foo.load(Ordering::SeqCst), -10);Run

pub fn fetch_and(&self, val: i32, order: Ordering) -> i32

🔬 This is a nightly-only experimental API. (integer_atomics #32976)

Bitwise "and" with the current value.

Performs a bitwise "and" operation on the current value and the argument val, and sets the new value to the result.

Returns the previous value.

Examples

use std::sync::atomic::{AtomicIsize, Ordering};

let foo = AtomicIsize::new(0b101101);
assert_eq!(foo.fetch_and(0b110011, Ordering::SeqCst), 0b101101);
assert_eq!(foo.load(Ordering::SeqCst), 0b100001);Run

pub fn fetch_or(&self, val: i32, order: Ordering) -> i32

🔬 This is a nightly-only experimental API. (integer_atomics #32976)

Bitwise "or" with the current value.

Performs a bitwise "or" operation on the current value and the argument val, and sets the new value to the result.

Returns the previous value.

Examples

use std::sync::atomic::{AtomicIsize, Ordering};

let foo = AtomicIsize::new(0b101101);
assert_eq!(foo.fetch_or(0b110011, Ordering::SeqCst), 0b101101);
assert_eq!(foo.load(Ordering::SeqCst), 0b111111);Run

pub fn fetch_xor(&self, val: i32, order: Ordering) -> i32

🔬 This is a nightly-only experimental API. (integer_atomics #32976)

Bitwise "xor" with the current value.

Performs a bitwise "xor" operation on the current value and the argument val, and sets the new value to the result.

Returns the previous value.

Examples

use std::sync::atomic::{AtomicIsize, Ordering};

let foo = AtomicIsize::new(0b101101);
assert_eq!(foo.fetch_xor(0b110011, Ordering::SeqCst), 0b101101);
assert_eq!(foo.load(Ordering::SeqCst), 0b011110);Run