Module core::task

Types and Traits for working with asynchronous tasks.

Macros

ready

Extracts the successful type of a Poll<T>.

Structs

Context

The context of an asynchronous task.

RawWaker

A RawWaker allows the implementor of a task executor to create a Waker or a LocalWaker which provides customized wakeup behavior.

RawWakerVTable

A virtual function pointer table (vtable) that specifies the behavior of a RawWaker.

Waker

A Waker is a handle for waking up a task by notifying its executor that it is ready to be run.

ContextBuilderExperimental

A Builder used to construct a Context instance with support for LocalWaker.

LocalWakerExperimental

A LocalWaker is analogous to a Waker, but it does not implement Send or Sync.

Enums

Poll

Indicates whether a value is available or if the current task has been scheduled to receive a wakeup instead.