Struct core::task::Waker

pub struct Waker { /* private fields */ }

A Waker is a handle for waking up a task by notifying its executor that it is ready to be run.

This handle encapsulates a RawWaker instance, which defines the executor-specific wakeup behavior.

The typical life of a Waker is that it is constructed by an executor, wrapped in a Context, then passed to Future::poll(). Then, if the future chooses to return Poll::Pending, it must also store the waker somehow and call Waker::wake() when the future should be polled again.

Implements Clone, Send, and Sync; therefore, a waker may be invoked from any thread, including ones not in any way managed by the executor. For example, this might be done to wake a future when a blocking function call completes on another thread.

Note that it is preferable to use waker.clone_from(&new_waker) instead of *waker = new_waker.clone(), as the former will avoid cloning the waker unnecessarily if the two wakers wake the same task.

Constructing a Waker from a RawWaker is unsafe. Implementing the Wake trait is a safe alternative that requires memory allocation.

Implementations

impl Waker

pub fn wake(self)

Wake up the task associated with this Waker.

As long as the executor keeps running and the task is not finished, it is guaranteed that each invocation of wake() (or wake_by_ref()) will be followed by at least one poll() of the task to which this Waker belongs. This makes it possible to temporarily yield to other tasks while running potentially unbounded processing loops.

Note that the above implies that multiple wake-ups may be coalesced into a single poll() invocation by the runtime.

Also note that yielding to competing tasks is not guaranteed: it is the executor’s choice which task to run and the executor may choose to run the current task again.

pub fn wake_by_ref(&self)

Wake up the task associated with this Waker without consuming the Waker.

This is similar to wake(), but may be slightly less efficient in the case where an owned Waker is available. This method should be preferred to calling waker.clone().wake().

pub fn will_wake(&self, other: &Waker) -> bool

Returns true if this Waker and another Waker would awake the same task.

This function works on a best-effort basis, and may return false even when the Wakers would awaken the same task. However, if this function returns true, it is guaranteed that the Wakers will awaken the same task.

This function is primarily used for optimization purposes — for example, this type’s clone_from implementation uses it to avoid cloning the waker when they would wake the same task anyway.

pub unsafe fn from_raw(waker: RawWaker) -> Waker

Creates a new Waker from RawWaker.

Safety

The behavior of the returned Waker is undefined if the contract defined in RawWaker’s and RawWakerVTable’s documentation is not upheld.

(Authors wishing to avoid unsafe code may implement the Wake trait instead, at the cost of a required heap allocation.)

pub const fn noop() -> &'static Waker

🔬This is a nightly-only experimental API. (noop_waker #98286)

Returns a reference to a Waker that does nothing when used.

This is mostly useful for writing tests that need a Context to poll some futures, but are not expecting those futures to wake the waker or do not need to do anything specific if it happens.

If an owned Waker is needed, clone() this one.

Examples

#![feature(noop_waker)]

use std::future::Future;
use std::task;

let mut cx = task::Context::from_waker(task::Waker::noop());

let mut future = Box::pin(async { 10 });
assert_eq!(future.as_mut().poll(&mut cx), task::Poll::Ready(10));
Run

pub fn as_raw(&self) -> &RawWaker

🔬This is a nightly-only experimental API. (waker_getters #96992)

Get a reference to the underlying RawWaker.