Struct core::time::FromFloatSecsError

pub struct FromFloatSecsError { /* private fields */ }

🔬This is a nightly-only experimental API. (duration_checked_float #83400)

An error which can be returned when converting a floating-point value of seconds into a Duration.

This error is used as the error type for Duration::try_from_secs_f32 and Duration::try_from_secs_f64.

Example

#![feature(duration_checked_float)]
use std::time::Duration;

if let Err(e) = Duration::try_from_secs_f32(-1.0) {
    println!("Failed conversion to Duration: {e}");
}
Run