Module core::usize

ğŸ‘ Deprecating in a future Rust version:
all constants in this module replaced by associated constants on usize

æŒ‡é’ˆå¤å°çš„æ— ç¬¦å·æ•´æ•°ç±»å‹çš„å¸¸é‡ã€‚

See also the usize primitive type.

æ–°ä»£ç åº”ç›´æ¥åœ¨åŸå‹ç±»å‹ä¸Šä½¿ç”¨å…³è”çš„å¸¸é‡ã€‚

Constants

MAX	Deprecation planned
The largest value that can be represented by this integer type. Use usize::MAX instead.

MIN	Deprecation planned
The smallest value that can be represented by this integer type. Use usize::MIN instead.