Crate core

The Rust Core Library

The Rust Core Library is the dependency-free1 foundation of The Rust Standard Library. It is the portable glue between the language and its libraries, defining the intrinsic and primitive building blocks of all Rust code. It links to no upstream libraries, no system libraries, and no libc.

The core library is minimal: it isn't even aware of heap allocation, nor does it provide concurrency or I/O. These things require platform integration, and this library is platform-agnostic.

How to use the core library

Please note that all of these details are currently not considered stable.

This library is built on the assumption of a few existing symbols:

memcpy, memcmp, memset - These are core memory routines which are often generated by LLVM. Additionally, this library can make explicit calls to these functions. Their signatures are the same as found in C. These functions are often provided by the system libc, but can also be provided by the compiler-builtins crate.

rust_begin_panic - This function takes four arguments, a fmt::Arguments, a &'static str, and two u32's. These four arguments dictate the panic message, the file at which panic was invoked, and the line and column inside the file. It is up to consumers of this core library to define this panic function; it is only required to never return. This requires a lang attribute named panic_impl.

rust_eh_personality - is used by the failure mechanisms of the compiler. This is often mapped to GCC's personality function, but crates which do not trigger a panic can be assured that this function is never called. The lang attribute is called eh_personality.

Strictly speaking, there are some symbols which are needed but they aren't always necessary. ↩

Modules

alloc
内存分配APIs

any
该模块实现了Any特质，该特质允许通过运行时反射来动态键入任何'static 类型

arch
SIMD and vendor intrinsics module.

array
定长数组的 Eq 之类的实现. 限制到一定长度。 最终我们应该能够扩展到所有长度。

ascii
ASCII字符串和字符的操作

borrow
用于处理借入(borrowed)数据的模块

cell
共享可变容器

char
字符类型

clone
无法隐式复制类型的Clone特质

cmp
排序和比较的功能

convert
类型之间转换的特质

default
类型的默认值可能具有意义的Default特质

f32
该模块提供特定于f32浮点数据类型实现的常量

f64
该模块提供特定于f64浮点数据类型实现的常量

ffi
FFI绑定有关的实用工具

fmt
格式化和打印 String的实用工具

future
异步值

hash
通用哈希支持

hint
对编译器的提示，该提示会影响应如何发出或优化代码

i8
8位带符号整数类型

i16
16位带符号整数类型

i32
32位带符号整数类型

i64
64位带符号整数类型

i128
128位带符号整数类型

isize
指针大小的有符号整数类型

iter
可组合外部迭代

marker
原始特质和类型的类型基本属性表现

mem
处理内存的基本功能

num
内置数字类型的数字特质和函数

ops
可重载运算符

option
可选值

pin
将数据固定到其在内存中位置的类型

prelude
Rust libcore 前导

ptr
Manually manage memory through raw pointers.

result
Result 类型错误处理

slice
切片管理和操作

str
字符串操作

sync
同步原语

task
用于处理异步任务的类型和特质

time
时间量化

u8
8位无符号整数类型

u16
16位无符号整数类型

u32
32位无符号整数类型

u64
64位无符号整数类型

u128
128位无符号整数类型

usize
指针大小的无符号整数类型

intrinsics	Experimental
Compiler intrinsics.

panic	Experimental
标准库中的恐慌/故障支持

panicking	Experimental
libcore 的Panic支持

raw	Experimental
包含用于编译器内置类型布局的结构定义

unicode	Experimental

Macros

Clone
Derive macro generating an impl of the trait Clone.

Copy
Derive macro generating an impl of the trait Copy.

Debug
Derive macro generating an impl of the trait Debug.

Default
Derive macro generating an impl of the trait Default.

Eq
Derive macro generating an impl of the trait Eq.

Hash
Derive macro generating an impl of the trait Hash.

Ord
Derive macro generating an impl of the trait Ord.

PartialEq
Derive macro generating an impl of the trait PartialEq.

PartialOrd
Derive macro generating an impl of the trait PartialOrd.

RustcDecodable
Unstable implementation detail of the rustc compiler, do not use.

RustcEncodable
Unstable implementation detail of the rustc compiler, do not use.

assert
运行时断言布尔表达式在为 true

assert_eq
断言两个表达式彼此相等（使用 PartialEq）

assert_ne
断言两个表达式彼此不相等（使用PartialEq）

cfg
在编译时评估配置标志的布尔组合

column
扩展为调用它的列号

compile_error
当导致编译失败时,给定的错误消息

concat
Concatenates literals into a static string slice.

debug_assert
运行时断言布尔表达式在为 true

debug_assert_eq
断言两个表达式彼此相等

debug_assert_ne
断言两个表达式彼此不相等

env
在编译时检查环境变量.

file
扩展为调用它的文件名

format_args
字符串格式宏的构造参数

global_allocator
Attribute macro applied to a static to register it as a global allocator.

include
根据上下文将文件解析为表达式或项

include_bytes
包含一个文件，作为对字节数组的引用

include_str
包含一个utf8编码的文件作为字符串

line
Expands to the line number on which it was invoked.

module_path
扩展为表示当前模块路径的字符串

option_env
可选的在编译时检查环境变量。

panic
恐慌/Panics当前线程

stringify
对其参数进行字符串化

test
Attribute macro applied to a function to turn it into a unit test.

todo
表示未完成的代码

try	Deprecated
展开result或传播其错误

unimplemented
Panic并显示“尚未实现”信息，表示未完成的代码

unreachable
表示无法访问的代码

write
将格式化的数据写入缓冲区

writeln
将格式化的数据写入缓冲区并换行

asm	Experimental
内联汇编

bench	Experimental
Attribute macro applied to a function to turn it into a benchmark test.

concat_idents	Experimental
将多个标识符连接成一个标识符

format_args_nl	Experimental
与format_args相同，但最后添加一个换行符

global_asm	Experimental
模块级内联汇编

log_syntax	Experimental
将传递的令牌打印到标准输出中

matches	Experimental
返回与任何给定模式匹配的表达式 Returns whether the given expression matches any of the given patterns.

test_case	Experimental
An implementation detail of the #[test] and #[bench] macros.

trace_macros	Experimental
启用或禁用用于调试其他宏的跟踪功能