Crate core

Structs

BevyBones ğŸ”’

Marker component for entities that are rendered for bones.

CoreMetaHandle ğŸ”’

Session ğŸ”’

Snapshot ğŸ”’

Statics

GLOBAL ğŸ”’

Functions

load ğŸ”’

Wait for game meta to load, and then start the game session.

main

play_sounds ğŸ”’

setup ğŸ”’

Setup the game, loading the metadata and starting the game session.

snapshot_restore ğŸ”’

update_game ğŸ”’

Update the game simulation.

update_input ğŸ”’

Update the game session input.