Crate core

The Rust Core Library

The Rust Core Library is the dependency-free1 foundation of The Rust Standard Library. It is the portable glue between the language and its libraries, defining the intrinsic and primitive building blocks of all Rust code. It links to no upstream libraries, no system libraries, and no libc.

The core library is minimal: it isn't even aware of heap allocation, nor does it provide concurrency or I/O. These things require platform integration, and this library is platform-agnostic.

How to use the core library

Please note that all of these details are currently not considered stable.

This library is built on the assumption of a few existing symbols:

memcpy, memcmp, memset - These are core memory routines which are often generated by LLVM. Additionally, this library can make explicit calls to these functions. Their signatures are the same as found in C. These functions are often provided by the system libc, but can also be provided by the rlibc crate.

rust_begin_panic - This function takes four arguments, a fmt::Arguments, a &'static str, and two u32's. These four arguments dictate the panic message, the file at which panic was invoked, and the line and column inside the file. It is up to consumers of this core library to define this panic function; it is only required to never return. This requires a lang attribute named panic_fmt.

rust_eh_personality - is used by the failure mechanisms of the compiler. This is often mapped to GCC's personality function, but crates which do not trigger a panic can be assured that this function is never called. The lang attribute is called eh_personality.

Strictly speaking, there are some symbols which are needed but they aren't always necessary. ↩

Modules

any
This module implements the Any trait, which enables dynamic typing of any 'static type through runtime reflection.

borrow
A module for working with borrowed data.

cell
Shareable mutable containers.

char
Character manipulation.

clone
The Clone trait for types that cannot be 'implicitly copied'.

cmp
Functionality for ordering and comparison.

convert
Traits for conversions between types.

default
The Default trait for types which may have meaningful default values.

f32
Operations and constants for 32-bits floats (f32 type)

f64
Operations and constants for 64-bits floats (f64 type)

fmt
Utilities for formatting and printing strings.

hash
Generic hashing support.

i8
The 8-bit signed integer type.

i16
The 16-bit signed integer type.

i32
The 32-bit signed integer type.

i64
The 64-bit signed integer type.

isize
The pointer-sized signed integer type.

iter
Composable external iteration.

marker
Primitive traits and types representing basic properties of types.

mem
Basic functions for dealing with memory.

num
Numeric traits and functions for the built-in numeric types.

ops
Overloadable operators.

option
Optional values.

prelude
The libcore prelude

ptr
Raw, unsafe pointers, *const T, and *mut T.

result
Error handling with the Result type.

slice
Slice management and manipulation

str
String manipulation

sync
Synchronization primitives

u8
The 8-bit unsigned integer type.

u16
The 16-bit unsigned integer type.

u32
The 32-bit unsigned integer type.

u64
The 64-bit unsigned integer type.

usize
The pointer-sized unsigned integer type.

array 	 [
Experimental
]
Implementations of things like Eq for fixed-length arrays up to a certain length. Eventually we should able to generalize to all lengths.

i128 	 [
Experimental
]
The 128-bit signed integer type.

intrinsics 	 [
Experimental
]
rustc compiler intrinsics.

nonzero 	 [
Experimental
]
Exposes the NonZero lang item which provides optimization hints.

panicking 	 [
Experimental
]
Panic support for libcore

raw 	 [
Experimental
]
Contains struct definitions for the layout of compiler built-in types.

u128 	 [
Experimental
]
The 128-bit unsigned integer type.

Macros

assert
Ensure that a boolean expression is true at runtime.

assert_eq
Asserts that two expressions are equal to each other (using PartialEq).

assert_ne
Asserts that two expressions are not equal to each other (using PartialEq).

cfg
Boolean evaluation of configuration flags.

column
A macro which expands to the column number on which it was invoked.

compile_error
Unconditionally causes compilation to fail with the given error message when encountered.

concat
Concatenates literals into a static string slice.

debug_assert
Ensure that a boolean expression is true at runtime.

debug_assert_eq
Asserts that two expressions are equal to each other.

debug_assert_ne
Asserts that two expressions are not equal to each other.

env
Inspect an environment variable at compile time.

file
A macro which expands to the file name from which it was invoked.

format_args
The core macro for formatted string creation & output.

include
Parse a file as an expression or an item according to the context.

include_bytes
Includes a file as a reference to a byte array.

include_str
Includes a utf8-encoded file as a string.

line
A macro which expands to the line number on which it was invoked.

module_path
Expands to a string that represents the current module path.

option_env
Optionally inspect an environment variable at compile time.

panic
Entry point of thread panic, for details, see std::macros

stringify
A macro which stringifies its argument.

try
Helper macro for reducing boilerplate code for matching Result together with converting downstream errors.

unimplemented
A standardized placeholder for marking unfinished code.

unreachable
A utility macro for indicating unreachable code.

write
Write formatted data into a buffer

writeln
Write formatted data into a buffer, with a newline appended.

concat_idents 	 [
Experimental
]
Concatenate identifiers into one identifier.