Crate core

This crate contains the core components and functionalities of the feonix software. It is designed around the actor framework (see https://ryhl.io/blog/actors-with-tokio/), with all main functionalities being implemented as actors.

The file main.rs contains the entry point for the application.

Modules

camera 🔒

Camera Actor is an actor which is responsible for taking photos and sending the filepath to the Sauron Actor and is the start of the whole image-processing pipeline.

database 🔒

Utility functions for managing the PostgreSQL database.

mailbox 🔒

Mailbox is an actor to handle message routing and persistence.

sauron 🔒

Sauron Actor is an actor which implements the object detection and localization pipeline.

utils 🔒

General utilities for the core crate.

Enums

PostgresTypes 🔒

An enum representing the active PostgreSQL instance. Currently supports either an embedded server or a Docker container.

Functions

main 🔒