Crate core

Unstable (core #27701)
: the libcore library has not yet been scrutinized for stabilization in terms of structure and naming

The Rust Core Library

The Rust Core Library is the dependency-free foundation of The Rust Standard Library. It is the portable glue between the language and its libraries, defining the intrinsic and primitive building blocks of all Rust code. It links to no upstream libraries, no system libraries, and no libc.

The core library is minimal: it isn't even aware of heap allocation, nor does it provide concurrency or I/O. These things require platform integration, and this library is platform-agnostic.

It is not recommended to use the core library. The stable functionality of libcore is reexported from the standard library. The composition of this library is subject to change over time; only the interface exposed through libstd is intended to be stable.

How to use the core library

This library is built on the assumption of a few existing symbols:

memcpy, memcmp, memset - These are core memory routines which are often generated by LLVM. Additionally, this library can make explicit calls to these functions. Their signatures are the same as found in C. These functions are often provided by the system libc, but can also be provided by the rlibc crate.

rust_begin_unwind - This function takes three arguments, a fmt::Arguments, a &str, and a u32. These three arguments dictate the panic message, the file at which panic was invoked, and the line. It is up to consumers of this core library to define this panic function; it is only required to never return.

Modules

any
This module implements the Any trait, which enables dynamic typing of any 'static type through runtime reflection.

borrow
A module for working with borrowed data.

cell
Shareable mutable containers.

char
Character manipulation.

clone
The Clone trait for types that cannot be 'implicitly copied'

cmp
Functionality for ordering and comparison.

convert
Traits for conversions between types.

default
The Default trait for types which may have meaningful default values.

f32
Operations and constants for 32-bits floats (f32 type)

f64
Operations and constants for 64-bits floats (f64 type)

fmt
Utilities for formatting and printing strings

hash
Generic hashing support.

i16
The 16-bit signed integer type.

i32
The 32-bit signed integer type.

i64
The 64-bit signed integer type.

i8
The 8-bit signed integer type.

isize
The pointer-sized signed integer type.

iter
Composable external iterators

marker
Primitive traits and marker types representing basic 'kinds' of types.

mem
Basic functions for dealing with memory

num
Numeric traits and functions for the built-in numeric types.

ops
Overloadable operators

option
Optional values

prelude
The libcore prelude

ptr
Raw, unsafe pointers, *const T, and *mut T

result
Error handling with the Result type

slice
Slice management and manipulation

str
String manipulation

sync
Synchronization primitives

u16
The 16-bit unsigned integer type.

u32
The 32-bit unsigned integer type.

u64
The 64-bit unsigned integer type.

u8
The 8-bit unsigned integer type.

usize
The pointer-sized unsigned integer type.

array 	 [Unstable]
Implementations of things like Eq for fixed-length arrays up to a certain length. Eventually we should able to generalize to all lengths.

intrinsics 	 [Unstable]
rustc compiler intrinsics.

nonzero 	 [Unstable]
Exposes the NonZero lang item which provides optimization hints.

panicking 	 [Unstable]
Panic support for libcore

raw 	 [Unstable]
Contains struct definitions for the layout of compiler built-in types.

simd 	 [Deprecated]
SIMD vectors.

Macros

__impl_slice_eq1!

__impl_slice_eq2!

assert!
Ensure that a boolean expression is true at runtime.

assert_eq!
Asserts that two expressions are equal to each other.

debug_assert!
Ensure that a boolean expression is true at runtime.

debug_assert_eq!
Asserts that two expressions are equal to each other, testing equality in both directions.

panic!
Entry point of thread panic, for details, see std::macros

try!
Short circuiting evaluation on Err

unimplemented!
A standardised placeholder for marking unfinished code. It panics with the message "not yet implemented" when executed.

unreachable!
A utility macro for indicating unreachable code.

write!
Use the format! syntax to write data into a buffer of type &mut Write. See std::fmt for more information.

writeln!
Equivalent to the write! macro, except that a newline is appended after the message is written.