Module core::alloc

Memory allocation APIs

Structs

AllocErrorExperimental

The AllocError error indicates an allocation failure that may be due to resource exhaustion or to something wrong when combining the given input arguments with this allocator.

Layout

Layout of a block of memory.

LayoutError

The parameters given to Layout::from_size_align or some other Layout constructor do not satisfy its documented constraints.

Traits

AllocatorExperimental

An implementation of Allocator can allocate, grow, shrink, and deallocate arbitrary blocks of data described via Layout.

GlobalAlloc

A memory allocator that can be registered as the standard library’s default through the #[global_allocator] attribute.

Type Definitions

LayoutErrDeprecated

This documentation is an old archive. Please see https://rust.docs.kernel.org instead.