Struct core::any::TypeId

pub struct TypeId { /* fields omitted */ }

A TypeId represents a globally unique identifier for a type.

Each TypeId is an opaque object which does not allow inspection of what's inside but does allow basic operations such as cloning, comparison, printing, and showing.

A TypeId is currently only available for types which ascribe to 'static, but this limitation may be removed in the future.

While TypeId implements Hash, PartialOrd, and Ord, it is worth noting that the hashes and ordering will vary between Rust releases. Beware of relying on them outside of your code!

Methods

impl TypeId

fn of<T: ?Sized + 'static>() -> TypeId

Returns the TypeId of the type this generic function has been instantiated with.

Examples

use std::any::{Any, TypeId};

fn is_string<T: ?Sized + Any>(_s: &T) -> bool {
    TypeId::of::<String>() == TypeId::of::<T>()
}

fn main() {
    assert_eq!(is_string(&0), false);
    assert_eq!(is_string(&"cookie monster".to_string()), true);
}Run