Struct core::any::TypeId

pub struct TypeId {
    // some fields omitted
}

A TypeId represents a globally unique identifier for a type.

Each TypeId is an opaque object which does not allow inspection of what's inside but does allow basic operations such as cloning, comparison, printing, and showing.

A TypeId is currently only available for types which ascribe to 'static, but this limitation may be removed in the future.

Methods

impl TypeId

fn of<T: ?Sized + Reflect + 'static>() -> TypeId

Returns the TypeId of the type this generic function has been instantiated with