Trait core::any::Any

pub trait Any: 'static {
    fn type_id(&self) -> TypeId;
}

A type to emulate dynamic typing.

Most types implement Any. However, any type which contains a non-'static reference does not. See the module-level documentation for more details.

Required methods

fn type_id(&self) -> TypeId

Gets the TypeId of self.

Examples

use std::any::{Any, TypeId};

fn is_string(s: &dyn Any) -> bool {
    TypeId::of::<String>() == s.type_id()
}

fn main() {
    assert_eq!(is_string(&0), false);
    assert_eq!(is_string(&"cookie monster".to_string()), true);
}Run

Loading content...
Methods

impl dyn Any

pub fn is<T: Any>(&self) -> bool

Returns true if the boxed type is the same as T.

Examples

use std::any::Any;

fn is_string(s: &dyn Any) {
    if s.is::<String>() {
        println!("It's a string!");
    } else {
        println!("Not a string...");
    }
}

fn main() {
    is_string(&0);
    is_string(&"cookie monster".to_string());
}Run

pub fn downcast_ref<T: Any>(&self) -> Option<&T>

Returns some reference to the boxed value if it is of type T, or None if it isn't.

Examples

use std::any::Any;

fn print_if_string(s: &dyn Any) {
    if let Some(string) = s.downcast_ref::<String>() {
        println!("It's a string({}): '{}'", string.len(), string);
    } else {
        println!("Not a string...");
    }
}

fn main() {
    print_if_string(&0);
    print_if_string(&"cookie monster".to_string());
}Run

pub fn downcast_mut<T: Any>(&mut self) -> Option<&mut T>

Returns some mutable reference to the boxed value if it is of type T, or None if it isn't.

Examples

use std::any::Any;

fn modify_if_u32(s: &mut dyn Any) {
    if let Some(num) = s.downcast_mut::<u32>() {
        *num = 42;
    }
}

fn main() {
    let mut x = 10u32;
    let mut s = "starlord".to_string();

    modify_if_u32(&mut x);
    modify_if_u32(&mut s);

    assert_eq!(x, 42);
    assert_eq!(&s, "starlord");
}Run

impl dyn Any + Send

pub fn is<T: Any>(&self) -> bool

Forwards to the method defined on the type Any.

Examples

use std::any::Any;

fn is_string(s: &(dyn Any + Send)) {
    if s.is::<String>() {
        println!("It's a string!");
    } else {
        println!("Not a string...");
    }
}

fn main() {
    is_string(&0);
    is_string(&"cookie monster".to_string());
}Run

pub fn downcast_ref<T: Any>(&self) -> Option<&T>

Forwards to the method defined on the type Any.

Examples

use std::any::Any;

fn print_if_string(s: &(dyn Any + Send)) {
    if let Some(string) = s.downcast_ref::<String>() {
        println!("It's a string({}): '{}'", string.len(), string);
    } else {
        println!("Not a string...");
    }
}

fn main() {
    print_if_string(&0);
    print_if_string(&"cookie monster".to_string());
}Run

pub fn downcast_mut<T: Any>(&mut self) -> Option<&mut T>

Forwards to the method defined on the type Any.

Examples

use std::any::Any;

fn modify_if_u32(s: &mut (dyn Any + Send)) {
    if let Some(num) = s.downcast_mut::<u32>() {
        *num = 42;
    }
}

fn main() {
    let mut x = 10u32;
    let mut s = "starlord".to_string();

    modify_if_u32(&mut x);
    modify_if_u32(&mut s);

    assert_eq!(x, 42);
    assert_eq!(&s, "starlord");
}Run

impl dyn Any + Send + Sync

pub fn is<T: Any>(&self) -> bool

Forwards to the method defined on the type Any.

Examples

use std::any::Any;

fn is_string(s: &(dyn Any + Send + Sync)) {
    if s.is::<String>() {
        println!("It's a string!");
    } else {
        println!("Not a string...");
    }
}

fn main() {
    is_string(&0);
    is_string(&"cookie monster".to_string());
}Run

pub fn downcast_ref<T: Any>(&self) -> Option<&T>

Forwards to the method defined on the type Any.

Examples

use std::any::Any;

fn print_if_string(s: &(dyn Any + Send + Sync)) {
    if let Some(string) = s.downcast_ref::<String>() {
        println!("It's a string({}): '{}'", string.len(), string);
    } else {
        println!("Not a string...");
    }
}

fn main() {
    print_if_string(&0);
    print_if_string(&"cookie monster".to_string());
}Run

pub fn downcast_mut<T: Any>(&mut self) -> Option<&mut T>

Forwards to the method defined on the type Any.

Examples

use std::any::Any;

fn modify_if_u32(s: &mut (dyn Any + Send + Sync)) {
    if let Some(num) = s.downcast_mut::<u32>() {
        *num = 42;
    }
}

fn main() {
    let mut x = 10u32;
    let mut s = "starlord".to_string();

    modify_if_u32(&mut x);
    modify_if_u32(&mut s);

    assert_eq!(x, 42);
    assert_eq!(&s, "starlord");
}Run