Function core::arch::riscv64::frflags

pub fn frflags() -> u32

🔬This is a nightly-only experimental API. (stdsimd #27731)

Reads the floating-point accrued exception flags register fflags

The accrued exception flags indicate the exception conditions that have arisen on any floating-point arithmetic instruction since the field was last reset by software.

According to “F” Standard Extension for Single-Precision Floating-Point, Version 2.2, the accrued exception flags is defined as a bit vector of 5 bits. The meaning of each binary bit is listed in the table below.

Bit index	Mnemonic	Meaning

4	NV	Invalid Operation

3	DZ	Divide by Zero

2	OF	Overflow

1	UF	Underflow

0	NX	Inexact