core::arch

Module s390x

🔬This is a nightly-only experimental API. (stdarch_s390x #135681)

Available on s390x only.

Platform-specific intrinsics for the s390x platform.

See the module documentation for more details.

Structs

vector_bool_charExperimental

s390x-specific 128-bit wide vector mask of sixteen packed elements

vector_bool_intExperimental

s390x-specific 128-bit wide vector mask of four packed elements

vector_bool_long_longExperimental

s390x-specific 128-bit wide vector mask of two packed elements

vector_bool_shortExperimental

s390x-specific 128-bit wide vector mask of eight packed elements

vector_doubleExperimental

s390x-specific 128-bit wide vector of two packed f64

vector_floatExperimental

s390x-specific 128-bit wide vector of four packed f32

vector_signed_charExperimental

s390x-specific 128-bit wide vector of sixteen packed i8

vector_signed_intExperimental

s390x-specific 128-bit wide vector of four packed i32

vector_signed_long_longExperimental

s390x-specific 128-bit wide vector of two packed i64

vector_signed_shortExperimental

s390x-specific 128-bit wide vector of eight packed i16

vector_unsigned_charExperimental

s390x-specific 128-bit wide vector of sixteen packed u8

vector_unsigned_intExperimental

s390x-specific 128-bit wide vector of four packed u32

vector_unsigned_long_longExperimental

s390x-specific 128-bit wide vector of two packed u64

vector_unsigned_shortExperimental

s390x-specific 128-bit wide vector of eight packed u16

Constants

__VEC_CLASS_FP_INFINITYExperimental

__VEC_CLASS_FP_INFINITY_NExperimental

__VEC_CLASS_FP_INFINITY_PExperimental

__VEC_CLASS_FP_NANExperimental

__VEC_CLASS_FP_NORMALExperimental

__VEC_CLASS_FP_NORMAL_NExperimental

__VEC_CLASS_FP_NORMAL_PExperimental

__VEC_CLASS_FP_NOT_NORMALExperimental

__VEC_CLASS_FP_QNANExperimental

__VEC_CLASS_FP_QNAN_NExperimental

__VEC_CLASS_FP_QNAN_PExperimental

__VEC_CLASS_FP_SNANExperimental

__VEC_CLASS_FP_SNAN_NExperimental

__VEC_CLASS_FP_SNAN_PExperimental

__VEC_CLASS_FP_SUBNORMALExperimental

__VEC_CLASS_FP_SUBNORMAL_NExperimental

__VEC_CLASS_FP_SUBNORMAL_PExperimental

__VEC_CLASS_FP_ZEROExperimental

__VEC_CLASS_FP_ZERO_NExperimental

__VEC_CLASS_FP_ZERO_PExperimental

Functions

vec_abs⚠Experimentalvector

Vector Absolute

vec_add⚠Experimentalvector

Vector Add

vec_add_u128⚠Experimentalvector

Vector Addition unsigned 128-bits

vec_addc_u128⚠Experimentalvector

Vector Add Compute Carryout unsigned 128-bits

vec_adde_u128⚠Experimentalvector

Vector Add With Carry unsigned 128-bits

vec_addec_u128⚠Experimentalvector

Vector Add With Carry Compute Carry unsigned 128-bits

vec_all_eq⚠Experimentalvector

All Elements Equal

vec_all_ge⚠Experimentalvector

All Elements Greater Than or Equal

vec_all_gt⚠Experimentalvector

All Elements Greater Than

vec_all_le⚠Experimentalvector

All Elements Less Than or Equal

vec_all_lt⚠Experimentalvector

All Elements Less Than

vec_all_nan⚠Experimentalvector

All Elements Not a Number

vec_all_ne⚠Experimentalvector

All Elements Not Equal

vec_all_nge⚠Experimentalvector

All Elements Not Greater Than or Equal

vec_all_ngt⚠Experimentalvector

All Elements Not Greater Than

vec_all_nle⚠Experimentalvector

All Elements Not Less Than or Equal

vec_all_nlt⚠Experimentalvector

All Elements Not Less Than

vec_all_numeric⚠Experimentalvector

All Elements Numeric

vec_and⚠Experimentalvector

Vector AND

vec_andc⚠Experimentalvector

Vector ANDC

vec_any_eq⚠Experimentalvector

Any Element Equal

vec_any_ge⚠Experimentalvector

Any Elements Greater Than or Equal

vec_any_gt⚠Experimentalvector

Any Elements Greater Than

vec_any_le⚠Experimentalvector

Any Elements Less Than or Equal

vec_any_lt⚠Experimentalvector

Any Elements Less Than

vec_any_nan⚠Experimentalvector

Any Elements Not a Number

vec_any_ne⚠Experimentalvector

Any Element Not Equal

vec_any_nge⚠Experimentalvector

Any Elements Not Greater Than or Equal

vec_any_ngt⚠Experimentalvector

Any Elements Not Greater Than

vec_any_nle⚠Experimentalvector

Any Elements Not Less Than or Equal

vec_any_nlt⚠Experimentalvector

Any Elements Not Less Than

vec_any_numeric⚠Experimentalvector

Any Elements Numeric

vec_avg⚠Experimentalvector

Vector Average

vec_bperm_u128⚠Experimentalvector-enhancements-1

Vector Bit Permute

vec_ceil⚠Experimentalvector

Vector Ceil

vec_checksum⚠Experimentalvector

Vector Checksum

vec_cmpeq⚠Experimentalvector

Vector Compare Equal

vec_cmpeq_idx⚠Experimentalvector

Vector Compare Equal Index

vec_cmpeq_idx_cc⚠Experimentalvector

Vector Compare Equal Index with Condition Code

vec_cmpeq_or_0_idx⚠Experimentalvector

Vector Compare Equal or Zero Index

vec_cmpeq_or_0_idx_cc⚠Experimentalvector

Vector Compare Equal or Zero Index with Condition Code

vec_cmpge⚠Experimentalvector

Vector Compare Greater Than or Equal

vec_cmpgt⚠Experimentalvector

Vector Compare Greater Than

vec_cmple⚠Experimentalvector

Vector Compare Less Than or Equal

vec_cmplt⚠Experimentalvector

Vector Compare Less

vec_cmpne⚠Experimentalvector

Vector Compare Not Equal

vec_cmpne_idx⚠Experimentalvector

Vector Compare Not Equal Index

vec_cmpne_idx_cc⚠Experimentalvector

Vector Compare Not Equal Index with Condition Code

vec_cmpne_or_0_idx⚠Experimentalvector

Vector Compare Not Equal or Zero Index

vec_cmpne_or_0_idx_cc⚠Experimentalvector

Vector Compare Not Equal or Zero Index with Condition Code

vec_cmpnrg⚠Experimentalvector

Vector Compare Not in Ranges

vec_cmpnrg_cc⚠Experimentalvector

Vector Compare Not in Ranges with Condition Code

vec_cmpnrg_idx⚠Experimentalvector

Vector Compare Not in Ranges Index

vec_cmpnrg_idx_cc⚠Experimentalvector

Vector Compare Not in Ranges Index with Condition Code

vec_cmpnrg_or_0_idx⚠Experimentalvector

Vector Compare Not in Ranges or Zero Index

vec_cmpnrg_or_0_idx_cc⚠Experimentalvector

Vector Compare Not in Ranges or Zero Index with Condition Code

vec_cmprg⚠Experimentalvector

Vector Compare Ranges

vec_cmprg_cc⚠Experimentalvector

Vector Compare Ranges with Condition Code

vec_cmprg_idx⚠Experimentalvector

Vector Compare Ranges Index

vec_cmprg_idx_cc⚠Experimentalvector

Vector Compare Ranges Index with Condition Code

vec_cmprg_or_0_idx⚠Experimentalvector

Vector Compare Ranges or Zero Index

vec_cmprg_or_0_idx_cc⚠Experimentalvector

Vector Compare Ranges or Zero Index with Condition Code

vec_cntlz⚠Experimentalvector

Vector Count Leading Zeros

vec_cnttz⚠Experimentalvector

Vector Count Trailing Zeros

vec_cp_until_zero⚠Experimentalvector

Vector Copy Until Zero

vec_cp_until_zero_cc⚠Experimentalvector

Vector Copy Until Zero

vec_double⚠Experimentalvector

Vector Convert from long long to double

vec_doublee⚠Experimentalvector-enhancements-1

Vector Convert from float (even elements) to double

vec_eqv⚠Experimentalvector

Vector XNOR

vec_extend_s64⚠Experimentalvector

Vector Sign Extend to Doubleword

vec_extract⚠Experimentalvector

Vector Extract

vec_find_any_eq⚠Experimentalvector

Vector Find Any Element Equal with Condition Code

vec_find_any_eq_cc⚠Experimentalvector

Vector Find Any Element Equal with Condition Code

vec_find_any_eq_idx⚠Experimentalvector

Vector Find Any Element Equal Index with Condition Code

vec_find_any_eq_idx_cc⚠Experimentalvector

Vector Find Any Element Equal Index with Condition Code

vec_find_any_eq_or_0_idx⚠Experimentalvector

Vector Find Any Element Equal or Zero Index with Condition Code

vec_find_any_eq_or_0_idx_cc⚠Experimentalvector

Vector Find Any Element Equal or Zero Index with Condition Code

vec_find_any_ne⚠Experimentalvector

Vector Find Any Element Not Equal with Condition Code

vec_find_any_ne_cc⚠Experimentalvector

Vector Find Any Element Not Equal with Condition Code

vec_find_any_ne_idx⚠Experimentalvector

Vector Find Any Element Not Equal Index with Condition Code

vec_find_any_ne_idx_cc⚠Experimentalvector

Vector Find Any Element Not Equal Index with Condition Code

vec_find_any_ne_or_0_idx⚠Experimentalvector

Vector Find Any Element Not Equal or Zero Index with Condition Code

vec_find_any_ne_or_0_idx_cc⚠Experimentalvector

Vector Find Any Element Not Equal or Zero Index with Condition Code

vec_float⚠Experimentalvector

Vector Convert from int to float

vec_floate⚠Experimentalvector-enhancements-1

Vector Convert from double to float (even elements)

vec_floor⚠Experimentalvector

Vector Floor

vec_fp_test_data_class⚠Experimentalvector

Vector Floating-Point Test Data Class

vec_gather_element⚠Experimentalvector

Vector Gather Element

vec_genmask⚠Experimentalvector

Vector Generate Byte Mask

vec_genmasks_8⚠Experimentalvector

Vector Generate Mask (Byte)

vec_genmasks_16⚠Experimentalvector

Vector Generate Mask (Halfword)

vec_genmasks_32⚠Experimentalvector

Vector Generate Mask (Word)

vec_genmasks_64⚠Experimentalvector

Vector Generate Mask (Doubleword)

vec_gfmsum⚠Experimentalvector

Vector Galois Field Multiply Sum

vec_gfmsum_128⚠Experimentalvector

Vector Galois Field Multiply Sum 128-bits

vec_gfmsum_accum⚠Experimentalvector

Vector Galois Field Multiply Sum

vec_gfmsum_accum_128⚠Experimentalvector

Vector Galois Field Multiply Sum and Accumulate 128-bits

vec_insert⚠Experimentalvector

Vector Insert

vec_insert_and_zero⚠Experimentalvector

Vector Insert and Zero

vec_load_bndry⚠Experimentalvector

Vector Load to Block Boundary

vec_load_len⚠Experimentalvector

Vector Load with Length

vec_load_len_r⚠Experimentalvector-packed-decimal

Vector Load Rightmost with Length

vec_load_pair⚠Experimentalvector

Vector Load Pair

vec_madd⚠Experimentalvector-packed-decimal

Vector Multiply Add

vec_max⚠Experimentalvector

Vector Maximum

vec_meadd⚠Experimentalvector-packed-decimal

Vector Multiply and Add Even

vec_mergeh⚠Experimentalvector

Vector Merge High

vec_mergel⚠Experimentalvector

Vector Merge Low

vec_mhadd⚠Experimentalvector-packed-decimal

Vector Multiply and Add High

vec_min⚠Experimentalvector

Vector Minimum

vec_mladd⚠Experimentalvector-packed-decimal

Vector Multiply and Add Low

vec_moadd⚠Experimentalvector-packed-decimal

Vector Multiply and Add Odd

vec_msub⚠Experimentalvector-packed-decimal

Vector Multiply Add

vec_msum_u128⚠Experimentalvector-enhancements-1

Vector Multiply Sum Logical

vec_mul⚠Experimentalvector

Vector Multiply

vec_mule⚠Experimentalvector

Vector Multiply Even

vec_mulh⚠Experimentalvector

Vector Multiply High

vec_mulo⚠Experimentalvector

Vector Multiply Odd

vec_nabs⚠Experimentalvector

Vector Negative Absolute

vec_nand⚠Experimentalvector

Vector NAND

vec_nmadd⚠Experimentalvector

Vector Negative Multiply Add

vec_nmsub⚠Experimentalvector

Vector Negative Multiply Subtract

vec_nor⚠Experimentalvector

Vector NOR

vec_or⚠Experimentalvector

Vector OR

vec_orc⚠Experimentalvector

Vector OR with Complement

vec_pack⚠Experimentalvector

Vector Pack

vec_packs⚠Experimentalvector

Vector Pack Saturated

vec_packs_cc⚠Experimentalvector

Vector Pack Saturated Condition Code

vec_packsu⚠Experimentalvector

Vector Pack Saturated Unsigned

vec_packsu_cc⚠Experimentalvector

Vector Pack Saturated Unsigned Condition Code

vec_perm⚠Experimentalvector

Vector Permute

vec_popcnt⚠Experimentalvector

Vector Population Count

vec_promote⚠Experimentalvector

Vector Promote

vec_revb⚠Experimentalvector

Vector Byte Reverse

vec_reve⚠Experimentalvector

Vector Reverse Elements

vec_rint⚠Experimentalvector

Vector Round to Integer

vec_rl⚠Experimentalvector

Vector Element Rotate Left

vec_rli⚠Experimentalvector

Vector Element Rotate Left Immediate

vec_round⚠Experimentalvector

Vector Round

vec_roundc⚠Experimentalvector

Vector Round to Current

vec_roundm⚠Experimentalvector

Vector Round toward Negative Infinity

vec_roundp⚠Experimentalvector

Vector Round toward Positive Infinity

vec_roundz⚠Experimentalvector

Vector Round toward Zero

vec_search_string_cc⚠Experimentalvector

Vector Search String

vec_search_string_until_zero_cc⚠Experimentalvector

Vector Search String Until Zero

vec_sel⚠Experimentalvector

Vector Select

vec_signed⚠Experimentalvector

Vector Convert floating point to signed

vec_sl⚠Experimentalvector

Vector Shift Left

vec_slb⚠Experimentalvector

Vector Shift Left by Byte

vec_sld⚠Experimentalvector

Vector Shift Left Double by Byte

vec_sldb⚠Experimentalvector-enhancements-2

Vector Shift Left Double by Bit

vec_sldw⚠Experimentalvector

Vector Shift Left Double by Word

vec_sll⚠Experimentalvector

Vector Shift Left

vec_splat⚠Experimentalvector

Vector Splat

vec_splat_s8⚠Experimentalvector

Vector Splat Signed Byte

vec_splat_s16⚠Experimentalvector

Vector Splat Signed Halfword

vec_splat_s32⚠Experimentalvector

Vector Splat Signed Word

vec_splat_s64⚠Experimentalvector

Vector Splat Signed Doubleword

vec_splat_u8⚠Experimentalvector

Vector Splat Unsigned Byte

vec_splat_u16⚠Experimentalvector

Vector Splat Unsigned Halfword

vec_splat_u32⚠Experimentalvector

Vector Splat Unsigned Word

vec_splat_u64⚠Experimentalvector

Vector Splat Unsigned Doubleword

vec_splats⚠Experimentalvector

Vector Splats

vec_sqrt⚠Experimentalvector

Vector Square Root

vec_sr⚠Experimentalvector

Vector Shift Right

vec_sra⚠Experimentalvector

Vector Shift Right Algebraic

vec_srab⚠Experimentalvector

Vector Shift Right Algebraic by Byte

vec_sral⚠Experimentalvector

Vector Shift Right Arithmetic

vec_srb⚠Experimentalvector

Vector Shift Right by Byte

vec_srdb⚠Experimentalvector-enhancements-2

Vector Shift Right Double by Bit

vec_srl⚠Experimentalvector

Vector Shift Right

vec_store_len⚠Experimentalvector

Vector Store with Length

vec_store_len_r⚠Experimentalvector-packed-decimal

Vector Store Rightmost with Length

vec_sub⚠Experimentalvector

Vector Subtract

vec_sub_u128⚠Experimentalvector

Vector Subtract unsigned 128-bits

vec_subc⚠Experimentalvector

Vector Subtract Carryout

vec_subc_u128⚠Experimentalvector

Vector Subtract Carryout unsigned 128-bits

vec_sube_u128⚠Experimentalvector

Vector Subtract with Carryout

vec_subec_u128⚠Experimentalvector

Vector Subtract with Carryout, Carryout

vec_sum2⚠Experimentalvector

Vector Sum Across Doubleword

vec_sum4⚠Experimentalvector

Vector Sum Across Word

vec_sum_u128⚠Experimentalvector

Vector Sum Across Quadword

vec_test_mask⚠Experimentalvector

Vector Test under Mask

vec_trunc⚠Experimentalvector

Vector Truncate

vec_unpackh⚠Experimentalvector

Vector Unpack High

vec_unpackl⚠Experimentalvector

Vector Unpack Low

vec_unsigned⚠Experimentalvector

Vector Convert floating point to unsigned

vec_xl⚠Experimentalvector

Vector Load

vec_xor⚠Experimentalvector

Vector XOR

vec_xst⚠Experimentalvector

Vector Store