Function core::arch::wasm::v128_store

pub unsafe fn v128_store(m: *mut v128, a: v128)

This is supported on target_family="wasm" and target feature simd128 only.

Stores a v128 vector to the given heap address.

This intrinsic will emit a store with an alignment of 1. While this is provided for completeness it is not strictly necessary, you can also store the pointer directly:

let a: &mut v128 = ...;
unsafe { v128_store(a, value) };
// .. is the same as ..
*a = value;
Run

The alignment of the store can be configured by doing a manual store without this intrinsic.

Unsafety

This intrinsic is unsafe because it takes a raw pointer as an argument, and the pointer must be valid to store 16 bytes to. Note that there is no alignment requirement on this pointer since this intrinsic performs a 1-aligned store.