Struct core::arch::x86::__m512i

#[repr(simd)]
pub struct __m512i(_, _, _, _, _, _, _, _);

๐ฌThis is a nightly-only experimental API. (stdsimd #48556)

Available on x86 and (x86 or x86-64) only.

512-bit wide integer vector type, x86-specific

This type is the same as the __m512i type defined by Intel, representing a 512-bit SIMD register. Usage of this type typically corresponds to the avx512* and up target features for x86/x86_64.

Internally this type may be viewed as:

i8x64 - sixty-four i8 variables packed together

i16x32 - thirty-two i16 variables packed together

i32x16 - sixteen i32 variables packed together

i64x8 - eight i64 variables packed together

(as well as unsigned versions). Each intrinsic may interpret the internal bits differently, check the documentation of the intrinsic to see how itโs being used.

Note that this means that an instance of __m512i typically just means a โbag of bitsโ which is left up to interpretation at the point of use.