Trait core::borrow::Borrow

pub trait Borrow<Borrowed: ?Sized> {
    fn borrow(&self) -> &Borrowed;
}

借用数据的Trait。

A trait for borrowing data.

在 Rust 中，为不同的用例提供不同类型的表示是很常见的。 例如，可以通过诸如 Box 或 Rc 之类的指针类型，根据特定用途专门选择一个值的存储位置和管理。 除了可以与任何类型一起使用的这些通用包装器之外，某些类型还提供了可选方面，这些方面提供了潜在的昂贵功能。 这种类型的一个例子是 String，它增加了将字符串扩展到基本 str 的能力。 这需要为简单的、不可变的字符串保留不必要的附加信息。

In Rust, it is common to provide different representations of a type for different use cases. For instance, storage location and management for a value can be specifically chosen as appropriate for a particular use via pointer types such as Box<T> or Rc<T>. Beyond these generic wrappers that can be used with any type, some types provide optional facets providing potentially costly functionality. An example for such a type is String which adds the ability to extend a string to the basic str. This requires keeping additional information unnecessary for a simple, immutable string.

这些类型通过引用该数据的类型提供对基础数据的访问。 据说它们是“借来的”那种类型。 例如，Box 可以作为 T 借用，而 String 可以作为 str 借用。

These types provide access to the underlying data through references to the type of that data. They are said to be ‘borrowed as’ that type. For instance, a Box<T> can be borrowed as T while a String can be borrowed as str.

类型表示可以通过实现 Borrow 将它们作为某种类型 T 借用，在 trait 的借用方法中提供对 T 的引用。 一个类型可以作为几种不同的类型自由借用。 如果它希望可变地借用类型——允许修改底层数据，它可以另外实现 BorrowMut。

Types express that they can be borrowed as some type T by implementing Borrow<T>, providing a reference to a T in the trait’s borrow method. A type is free to borrow as several different types. If it wishes to mutably borrow as the type – allowing the underlying data to be modified, it can additionally implement BorrowMut<T>.

此外，在为其他Trait提供实现时，需要考虑它们是否应该与底层类型的行为相同，因为它们是作为底层类型的表示的结果。 当通用代码依赖于这些附加Trait实现的相同行为时，它通常使用 Borrow。 这些Trait可能会作为额外的Trait界限出现。

Further, when providing implementations for additional traits, it needs to be considered whether they should behave identical to those of the underlying type as a consequence of acting as a representation of that underlying type. Generic code typically uses Borrow<T> when it relies on the identical behavior of these additional trait implementations. These traits will likely appear as additional trait bounds.

特别是 Eq、Ord 和 Hash 对于借用和拥有的值必须是等价的：x.borrow() == y.borrow() 应该给出与 x == y 相同的结果。

In particular Eq, Ord and Hash must be equivalent for borrowed and owned values: x.borrow() == y.borrow() should give the same result as x == y.

如果泛型代码只需要适用于所有可以提供对相关类型 T 的引用的类型，那么使用 AsRef 通常会更好，因为更多类型可以安全地实现它。

If generic code merely needs to work for all types that can provide a reference to related type T, it is often better to use AsRef<T> as more types can safely implement it.

Examples

作为一个数据集合，HashMap 拥有键和值。 如果键的实际数据被包装在某种管理类型中，那么它应该仍然可以使用对键数据的引用来搜索值。 例如，如果键是字符串，那么它很可能与哈希映射一起存储为字符串，而应该可以使用 &str 进行搜索。 因此，insert 需要对 String 进行操作，而 get 需要能够使用 &str。

As a data collection, HashMap<K, V> owns both keys and values. If the key’s actual data is wrapped in a managing type of some kind, it should, however, still be possible to search for a value using a reference to the key’s data. For instance, if the key is a string, then it is likely stored with the hash map as a String, while it should be possible to search using a &str. Thus, insert needs to operate on a String while get needs to be able to use a &str.

稍微简化一下，HashMap 的相关部分如下所示：

Slightly simplified, the relevant parts of HashMap<K, V> look like this:

use std::borrow::Borrow;
use std::hash::Hash;

pub struct HashMap<K, V> {
    // fields omitted
}

impl<K, V> HashMap<K, V> {
    pub fn insert(&self, key: K, value: V) -> Option<V>
    where K: Hash + Eq
    {
        // ...
    }

    pub fn get<Q>(&self, k: &Q) -> Option<&V>
    where
        K: Borrow<Q>,
        Q: Hash + Eq + ?Sized
    {
        // ...
    }
}
Run

整个散列映射在键类型 K 上是通用的。因为这些键与散列映射一起存储，所以该类型必须拥有键的数据。 插入键值对时，给映射一个这样的K，需要找到正确的散列桶，并根据该K检查键是否已经存在。因此需要K：Hash + Eq。

The entire hash map is generic over a key type K. Because these keys are stored with the hash map, this type has to own the key’s data. When inserting a key-value pair, the map is given such a K and needs to find the correct hash bucket and check if the key is already present based on that K. It therefore requires K: Hash + Eq.

但是，在映射中搜索值时，必须提供对 K 的引用作为搜索的键，这将需要始终创建这样的拥有值。 对于字符串键，这意味着只需要创建一个字符串值来搜索只有 str 可用的情况。

When searching for a value in the map, however, having to provide a reference to a K as the key to search for would require to always create such an owned value. For string keys, this would mean a String value needs to be created just for the search for cases where only a str is available.

相反，get 方法对底层密钥数据的类型是通用的，在上面的方法签名中称为 Q。 它通过要求 K: Borrow 来声明 K 作为 Q 借用。 通过额外要求 Q: Hash + Eq，它表明要求 K 和 Q 具有产生相同结果的 Hash 和 Eq Trait的实现。

Instead, the get method is generic over the type of the underlying key data, called Q in the method signature above. It states that K borrows as a Q by requiring that K: Borrow<Q>. By additionally requiring Q: Hash + Eq, it signals the requirement that K and Q have implementations of the Hash and Eq traits that produce identical results.

get 的实现尤其依赖于 Hash 的相同实现，即通过在 Q 值上调用 Hash::hash 来确定键的哈希桶，即使它基于从 K 值计算的哈希值插入键。

The implementation of get relies in particular on identical implementations of Hash by determining the key’s hash bucket by calling Hash::hash on the Q value even though it inserted the key based on the hash value calculated from the K value.

因此，如果包含 Q 值的 K 产生与 Q 不同的散列，散列映射就会中断。例如，假设您有一个包含字符串但比较 ASCII 字母忽略其大小写的类型：

As a consequence, the hash map breaks if a K wrapping a Q value produces a different hash than Q. For instance, imagine you have a type that wraps a string but compares ASCII letters ignoring their case:

pub struct CaseInsensitiveString(String);

impl PartialEq for CaseInsensitiveString {
    fn eq(&self, other: &Self) -> bool {
        self.0.eq_ignore_ascii_case(&other.0)
    }
}

impl Eq for CaseInsensitiveString { }
Run

因为两个相等的值需要产生相同的哈希值，所以 Hash 的实现也需要忽略 ASCII 大小写：

Because two equal values need to produce the same hash value, the implementation of Hash needs to ignore ASCII case, too:

impl Hash for CaseInsensitiveString {
    fn hash<H: Hasher>(&self, state: &mut H) {
        for c in self.0.as_bytes() {
            c.to_ascii_lowercase().hash(state)
        }
    }
}
Run

CaseInsensitiveString 可以实现 Borrow 吗？ 它当然可以通过其包含的自有字符串提供对字符串切片的引用。 但是因为它的 Hash 实现不同，它的行为与 str 不同，因此实际上不能实现 Borrow。 如果它想允许其他人访问底层的 str，它可以通过 AsRef 来实现，它没有任何额外的要求。

Can CaseInsensitiveString implement Borrow<str>? It certainly can provide a reference to a string slice via its contained owned string. But because its Hash implementation differs, it behaves differently from str and therefore must not, in fact, implement Borrow<str>. If it wants to allow others access to the underlying str, it can do that via AsRef<str> which doesn’t carry any extra requirements.

Required methods

fn borrow(&self) -> &Borrowed

不可变地从拥有的价值中借用。

Immutably borrows from an owned value.

Examples

use std::borrow::Borrow;

fn check<T: Borrow<str>>(s: T) {
    assert_eq!("Hello", s.borrow());
}

let s = "Hello".to_string();

check(s);

let s = "Hello";

check(s);
Run