Struct core::cell::LazyCell

pub struct LazyCell<T, F = fn() -> T> { /* private fields */ }

๐ฌThis is a nightly-only experimental API. (lazy_cell #109736)

A value which is initialized on the first access.

For a thread-safe version of this struct, see std::sync::LazyLock.

Examples

#![feature(lazy_cell)]

use std::cell::LazyCell;

let lazy: LazyCell<i32> = LazyCell::new(|| {
    println!("initializing");
    92
});
println!("ready");
println!("{}", *lazy);
println!("{}", *lazy);

// Prints:
//   ready
//   initializing
//   92
//   92
Run

Implementations

impl<T, F: FnOnce() -> T> LazyCell<T, F>

pub const fn new(f: F) -> LazyCell<T, F>

๐ฌThis is a nightly-only experimental API. (lazy_cell #109736)

Creates a new lazy value with the given initializing function.

Examples

#![feature(lazy_cell)]

use std::cell::LazyCell;

let hello = "Hello, World!".to_string();

let lazy = LazyCell::new(|| hello.to_uppercase());

assert_eq!(&*lazy, "HELLO, WORLD!");
Run

pub fn into_inner(this: Self) -> Result<T, F>

๐ฌThis is a nightly-only experimental API. (lazy_cell_consume #109736)

Consumes this LazyCell returning the stored value.

Returns Ok(value) if Lazy is initialized and Err(f) otherwise.

Examples

#![feature(lazy_cell)]
#![feature(lazy_cell_consume)]

use std::cell::LazyCell;

let hello = "Hello, World!".to_string();

let lazy = LazyCell::new(|| hello.to_uppercase());

assert_eq!(&*lazy, "HELLO, WORLD!");
assert_eq!(LazyCell::into_inner(lazy).ok(), Some("HELLO, WORLD!".to_string()));
Run

pub fn force(this: &LazyCell<T, F>) -> &T

๐ฌThis is a nightly-only experimental API. (lazy_cell #109736)

Forces the evaluation of this lazy value and returns a reference to the result.

This is equivalent to the Deref impl, but is explicit.

Examples

#![feature(lazy_cell)]

use std::cell::LazyCell;

let lazy = LazyCell::new(|| 92);

assert_eq!(LazyCell::force(&lazy), &92);
assert_eq!(&*lazy, &92);
Run