Struct core::cell::RefStable

pub struct Ref<'b, T: 'b> {
    // some fields omitted
}

Wraps a borrowed reference to a value in a RefCell box. A wrapper type for an immutably borrowed value from a RefCell<T>.

See the module-level documentation for more.