Module core::char

Character manipulation.

For more details, see ::rustc_unicode::char (a.k.a. std::char)

Reexports

use iter::Iterator;

use mem::transmute;

use option::Option::{None, Some};

use option::Option;

use slice::SliceExt;

Structs

EscapeDefault
An iterator that yields the literal escape code of a char.

EscapeUnicode
Returns an iterator that yields the hexadecimal Unicode escape of a character, as chars.

Enums

EscapeDefaultState

EscapeUnicodeState

Constants

MAX
The highest valid code point a char can have.

MAX_ONE_B

MAX_THREE_B

MAX_TWO_B

TAG_CONT

TAG_FOUR_B

TAG_THREE_B

TAG_TWO_B

Traits

CharExt 	 [Unstable]

Functions

from_digit
Converts a digit in the given radix to a char.

from_u32
Converts a u32 to a char.

from_u32_unchecked
Converts a u32 to a char, ignoring validity.

encode_utf16_raw 	 [Unstable]
Encodes a raw u32 value as UTF-16 into the provided u16 buffer, and then returns the number of u16s written.

encode_utf8_raw 	 [Unstable]
Encodes a raw u32 value as UTF-8 into the provided byte buffer, and then returns the number of bytes written.