Module core::cmp

订购和比较的功能。

Functionality for ordering and comparison.

该模块包含用于排序和比较值的各种工具。 总之：

This module contains various tools for ordering and comparing values. In summary:

Eq 和 PartialEq 是允许您分别定义值之间的完全相等和部分相等的Trait。 实现它们会重载 == 和 != 运算符。

Eq and PartialEq are traits that allow you to define total and partial equality between values, respectively. Implementing them overloads the == and != operators.

Ord 和 PartialOrd 是允许您分别定义值之间的总排序和部分排序的Trait。 实现它们会重载 <、<=、> 和 >= 运算符。

Ord and PartialOrd are traits that allow you to define total and partial orderings between values, respectively. Implementing them overloads the <, <=, >, and >= operators.

Ordering 是 Ord 和 PartialOrd 的主要函数返回的枚举，描述了一个排序。

Ordering is an enum returned by the main functions of Ord and PartialOrd, and describes an ordering.

Reverse 是一个允许您轻松反转排序的结构。

Reverse is a struct that allows you to easily reverse an ordering.

max 和 min 是基于 Ord 构建的函数，允许您找到两个值的最大值或最小值。

max and min are functions that build off of Ord and allow you to find the maximum or minimum of two values.

有关更多详细信息，请参阅列表中每个项目的相应文档。

For more details, see the respective documentation of each item in the list.

Macros

Eq

Derive macro generating an impl of the trait Eq.

Ord

Derive macro generating an impl of the trait Ord.

PartialEq

Derive macro generating an impl of the trait PartialEq.

PartialOrd

Derive macro generating an impl of the trait PartialOrd.

Structs

Reverse

A helper struct for reverse ordering.

Enums

Ordering

An Ordering is the result of a comparison between two values.

Traits

Eq

Trait for equality comparisons which are equivalence relations.

Ord

Trait for types that form a total order.

PartialEq

Trait for equality comparisons which are partial equivalence relations.

PartialOrd

Trait for values that can be compared for a sort-order.

Functions

max

Compares and returns the maximum of two values.

max_by

Returns the maximum of two values with respect to the specified comparison function.

max_by_key

Returns the element that gives the maximum value from the specified function.

min

Compares and returns the minimum of two values.

min_by

Returns the minimum of two values with respect to the specified comparison function.

min_by_key

Returns the element that gives the minimum value from the specified function.