Trait core::cmp::OrdStable

pub trait Ord: Eq + PartialOrd<Self> {
    fn cmp(&self, other: &Self) -> Ordering;
}

Trait for types that form a total order.

An order is a total order if it is (for all a, b and c):

total and antisymmetric: exactly one of a < b, a == b or a > b is true; and

transitive, a < b and b < c implies a < c. The same must hold for both == and >.

Required Methods

fn cmp(&self, other: &Self) -> Ordering

This method returns an ordering between self and other values.

By convention, self.cmp(&other) returns the ordering matching the expression self <operator> other if true. For example:
fn main() { use std::cmp::Ordering::{Less, Equal, Greater}; assert_eq!( 5.cmp(&10), Less); // because 5 < 10 assert_eq!(10.cmp(&5), Greater); // because 10 > 5 assert_eq!( 5.cmp(&5), Equal); // because 5 == 5 }

use std::cmp::Ordering::{Less, Equal, Greater};

assert_eq!( 5.cmp(&10), Less);     // because 5 < 10
assert_eq!(10.cmp(&5),  Greater);  // because 10 > 5
assert_eq!( 5.cmp(&5),  Equal);    // because 5 == 5