Trait core::cmp::PartialEq

pub trait PartialEq<Rhs: ?Sized = Self> {
    fn eq(&self, other: &Rhs) -> bool;

    fn ne(&self, other: &Rhs) -> bool { ... }
}

等价比较的Trait，它是部分等价关系。

Trait for equality comparisons which are partial equivalence relations.

x.eq(y) 也可以写成 x == y，而 x.ne(y) 可以写成 x != y。 我们在本文档的其余部分使用更易于阅读的中缀表示法。

x.eq(y) can also be written x == y, and x.ne(y) can be written x != y. We use the easier-to-read infix notation in the remainder of this documentation.

对于不具有完全等价关系的类型，此Trait允许部分相等。 例如，在浮点数中 NaN != NaN，所以浮点类型实现 PartialEq 但不是 Eq。

This trait allows for partial equality, for types that do not have a full equivalence relation. For example, in floating point numbers NaN != NaN, so floating point types implement PartialEq but not Eq.

实现必须确保 eq 和 ne 彼此一致：

Implementations must ensure that eq and ne are consistent with each other:

a != b 当且仅当 !(a == b) （由默认实现确保）。

a != b if and only if !(a == b) (ensured by the default implementation).

如果还为 Self 和 Rhs 实现了 PartialOrd 或 Ord，则它们的方法也必须与 PartialEq 一致（有关确切要求，请参阅这些Trait的文档）。 通过派生一些Trait并手动实现其他Trait，很容易意外地使他们不同意。

If PartialOrd or Ord are also implemented for Self and Rhs, their methods must also be consistent with PartialEq (see the documentation of those traits for the exact requirements). It’s easy to accidentally make them disagree by deriving some of the traits and manually implementing others.

等式关系 == 必须满足以下条件（对于类型 A、B、C 的所有 a、b、c）：

The equality relation == must satisfy the following conditions (for all a, b, c of type A, B, C):

对称：如果 A：PartialEq 和 B：PartialEq，则 a == b 意味着 b == a； 和

Symmetric: if A: PartialEq<B> and B: PartialEq<A>, then a == b implies b == a; and

传递：如果 A：PartialEq 和 B：PartialEq 和 A：PartialEq，则 a == b 和 b == c 意味着 a == c。

Transitive: if A: PartialEq<B> and B: PartialEq<C> and A: PartialEq<C>, then a == b and b == c implies a == c.

请注意，B: PartialEq（对称）和 A:PartialEq（传递）impls 不是强制存在的，但只要它们确实存在，这些要求就适用。

Note that the B: PartialEq<A> (symmetric) and A: PartialEq<C> (transitive) impls are not forced to exist, but these requirements apply whenever they do exist.

Derivable

此Trait可与#[derive] 一起使用。 在结构上派生时，如果所有字段相等，则两个实例相等，如果任何字段不相等，则不相等。 在枚举上派生时，每个变体都等于自身，而不等于其他变体。

This trait can be used with #[derive]. When derived on structs, two instances are equal if all fields are equal, and not equal if any fields are not equal. When derived on enums, each variant is equal to itself and not equal to the other variants.

How can I implement PartialEq?

一个域的示例实现，其中两本书如果 ISBN 匹配则被视为同一本书，即使格式不同：

An example implementation for a domain in which two books are considered the same book if their ISBN matches, even if the formats differ:

enum BookFormat {
    Paperback,
    Hardback,
    Ebook,
}

struct Book {
    isbn: i32,
    format: BookFormat,
}

impl PartialEq for Book {
    fn eq(&self, other: &Self) -> bool {
        self.isbn == other.isbn
    }
}

let b1 = Book { isbn: 3, format: BookFormat::Paperback };
let b2 = Book { isbn: 3, format: BookFormat::Ebook };
let b3 = Book { isbn: 10, format: BookFormat::Paperback };

assert!(b1 == b2);
assert!(b1 != b3);
Run

How can I compare two different types?

您可以比较的类型由 PartialEq 的类型参数控制。 例如，让我们稍微调整一下我们之前的代码：

The type you can compare with is controlled by PartialEq’s type parameter. For example, let’s tweak our previous code a bit:

// The derive implements <BookFormat> == <BookFormat> comparisons
#[derive(PartialEq)]
enum BookFormat {
    Paperback,
    Hardback,
    Ebook,
}

struct Book {
    isbn: i32,
    format: BookFormat,
}

// Implement <Book> == <BookFormat> comparisons
impl PartialEq<BookFormat> for Book {
    fn eq(&self, other: &BookFormat) -> bool {
        self.format == *other
    }
}

// Implement <BookFormat> == <Book> comparisons
impl PartialEq<Book> for BookFormat {
    fn eq(&self, other: &Book) -> bool {
        *self == other.format
    }
}

let b1 = Book { isbn: 3, format: BookFormat::Paperback };

assert!(b1 == BookFormat::Paperback);
assert!(BookFormat::Ebook != b1);
Run

通过将 Book 的 impl PartialEq 更改为 Book 的 impl PartialEq，我们允许将 BookFormats 与 Books 进行比较。

By changing impl PartialEq for Book to impl PartialEq<BookFormat> for Book, we allow BookFormats to be compared with Books.

像上面这样忽略结构的某些字段的比较可能很危险。 它很容易导致意外违反部分等价关系的要求。 例如，如果我们为 BookFormat 保留了 PartialEq 的上述实现，并为 Book 添加了 PartialEq 的实现（通过#[derive] 或通过第一个示例中的手动实现），那么结果将违反 传递性：

A comparison like the one above, which ignores some fields of the struct, can be dangerous. It can easily lead to an unintended violation of the requirements for a partial equivalence relation. For example, if we kept the above implementation of PartialEq<Book> for BookFormat and added an implementation of PartialEq<Book> for Book (either via a #[derive] or via the manual implementation from the first example) then the result would violate transitivity:

#[derive(PartialEq)]
enum BookFormat {
    Paperback,
    Hardback,
    Ebook,
}

#[derive(PartialEq)]
struct Book {
    isbn: i32,
    format: BookFormat,
}

impl PartialEq<BookFormat> for Book {
    fn eq(&self, other: &BookFormat) -> bool {
        self.format == *other
    }
}

impl PartialEq<Book> for BookFormat {
    fn eq(&self, other: &Book) -> bool {
        *self == other.format
    }
}

fn main() {
    let b1 = Book { isbn: 1, format: BookFormat::Paperback };
    let b2 = Book { isbn: 2, format: BookFormat::Paperback };

    assert!(b1 == BookFormat::Paperback);
    assert!(BookFormat::Paperback == b2);

    // The following should hold by transitivity but doesn't.
    assert!(b1 == b2); // <-- PANICS
}
Run

Examples

let x: u32 = 0;
let y: u32 = 1;

assert_eq!(x == y, false);
assert_eq!(x.eq(&y), false);
Run

Required methods

fn eq(&self, other: &Rhs) -> bool

此方法测试 self 和其他值是否相等，并由 == 使用。

This method tests for self and other values to be equal, and is used by ==.

Provided methods

fn ne(&self, other: &Rhs) -> bool

此方法测试!=。

This method tests for !=.