Enum core::convert::Infallible

pub enum Infallible {}

🔬 This is a nightly-only experimental API. (try_from #33417)

A type used as the error type for implementations of fallible conversion traits in cases where conversions cannot actually fail.

Because Infallible has no variants, a value of this type can never exist. It is used only to satisfy trait signatures that expect an error type, and signals to both the compiler and the user that the error case is impossible.