Trait core::convert::AsMut

pub trait AsMut<T: ?Sized> {
    fn as_mut(&mut self) -> &mut T;
}

用于进行廉价的可变到可变引用转换。

Used to do a cheap mutable-to-mutable reference conversion.

此Trait类似于“AsRef”，但用于在可变引用之间进行转换。 如果您需要进行昂贵的转换，最好使用 `&mut T` 类型实现“From”或编写自定义函数。

This trait is similar to AsRef but used for converting between mutable references. If you need to do a costly conversion it is better to implement From with type &mut T or write a custom function.

注意：这个特性不能失败。 如果转换失败，请使用返回“Option”或“Result”的专用方法。

Note: This trait must not fail. If the conversion can fail, use a dedicated method which returns an Option<T> or a Result<T, E>.

Generic Implementations

如果内部类型是可变引用，则 `AsMut` 自动取消引用（例如：如果 `foo` 具有 `&mut Foo` 或 `&mut &mut Foo` 类型，则 `foo.as_mut()` 的工作方式相同）

AsMut auto-dereferences if the inner type is a mutable reference (e.g.: foo.as_mut() will work the same if foo has type &mut Foo or &mut &mut Foo)

Examples

使用 `AsMut` 作为泛型函数的Trait绑定，我们可以接受所有可以转换为 `&mut T` 类型的可变引用。 因为“Box”实现了`AsMut`，我们可以编写一个函数`add_one`，它接受所有可以转换为`&mut u64`的参数。 因为 "Box" 实现了 `AsMut`，所以 `add_one` 也接受 `&mut Box` 类型的参数：

Using AsMut as trait bound for a generic function we can accept all mutable references that can be converted to type &mut T. Because Box<T> implements AsMut<T> we can write a function add_one that takes all arguments that can be converted to &mut u64. Because Box<T> implements AsMut<T>, add_one accepts arguments of type &mut Box<u64> as well:

fn add_one<T: AsMut<u64>>(num: &mut T) {
    *num.as_mut() += 1;
}

let mut boxed_num = Box::new(0);
add_one(&mut boxed_num);
assert_eq!(*boxed_num, 1);
Run

Required methods

fn as_mut(&mut self) -> &mut T

执行转换。

Performs the conversion.