Trait core::convert::AsMut

pub trait AsMut<T: ?Sized> {
    fn as_mut(&mut self) -> &mut T;
}

A cheap, mutable reference-to-mutable reference conversion.

This trait is similar to AsRef but used for converting between mutable references.

Note: this trait must not fail. If the conversion can fail, use a dedicated method which returns an Option<T> or a Result<T, E>.

Generic Implementations

AsMut auto-dereferences if the inner type is a mutable reference (e.g.: foo.as_mut() will work the same if foo has type &mut Foo or &mut &mut Foo)

Examples

Box<T> implements AsMut<T>:

fn add_one<T: AsMut<u64>>(num: &mut T) {
    *num.as_mut() += 1;
}

let mut boxed_num = Box::new(0);
add_one(&mut boxed_num);
assert_eq!(*boxed_num, 1);Run

Required methods

fn as_mut(&mut self) -> &mut T

Performs the conversion.

Loading content...