Struct core::ctx::Ctx

pub struct Ctx {
    types: HashMap<Ident, Rc<TypAbs>>,
    ctors: HashMap<Ident, Rc<Ctor>>,
    dtors: HashMap<Ident, Rc<Dtor>>,
    type_for_xtor: HashMap<Ident, Ident>,
    xtors_in_type: HashMap<Ident, HashSet<Ident>>,
    bound: Vec<Vec<Rc<Exp>>>,
}

Fields
types: HashMap<Ident, Rc<TypAbs>>ctors: HashMap<Ident, Rc<Ctor>>dtors: HashMap<Ident, Rc<Dtor>>type_for_xtor: HashMap<Ident, Ident>xtors_in_type: HashMap<Ident, HashSet<Ident>>bound: Vec<Vec<Rc<Exp>>>
Implementations

impl Ctx

pub fn empty() -> Self

pub fn build(prg: &Prg) -> Self

pub fn bound(&self, idx: Idx) -> Rc<Exp>

pub fn typ_for_xtor(&self, name: &Ident) -> &Ident

pub fn xtors_for_typ(&self, name: &Ident) -> &HashSet<Ident>

pub fn typ(&self, name: &Ident) -> Rc<TypAbs>

pub fn ctor(&self, name: &Ident) -> Rc<Ctor>

pub fn dtor(&self, name: &Ident) -> Rc<Dtor>

pub fn bind<T, F>(&mut self, typ: Rc<Exp>, f: F) -> T where
F: Fn(&mut Ctx) -> T,

Bind a single type

pub fn bind_fold<T, I: Iterator<Item = T>, O1, O2, F1, F2>(
&mut self,
iter: I,
acc: O1,
f_acc: F1,
f_inner: F2
) -> O2 where
T: Typed,
F1: Fn(&mut Ctx, O1, T) -> O1,
F2: FnOnce(&mut Ctx, O1) -> O2,

Bind an iterator iter of types

Fold the iterator and consume the result under the inner context with all binders in scope.

This is used for checking telescopes.

iter - An iterator of binders implementing Typed.

acc - Accumulator for folding the iterator

f_acc - Accumulator function run for each binder

f_inner - Inner function computing the final result under the context of all binders

fn map<F>(&self, f: F) -> Self where
F: Fn(&Rc<Exp>) -> Rc<Exp>,

fn shift(&mut self, by: (isize, isize))

fn shift_at_lvl(&mut self, lvl: usize, by: (isize, isize))

fn level_inc_fst(&mut self)

Increment the first component of the current De-Bruijn level

fn level_dec_fst(&mut self)

Decrement the first component of the current De-Bruijn level

fn push(&mut self, typ: Rc<Exp>)

Push a binder contained in a binder list, incrementing the second dimension of the current De Bruijn level

fn pop(&mut self)

Push a binder contained in a binder list, decrementing the second dimension of the current De Bruijn level