Module core::default

The Default trait for types with a default value.

Traits

Default

A trait for giving a type a useful default value.

Derive Macros

Default

Derive macro generating an impl of the trait Default.