Struct core::future::Ready

pub struct Ready<T>(_);

A future that is immediately ready with a value.

This struct is created by ready(). See its documentation for more.

Implementations

impl<T> Ready<T>

pub fn into_inner(self) -> T

🔬This is a nightly-only experimental API. (ready_into_inner #101196)

Consumes the Ready, returning the wrapped value.

Panics

Will panic if this Ready was already polled to completion.

Examples

#![feature(ready_into_inner)]
use std::future;

let a = future::ready(1);
assert_eq!(a.into_inner(), 1);
Run