Struct core::hash::BuildHasherDefault

pub struct BuildHasherDefault<H>(_);

Used to create a default BuildHasher instance for types that implement Hasher and Default.

BuildHasherDefault<H> can be used when a type H implements Hasher and Default, and you need a corresponding BuildHasher instance, but none is defined.

Any BuildHasherDefault is zero-sized. It can be created with default. When using BuildHasherDefault with HashMap or HashSet, this doesn’t need to be done, since they implement appropriate Default instances themselves.

Examples

Using BuildHasherDefault to specify a custom BuildHasher for HashMap:

use std::collections::HashMap;
use std::hash::{BuildHasherDefault, Hasher};

#[derive(Default)]
struct MyHasher;

impl Hasher for MyHasher {
    fn write(&mut self, bytes: &[u8]) {
        // Your hashing algorithm goes here!
       unimplemented!()
    }

    fn finish(&self) -> u64 {
        // Your hashing algorithm goes here!
        unimplemented!()
    }
}

type MyBuildHasher = BuildHasherDefault<MyHasher>;

let hash_map = HashMap::<u32, u32, MyBuildHasher>::default();
Run