Trait core::hash::Hasher

pub trait Hasher {

    fn finish(&self) -> u64;

    fn write(&mut self, bytes: &[u8]);

    fn write_u8(&mut self, i: u8) { ... }

    fn write_u16(&mut self, i: u16) { ... }

    fn write_u32(&mut self, i: u32) { ... }

    fn write_u64(&mut self, i: u64) { ... }

    fn write_u128(&mut self, i: u128) { ... }

    fn write_usize(&mut self, i: usize) { ... }

    fn write_i8(&mut self, i: i8) { ... }

    fn write_i16(&mut self, i: i16) { ... }

    fn write_i32(&mut self, i: i32) { ... }

    fn write_i64(&mut self, i: i64) { ... }

    fn write_i128(&mut self, i: i128) { ... }

    fn write_isize(&mut self, i: isize) { ... }

}

散列任意字节流的Trait。

A trait for hashing an arbitrary stream of bytes.

Hasher 的实例通常表示在散列数据时更改的状态。

Instances of Hasher usually represent state that is changed while hashing data.

Hasher 提供了一个相当基本的接口，用于检索生成的哈希（使用完成），并将整数和字节切片写入实例（使用 write 和 write_u8 等）。 大多数时候，Hasher 实例与 Hash trait 结合使用。

Hasher provides a fairly basic interface for retrieving the generated hash (with finish), and writing integers as well as slices of bytes into an instance (with write and write_u8 etc.). Most of the time, Hasher instances are used in conjunction with the Hash trait.

这个特性没有假设各种 write_* 方法是如何定义的，并且 Hash 的实现不应该假设它们以一种或另一种方式工作。 例如，您不能假设一个 write_u32 调用等同于四个 write_u8 调用。

This trait makes no assumptions about how the various write_* methods are defined and implementations of Hash should not assume that they work one way or another. You cannot assume, for example, that a write_u32 call is equivalent to four calls of write_u8.

Examples

use std::collections::hash_map::DefaultHasher;
use std::hash::Hasher;

let mut hasher = DefaultHasher::new();

hasher.write_u32(1989);
hasher.write_u8(11);
hasher.write_u8(9);
hasher.write(b"Huh?");

println!("Hash is {:x}!", hasher.finish());
Run

Required methods

fn finish(&self) -> u64

返回到目前为止写入的值的哈希值。

Returns the hash value for the values written so far.

尽管它的名字，该方法不会重置哈希器的内部状态。 其他写入将从当前值继续。 如果您需要开始一个新的哈希值，则必须创建一个新的哈希。

Despite its name, the method does not reset the hasher’s internal state. Additional writes will continue from the current value. If you need to start a fresh hash value, you will have to create a new hasher.

Examples

use std::collections::hash_map::DefaultHasher;
use std::hash::Hasher;

let mut hasher = DefaultHasher::new();
hasher.write(b"Cool!");

println!("Hash is {:x}!", hasher.finish());
Run

fn write(&mut self, bytes: &[u8])

将一些数据写入此哈希器。

Writes some data into this Hasher.

Examples

use std::collections::hash_map::DefaultHasher;
use std::hash::Hasher;

let mut hasher = DefaultHasher::new();
let data = [0x01, 0x23, 0x45, 0x67, 0x89, 0xab, 0xcd, 0xef];

hasher.write(&data);

println!("Hash is {:x}!", hasher.finish());
Run

Provided methods

fn write_u8(&mut self, i: u8)

将单个 u8 写入此哈希器。

Writes a single u8 into this hasher.

fn write_u16(&mut self, i: u16)

将单个 u16 写入此哈希器。

Writes a single u16 into this hasher.

fn write_u32(&mut self, i: u32)

将单个 u32 写入此哈希器。

Writes a single u32 into this hasher.

fn write_u64(&mut self, i: u64)

将单个 u64 写入此哈希器。

Writes a single u64 into this hasher.

fn write_u128(&mut self, i: u128)

将单个 u128 写入此哈希器。

Writes a single u128 into this hasher.

fn write_usize(&mut self, i: usize)

将单个 usize 写入此哈希器。

Writes a single usize into this hasher.

fn write_i8(&mut self, i: i8)

将单个 i8 写入此哈希器。

Writes a single i8 into this hasher.

fn write_i16(&mut self, i: i16)

将单个 i16 写入此哈希器。

Writes a single i16 into this hasher.

fn write_i32(&mut self, i: i32)

将单个 i32 写入此哈希器。

Writes a single i32 into this hasher.

fn write_i64(&mut self, i: i64)

将单个 i64 写入此哈希器。

Writes a single i64 into this hasher.

fn write_i128(&mut self, i: i128)

将单个 i128 写入此哈希器。

Writes a single i128 into this hasher.

fn write_isize(&mut self, i: isize)

将单个 isize 写入此哈希器。

Writes a single isize into this hasher.