Trait core::iter::DoubleEndedIteratorStable

pub trait DoubleEndedIterator: Iterator {
    fn next_back(&mut self) -> Option<Self::Item>;
}

A range iterator able to yield elements from both ends

A DoubleEndedIterator can be thought of as a deque in that next() and next_back() exhaust elements from the same range, and do not work independently of each other.

Required Methods

fn next_back(&mut self) -> Option<Self::Item>

Yield an element from the end of the range, returning None if the range is empty.