Trait core::iter::FusedIterator

pub trait FusedIterator: Iterator { }

Unstable (fused #35602)

An iterator that always continues to yield None when exhausted.

Calling next on a fused iterator that has returned None once is guaranteed to return None again. This trait is should be implemented by all iterators that behave this way because it allows for some significant optimizations.

Note: In general, you should not use FusedIterator in generic bounds if you need a fused iterator. Instead, you should just call Iterator::fuse() on the iterator. If the iterator is already fused, the additional Fuse wrapper will be a no-op with no performance penalty.