core::iter

Trait Iterator
·

pub trait Iterator {
    type Item;

    // Required method
    fn next(&mut self) -> Option<Self::Item>;

    // Provided methods
    fn next_chunk<const N: usize>(
        &mut self,
    ) -> Result<[Self::Item; N], IntoIter<Self::Item, N>>
       where Self: Sized { ... }
    fn size_hint(&self) -> (usize, Option<usize>) { ... }
    fn count(self) -> usize
       where Self: Sized { ... }
    fn last(self) -> Option<Self::Item>
       where Self: Sized { ... }
    fn advance_by(&mut self, n: usize) -> Result<(), NonZero<usize>> { ... }
    fn nth(&mut self, n: usize) -> Option<Self::Item> { ... }
    fn step_by(self, step: usize) -> StepBy<Self> 
       where Self: Sized { ... }
    fn chain<U>(self, other: U) -> Chain<Self, U::IntoIter> 
       where Self: Sized,
             U: IntoIterator<Item = Self::Item> { ... }
    fn zip<U>(self, other: U) -> Zip<Self, U::IntoIter> 
       where Self: Sized,
             U: IntoIterator { ... }
    fn intersperse(self, separator: Self::Item) -> Intersperse<Self> 
       where Self: Sized,
             Self::Item: Clone { ... }
    fn intersperse_with<G>(self, separator: G) -> IntersperseWith<Self, G> 
       where Self: Sized,
             G: FnMut() -> Self::Item { ... }
    fn map<B, F>(self, f: F) -> Map<Self, F> 
       where Self: Sized,
             F: FnMut(Self::Item) -> B { ... }
    fn for_each<F>(self, f: F)
       where Self: Sized,
             F: FnMut(Self::Item) { ... }
    fn filter<P>(self, predicate: P) -> Filter<Self, P> 
       where Self: Sized,
             P: FnMut(&Self::Item) -> bool { ... }
    fn filter_map<B, F>(self, f: F) -> FilterMap<Self, F> 
       where Self: Sized,
             F: FnMut(Self::Item) -> Option<B> { ... }
    fn enumerate(self) -> Enumerate<Self> 
       where Self: Sized { ... }
    fn peekable(self) -> Peekable<Self> 
       where Self: Sized { ... }
    fn skip_while<P>(self, predicate: P) -> SkipWhile<Self, P> 
       where Self: Sized,
             P: FnMut(&Self::Item) -> bool { ... }
    fn take_while<P>(self, predicate: P) -> TakeWhile<Self, P> 
       where Self: Sized,
             P: FnMut(&Self::Item) -> bool { ... }
    fn map_while<B, P>(self, predicate: P) -> MapWhile<Self, P> 
       where Self: Sized,
             P: FnMut(Self::Item) -> Option<B> { ... }
    fn skip(self, n: usize) -> Skip<Self> 
       where Self: Sized { ... }
    fn take(self, n: usize) -> Take<Self> 
       where Self: Sized { ... }
    fn scan<St, B, F>(self, initial_state: St, f: F) -> Scan<Self, St, F> 
       where Self: Sized,
             F: FnMut(&mut St, Self::Item) -> Option<B> { ... }
    fn flat_map<U, F>(self, f: F) -> FlatMap<Self, U, F> 
       where Self: Sized,
             U: IntoIterator,
             F: FnMut(Self::Item) -> U { ... }
    fn flatten(self) -> Flatten<Self> 
       where Self: Sized,
             Self::Item: IntoIterator { ... }
    fn map_windows<F, R, const N: usize>(self, f: F) -> MapWindows<Self, F, N> 
       where Self: Sized,
             F: FnMut(&[Self::Item; N]) -> R { ... }
    fn fuse(self) -> Fuse<Self> 
       where Self: Sized { ... }
    fn inspect<F>(self, f: F) -> Inspect<Self, F> 
       where Self: Sized,
             F: FnMut(&Self::Item) { ... }
    fn by_ref(&mut self) -> &mut Self
       where Self: Sized { ... }
    fn collect<B: FromIterator<Self::Item>>(self) -> B
       where Self: Sized { ... }
    fn try_collect<B>(
        &mut self,
    ) -> <Self::Item::Residual as Residual<B>>::TryType
       where Self: Sized,
             Self::Item: Try<Residual: Residual<B>>,
             B: FromIterator<<Self::Item as Try>::Output> { ... }
    fn collect_into<E: Extend<Self::Item>>(self, collection: &mut E) -> &mut E
       where Self: Sized { ... }
    fn partition<B, F>(self, f: F) -> (B, B)
       where Self: Sized,
             B: Default + Extend<Self::Item>,
             F: FnMut(&Self::Item) -> bool { ... }
    fn partition_in_place<'a, T: 'a, P>(self, predicate: P) -> usize
       where Self: Sized + DoubleEndedIterator<Item = &'a mut T>,
             P: FnMut(&T) -> bool { ... }
    fn is_partitioned<P>(self, predicate: P) -> bool
       where Self: Sized,
             P: FnMut(Self::Item) -> bool { ... }
    fn try_fold<B, F, R>(&mut self, init: B, f: F) -> R
       where Self: Sized,
             F: FnMut(B, Self::Item) -> R,
             R: Try<Output = B> { ... }
    fn try_for_each<F, R>(&mut self, f: F) -> R
       where Self: Sized,
             F: FnMut(Self::Item) -> R,
             R: Try<Output = ()> { ... }
    fn fold<B, F>(self, init: B, f: F) -> B
       where Self: Sized,
             F: FnMut(B, Self::Item) -> B { ... }
    fn reduce<F>(self, f: F) -> Option<Self::Item>
       where Self: Sized,
             F: FnMut(Self::Item, Self::Item) -> Self::Item { ... }
    fn try_reduce<R>(
        &mut self,
        f: impl FnMut(Self::Item, Self::Item) -> R,
    ) -> <R::Residual as Residual<Option<R::Output>>>::TryType
       where Self: Sized,
             R: Try<Output = Self::Item, Residual: Residual<Option<Self::Item>>> { ... }
    fn all<F>(&mut self, f: F) -> bool
       where Self: Sized,
             F: FnMut(Self::Item) -> bool { ... }
    fn any<F>(&mut self, f: F) -> bool
       where Self: Sized,
             F: FnMut(Self::Item) -> bool { ... }
    fn find<P>(&mut self, predicate: P) -> Option<Self::Item>
       where Self: Sized,
             P: FnMut(&Self::Item) -> bool { ... }
    fn find_map<B, F>(&mut self, f: F) -> Option<B>
       where Self: Sized,
             F: FnMut(Self::Item) -> Option<B> { ... }
    fn try_find<R>(
        &mut self,
        f: impl FnMut(&Self::Item) -> R,
    ) -> <R::Residual as Residual<Option<Self::Item>>>::TryType
       where Self: Sized,
             R: Try<Output = bool, Residual: Residual<Option<Self::Item>>> { ... }
    fn position<P>(&mut self, predicate: P) -> Option<usize>
       where Self: Sized,
             P: FnMut(Self::Item) -> bool { ... }
    fn rposition<P>(&mut self, predicate: P) -> Option<usize>
       where P: FnMut(Self::Item) -> bool,
             Self: Sized + ExactSizeIterator + DoubleEndedIterator { ... }
    fn max(self) -> Option<Self::Item>
       where Self: Sized,
             Self::Item: Ord { ... }
    fn min(self) -> Option<Self::Item>
       where Self: Sized,
             Self::Item: Ord { ... }
    fn max_by_key<B: Ord, F>(self, f: F) -> Option<Self::Item>
       where Self: Sized,
             F: FnMut(&Self::Item) -> B { ... }
    fn max_by<F>(self, compare: F) -> Option<Self::Item>
       where Self: Sized,
             F: FnMut(&Self::Item, &Self::Item) -> Ordering { ... }
    fn min_by_key<B: Ord, F>(self, f: F) -> Option<Self::Item>
       where Self: Sized,
             F: FnMut(&Self::Item) -> B { ... }
    fn min_by<F>(self, compare: F) -> Option<Self::Item>
       where Self: Sized,
             F: FnMut(&Self::Item, &Self::Item) -> Ordering { ... }
    fn rev(self) -> Rev<Self> 
       where Self: Sized + DoubleEndedIterator { ... }
    fn unzip<A, B, FromA, FromB>(self) -> (FromA, FromB)
       where FromA: Default + Extend<A>,
             FromB: Default + Extend<B>,
             Self: Sized + Iterator<Item = (A, B)> { ... }
    fn copied<'a, T>(self) -> Copied<Self> 
       where T: Copy + 'a,
             Self: Sized + Iterator<Item = &'a T> { ... }
    fn cloned<'a, T>(self) -> Cloned<Self> 
       where T: Clone + 'a,
             Self: Sized + Iterator<Item = &'a T> { ... }
    fn cycle(self) -> Cycle<Self> 
       where Self: Sized + Clone { ... }
    fn array_chunks<const N: usize>(self) -> ArrayChunks<Self, N> 
       where Self: Sized { ... }
    fn sum<S>(self) -> S
       where Self: Sized,
             S: Sum<Self::Item> { ... }
    fn product<P>(self) -> P
       where Self: Sized,
             P: Product<Self::Item> { ... }
    fn cmp<I>(self, other: I) -> Ordering
       where I: IntoIterator<Item = Self::Item>,
             Self::Item: Ord,
             Self: Sized { ... }
    fn cmp_by<I, F>(self, other: I, cmp: F) -> Ordering
       where Self: Sized,
             I: IntoIterator,
             F: FnMut(Self::Item, I::Item) -> Ordering { ... }
    fn partial_cmp<I>(self, other: I) -> Option<Ordering>
       where I: IntoIterator,
             Self::Item: PartialOrd<I::Item>,
             Self: Sized { ... }
    fn partial_cmp_by<I, F>(self, other: I, partial_cmp: F) -> Option<Ordering>
       where Self: Sized,
             I: IntoIterator,
             F: FnMut(Self::Item, I::Item) -> Option<Ordering> { ... }
    fn eq<I>(self, other: I) -> bool
       where I: IntoIterator,
             Self::Item: PartialEq<I::Item>,
             Self: Sized { ... }
    fn eq_by<I, F>(self, other: I, eq: F) -> bool
       where Self: Sized,
             I: IntoIterator,
             F: FnMut(Self::Item, I::Item) -> bool { ... }
    fn ne<I>(self, other: I) -> bool
       where I: IntoIterator,
             Self::Item: PartialEq<I::Item>,
             Self: Sized { ... }
    fn lt<I>(self, other: I) -> bool
       where I: IntoIterator,
             Self::Item: PartialOrd<I::Item>,
             Self: Sized { ... }
    fn le<I>(self, other: I) -> bool
       where I: IntoIterator,
             Self::Item: PartialOrd<I::Item>,
             Self: Sized { ... }
    fn gt<I>(self, other: I) -> bool
       where I: IntoIterator,
             Self::Item: PartialOrd<I::Item>,
             Self: Sized { ... }
    fn ge<I>(self, other: I) -> bool
       where I: IntoIterator,
             Self::Item: PartialOrd<I::Item>,
             Self: Sized { ... }
    fn is_sorted(self) -> bool
       where Self: Sized,
             Self::Item: PartialOrd { ... }
    fn is_sorted_by<F>(self, compare: F) -> bool
       where Self: Sized,
             F: FnMut(&Self::Item, &Self::Item) -> bool { ... }
    fn is_sorted_by_key<F, K>(self, f: F) -> bool
       where Self: Sized,
             F: FnMut(Self::Item) -> K,
             K: PartialOrd { ... }

}

A trait for dealing with iterators.

This is the main iterator trait. For more about the concept of iterators generally, please see the module-level documentation. In particular, you may want to know how to implement Iterator.

Required Associated Types

type Item

The type of the elements being iterated over.

Required Methods

fn next(&mut self) -> Option<Self::Item>

Advances the iterator and returns the next value.

Returns None when iteration is finished. Individual iterator implementations may choose to resume iteration, and so calling next() again may or may not eventually start returning Some(Item) again at some point.

Examples

let a = [1, 2, 3];

let mut iter = a.into_iter();

// A call to next() returns the next value...
assert_eq!(Some(1), iter.next());
assert_eq!(Some(2), iter.next());
assert_eq!(Some(3), iter.next());

// ... and then None once it's over.
assert_eq!(None, iter.next());

// More calls may or may not return `None`. Here, they always will.
assert_eq!(None, iter.next());
assert_eq!(None, iter.next());

Provided Methods

fn next_chunk<const N: usize>( &mut self, ) -> Result<[Self::Item; N], IntoIter<Self::Item, N>>
where Self: Sized,

🔬This is a nightly-only experimental API. (iter_next_chunk #98326)

Advances the iterator and returns an array containing the next N values.

If there are not enough elements to fill the array then Err is returned containing an iterator over the remaining elements.

Examples

Basic usage:

#![feature(iter_next_chunk)]

let mut iter = "lorem".chars();

assert_eq!(iter.next_chunk().unwrap(), ['l', 'o']);              // N is inferred as 2
assert_eq!(iter.next_chunk().unwrap(), ['r', 'e', 'm']);         // N is inferred as 3
assert_eq!(iter.next_chunk::<4>().unwrap_err().as_slice(), &[]); // N is explicitly 4

Split a string and get the first three items.

#![feature(iter_next_chunk)]

let quote = "not all those who wander are lost";
let [first, second, third] = quote.split_whitespace().next_chunk().unwrap();
assert_eq!(first, "not");
assert_eq!(second, "all");
assert_eq!(third, "those");

fn size_hint(&self) -> (usize, Option<usize>)

Returns the bounds on the remaining length of the iterator.

Specifically, size_hint() returns a tuple where the first element is the lower bound, and the second element is the upper bound.

The second half of the tuple that is returned is an Option<usize>. A None here means that either there is no known upper bound, or the upper bound is larger than usize.

Implementation notes

It is not enforced that an iterator implementation yields the declared number of elements. A buggy iterator may yield less than the lower bound or more than the upper bound of elements.

size_hint() is primarily intended to be used for optimizations such as reserving space for the elements of the iterator, but must not be trusted to e.g., omit bounds checks in unsafe code. An incorrect implementation of size_hint() should not lead to memory safety violations.

That said, the implementation should provide a correct estimation, because otherwise it would be a violation of the trait’s protocol.

The default implementation returns (0, None) which is correct for any iterator.

Examples

Basic usage:

let a = [1, 2, 3];
let mut iter = a.iter();

assert_eq!((3, Some(3)), iter.size_hint());
let _ = iter.next();
assert_eq!((2, Some(2)), iter.size_hint());

A more complex example:

// The even numbers in the range of zero to nine.
let iter = (0..10).filter(|x| x % 2 == 0);

// We might iterate from zero to ten times. Knowing that it's five
// exactly wouldn't be possible without executing filter().
assert_eq!((0, Some(10)), iter.size_hint());

// Let's add five more numbers with chain()
let iter = (0..10).filter(|x| x % 2 == 0).chain(15..20);

// now both bounds are increased by five
assert_eq!((5, Some(15)), iter.size_hint());

Returning None for an upper bound:

// an infinite iterator has no upper bound
// and the maximum possible lower bound
let iter = 0..;

assert_eq!((usize::MAX, None), iter.size_hint());

fn count(self) -> usize
where Self: Sized,

Consumes the iterator, counting the number of iterations and returning it.

This method will call next repeatedly until None is encountered, returning the number of times it saw Some. Note that next has to be called at least once even if the iterator does not have any elements.

Overflow Behavior

The method does no guarding against overflows, so counting elements of an iterator with more than usize::MAX elements either produces the wrong result or panics. If overflow checks are enabled, a panic is guaranteed.

Panics

This function might panic if the iterator has more than usize::MAX elements.

Examples

let a = [1, 2, 3];
assert_eq!(a.iter().count(), 3);

let a = [1, 2, 3, 4, 5];
assert_eq!(a.iter().count(), 5);

fn last(self) -> Option<Self::Item>
where Self: Sized,

Consumes the iterator, returning the last element.

This method will evaluate the iterator until it returns None. While doing so, it keeps track of the current element. After None is returned, last() will then return the last element it saw.

Examples

let a = [1, 2, 3];
assert_eq!(a.into_iter().last(), Some(3));

let a = [1, 2, 3, 4, 5];
assert_eq!(a.into_iter().last(), Some(5));

fn advance_by(&mut self, n: usize) -> Result<(), NonZero<usize>>

🔬This is a nightly-only experimental API. (iter_advance_by #77404)

Advances the iterator by n elements.

This method will eagerly skip n elements by calling next up to n times until None is encountered.

advance_by(n) will return Ok(()) if the iterator successfully advances by n elements, or a Err(NonZero<usize>) with value k if None is encountered, where k is remaining number of steps that could not be advanced because the iterator ran out. If self is empty and n is non-zero, then this returns Err(n). Otherwise, k is always less than n.

Calling advance_by(0) can do meaningful work, for example Flatten can advance its outer iterator until it finds an inner iterator that is not empty, which then often allows it to return a more accurate size_hint() than in its initial state.

Examples

#![feature(iter_advance_by)]

use std::num::NonZero;

let a = [1, 2, 3, 4];
let mut iter = a.into_iter();

assert_eq!(iter.advance_by(2), Ok(()));
assert_eq!(iter.next(), Some(3));
assert_eq!(iter.advance_by(0), Ok(()));
assert_eq!(iter.advance_by(100), Err(NonZero::new(99).unwrap())); // only `4` was skipped

fn nth(&mut self, n: usize) -> Option<Self::Item>

Returns the nth element of the iterator.

Like most indexing operations, the count starts from zero, so nth(0) returns the first value, nth(1) the second, and so on.

Note that all preceding elements, as well as the returned element, will be consumed from the iterator. That means that the preceding elements will be discarded, and also that calling nth(0) multiple times on the same iterator will return different elements.

nth() will return None if n is greater than or equal to the length of the iterator.

Examples

Basic usage:

let a = [1, 2, 3];
assert_eq!(a.into_iter().nth(1), Some(2));

Calling nth() multiple times doesn’t rewind the iterator:

let a = [1, 2, 3];

let mut iter = a.into_iter();

assert_eq!(iter.nth(1), Some(2));
assert_eq!(iter.nth(1), None);

Returning None if there are less than n + 1 elements:

let a = [1, 2, 3];
assert_eq!(a.into_iter().nth(10), None);

fn step_by(self, step: usize) -> StepBy<Self>
where Self: Sized,

Creates an iterator starting at the same point, but stepping by the given amount at each iteration.

Note 1: The first element of the iterator will always be returned, regardless of the step given.

Note 2: The time at which ignored elements are pulled is not fixed. StepBy behaves like the sequence self.next(), self.nth(step-1), self.nth(step-1), …, but is also free to behave like the sequence advance_n_and_return_first(&mut self, step), advance_n_and_return_first(&mut self, step), … Which way is used may change for some iterators for performance reasons. The second way will advance the iterator earlier and may consume more items.

advance_n_and_return_first is the equivalent of:

fn advance_n_and_return_first<I>(iter: &mut I, n: usize) -> Option<I::Item>
where
    I: Iterator,
{
    let next = iter.next();
    if n > 1 {
        iter.nth(n - 2);
    }
    next
}

Panics

The method will panic if the given step is 0.

Examples

let a = [0, 1, 2, 3, 4, 5];
let mut iter = a.into_iter().step_by(2);

assert_eq!(iter.next(), Some(0));
assert_eq!(iter.next(), Some(2));
assert_eq!(iter.next(), Some(4));
assert_eq!(iter.next(), None);

fn chain<U>(self, other: U) -> Chain<Self, U::IntoIter>
where Self: Sized, U: IntoIterator<Item = Self::Item>,

Takes two iterators and creates a new iterator over both in sequence.

chain() will return a new iterator which will first iterate over values from the first iterator and then over values from the second iterator.

In other words, it links two iterators together, in a chain. 🔗

once is commonly used to adapt a single value into a chain of other kinds of iteration.

Examples

Basic usage:

let s1 = "abc".chars();
let s2 = "def".chars();

let mut iter = s1.chain(s2);

assert_eq!(iter.next(), Some('a'));
assert_eq!(iter.next(), Some('b'));
assert_eq!(iter.next(), Some('c'));
assert_eq!(iter.next(), Some('d'));
assert_eq!(iter.next(), Some('e'));
assert_eq!(iter.next(), Some('f'));
assert_eq!(iter.next(), None);

Since the argument to chain() uses IntoIterator, we can pass anything that can be converted into an Iterator, not just an Iterator itself. For example, arrays ([T]) implement IntoIterator, and so can be passed to chain() directly:

let a1 = [1, 2, 3];
let a2 = [4, 5, 6];

let mut iter = a1.into_iter().chain(a2);

assert_eq!(iter.next(), Some(1));
assert_eq!(iter.next(), Some(2));
assert_eq!(iter.next(), Some(3));
assert_eq!(iter.next(), Some(4));
assert_eq!(iter.next(), Some(5));
assert_eq!(iter.next(), Some(6));
assert_eq!(iter.next(), None);

If you work with Windows API, you may wish to convert OsStr to Vec<u16>:

#[cfg(windows)]
fn os_str_to_utf16(s: &std::ffi::OsStr) -> Vec<u16> {
    use std::os::windows::ffi::OsStrExt;
    s.encode_wide().chain(std::iter::once(0)).collect()
}

fn zip<U>(self, other: U) -> Zip<Self, U::IntoIter>
where Self: Sized, U: IntoIterator,

‘Zips up’ two iterators into a single iterator of pairs.

zip() returns a new iterator that will iterate over two other iterators, returning a tuple where the first element comes from the first iterator, and the second element comes from the second iterator.

In other words, it zips two iterators together, into a single one.

If either iterator returns None, next from the zipped iterator will return None. If the zipped iterator has no more elements to return then each further attempt to advance it will first try to advance the first iterator at most one time and if it still yielded an item try to advance the second iterator at most one time.

To ‘undo’ the result of zipping up two iterators, see unzip.

Examples

Basic usage:

let s1 = "abc".chars();
let s2 = "def".chars();

let mut iter = s1.zip(s2);

assert_eq!(iter.next(), Some(('a', 'd')));
assert_eq!(iter.next(), Some(('b', 'e')));
assert_eq!(iter.next(), Some(('c', 'f')));
assert_eq!(iter.next(), None);

Since the argument to zip() uses IntoIterator, we can pass anything that can be converted into an Iterator, not just an Iterator itself. For example, arrays ([T]) implement IntoIterator, and so can be passed to zip() directly:

let a1 = [1, 2, 3];
let a2 = [4, 5, 6];

let mut iter = a1.into_iter().zip(a2);

assert_eq!(iter.next(), Some((1, 4)));
assert_eq!(iter.next(), Some((2, 5)));
assert_eq!(iter.next(), Some((3, 6)));
assert_eq!(iter.next(), None);

zip() is often used to zip an infinite iterator to a finite one. This works because the finite iterator will eventually return None, ending the zipper. Zipping with (0..) can look a lot like enumerate:

let enumerate: Vec<_> = "foo".chars().enumerate().collect();

let zipper: Vec<_> = (0..).zip("foo".chars()).collect();

assert_eq!((0, 'f'), enumerate[0]);
assert_eq!((0, 'f'), zipper[0]);

assert_eq!((1, 'o'), enumerate[1]);
assert_eq!((1, 'o'), zipper[1]);

assert_eq!((2, 'o'), enumerate[2]);
assert_eq!((2, 'o'), zipper[2]);

If both iterators have roughly equivalent syntax, it may be more readable to use zip:

use std::iter::zip;

let a = [1, 2, 3];
let b = [2, 3, 4];

let mut zipped = zip(
    a.into_iter().map(|x| x * 2).skip(1),
    b.into_iter().map(|x| x * 2).skip(1),
);

assert_eq!(zipped.next(), Some((4, 6)));
assert_eq!(zipped.next(), Some((6, 8)));
assert_eq!(zipped.next(), None);

compared to:

let mut zipped = a
    .into_iter()
    .map(|x| x * 2)
    .skip(1)
    .zip(b.into_iter().map(|x| x * 2).skip(1));

fn intersperse(self, separator: Self::Item) -> Intersperse<Self>
where Self: Sized, Self::Item: Clone,

🔬This is a nightly-only experimental API. (iter_intersperse #79524)

Creates a new iterator which places a copy of separator between adjacent items of the original iterator.

In case separator does not implement Clone or needs to be computed every time, use intersperse_with.

Examples

Basic usage:

#![feature(iter_intersperse)]

let mut a = [0, 1, 2].into_iter().intersperse(100);
assert_eq!(a.next(), Some(0));   // The first element from `a`.
assert_eq!(a.next(), Some(100)); // The separator.
assert_eq!(a.next(), Some(1));   // The next element from `a`.
assert_eq!(a.next(), Some(100)); // The separator.
assert_eq!(a.next(), Some(2));   // The last element from `a`.
assert_eq!(a.next(), None);       // The iterator is finished.

intersperse can be very useful to join an iterator’s items using a common element:

#![feature(iter_intersperse)]

let words = ["Hello", "World", "!"];
let hello: String = words.into_iter().intersperse(" ").collect();
assert_eq!(hello, "Hello World !");

fn intersperse_with<G>(self, separator: G) -> IntersperseWith<Self, G>
where Self: Sized, G: FnMut() -> Self::Item,

🔬This is a nightly-only experimental API. (iter_intersperse #79524)

Creates a new iterator which places an item generated by separator between adjacent items of the original iterator.

The closure will be called exactly once each time an item is placed between two adjacent items from the underlying iterator; specifically, the closure is not called if the underlying iterator yields less than two items and after the last item is yielded.

If the iterator’s item implements Clone, it may be easier to use intersperse.

Examples

Basic usage:

#![feature(iter_intersperse)]

#[derive(PartialEq, Debug)]
struct NotClone(usize);

let v = [NotClone(0), NotClone(1), NotClone(2)];
let mut it = v.into_iter().intersperse_with(|| NotClone(99));

assert_eq!(it.next(), Some(NotClone(0)));  // The first element from `v`.
assert_eq!(it.next(), Some(NotClone(99))); // The separator.
assert_eq!(it.next(), Some(NotClone(1)));  // The next element from `v`.
assert_eq!(it.next(), Some(NotClone(99))); // The separator.
assert_eq!(it.next(), Some(NotClone(2)));  // The last element from `v`.
assert_eq!(it.next(), None);               // The iterator is finished.

intersperse_with can be used in situations where the separator needs to be computed:

#![feature(iter_intersperse)]

let src = ["Hello", "to", "all", "people", "!!"].iter().copied();

// The closure mutably borrows its context to generate an item.
let mut happy_emojis = [" ❤️ ", " 😀 "].into_iter();
let separator = || happy_emojis.next().unwrap_or(" 🦀 ");

let result = src.intersperse_with(separator).collect::<String>();
assert_eq!(result, "Hello ❤️ to 😀 all 🦀 people 🦀 !!");

fn map<B, F>(self, f: F) -> Map<Self, F>
where Self: Sized, F: FnMut(Self::Item) -> B,

Takes a closure and creates an iterator which calls that closure on each element.

map() transforms one iterator into another, by means of its argument: something that implements FnMut. It produces a new iterator which calls this closure on each element of the original iterator.

If you are good at thinking in types, you can think of map() like this: If you have an iterator that gives you elements of some type A, and you want an iterator of some other type B, you can use map(), passing a closure that takes an A and returns a B.

map() is conceptually similar to a for loop. However, as map() is lazy, it is best used when you’re already working with other iterators. If you’re doing some sort of looping for a side effect, it’s considered more idiomatic to use for than map().

Examples

Basic usage:

let a = [1, 2, 3];

let mut iter = a.iter().map(|x| 2 * x);

assert_eq!(iter.next(), Some(2));
assert_eq!(iter.next(), Some(4));
assert_eq!(iter.next(), Some(6));
assert_eq!(iter.next(), None);

If you’re doing some sort of side effect, prefer for to map():

// don't do this:
(0..5).map(|x| println!("{x}"));

// it won't even execute, as it is lazy. Rust will warn you about this.

// Instead, use a for-loop:
for x in 0..5 {
    println!("{x}");
}

fn for_each<F>(self, f: F)
where Self: Sized, F: FnMut(Self::Item),

Calls a closure on each element of an iterator.

This is equivalent to using a for loop on the iterator, although break and continue are not possible from a closure. It’s generally more idiomatic to use a for loop, but for_each may be more legible when processing items at the end of longer iterator chains. In some cases for_each may also be faster than a loop, because it will use internal iteration on adapters like Chain.

Examples

Basic usage:

use std::sync::mpsc::channel;

let (tx, rx) = channel();
(0..5).map(|x| x * 2 + 1)
      .for_each(move |x| tx.send(x).unwrap());

let v: Vec<_> = rx.iter().collect();
assert_eq!(v, vec![1, 3, 5, 7, 9]);

For such a small example, a for loop may be cleaner, but for_each might be preferable to keep a functional style with longer iterators:

(0..5).flat_map(|x| (x * 100)..(x * 110))
      .enumerate()
      .filter(|&(i, x)| (i + x) % 3 == 0)
      .for_each(|(i, x)| println!("{i}:{x}"));

fn filter<P>(self, predicate: P) -> Filter<Self, P>
where Self: Sized, P: FnMut(&Self::Item) -> bool,

Creates an iterator which uses a closure to determine if an element should be yielded.

Given an element the closure must return true or false. The returned iterator will yield only the elements for which the closure returns true.

Examples

Basic usage:

let a = [0i32, 1, 2];

let mut iter = a.into_iter().filter(|x| x.is_positive());

assert_eq!(iter.next(), Some(1));
assert_eq!(iter.next(), Some(2));
assert_eq!(iter.next(), None);

Because the closure passed to filter() takes a reference, and many iterators iterate over references, this leads to a possibly confusing situation, where the type of the closure is a double reference:

let s = &[0, 1, 2];

let mut iter = s.iter().filter(|x| **x > 1); // needs two *s!

assert_eq!(iter.next(), Some(&2));
assert_eq!(iter.next(), None);

It’s common to instead use destructuring on the argument to strip away one:

let s = &[0, 1, 2];

let mut iter = s.iter().filter(|&x| *x > 1); // both & and *

assert_eq!(iter.next(), Some(&2));
assert_eq!(iter.next(), None);

or both:

let s = &[0, 1, 2];

let mut iter = s.iter().filter(|&&x| x > 1); // two &s

assert_eq!(iter.next(), Some(&2));
assert_eq!(iter.next(), None);

of these layers.

Note that iter.filter(f).next() is equivalent to iter.find(f).

fn filter_map<B, F>(self, f: F) -> FilterMap<Self, F>
where Self: Sized, F: FnMut(Self::Item) -> Option<B>,

Creates an iterator that both filters and maps.

The returned iterator yields only the values for which the supplied closure returns Some(value).

filter_map can be used to make chains of filter and map more concise. The example below shows how a map().filter().map() can be shortened to a single call to filter_map.

Examples

Basic usage:

let a = ["1", "two", "NaN", "four", "5"];

let mut iter = a.iter().filter_map(|s| s.parse().ok());

assert_eq!(iter.next(), Some(1));
assert_eq!(iter.next(), Some(5));
assert_eq!(iter.next(), None);

Here’s the same example, but with filter and map:

let a = ["1", "two", "NaN", "four", "5"];
let mut iter = a.iter().map(|s| s.parse()).filter(|s| s.is_ok()).map(|s| s.unwrap());
assert_eq!(iter.next(), Some(1));
assert_eq!(iter.next(), Some(5));
assert_eq!(iter.next(), None);

fn enumerate(self) -> Enumerate<Self>
where Self: Sized,

Creates an iterator which gives the current iteration count as well as the next value.

The iterator returned yields pairs (i, val), where i is the current index of iteration and val is the value returned by the iterator.

enumerate() keeps its count as a usize. If you want to count by a different sized integer, the zip function provides similar functionality.

Overflow Behavior

The method does no guarding against overflows, so enumerating more than usize::MAX elements either produces the wrong result or panics. If overflow checks are enabled, a panic is guaranteed.

Panics

The returned iterator might panic if the to-be-returned index would overflow a usize.

Examples

let a = ['a', 'b', 'c'];

let mut iter = a.into_iter().enumerate();

assert_eq!(iter.next(), Some((0, 'a')));
assert_eq!(iter.next(), Some((1, 'b')));
assert_eq!(iter.next(), Some((2, 'c')));
assert_eq!(iter.next(), None);

fn peekable(self) -> Peekable<Self>
where Self: Sized,

Creates an iterator which can use the peek and peek_mut methods to look at the next element of the iterator without consuming it. See their documentation for more information.

Note that the underlying iterator is still advanced when peek or peek_mut are called for the first time: In order to retrieve the next element, next is called on the underlying iterator, hence any side effects (i.e. anything other than fetching the next value) of the next method will occur.

Examples

Basic usage:

let xs = [1, 2, 3];

let mut iter = xs.into_iter().peekable();

// peek() lets us see into the future
assert_eq!(iter.peek(), Some(&1));
assert_eq!(iter.next(), Some(1));

assert_eq!(iter.next(), Some(2));

// we can peek() multiple times, the iterator won't advance
assert_eq!(iter.peek(), Some(&3));
assert_eq!(iter.peek(), Some(&3));

assert_eq!(iter.next(), Some(3));

// after the iterator is finished, so is peek()
assert_eq!(iter.peek(), None);
assert_eq!(iter.next(), None);

Using peek_mut to mutate the next item without advancing the iterator:

let xs = [1, 2, 3];

let mut iter = xs.into_iter().peekable();

// `peek_mut()` lets us see into the future
assert_eq!(iter.peek_mut(), Some(&mut 1));
assert_eq!(iter.peek_mut(), Some(&mut 1));
assert_eq!(iter.next(), Some(1));

if let Some(p) = iter.peek_mut() {
    assert_eq!(*p, 2);
    // put a value into the iterator
    *p = 1000;
}

// The value reappears as the iterator continues
assert_eq!(iter.collect::<Vec<_>>(), vec![1000, 3]);

fn skip_while<P>(self, predicate: P) -> SkipWhile<Self, P>
where Self: Sized, P: FnMut(&Self::Item) -> bool,

Creates an iterator that skips elements based on a predicate.

skip_while() takes a closure as an argument. It will call this closure on each element of the iterator, and ignore elements until it returns false.

After false is returned, skip_while()’s job is over, and the rest of the elements are yielded.

Examples

Basic usage:

let a = [-1i32, 0, 1];

let mut iter = a.into_iter().skip_while(|x| x.is_negative());

assert_eq!(iter.next(), Some(0));
assert_eq!(iter.next(), Some(1));
assert_eq!(iter.next(), None);

Because the closure passed to skip_while() takes a reference, and many iterators iterate over references, this leads to a possibly confusing situation, where the type of the closure argument is a double reference:

let s = &[-1, 0, 1];

let mut iter = s.iter().skip_while(|x| **x < 0); // need two *s!

assert_eq!(iter.next(), Some(&0));
assert_eq!(iter.next(), Some(&1));
assert_eq!(iter.next(), None);

Stopping after an initial false:

let a = [-1, 0, 1, -2];

let mut iter = a.into_iter().skip_while(|&x| x < 0);

assert_eq!(iter.next(), Some(0));
assert_eq!(iter.next(), Some(1));

// while this would have been false, since we already got a false,
// skip_while() isn't used any more
assert_eq!(iter.next(), Some(-2));

assert_eq!(iter.next(), None);

fn take_while<P>(self, predicate: P) -> TakeWhile<Self, P>
where Self: Sized, P: FnMut(&Self::Item) -> bool,

Creates an iterator that yields elements based on a predicate.

take_while() takes a closure as an argument. It will call this closure on each element of the iterator, and yield elements while it returns true.

After false is returned, take_while()’s job is over, and the rest of the elements are ignored.

Examples

Basic usage:

let a = [-1i32, 0, 1];

let mut iter = a.into_iter().take_while(|x| x.is_negative());

assert_eq!(iter.next(), Some(-1));
assert_eq!(iter.next(), None);

Because the closure passed to take_while() takes a reference, and many iterators iterate over references, this leads to a possibly confusing situation, where the type of the closure is a double reference:

let s = &[-1, 0, 1];

let mut iter = s.iter().take_while(|x| **x < 0); // need two *s!

assert_eq!(iter.next(), Some(&-1));
assert_eq!(iter.next(), None);

Stopping after an initial false:

let a = [-1, 0, 1, -2];

let mut iter = a.into_iter().take_while(|&x| x < 0);

assert_eq!(iter.next(), Some(-1));

// We have more elements that are less than zero, but since we already
// got a false, take_while() ignores the remaining elements.
assert_eq!(iter.next(), None);

Because take_while() needs to look at the value in order to see if it should be included or not, consuming iterators will see that it is removed:

let a = [1, 2, 3, 4];
let mut iter = a.into_iter();

let result: Vec<i32> = iter.by_ref().take_while(|&n| n != 3).collect();

assert_eq!(result, [1, 2]);

let result: Vec<i32> = iter.collect();

assert_eq!(result, [4]);

The 3 is no longer there, because it was consumed in order to see if the iteration should stop, but wasn’t placed back into the iterator.

fn map_while<B, P>(self, predicate: P) -> MapWhile<Self, P>
where Self: Sized, P: FnMut(Self::Item) -> Option<B>,

Creates an iterator that both yields elements based on a predicate and maps.

map_while() takes a closure as an argument. It will call this closure on each element of the iterator, and yield elements while it returns Some(_).

Examples

Basic usage:

let a = [-1i32, 4, 0, 1];

let mut iter = a.into_iter().map_while(|x| 16i32.checked_div(x));

assert_eq!(iter.next(), Some(-16));
assert_eq!(iter.next(), Some(4));
assert_eq!(iter.next(), None);

Here’s the same example, but with take_while and map:

let a = [-1i32, 4, 0, 1];

let mut iter = a.into_iter()
                .map(|x| 16i32.checked_div(x))
                .take_while(|x| x.is_some())
                .map(|x| x.unwrap());

assert_eq!(iter.next(), Some(-16));
assert_eq!(iter.next(), Some(4));
assert_eq!(iter.next(), None);

Stopping after an initial None:

let a = [0, 1, 2, -3, 4, 5, -6];

let iter = a.into_iter().map_while(|x| u32::try_from(x).ok());
let vec: Vec<_> = iter.collect();

// We have more elements that could fit in u32 (such as 4, 5), but `map_while` returned `None` for `-3`
// (as the `predicate` returned `None`) and `collect` stops at the first `None` encountered.
assert_eq!(vec, [0, 1, 2]);

Because map_while() needs to look at the value in order to see if it should be included or not, consuming iterators will see that it is removed:

let a = [1, 2, -3, 4];
let mut iter = a.into_iter();

let result: Vec<u32> = iter.by_ref()
                           .map_while(|n| u32::try_from(n).ok())
                           .collect();

assert_eq!(result, [1, 2]);

let result: Vec<i32> = iter.collect();

assert_eq!(result, [4]);

The -3 is no longer there, because it was consumed in order to see if the iteration should stop, but wasn’t placed back into the iterator.

Note that unlike take_while this iterator is not fused. It is also not specified what this iterator returns after the first None is returned. If you need a fused iterator, use fuse.

fn skip(self, n: usize) -> Skip<Self>
where Self: Sized,

Creates an iterator that skips the first n elements.

skip(n) skips elements until n elements are skipped or the end of the iterator is reached (whichever happens first). After that, all the remaining elements are yielded. In particular, if the original iterator is too short, then the returned iterator is empty.

Rather than overriding this method directly, instead override the nth method.

Examples

let a = [1, 2, 3];

let mut iter = a.into_iter().skip(2);

assert_eq!(iter.next(), Some(3));
assert_eq!(iter.next(), None);

fn take(self, n: usize) -> Take<Self>
where Self: Sized,

Creates an iterator that yields the first n elements, or fewer if the underlying iterator ends sooner.

take(n) yields elements until n elements are yielded or the end of the iterator is reached (whichever happens first). The returned iterator is a prefix of length n if the original iterator contains at least n elements, otherwise it contains all of the (fewer than n) elements of the original iterator.

Examples

Basic usage:

let a = [1, 2, 3];

let mut iter = a.into_iter().take(2);

assert_eq!(iter.next(), Some(1));
assert_eq!(iter.next(), Some(2));
assert_eq!(iter.next(), None);

take() is often used with an infinite iterator, to make it finite:

let mut iter = (0..).take(3);

assert_eq!(iter.next(), Some(0));
assert_eq!(iter.next(), Some(1));
assert_eq!(iter.next(), Some(2));
assert_eq!(iter.next(), None);

If less than n elements are available, take will limit itself to the size of the underlying iterator:

let v = [1, 2];
let mut iter = v.into_iter().take(5);
assert_eq!(iter.next(), Some(1));
assert_eq!(iter.next(), Some(2));
assert_eq!(iter.next(), None);

Use by_ref to take from the iterator without consuming it, and then continue using the original iterator:

let mut words = ["hello", "world", "of", "Rust"].into_iter();

// Take the first two words.
let hello_world: Vec<_> = words.by_ref().take(2).collect();
assert_eq!(hello_world, vec!["hello", "world"]);

// Collect the rest of the words.
// We can only do this because we used `by_ref` earlier.
let of_rust: Vec<_> = words.collect();
assert_eq!(of_rust, vec!["of", "Rust"]);

fn scan<St, B, F>(self, initial_state: St, f: F) -> Scan<Self, St, F>
where Self: Sized, F: FnMut(&mut St, Self::Item) -> Option<B>,

An iterator adapter which, like fold, holds internal state, but unlike fold, produces a new iterator.

scan() takes two arguments: an initial value which seeds the internal state, and a closure with two arguments, the first being a mutable reference to the internal state and the second an iterator element. The closure can assign to the internal state to share state between iterations.

On iteration, the closure will be applied to each element of the iterator and the return value from the closure, an Option, is returned by the next method. Thus the closure can return Some(value) to yield value, or None to end the iteration.

Examples

let a = [1, 2, 3, 4];

let mut iter = a.into_iter().scan(1, |state, x| {
    // each iteration, we'll multiply the state by the element ...
    *state = *state * x;

    // ... and terminate if the state exceeds 6
    if *state > 6 {
        return None;
    }
    // ... else yield the negation of the state
    Some(-*state)
});

assert_eq!(iter.next(), Some(-1));
assert_eq!(iter.next(), Some(-2));
assert_eq!(iter.next(), Some(-6));
assert_eq!(iter.next(), None);

fn flat_map<U, F>(self, f: F) -> FlatMap<Self, U, F>
where Self: Sized, U: IntoIterator, F: FnMut(Self::Item) -> U,

Creates an iterator that works like map, but flattens nested structure.

The map adapter is very useful, but only when the closure argument produces values. If it produces an iterator instead, there’s an extra layer of indirection. flat_map() will remove this extra layer on its own.

You can think of flat_map(f) as the semantic equivalent of mapping, and then flattening as in map(f).flatten().

Another way of thinking about flat_map(): map’s closure returns one item for each element, and flat_map()’s closure returns an iterator for each element.

Examples

let words = ["alpha", "beta", "gamma"];

// chars() returns an iterator
let merged: String = words.iter()
                          .flat_map(|s| s.chars())
                          .collect();
assert_eq!(merged, "alphabetagamma");

fn flatten(self) -> Flatten<Self>
where Self: Sized, Self::Item: IntoIterator,

Creates an iterator that flattens nested structure.

This is useful when you have an iterator of iterators or an iterator of things that can be turned into iterators and you want to remove one level of indirection.

Examples

Basic usage:

let data = vec![vec![1, 2, 3, 4], vec![5, 6]];
let flattened: Vec<_> = data.into_iter().flatten().collect();
assert_eq!(flattened, [1, 2, 3, 4, 5, 6]);

Mapping and then flattening:

let words = ["alpha", "beta", "gamma"];

// chars() returns an iterator
let merged: String = words.iter()
                          .map(|s| s.chars())
                          .flatten()
                          .collect();
assert_eq!(merged, "alphabetagamma");

You can also rewrite this in terms of flat_map(), which is preferable in this case since it conveys intent more clearly:

let words = ["alpha", "beta", "gamma"];

// chars() returns an iterator
let merged: String = words.iter()
                          .flat_map(|s| s.chars())
                          .collect();
assert_eq!(merged, "alphabetagamma");

Flattening works on any IntoIterator type, including Option and Result:

let options = vec![Some(123), Some(321), None, Some(231)];
let flattened_options: Vec<_> = options.into_iter().flatten().collect();
assert_eq!(flattened_options, [123, 321, 231]);

let results = vec![Ok(123), Ok(321), Err(456), Ok(231)];
let flattened_results: Vec<_> = results.into_iter().flatten().collect();
assert_eq!(flattened_results, [123, 321, 231]);

Flattening only removes one level of nesting at a time:

let d3 = [[[1, 2], [3, 4]], [[5, 6], [7, 8]]];

let d2: Vec<_> = d3.into_iter().flatten().collect();
assert_eq!(d2, [[1, 2], [3, 4], [5, 6], [7, 8]]);

let d1: Vec<_> = d3.into_iter().flatten().flatten().collect();
assert_eq!(d1, [1, 2, 3, 4, 5, 6, 7, 8]);

Here we see that flatten() does not perform a “deep” flatten. Instead, only one level of nesting is removed. That is, if you flatten() a three-dimensional array, the result will be two-dimensional and not one-dimensional. To get a one-dimensional structure, you have to flatten() again.

fn map_windows<F, R, const N: usize>(self, f: F) -> MapWindows<Self, F, N>
where Self: Sized, F: FnMut(&[Self::Item; N]) -> R,

🔬This is a nightly-only experimental API. (iter_map_windows #87155)

Calls the given function f for each contiguous window of size N over self and returns an iterator over the outputs of f. Like slice::windows(), the windows during mapping overlap as well.

In the following example, the closure is called three times with the arguments &['a', 'b'], &['b', 'c'] and &['c', 'd'] respectively.

#![feature(iter_map_windows)]

let strings = "abcd".chars()
    .map_windows(|[x, y]| format!("{}+{}", x, y))
    .collect::<Vec<String>>();

assert_eq!(strings, vec!["a+b", "b+c", "c+d"]);

Note that the const parameter N is usually inferred by the destructured argument in the closure.

The returned iterator yields 𝑘 − N + 1 items (where 𝑘 is the number of items yielded by self). If 𝑘 is less than N, this method yields an empty iterator.

The returned iterator implements FusedIterator, because once self returns None, even if it returns a Some(T) again in the next iterations, we cannot put it into a contiguous array buffer, and thus the returned iterator should be fused.

Panics

Panics if N is zero. This check will most probably get changed to a compile time error before this method gets stabilized.

#![feature(iter_map_windows)]

let iter = std::iter::repeat(0).map_windows(|&[]| ());

Examples

Building the sums of neighboring numbers.

#![feature(iter_map_windows)]

let mut it = [1, 3, 8, 1].iter().map_windows(|&[a, b]| a + b);
assert_eq!(it.next(), Some(4));  // 1 + 3
assert_eq!(it.next(), Some(11)); // 3 + 8
assert_eq!(it.next(), Some(9));  // 8 + 1
assert_eq!(it.next(), None);

Since the elements in the following example implement Copy, we can just copy the array and get an iterator over the windows.

#![feature(iter_map_windows)]

let mut it = "ferris".chars().map_windows(|w: &[_; 3]| *w);
assert_eq!(it.next(), Some(['f', 'e', 'r']));
assert_eq!(it.next(), Some(['e', 'r', 'r']));
assert_eq!(it.next(), Some(['r', 'r', 'i']));
assert_eq!(it.next(), Some(['r', 'i', 's']));
assert_eq!(it.next(), None);

You can also use this function to check the sortedness of an iterator. For the simple case, rather use Iterator::is_sorted.

#![feature(iter_map_windows)]

let mut it = [0.5, 1.0, 3.5, 3.0, 8.5, 8.5, f32::NAN].iter()
    .map_windows(|[a, b]| a <= b);

assert_eq!(it.next(), Some(true));  // 0.5 <= 1.0
assert_eq!(it.next(), Some(true));  // 1.0 <= 3.5
assert_eq!(it.next(), Some(false)); // 3.5 <= 3.0
assert_eq!(it.next(), Some(true));  // 3.0 <= 8.5
assert_eq!(it.next(), Some(true));  // 8.5 <= 8.5
assert_eq!(it.next(), Some(false)); // 8.5 <= NAN
assert_eq!(it.next(), None);

For non-fused iterators, they are fused after map_windows.

#![feature(iter_map_windows)]

#[derive(Default)]
struct NonFusedIterator {
    state: i32,
}

impl Iterator for NonFusedIterator {
    type Item = i32;

    fn next(&mut self) -> Option<i32> {
        let val = self.state;
        self.state = self.state + 1;

        // yields `0..5` first, then only even numbers since `6..`.
        if val < 5 || val % 2 == 0 {
            Some(val)
        } else {
            None
        }
    }
}

let mut iter = NonFusedIterator::default();

// yields 0..5 first.
assert_eq!(iter.next(), Some(0));
assert_eq!(iter.next(), Some(1));
assert_eq!(iter.next(), Some(2));
assert_eq!(iter.next(), Some(3));
assert_eq!(iter.next(), Some(4));
// then we can see our iterator going back and forth
assert_eq!(iter.next(), None);
assert_eq!(iter.next(), Some(6));
assert_eq!(iter.next(), None);
assert_eq!(iter.next(), Some(8));
assert_eq!(iter.next(), None);

// however, with `.map_windows()`, it is fused.
let mut iter = NonFusedIterator::default()
    .map_windows(|arr: &[_; 2]| *arr);

assert_eq!(iter.next(), Some([0, 1]));
assert_eq!(iter.next(), Some([1, 2]));
assert_eq!(iter.next(), Some([2, 3]));
assert_eq!(iter.next(), Some([3, 4]));
assert_eq!(iter.next(), None);

// it will always return `None` after the first time.
assert_eq!(iter.next(), None);
assert_eq!(iter.next(), None);
assert_eq!(iter.next(), None);

fn fuse(self) -> Fuse<Self>
where Self: Sized,

Creates an iterator which ends after the first None.

After an iterator returns None, future calls may or may not yield Some(T) again. fuse() adapts an iterator, ensuring that after a None is given, it will always return None forever.

Note that the Fuse wrapper is a no-op on iterators that implement the FusedIterator trait. fuse() may therefore behave incorrectly if the FusedIterator trait is improperly implemented.

Examples

// an iterator which alternates between Some and None
struct Alternate {
    state: i32,
}

impl Iterator for Alternate {
    type Item = i32;

    fn next(&mut self) -> Option<i32> {
        let val = self.state;
        self.state = self.state + 1;

        // if it's even, Some(i32), else None
        (val % 2 == 0).then_some(val)
    }
}

let mut iter = Alternate { state: 0 };

// we can see our iterator going back and forth
assert_eq!(iter.next(), Some(0));
assert_eq!(iter.next(), None);
assert_eq!(iter.next(), Some(2));
assert_eq!(iter.next(), None);

// however, once we fuse it...
let mut iter = iter.fuse();

assert_eq!(iter.next(), Some(4));
assert_eq!(iter.next(), None);

// it will always return `None` after the first time.
assert_eq!(iter.next(), None);
assert_eq!(iter.next(), None);
assert_eq!(iter.next(), None);

fn inspect<F>(self, f: F) -> Inspect<Self, F>
where Self: Sized, F: FnMut(&Self::Item),

Does something with each element of an iterator, passing the value on.

When using iterators, you’ll often chain several of them together. While working on such code, you might want to check out what’s happening at various parts in the pipeline. To do that, insert a call to inspect().

It’s more common for inspect() to be used as a debugging tool than to exist in your final code, but applications may find it useful in certain situations when errors need to be logged before being discarded.

Examples

Basic usage:

let a = [1, 4, 2, 3];

// this iterator sequence is complex.
let sum = a.iter()
    .cloned()
    .filter(|x| x % 2 == 0)
    .fold(0, |sum, i| sum + i);

println!("{sum}");

// let's add some inspect() calls to investigate what's happening
let sum = a.iter()
    .cloned()
    .inspect(|x| println!("about to filter: {x}"))
    .filter(|x| x % 2 == 0)
    .inspect(|x| println!("made it through filter: {x}"))
    .fold(0, |sum, i| sum + i);

println!("{sum}");

This will print:

6
about to filter: 1
about to filter: 4
made it through filter: 4
about to filter: 2
made it through filter: 2
about to filter: 3
6

Logging errors before discarding them:

let lines = ["1", "2", "a"];

let sum: i32 = lines
    .iter()
    .map(|line| line.parse::<i32>())
    .inspect(|num| {
        if let Err(ref e) = *num {
            println!("Parsing error: {e}");
        }
    })
    .filter_map(Result::ok)
    .sum();

println!("Sum: {sum}");

This will print:

Parsing error: invalid digit found in string
Sum: 3

fn by_ref(&mut self) -> &mut Self
where Self: Sized,

Creates a “by reference” adapter for this instance of Iterator.

Consuming method calls (direct or indirect calls to next) on the “by reference” adapter will consume the original iterator, but ownership-taking methods (those with a self parameter) only take ownership of the “by reference” iterator.

This is useful for applying ownership-taking methods (such as take in the example below) without giving up ownership of the original iterator, so you can use the original iterator afterwards.

Uses impl<I: Iterator + ?Sized> Iterator for &mut I { type Item = I::Item; ...}.

Examples

let mut words = ["hello", "world", "of", "Rust"].into_iter();

// Take the first two words.
let hello_world: Vec<_> = words.by_ref().take(2).collect();
assert_eq!(hello_world, vec!["hello", "world"]);

// Collect the rest of the words.
// We can only do this because we used `by_ref` earlier.
let of_rust: Vec<_> = words.collect();
assert_eq!(of_rust, vec!["of", "Rust"]);

fn collect<B: FromIterator<Self::Item>>(self) -> B
where Self: Sized,

Transforms an iterator into a collection.

collect() can take anything iterable, and turn it into a relevant collection. This is one of the more powerful methods in the standard library, used in a variety of contexts.

The most basic pattern in which collect() is used is to turn one collection into another. You take a collection, call iter on it, do a bunch of transformations, and then collect() at the end.

collect() can also create instances of types that are not typical collections. For example, a String can be built from chars, and an iterator of Result<T, E> items can be collected into Result<Collection<T>, E>. See the examples below for more.

Because collect() is so general, it can cause problems with type inference. As such, collect() is one of the few times you’ll see the syntax affectionately known as the ‘turbofish’: ::<>. This helps the inference algorithm understand specifically which collection you’re trying to collect into.

Examples

Basic usage:

let a = [1, 2, 3];

let doubled: Vec<i32> = a.iter()
                         .map(|x| x * 2)
                         .collect();

assert_eq!(vec![2, 4, 6], doubled);

Note that we needed the : Vec<i32> on the left-hand side. This is because we could collect into, for example, a VecDeque<T> instead:

use std::collections::VecDeque;

let a = [1, 2, 3];

let doubled: VecDeque<i32> = a.iter().map(|x| x * 2).collect();

assert_eq!(2, doubled[0]);
assert_eq!(4, doubled[1]);
assert_eq!(6, doubled[2]);

Using the ‘turbofish’ instead of annotating doubled:

let a = [1, 2, 3];

let doubled = a.iter().map(|x| x * 2).collect::<Vec<i32>>();

assert_eq!(vec![2, 4, 6], doubled);

Because collect() only cares about what you’re collecting into, you can still use a partial type hint, _, with the turbofish:

let a = [1, 2, 3];

let doubled = a.iter().map(|x| x * 2).collect::<Vec<_>>();

assert_eq!(vec![2, 4, 6], doubled);

Using collect() to make a String:

let chars = ['g', 'd', 'k', 'k', 'n'];

let hello: String = chars.into_iter()
    .map(|x| x as u8)
    .map(|x| (x + 1) as char)
    .collect();

assert_eq!("hello", hello);

If you have a list of Result<T, E>s, you can use collect() to see if any of them failed:

let results = [Ok(1), Err("nope"), Ok(3), Err("bad")];

let result: Result<Vec<_>, &str> = results.into_iter().collect();

// gives us the first error
assert_eq!(Err("nope"), result);

let results = [Ok(1), Ok(3)];

let result: Result<Vec<_>, &str> = results.into_iter().collect();

// gives us the list of answers
assert_eq!(Ok(vec![1, 3]), result);

fn try_collect<B>(&mut self) -> <Self::Item::Residual as Residual<B>>::TryType
where Self: Sized, Self::Item: Try<Residual: Residual<B>>, B: FromIterator<<Self::Item as Try>::Output>,

🔬This is a nightly-only experimental API. (iterator_try_collect #94047)

Fallibly transforms an iterator into a collection, short circuiting if a failure is encountered.

try_collect() is a variation of collect() that allows fallible conversions during collection. Its main use case is simplifying conversions from iterators yielding Option<T> into Option<Collection<T>>, or similarly for other Try types (e.g. Result).

Importantly, try_collect() doesn’t require that the outer Try type also implements FromIterator; only the inner type produced on Try::Output must implement it. Concretely, this means that collecting into ControlFlow<_, Vec<i32>> is valid because Vec<i32> implements FromIterator, even though ControlFlow doesn’t.

Also, if a failure is encountered during try_collect(), the iterator is still valid and may continue to be used, in which case it will continue iterating starting after the element that triggered the failure. See the last example below for an example of how this works.

Examples

Successfully collecting an iterator of Option<i32> into Option<Vec<i32>>:

#![feature(iterator_try_collect)]

let u = vec![Some(1), Some(2), Some(3)];
let v = u.into_iter().try_collect::<Vec<i32>>();
assert_eq!(v, Some(vec![1, 2, 3]));

Failing to collect in the same way:

#![feature(iterator_try_collect)]

let u = vec![Some(1), Some(2), None, Some(3)];
let v = u.into_iter().try_collect::<Vec<i32>>();
assert_eq!(v, None);

A similar example, but with Result:

#![feature(iterator_try_collect)]

let u: Vec<Result<i32, ()>> = vec![Ok(1), Ok(2), Ok(3)];
let v = u.into_iter().try_collect::<Vec<i32>>();
assert_eq!(v, Ok(vec![1, 2, 3]));

let u = vec![Ok(1), Ok(2), Err(()), Ok(3)];
let v = u.into_iter().try_collect::<Vec<i32>>();
assert_eq!(v, Err(()));

Finally, even ControlFlow works, despite the fact that it doesn’t implement FromIterator. Note also that the iterator can continue to be used, even if a failure is encountered:

#![feature(iterator_try_collect)]

use core::ops::ControlFlow::{Break, Continue};

let u = [Continue(1), Continue(2), Break(3), Continue(4), Continue(5)];
let mut it = u.into_iter();

let v = it.try_collect::<Vec<_>>();
assert_eq!(v, Break(3));

let v = it.try_collect::<Vec<_>>();
assert_eq!(v, Continue(vec![4, 5]));

fn collect_into<E: Extend<Self::Item>>(self, collection: &mut E) -> &mut E
where Self: Sized,

🔬This is a nightly-only experimental API. (iter_collect_into #94780)

Collects all the items from an iterator into a collection.

This method consumes the iterator and adds all its items to the passed collection. The collection is then returned, so the call chain can be continued.

This is useful when you already have a collection and want to add the iterator items to it.

This method is a convenience method to call Extend::extend, but instead of being called on a collection, it’s called on an iterator.

Examples

Basic usage:

#![feature(iter_collect_into)]

let a = [1, 2, 3];
let mut vec: Vec::<i32> = vec![0, 1];

a.iter().map(|x| x * 2).collect_into(&mut vec);
a.iter().map(|x| x * 10).collect_into(&mut vec);

assert_eq!(vec, vec![0, 1, 2, 4, 6, 10, 20, 30]);

Vec can have a manual set capacity to avoid reallocating it:

#![feature(iter_collect_into)]

let a = [1, 2, 3];
let mut vec: Vec::<i32> = Vec::with_capacity(6);

a.iter().map(|x| x * 2).collect_into(&mut vec);
a.iter().map(|x| x * 10).collect_into(&mut vec);

assert_eq!(6, vec.capacity());
assert_eq!(vec, vec![2, 4, 6, 10, 20, 30]);

The returned mutable reference can be used to continue the call chain:

#![feature(iter_collect_into)]

let a = [1, 2, 3];
let mut vec: Vec::<i32> = Vec::with_capacity(6);

let count = a.iter().collect_into(&mut vec).iter().count();

assert_eq!(count, vec.len());
assert_eq!(vec, vec![1, 2, 3]);

let count = a.iter().collect_into(&mut vec).iter().count();

assert_eq!(count, vec.len());
assert_eq!(vec, vec![1, 2, 3, 1, 2, 3]);

fn partition<B, F>(self, f: F) -> (B, B)
where Self: Sized, B: Default + Extend<Self::Item>, F: FnMut(&Self::Item) -> bool,

Consumes an iterator, creating two collections from it.

The predicate passed to partition() can return true, or false. partition() returns a pair, all of the elements for which it returned true, and all of the elements for which it returned false.

See also is_partitioned() and partition_in_place().

Examples

let a = [1, 2, 3];

let (even, odd): (Vec<_>, Vec<_>) = a
    .into_iter()
    .partition(|n| n % 2 == 0);

assert_eq!(even, [2]);
assert_eq!(odd, [1, 3]);