Struct core::lazy::Lazy

pub struct Lazy<T, F = fn() -> T> { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (once_cell #74465)

A value which is initialized on the first access.

Examples

#![feature(once_cell)]

use std::lazy::Lazy;

let lazy: Lazy<i32> = Lazy::new(|| {
    println!("initializing");
    92
});
println!("ready");
println!("{}", *lazy);
println!("{}", *lazy);

// Prints:
//   ready
//   initializing
//   92
//   92
Run

Implementations

impl<T, F> Lazy<T, F>

pub const fn new(init: F) -> Lazy<T, F>

🔬 This is a nightly-only experimental API. (once_cell #74465)

Creates a new lazy value with the given initializing function.

Examples

#![feature(once_cell)]

use std::lazy::Lazy;

let hello = "Hello, World!".to_string();

let lazy = Lazy::new(|| hello.to_uppercase());

assert_eq!(&*lazy, "HELLO, WORLD!");
Run

impl<T, F: FnOnce() -> T> Lazy<T, F>

pub fn force(this: &Lazy<T, F>) -> &T

🔬 This is a nightly-only experimental API. (once_cell #74465)

Forces the evaluation of this lazy value and returns a reference to the result.

This is equivalent to the Deref impl, but is explicit.

Examples

#![feature(once_cell)]

use std::lazy::Lazy;

let lazy = Lazy::new(|| 92);

assert_eq!(Lazy::force(&lazy), &92);
assert_eq!(&*lazy, &92);
Run