Trait core::marker::DiscriminantKind

#[lang = "discriminant_kind"]pub trait DiscriminantKind {
    type Discriminant: Clone + Copy + Debug + Eq + PartialEq + Hash + Send + Sync + Unpin;
}

🔬 This is a nightly-only experimental API. (discriminant_kind)

this trait is unlikely to ever be stabilized, use mem::discriminant instead

Compiler-internal trait used to indicate the type of enum discriminants.

This trait is automatically implemented for every type and does not add any guarantees to mem::Discriminant. It is undefined behavior to transmute between DiscriminantKind::Discriminant and mem::Discriminant.

Associated Types

type Discriminant: Clone + Copy + Debug + Eq + PartialEq + Hash + Send + Sync + Unpin

🔬 This is a nightly-only experimental API. (discriminant_kind)

this trait is unlikely to ever be stabilized, use mem::discriminant instead

The type of the discriminant, which must satisfy the trait bounds required by mem::Discriminant.

Loading content...