Function core::mem::align_of_val_raw

pub unsafe fn align_of_val_raw<T: ?Sized>(val: *const T) -> usize

🔬 This is a nightly-only experimental API. (layout_for_ptr #69835)

返回 val 指向的值类型的 ABI 要求的最小对齐方式。

Returns the ABI-required minimum alignment of the type of the value that val points to.

每个对 T 类型值的引用都必须是该数字的倍数。

Every reference to a value of the type T must be a multiple of this number.

Safety

只有在以下条件成立时，才能安全调用此函数：

This function is only safe to call if the following conditions hold:

如果 T 是 Sized，则此函数始终可以安全调用。

If T is Sized, this function is always safe to call.

如果 T 的未调整大小的尾部是：

If the unsized tail of T is:

一个切片，那么切片尾部的长度必须是一个初始化的整数，并且整个值的大小（动态尾部长度+静态大小的前缀）必须适合 isize。

a slice, then the length of the slice tail must be an initialized integer, and the size of the entire value (dynamic tail length + statically sized prefix) must fit in isize.

一个 trait 对象，那么指针的 vtable 部分必须指向一个通过 unsizing coercion 获得的有效 vtable，并且整个值的大小（动态尾部长度 + 静态大小的前缀）必须适合 isize。

a trait object, then the vtable part of the pointer must point to a valid vtable acquired by an unsizing coercion, and the size of the entire value (dynamic tail length + statically sized prefix) must fit in isize.

一个（不稳定的）外部类型，那么这个函数总是可以安全地调用，但可能会panic或返回错误的值，因为外部类型的布局是未知的。 这与 align_of_val 在引用具有外部类型尾部的类型时的行为相同。

an (unstable) extern type, then this function is always safe to call, but may panic or otherwise return the wrong value, as the extern type’s layout is not known. This is the same behavior as align_of_val on a reference to a type with an extern type tail.

否则，保守地不允许调用该函数。

otherwise, it is conservatively not allowed to call this function.

Examples

#![feature(layout_for_ptr)]
use std::mem;

assert_eq!(4, unsafe { mem::align_of_val_raw(&5i32) });
Run