Function core::mem::drop

pub fn drop<T>(_x: T)

处理一个值。

Disposes of a value.

这是通过调用 Drop 的参数实现来实现的。

This does so by calling the argument’s implementation of Drop.

这实际上对实现 Copy 的类型没有任何作用，例如 整数。 此类值被复制然后移动到函数中，因此该值在此函数调用后仍然存在。

This effectively does nothing for types which implement Copy, e.g. integers. Such values are copied and then moved into the function, so the value persists after this function call.

这个功能并不神奇； 它的字面意思是

This function is not magic; it is literally defined as

pub fn drop<T>(_x: T) { }
Run

因为 _x 被移动到函数中，所以它在函数返回之前被自动删除。

Because _x is moved into the function, it is automatically dropped before the function returns.

Examples

基本用法：

Basic usage:

let v = vec![1, 2, 3];

drop(v); // explicitly drop the vector
Run

由于 RefCell 在运行时强制执行借用规则，因此 drop 可以释放 RefCell 借用：

Since RefCell enforces the borrow rules at runtime, drop can release a RefCell borrow:

use std::cell::RefCell;

let x = RefCell::new(1);

let mut mutable_borrow = x.borrow_mut();
*mutable_borrow = 1;

drop(mutable_borrow); // relinquish the mutable borrow on this slot

let borrow = x.borrow();
println!("{}", *borrow);
Run

实现 Copy 的整数和其他类型不受 drop 的影响。

Integers and other types implementing Copy are unaffected by drop.

#[derive(Copy, Clone)]
struct Foo(u8);

let x = 1;
let y = Foo(2);
drop(x); // a copy of `x` is moved and dropped
drop(y); // a copy of `y` is moved and dropped

println!("x: {}, y: {}", x, y.0); // still available
Run