Function core::mem::forget

pub const fn forget<T>(t: T)

在不运行析构函数的情况下获得所有权并“忘记”价值。

Takes ownership and “forgets” about the value without running its destructor.

该值管理的任何资源，例如堆内存或文件句柄，将永远处于无法访问的状态。 但是，它不保证指向该内存的指针将保持有效。

Any resources the value manages, such as heap memory or a file handle, will linger forever in an unreachable state. However, it does not guarantee that pointers to this memory will remain valid.

如果要泄漏内存，请参阅“Box::leak”。

If you want to leak memory, see Box::leak.

如果要获取指向内存的原始指针，请参阅“Box::into_raw”。

If you want to obtain a raw pointer to the memory, see Box::into_raw.

如果你想正确地处理一个值，运行它的析构函数，参见“mem::drop”。

If you want to dispose of a value properly, running its destructor, see mem::drop.

Safety

`forget` 没有被标记为 `unsafe`，因为 Rust 的安全保证不包括析构函数将始终运行的保证。 例如，一个程序可以使用“Rc”创建一个引用循环，或者调用“process::exit”退出而不运行析构函数。 因此，允许从安全代码中使用 `mem::forget` 并不会从根本上改变 Rust 的安全保证。

forget is not marked as unsafe, because Rust’s safety guarantees do not include a guarantee that destructors will always run. For example, a program can create a reference cycle using Rc, or call process::exit to exit without running destructors. Thus, allowing mem::forget from safe code does not fundamentally change Rust’s safety guarantees.

也就是说，泄漏资源（例如内存或 I/O 对象）通常是不可取的。 在 FFI 或不安全代码的一些特殊用例中出现了需求，但即便如此，“ManuallyDrop”通常是首选。

That said, leaking resources such as memory or I/O objects is usually undesirable. The need comes up in some specialized use cases for FFI or unsafe code, but even then, ManuallyDrop is typically preferred.

因为允许忘记一个值，所以您编写的任何“不安全”代码都必须允许这种可能性。 你不能返回一个值并期望调用者一定会运行该值的析构函数。

Because forgetting a value is allowed, any unsafe code you write must allow for this possibility. You cannot return a value and expect that the caller will necessarily run the value’s destructor.

Examples

`mem::forget` 的规范安全使用是规避由 `Drop` Trait实现的值的析构函数。 例如，这将泄漏一个“文件”，即回收变量占用的空间，但永远不会关闭底层系统资源：

The canonical safe use of mem::forget is to circumvent a value’s destructor implemented by the Drop trait. For example, this will leak a File, i.e. reclaim the space taken by the variable but never close the underlying system resource:

use std::mem;
use std::fs::File;

let file = File::open("foo.txt").unwrap();
mem::forget(file);
Run

当底层资源的所有权先前被转移到 Rust 之外的代码时，这很有用，例如通过将原始文件描述符传输到 C 代码。

This is useful when the ownership of the underlying resource was previously transferred to code outside of Rust, for example by transmitting the raw file descriptor to C code.

Relationship with ManuallyDrop

虽然 `mem::forget` 也可用于转移内存所有权，但这样做很容易出错。 应改为使用“ManuallyDrop”。 例如，考虑以下代码：

While mem::forget can also be used to transfer memory ownership, doing so is error-prone. ManuallyDrop should be used instead. Consider, for example, this code:

use std::mem;

let mut v = vec![65, 122];
// Build a `String` using the contents of `v`
let s = unsafe { String::from_raw_parts(v.as_mut_ptr(), v.len(), v.capacity()) };
// leak `v` because its memory is now managed by `s`
mem::forget(v);  // ERROR - v is invalid and must not be passed to a function
assert_eq!(s, "Az");
// `s` is implicitly dropped and its memory deallocated.
Run

上面的例子有两个问题：

There are two issues with the above example:

如果在构造 `String` 和调用 `mem::forget()` 之间添加更多代码，其中的panic会导致双重释放，因为相同的内存由 `v` 和 `s` 处理。

If more code were added between the construction of String and the invocation of mem::forget(), a panic within it would cause a double free because the same memory is handled by both v and s.

调用 `v.as_mut_ptr()` 并将数据的所有权传递给 `s` 后，`v` 值无效。 即使一个值只是被移动到 `mem::forget` （它不会检查它），一些类型对它们的值有严格的要求，这使得它们在悬空或不再拥有时无效。 以任何方式使用无效值，包括将它们传递给函数或从函数返回它们，都会构成未定义的行为，并且可能会破坏编译器所做的假设。

After calling v.as_mut_ptr() and transmitting the ownership of the data to s, the v value is invalid. Even when a value is just moved to mem::forget (which won’t inspect it), some types have strict requirements on their values that make them invalid when dangling or no longer owned. Using invalid values in any way, including passing them to or returning them from functions, constitutes undefined behavior and may break the assumptions made by the compiler.

切换到“ManuallyDrop”可以避免这两个问题：

Switching to ManuallyDrop avoids both issues:

use std::mem::ManuallyDrop;

let v = vec![65, 122];
// Before we disassemble `v` into its raw parts, make sure it
// does not get dropped!
let mut v = ManuallyDrop::new(v);
// Now disassemble `v`. These operations cannot panic, so there cannot be a leak.
let (ptr, len, cap) = (v.as_mut_ptr(), v.len(), v.capacity());
// Finally, build a `String`.
let s = unsafe { String::from_raw_parts(ptr, len, cap) };
assert_eq!(s, "Az");
// `s` is implicitly dropped and its memory deallocated.
Run

`ManuallyDrop` 稳健地防止了双重释放，因为我们在执行任何其他操作之前禁用了 `v` 的析构函数。 `mem::forget()` 不允许这样做，因为它消耗了它的参数，迫使我们只有在从 `v` 中提取了我们需要的任何东西之后才调用它。 即使在构建“ManuallyDrop”和构建字符串之间引入了panic（这在代码中不会发生），它也会导致泄漏而不是双重释放。 换句话说，“ManuallyDrop”会在泄漏方面出错，而不是在（双）丢弃方面出错。

ManuallyDrop robustly prevents double-free because we disable v’s destructor before doing anything else. mem::forget() doesn’t allow this because it consumes its argument, forcing us to call it only after extracting anything we need from v. Even if a panic were introduced between construction of ManuallyDrop and building the string (which cannot happen in the code as shown), it would result in a leak and not a double free. In other words, ManuallyDrop errs on the side of leaking instead of erring on the side of (double-)dropping.

此外，`ManuallyDrop` 可以防止我们在将所有权转移给`s` 后不得不“触摸”`v`——完全避免了与`v` 交互以在不运行其析构函数的情况下处理它的最后一步。

Also, ManuallyDrop prevents us from having to “touch” v after transferring the ownership to s — the final step of interacting with v to dispose of it without running its destructor is entirely avoided.