Module core::num

Numeric traits and functions for the built-in numeric types.

Reexports

pub use num::dec2flt::ParseFloatError;

Modules

wrapping 	 [Unstable]

Structs

ParseIntError
An error which can be returned when parsing an integer.

Wrapping
Provides intentionally-wrapped arithmetic on T.

Enums

FpCategory
Used for representing the classification of floating point numbers

Traits

One 	 [Unstable]
Types that have a "one" value.

Zero 	 [Unstable]
Types that have a "zero" value.