Module core::num

Numeric traits and functions for the built-in numeric types.

Reexports

use char::CharExt;

use cmp::{Eq, PartialOrd};

use convert::From;

use fmt;

use intrinsics;

use marker::{Copy, Sized};

use mem::size_of;

use option::Option::{self, Some, None};

use result::Result::{self, Ok, Err};

use str::{FromStr, StrExt};

use slice::SliceExt;

pub use num::dec2flt::ParseFloatError;

Modules

bignum 	 [Unstable]
Custom arbitrary-precision number (bignum) implementation.

dec2flt 	 [Unstable]
Converting decimal strings into IEEE 754 binary floating point numbers.

diy_float 	 [Unstable]
Extended precision "soft float", for internal use only.

flt2dec 	 [Unstable]
Floating-point number to decimal conversion routines.

wrapping 	 [Unstable]

Structs

ParseIntError
An error which can be returned when parsing an integer.

Wrapping
Provides intentionally-wrapped arithmetic on T.

Enums

FpCategory
Used for representing the classification of floating point numbers

IntErrorKind

Traits

Float 	 [Unstable]
A built-in floating point number.

FromStrRadixHelper

One 	 [Unstable]
Types that have a "one" value.

Zero 	 [Unstable]
Types that have a "zero" value.

Functions

from_str_radix