Module core::numStable

Numeric traits and functions for the built-in numeric types.

Structs

ParseFloatError
An error which can be returned when parsing a float.

ParseIntError
An error which can be returned when parsing an integer.

Enums

FpCategory
Used for representing the classification of floating point numbers

Traits

Float
A built-in floating point number.

FromPrimitive
A generic trait for converting a number to a value.

FromStrRadix
A generic trait for converting a string with a radix (base) to a value

Int
A built-in signed or unsigned integer.

NumCast
An interface for casting between machine scalars.

SignedInt
A built-in two's complement integer.

ToPrimitive
A generic trait for converting a value to a number.

UnsignedInt
A built-in unsigned integer.

Functions

cast
Cast from one machine scalar to another.

from_f32
A utility function that just calls FromPrimitive::from_f32.

from_f64
A utility function that just calls FromPrimitive::from_f64.

from_i16
A utility function that just calls FromPrimitive::from_i16.

from_i32
A utility function that just calls FromPrimitive::from_i32.

from_i64
A utility function that just calls FromPrimitive::from_i64.

from_i8
A utility function that just calls FromPrimitive::from_i8.

from_int
A utility function that just calls FromPrimitive::from_int.

from_str_radix
A utility function that just calls FromStrRadix::from_str_radix.

from_u16
A utility function that just calls FromPrimitive::from_u16.

from_u32
A utility function that just calls FromPrimitive::from_u32.

from_u64
A utility function that just calls FromPrimitive::from_u64.

from_u8
A utility function that just calls FromPrimitive::from_u8.

from_uint
A utility function that just calls FromPrimitive::from_uint.