Struct core::num::ParseIntError

pub struct ParseIntError { /* private fields */ }

An error which can be returned when parsing an integer.

This error is used as the error type for the from_str_radix() functions on the primitive integer types, such as i8::from_str_radix.

Potential causes

Among other causes, ParseIntError can be thrown because of leading or trailing whitespace in the string e.g., when it is obtained from the standard input. Using the str::trim() method ensures that no whitespace remains before parsing.

Example

if let Err(e) = i32::from_str_radix("a12", 10) {
    println!("Failed conversion to i32: {e}");
}
Run

Implementations

impl ParseIntError

pub fn kind(&self) -> &IntErrorKind

Outputs the detailed cause of parsing an integer failing.