Enum core::ops::Bound

pub enum Bound<T> {
    Included(T),
    Excluded(T),
    Unbounded,
}

An endpoint of a range of keys.

Examples

Bounds are range endpoints:

use std::ops::Bound::*;
use std::ops::RangeBounds;

assert_eq!((..100).start_bound(), Unbounded);
assert_eq!((1..12).start_bound(), Included(&1));
assert_eq!((1..12).end_bound(), Excluded(&12));Run

Using a tuple of Bounds as an argument to BTreeMap::range. Note that in most cases, it's better to use range syntax (1..5) instead.

use std::collections::BTreeMap;
use std::ops::Bound::{Excluded, Included, Unbounded};

let mut map = BTreeMap::new();
map.insert(3, "a");
map.insert(5, "b");
map.insert(8, "c");

for (key, value) in map.range((Excluded(3), Included(8))) {
    println!("{}: {}", key, value);
}

assert_eq!(Some((&3, &"a")), map.range((Unbounded, Included(5))).next());Run

Variants
Included(T)

An inclusive bound.

Excluded(T)

An exclusive bound.

Unbounded

An infinite endpoint. Indicates that there is no bound in this direction.