Struct core::ops::RangeToInclusive

pub struct RangeToInclusive<Idx> {
    pub end: Idx,
}

🔬 This is a nightly-only experimental API. (inclusive_range #28237)

recently added, follows RFC

A range only bounded inclusively above (..=end).

The RangeToInclusive ..=end contains all values with x <= end. It cannot serve as an Iterator because it doesn't have a starting point.

Examples

The ..=end syntax is a RangeToInclusive:

#![feature(inclusive_range,inclusive_range_syntax)]
assert_eq!((..=5), std::ops::RangeToInclusive{ end: 5 });Run

It does not have an IntoIterator implementation, so you can't use it in a for loop directly. This won't compile:

#![feature(inclusive_range_syntax)]

// error[E0277]: the trait bound `std::ops::RangeToInclusive<{integer}>:
// std::iter::Iterator` is not satisfied
for i in ..=5 {
    // ...
}Run

When used as a slicing index, RangeToInclusive produces a slice of all array elements up to and including the index indicated by end.

#![feature(inclusive_range_syntax)]

let arr = [0, 1, 2, 3];
assert_eq!(arr[ ..=2], [0,1,2  ]);  // RangeToInclusive
assert_eq!(arr[1..=2], [  1,2  ]);Run

Fields
end: Idx

🔬 This is a nightly-only experimental API. (inclusive_range #28237)

recently added, follows RFC

The upper bound of the range (inclusive)

Methods

impl<Idx: PartialOrd<Idx>> RangeToInclusive<Idx>

pub fn contains(&self, item: Idx) -> bool

🔬 This is a nightly-only experimental API. (range_contains #32311)

recently added as per RFC

Returns true if item is contained in the range.

Examples

#![feature(range_contains,inclusive_range_syntax)]

assert!( (..=5).contains(-1_000_000_000));
assert!( (..=5).contains(5));
assert!(!(..=5).contains(6));Run