Trait core::ops::CoerceUnsized

pub trait CoerceUnsized<T: ?Sized> { }

🔬 This is a nightly-only experimental API. (coerce_unsized #27732)

表明这是一个指针或一个包装器的Trait，其中可以在指针对象上执行大小调整。

Trait that indicates that this is a pointer or a wrapper for one, where unsizing can be performed on the pointee.

有关更多详细信息，请参阅 DST 强制 RFC 和有关强制的 nomicon 条目。

See the DST coercion RFC and the nomicon entry on coercion for more details.

对于内置指针类型，如果 T: Unsize 通过从瘦指针转换为胖指针，指向 T 的指针将强制转换为指向 U 的指针。

For builtin pointer types, pointers to T will coerce to pointers to U if T: Unsize<U> by converting from a thin pointer to a fat pointer.

对于自定义类型，此处的强制通过将 Foo 强制为 Foo 来工作，提供了 CoerceUnsized<>> for Foo 的 impl 存在。 只有当 Foo 只有一个涉及 T 的非幻像数据字段时才能编写这样的 impl。如果该字段的类型是 Bar，则为 Bar 的 CoerceUnsized<>> 的实现 必须存在。 强制将通过将 Bar 字段强制转换为 Bar 并填充 Foo 中的其余字段以创建 Foo 来起作用。 这将有效地深入到一个指针字段并强制它。

For custom types, the coercion here works by coercing Foo<T> to Foo<U> provided an impl of CoerceUnsized<Foo<U>> for Foo<T> exists. Such an impl can only be written if Foo<T> has only a single non-phantomdata field involving T. If the type of that field is Bar<T>, an implementation of CoerceUnsized<Bar<U>> for Bar<T> must exist. The coercion will work by coercing the Bar<T> field into Bar<U> and filling in the rest of the fields from Foo<T> to create a Foo<U>. This will effectively drill down to a pointer field and coerce that.

通常，对于智能指针，您将为 Ptr 实现 CoerceUnsized<>> 其中 T: Unsize, U: ?Sized，在 T 本身上有一个可选的 ?Sized 边界。 对于像 Cell 和 RefCell 这样直接嵌入 T 的包装器类型，您可以直接为 Wrap 实现 CoerceUnsized<>> 其中 T: CoerceUnsized。 这将使 Cell<>> 等类型的强制起作用。

Generally, for smart pointers you will implement CoerceUnsized<Ptr<U>> for Ptr<T> where T: Unsize<U>, U: ?Sized, with an optional ?Sized bound on T itself. For wrapper types that directly embed T like Cell<T> and RefCell<T>, you can directly implement CoerceUnsized<Wrap<U>> for Wrap<T> where T: CoerceUnsized<U>. This will let coercions of types like Cell<Box<T>> work.

Unsize 用于标记如果在指针后面可以强制转换为 DST 的类型。 它由编译器自动实现。

Unsize is used to mark types which can be coerced to DSTs if behind pointers. It is implemented automatically by the compiler.