Trait core::ops::DropStable

pub trait Drop {
    fn drop(&mut self);
}

The Drop trait is used to run some code when a value goes out of scope. This is sometimes called a 'destructor'.

Example

A trivial implementation of Drop. The drop method is called when _x goes out of scope, and therefore main prints Dropping!.
struct HasDrop; impl Drop for HasDrop { fn drop(&mut self) { println!("Dropping!"); } } fn main() { let _x = HasDrop; }

struct HasDrop;

impl Drop for HasDrop {
    fn drop(&mut self) {
        println!("Dropping!");
    }
}

fn main() {
    let _x = HasDrop;
}

Required Methods

fn drop(&mut self)

The drop method, called when the value goes out of scope.