Trait core::ops::NotStable

pub trait Not {
    type Output;

    fn not(self) -> Self::Output;
}

The Not trait is used to specify the functionality of unary !.

Example

A trivial implementation of Not. When !Foo happens, it ends up calling not, and therefore, main prints Not-ing!.
use std::ops::Not; #[derive(Copy)] struct Foo; impl Not for Foo { type Output = Foo; fn not(self) -> Foo { println!("Not-ing!"); self } } fn main() { !Foo; }

use std::ops::Not;

#[derive(Copy)]
struct Foo;

impl Not for Foo {
    type Output = Foo;

    fn not(self) -> Foo {
        println!("Not-ing!");
        self
    }
}

fn main() {
    !Foo;
}

Associated Types

type Output

The resulting type after applying the ! operator

Required Methods

fn not(self) -> Self::Output

The method for the unary ! operator