Enum core::option::Option

pub enum Option<T> {
    None,
    Some(T),
}

Option ç±»å‹ã€‚æœ‰å…³æ›´å¤šä¿¡æ¯ï¼Œè¯·å‚è æ¨¡å—çºæ–‡æ¡£ã€‚

Variants

None

æ²¡æœ‰å€¼

Some(T)

ä¸€äº›å€¼ T

Implementations

impl<T> Option<T>

#[must_use = "if you intended to assert that this has a value, consider `.unwrap()` instead"]
pub const fn is_some(&self) -> bool

å¦‚æœé€‰é¡¹æ˜¯ Some å€¼ï¼Œåˆ™è¿”å› trueã€‚

Examples

let x: Option<u32> = Some(2);
assert_eq!(x.is_some(), true);

let x: Option<u32> = None;
assert_eq!(x.is_some(), false);
Run

#[must_use = "if you intended to assert that this doesn't have a value, consider \ `.and_then(|_| panic!(\"`Option` had a value when expected `None`\"))` instead"]
pub const fn is_none(&self) -> bool

å¦‚æœé€‰é¡¹æ˜¯ None å€¼ï¼Œåˆ™è¿”å› trueã€‚

Examples

let x: Option<u32> = Some(2);
assert_eq!(x.is_none(), false);

let x: Option<u32> = None;
assert_eq!(x.is_none(), true);
Run

#[must_use]
pub fn contains<U>(&self, x: &U) -> bool where
U: PartialEq<T>,

ğŸ”¬ This is a nightly-only experimental API. (option_result_contains #62358)

å¦‚æœé€‰é¡¹æ˜¯åŒ…å«ç»™å®šå€¼çš„ Some å€¼ï¼Œåˆ™è¿”å› trueã€‚

Examples

#![feature(option_result_contains)]

let x: Option<u32> = Some(2);
assert_eq!(x.contains(&2), true);

let x: Option<u32> = Some(3);
assert_eq!(x.contains(&2), false);

let x: Option<u32> = None;
assert_eq!(x.contains(&2), false);
Run

pub const fn as_ref(&self) -> Option<&T>

ä» &Option<T> è½¬æ¢ä¸º Option<&T>ã€‚

Examples

å°† Option<String> è½¬æ¢ä¸º Option<usize>ï¼Œå¹¶ä¿ç•™åŸå‹æ ¼å¼ã€‚ map æ–¹æ³•æŒ‰å€¼ä½¿ç”¨ self å‚æ•°ï¼Œä»è€Œæ¶ˆè€—äº†åŸå‹æ–‡ä»¶ï¼Œå› æ­¤è¯¥æŠ€æœ¯ä½¿ç”¨ as_ref é¦–å…ˆå°† Option å¼•ç”¨ç»™åŸå‹æ–‡ä»¶ä¸­çš„å€¼ã€‚

let text: Option<String> = Some("Hello, world!".to_string());
// é¦–å…ˆï¼Œä½¿ç”¨ `as_ref` å°† `Option<String>` è½¬æ¢ä¸º `Option<&String>`ï¼Œç„¶åä½¿ç”¨ `map` æ¶ˆè€— *that*ï¼Œå°† `text` ä¿ç•™åœ¨æ ˆä¸­ã€‚
let text_length: Option<usize> = text.as_ref().map(|s| s.len());
println!("still can print text: {:?}", text);
Run

pub fn as_mut(&mut self) -> Option<&mut T>

ä» &mut Option<T> è½¬æ¢ä¸º Option<&mut T>ã€‚

Examples

let mut x = Some(2);
match x.as_mut() {
    Some(v) => *v = 42,
    None => {},
}
assert_eq!(x, Some(42));
Run

pub fn as_pin_ref(self: Pin<&Self>) -> Option<Pin<&T>>

ä» Pin<&Option<T>> è½¬æ¢ä¸º Option<Pin<&T>>ã€‚

pub fn as_pin_mut(self: Pin<&mut Self>) -> Option<Pin<&mut T>>

ä» Pin<&mut Option<T>> è½¬æ¢ä¸º Option <Pin<&mut T>>ã€‚

pub fn expect(self, msg: &str) -> T

è¿”å›åŒ…å« self å€¼çš„åŒ…å«çš„ Some å€¼ã€‚

Panics

Panics (å¦‚æœå€¼æ˜¯å…·æœ‰ msg æä¾›çš„è‡ªå®šä¹‰ panic æ¶ˆæ¯çš„ None)ã€‚

Examples

let x = Some("value");
assert_eq!(x.expect("fruits are healthy"), "value");
Run

let x: Option<&str> = None;
x.expect("fruits are healthy"); // `fruits are healthy` çš„ panics
Run

pub fn unwrap(self) -> T

è¿”å›åŒ…å« self å€¼çš„åŒ…å«çš„ Some å€¼ã€‚

ç”±äºæ­¤å‡½æ•°å¯èƒ½ä¸º panicï¼Œå› æ­¤é€šå¸¸ä¸å»ºè®®ä½¿ç”¨è¯¥å‡½æ•°ã€‚ ç›¸åï¼Œæ›´å–œæ¬¢ä½¿ç”¨æ¨¡å¼åŒ¹é…å¹¶æ˜¾å¼å¤„ç† None å¤å°å†™ï¼Œæˆ–è€…è°ƒç”¨ unwrap_orï¼Œunwrap_or_else æˆ– unwrap_or_defaultã€‚

Panics

å¦‚æœè‡ªèº«å€¼ç­‰äº Noneï¼Œåˆ™ä¸º Panicsã€‚

Examples

let x = Some("air");
assert_eq!(x.unwrap(), "air");
Run

let x: Option<&str> = None;
assert_eq!(x.unwrap(), "air"); // fails
Run

pub fn unwrap_or(self, default: T) -> T

è¿”å›åŒ…å«çš„ Some å€¼æˆ–æä¾›çš„é»˜è®¤å€¼ã€‚

æ€¥åˆ‡åœ°è¯„ä¼°ä¼ é€’ç»™ unwrap_or çš„å‚æ•°; å¦‚æœè¦ä¼ é€’å‡½æ•°è°ƒç”¨çš„ç»“æœï¼Œå»ºè®®ä½¿ç”¨ unwrap_or_elseï¼Œå®ƒæ˜¯æƒ°æ€æ±‚å€¼çš„ã€‚

Examples

assert_eq!(Some("car").unwrap_or("bike"), "car");
assert_eq!(None.unwrap_or("bike"), "bike");
Run

pub fn unwrap_or_else<F: FnOnce() -> T>(self, f: F) -> T

è¿”å›åŒ…å«çš„ Some å€¼æˆ–ä»é—­åŒ…ä¸­è®¡ç®—å¾—å‡ºã€‚

Examples

let k = 10;
assert_eq!(Some(4).unwrap_or_else(|| 2 * k), 4);
assert_eq!(None.unwrap_or_else(|| 2 * k), 20);
Run

pub unsafe fn unwrap_unchecked(self) -> T

ğŸ”¬ This is a nightly-only experimental API. (option_result_unwrap_unchecked #81383)

newly added

è¿”å›åŒ…å« self å€¼çš„åŒ…å«çš„ Some å€¼ï¼Œè€Œä¸æ£€æŸ¥è¯¥å€¼æ˜¯å¦ä¸æ˜¯ Noneã€‚

Safety

åœ¨ None ä¸Šè°ƒç”¨æ­¤æ–¹æ³•æ˜¯ undefined behaviorã€‚

Examples

#![feature(option_result_unwrap_unchecked)]
let x = Some("air");
assert_eq!(unsafe { x.unwrap_unchecked() }, "air");
Run

#![feature(option_result_unwrap_unchecked)]
let x: Option<&str> = None;
assert_eq!(unsafe { x.unwrap_unchecked() }, "air"); // æœªå®šä¹‰çš„è¡Œä¸º!
Run

pub fn map<U, F: FnOnce(T) -> U>(self, f: F) -> Option<U>

é€šè¿‡å°†å‡½æ•°åº”ç”¨äºåŒ…å«çš„å€¼ï¼ŒMaps å°† Option<T> è½¬æ¢ä¸º Option<U>ã€‚

Examples

ä½¿ç”¨åŸå‹æ–‡ä»¶å°† Option<String> è½¬æ¢ä¸º Option<usize>:

let maybe_some_string = Some(String::from("Hello, World!"));
// `Option::map` éœ€è¦ self æŒ‰ç…å€¼æ¶ˆè€— `maybe_some_string`
let maybe_some_len = maybe_some_string.map(|s| s.len());

assert_eq!(maybe_some_len, Some(13));
Run

pub fn map_or<U, F: FnOnce(T) -> U>(self, default: U, f: F) -> U

è¿”å›æä¾›çš„é»˜è®¤ç»“æœ (å¦‚æœæ²¡æœ‰)ï¼Œæˆ–å°†å‡½æ•°åº”ç”¨äºåŒ…å«çš„å€¼ (å¦‚æœæœ‰)ã€‚

æ€¥åˆ‡åœ°è¯„ä¼°ä¼ é€’ç»™ map_or çš„å‚æ•°; å¦‚æœè¦ä¼ é€’å‡½æ•°è°ƒç”¨çš„ç»“æœï¼Œå»ºè®®ä½¿ç”¨ map_or_elseï¼Œå®ƒæ˜¯å»¶è¿Ÿè®¡ç®—çš„ã€‚

Examples

let x = Some("foo");
assert_eq!(x.map_or(42, |v| v.len()), 3);

let x: Option<&str> = None;
assert_eq!(x.map_or(42, |v| v.len()), 42);
Run

pub fn map_or_else<U, D: FnOnce() -> U, F: FnOnce(T) -> U>(
self,
default: D,
f: F
) -> U

è®¡ç®— default å‡½æ•°çš„ç»“æœ (å¦‚æœæ²¡æœ‰)ï¼Œæˆ–å°†ä¸åŒçš„å‡½æ•°åº”ç”¨äºåŒ…å«çš„å€¼ (å¦‚æœæœ‰)ã€‚

Examples

let k = 21;

let x = Some("foo");
assert_eq!(x.map_or_else(|| 2 * k, |v| v.len()), 3);

let x: Option<&str> = None;
assert_eq!(x.map_or_else(|| 2 * k, |v| v.len()), 42);
Run

pub fn ok_or<E>(self, err: E) -> Result<T, E>

å°† Option<T> è½¬æ¢ä¸º Result<T, E>ï¼Œå°† Some(v) æ˜ å°„åˆ° Ok(v)ï¼Œå°† None æ˜ å°„åˆ° Err(err)ã€‚

æ€¥åˆ‡åœ°è¯„ä¼°ä¼ é€’ç»™ ok_or çš„å‚æ•°; å¦‚æœè¦ä¼ é€’å‡½æ•°è°ƒç”¨çš„ç»“æœï¼Œå»ºè®®ä½¿ç”¨ ok_or_elseï¼Œå®ƒæ˜¯å»¶è¿Ÿè®¡ç®—çš„ã€‚

Examples

let x = Some("foo");
assert_eq!(x.ok_or(0), Ok("foo"));

let x: Option<&str> = None;
assert_eq!(x.ok_or(0), Err(0));
Run

pub fn ok_or_else<E, F: FnOnce() -> E>(self, err: F) -> Result<T, E>

å°† Option<T> è½¬æ¢ä¸º Result<T, E>ï¼Œå°† Some(v) æ˜ å°„åˆ° Ok(v)ï¼Œå°† None æ˜ å°„åˆ° Err(err())ã€‚

Examples

let x = Some("foo");
assert_eq!(x.ok_or_else(|| 0), Ok("foo"));

let x: Option<&str> = None;
assert_eq!(x.ok_or_else(|| 0), Err(0));
Run

pub fn iter(&self) -> Iter<'_, T>

è¿”å›å¯èƒ½åŒ…å«çš„å€¼çš„è¿­ä»£å™¨ã€‚

Examples

let x = Some(4);
assert_eq!(x.iter().next(), Some(&4));

let x: Option<u32> = None;
assert_eq!(x.iter().next(), None);
Run

pub fn iter_mut(&mut self) -> IterMut<'_, T>

è¿”å›å¯èƒ½åŒ…å«çš„å€¼çš„å¯å˜è¿­ä»£å™¨ã€‚

Examples

let mut x = Some(4);
match x.iter_mut().next() {
    Some(v) => *v = 42,
    None => {},
}
assert_eq!(x, Some(42));

let mut x: Option<u32> = None;
assert_eq!(x.iter_mut().next(), None);
Run

pub fn and<U>(self, optb: Option<U>) -> Option<U>

å¦‚æœé€‰é¡¹ä¸º Noneï¼Œåˆ™è¿”å› None; å¦åˆ™ï¼Œè¿”å› optbã€‚

Examples

let x = Some(2);
let y: Option<&str> = None;
assert_eq!(x.and(y), None);

let x: Option<u32> = None;
let y = Some("foo");
assert_eq!(x.and(y), None);

let x = Some(2);
let y = Some("foo");
assert_eq!(x.and(y), Some("foo"));

let x: Option<u32> = None;
let y: Option<&str> = None;
assert_eq!(x.and(y), None);
Run

pub fn and_then<U, F: FnOnce(T) -> Option<U>>(self, f: F) -> Option<U>

å¦‚æœé€‰é¡¹ä¸º Noneï¼Œåˆ™è¿”å› None; å¦åˆ™ï¼Œä½¿ç”¨åŒ…è£…çš„å€¼è°ƒç”¨ fï¼Œå¹¶è¿”å›ç»“æœã€‚

ä¸€äº›è¯­è¨€è°ƒç”¨æ­¤æ“ä½œå¹³é¢å›¾ã€‚

Examples

fn sq(x: u32) -> Option<u32> { Some(x * x) }
fn nope(_: u32) -> Option<u32> { None }

assert_eq!(Some(2).and_then(sq).and_then(sq), Some(16));
assert_eq!(Some(2).and_then(sq).and_then(nope), None);
assert_eq!(Some(2).and_then(nope).and_then(sq), None);
assert_eq!(None.and_then(sq).and_then(sq), None);
Run

pub fn filter<P: FnOnce(&T) -> bool>(self, predicate: P) -> Self

å¦‚æœé€‰é¡¹ä¸º Noneï¼Œåˆ™è¿”å› None; å¦åˆ™ï¼Œä½¿ç”¨åŒ…è£…çš„å€¼è°ƒç”¨ predicate å¹¶è¿”å›:

Some(t) å¦‚æœ predicate è¿”å› true (å…¶ä¸­ t æ˜¯åŒ…è£…çš„å€¼)ï¼Œå¹¶ä¸”

None å¦‚æœ predicate è¿”å› falseã€‚

è¯¥å‡½æ•°çš„å·¥ä½œæ–¹å¼ç±»ä¼¼äº Iterator::filter()ã€‚ æ‚¨å¯ä»¥æƒ³è±¡ Option<T> æ˜¯ä¸€ä¸ªæˆ–é›¶ä¸ªå…ƒç´ ä¸Šçš„è¿­ä»£å™¨ã€‚ filter() è®©æ‚¨å†³å®šè¦ä¿ç•™å“ªäº›å…ƒç´ ã€‚

Examples

fn is_even(n: &i32) -> bool {
    n % 2 == 0
}

assert_eq!(None.filter(is_even), None);
assert_eq!(Some(3).filter(is_even), None);
assert_eq!(Some(4).filter(is_even), Some(4));
Run

pub fn or(self, optb: Option<T>) -> Option<T>

å¦‚æœåŒ…å«å€¼ï¼Œåˆ™è¿”å›é€‰é¡¹ï¼Œå¦åˆ™è¿”å› optbã€‚

æ€¥åˆ‡åœ°è¯„ä¼°ä¼ é€’ç»™ or çš„å‚æ•°; å¦‚æœè¦ä¼ é€’å‡½æ•°è°ƒç”¨çš„ç»“æœï¼Œå»ºè®®ä½¿ç”¨ or_elseï¼Œå®ƒæ˜¯å»¶è¿Ÿè®¡ç®—çš„ã€‚

Examples

let x = Some(2);
let y = None;
assert_eq!(x.or(y), Some(2));

let x = None;
let y = Some(100);
assert_eq!(x.or(y), Some(100));

let x = Some(2);
let y = Some(100);
assert_eq!(x.or(y), Some(2));

let x: Option<u32> = None;
let y = None;
assert_eq!(x.or(y), None);
Run

pub fn or_else<F: FnOnce() -> Option<T>>(self, f: F) -> Option<T>

å¦‚æœé€‰é¡¹åŒ…å«å€¼ï¼Œåˆ™è¿”å›è¯¥é€‰é¡¹ï¼Œå¦åˆ™è°ƒç”¨ f å¹¶è¿”å›ç»“æœã€‚

Examples

fn nobody() -> Option<&'static str> { None }
fn vikings() -> Option<&'static str> { Some("vikings") }

assert_eq!(Some("barbarians").or_else(vikings), Some("barbarians"));
assert_eq!(None.or_else(vikings), Some("vikings"));
assert_eq!(None.or_else(nobody), None);
Run

pub fn xor(self, optb: Option<T>) -> Option<T>

å¦‚æœ selfï¼Œoptb ä¹‹ä¸€æ°å¥½æ˜¯ Someï¼Œåˆ™è¿”å› Someï¼Œå¦åˆ™è¿”å› Noneã€‚

Examples

let x = Some(2);
let y: Option<u32> = None;
assert_eq!(x.xor(y), Some(2));

let x: Option<u32> = None;
let y = Some(2);
assert_eq!(x.xor(y), Some(2));

let x = Some(2);
let y = Some(2);
assert_eq!(x.xor(y), None);

let x: Option<u32> = None;
let y: Option<u32> = None;
assert_eq!(x.xor(y), None);
Run

#[must_use = "if you intended to set a value, consider assignment instead"]
pub fn insert(&mut self, value: T) -> &mut T

å°† value æ’å…¥é€‰é¡¹ï¼Œç„¶åå‘å…¶è¿”å›å˜é‡å¼•ç”¨ã€‚

å¦‚æœè¯¥é€‰é¡¹å·²åŒ…å«å€¼ï¼Œåˆ™å°†ä¸¢å¼ƒæ—å€¼ã€‚

å¦è¯·å‚è Option::get_or_insertï¼Œå¦‚æœé€‰é¡¹å·²åŒ…å« Someï¼Œåˆ™ä¸ä¼šæ›´æ–°å€¼ã€‚

Example

let mut opt = None;
let val = opt.insert(1);
assert_eq!(*val, 1);
assert_eq!(opt.unwrap(), 1);
let val = opt.insert(2);
assert_eq!(*val, 2);
*val = 3;
assert_eq!(opt.unwrap(), 3);
Run

pub fn get_or_insert(&mut self, value: T) -> &mut T

å¦‚æœä¸º Noneï¼Œåˆ™å°† value æ’å…¥åˆ°é€‰é¡¹ä¸­ï¼Œç„¶åè¿”å›æ‰€åŒ…å«çš„å€¼çš„å˜é‡å¼•ç”¨ã€‚

å¦è¯·å‚è Option::insertï¼Œå³ä½¿é€‰é¡¹å·²åŒ…å« Someï¼Œå®ƒä¹Ÿä¼šæ›´æ–°å€¼ã€‚

Examples

let mut x = None;

{
    let y: &mut u32 = x.get_or_insert(5);
    assert_eq!(y, &5);

    *y = 7;
}

assert_eq!(x, Some(7));
Run

pub fn get_or_insert_default(&mut self) -> &mut T where
T: Default,

ğŸ”¬ This is a nightly-only experimental API. (option_get_or_insert_default #82901)

å¦‚æœé»˜è®¤å€¼ä¸º Noneï¼Œåˆ™å°†å…¶æ’å…¥é€‰é¡¹ä¸­ï¼Œç„¶åå°†æ‰€åŒ…å«çš„å€¼è¿”å›å˜é‡å¼•ç”¨ã€‚

Examples

#![feature(option_get_or_insert_default)]

let mut x = None;

{
    let y: &mut u32 = x.get_or_insert_default();
    assert_eq!(y, &0);

    *y = 7;
}

assert_eq!(x, Some(7));
Run

pub fn get_or_insert_with<F: FnOnce() -> T>(&mut self, f: F) -> &mut T

å¦‚æœä» f è®¡ç®—å¾—å‡ºçš„å€¼æ˜¯ Noneï¼Œåˆ™å°†å…¶æ’å…¥é€‰é¡¹ä¸­ï¼Œç„¶åå°†æ‰€åŒ…å«çš„å€¼è¿”å›å¯å˜å¼•ç”¨ã€‚

Examples

let mut x = None;

{
    let y: &mut u32 = x.get_or_insert_with(|| 5);
    assert_eq!(y, &5);

    *y = 7;
}

assert_eq!(x, Some(7));
Run

pub fn take(&mut self) -> Option<T>

ä»é€‰é¡¹ä¸­å–å‡ºå€¼ï¼Œå°† None ç•™åœ¨å…¶ä½ç½®ã€‚

Examples

let mut x = Some(2);
let y = x.take();
assert_eq!(x, None);
assert_eq!(y, Some(2));

let mut x: Option<u32> = None;
let y = x.take();
assert_eq!(x, None);
assert_eq!(y, None);
Run

pub fn replace(&mut self, value: T) -> Option<T>

ç”¨å‚æ•°ä¸­ç»™å®šçš„å€¼æ›¿æ¢é€‰é¡¹ä¸­çš„å®é™…å€¼ï¼Œå¦‚æœå­˜åœ¨åˆ™è¿”å›æ—å€¼ï¼Œå°† Some ä¿ç•™åœ¨å…¶ä½ç½®ï¼Œè€Œä¸ç”¨å¯¹å…¶ä¸­ä¸€ä¸ªè¿›è¡Œåˆå‹åŒ–ã€‚

Examples

let mut x = Some(2);
let old = x.replace(5);
assert_eq!(x, Some(5));
assert_eq!(old, Some(2));

let mut x = None;
let old = x.replace(3);
assert_eq!(x, Some(3));
assert_eq!(old, None);
Run

pub fn zip<U>(self, other: Option<U>) -> Option<(T, U)>

ç”¨å¦ä¸€ä¸ª Option å‹ç¼© selfã€‚

å¦‚æœ self æ˜¯ Some(s)ï¼Œè€Œ other æ˜¯ Some(o)ï¼Œåˆ™æ­¤æ–¹æ³•è¿”å› Some((s, o))ã€‚ å¦åˆ™ï¼Œè¿”å› Noneã€‚

Examples

let x = Some(1);
let y = Some("hi");
let z = None::<u8>;

assert_eq!(x.zip(y), Some((1, "hi")));
assert_eq!(x.zip(z), None);
Run

pub fn zip_with<U, F, R>(self, other: Option<U>, f: F) -> Option<R> where
F: FnOnce(T, U) -> R,

ğŸ”¬ This is a nightly-only experimental API. (option_zip #70086)

ä½¿ç”¨å‡½æ•° f å‹ç¼© self å’Œå¦ä¸€ä¸ª Optionã€‚

å¦‚æœ self æ˜¯ Some(s)ï¼Œè€Œ other æ˜¯ Some(o)ï¼Œåˆ™æ­¤æ–¹æ³•è¿”å› Some(f(s, o))ã€‚ å¦åˆ™ï¼Œè¿”å› Noneã€‚

Examples

#![feature(option_zip)]

#[derive(Debug, PartialEq)]
struct Point {
    x: f64,
    y: f64,
}

impl Point {
    fn new(x: f64, y: f64) -> Self {
        Self { x, y }
    }
}

let x = Some(17.5);
let y = Some(42.7);

assert_eq!(x.zip_with(y, Point::new), Some(Point { x: 17.5, y: 42.7 }));
assert_eq!(x.zip_with(None, Point::new), None);
Run

impl<T: Copy> Option<&T>

pub fn copied(self) -> Option<T>

é€šè¿‡å¤åˆ¶é€‰é¡¹çš„å†…å®¹å°† Option<&T> çš„ Maps è½¬æ¢ä¸º Option<T>ã€‚

Examples

let x = 12;
let opt_x = Some(&x);
assert_eq!(opt_x, Some(&12));
let copied = opt_x.copied();
assert_eq!(copied, Some(12));
Run

impl<T: Copy> Option<&mut T>

pub fn copied(self) -> Option<T>

é€šè¿‡å¤åˆ¶é€‰é¡¹çš„å†…å®¹å°† Option<&mut T> çš„ Maps è½¬æ¢ä¸º Option<T>ã€‚

Examples

let mut x = 12;
let opt_x = Some(&mut x);
assert_eq!(opt_x, Some(&mut 12));
let copied = opt_x.copied();
assert_eq!(copied, Some(12));
Run

impl<T: Clone> Option<&T>

pub fn cloned(self) -> Option<T>

é€šè¿‡å…‹éš†é€‰é¡¹çš„å†…å®¹å°† Option<&T> Maps è½¬æ¢ä¸º Option<T>ã€‚

Examples

let x = 12;
let opt_x = Some(&x);
assert_eq!(opt_x, Some(&12));
let cloned = opt_x.cloned();
assert_eq!(cloned, Some(12));
Run

impl<T: Clone> Option<&mut T>

pub fn cloned(self) -> Option<T>

é€šè¿‡å…‹éš†é€‰é¡¹çš„å†…å®¹å°† Option<&mut T> Maps è½¬æ¢ä¸º Option<T>ã€‚

Examples

let mut x = 12;
let opt_x = Some(&mut x);
assert_eq!(opt_x, Some(&mut 12));
let cloned = opt_x.cloned();
assert_eq!(cloned, Some(12));
Run

impl<T: Default> Option<T>

pub fn unwrap_or_default(self) -> T

è¿”å›åŒ…å«çš„ Some å€¼æˆ–é»˜è®¤å€¼

æ¶ˆè´¹ self å‚æ•°ï¼Œå¦‚æœ Someï¼Œåˆ™è¿”å›æ‰€åŒ…å«çš„å€¼ï¼Œå¦åˆ™ï¼Œå¦‚æœ Noneï¼Œåˆ™è¿”å›è¯¥ç±»å‹çš„ default valueã€‚

Examples

å°†å­—ç¬¦ä¸²è½¬æ¢ä¸ºæ•´æ•°ï¼Œå°†æ ¼å¼ä¸æ­£ç¡®çš„å­—ç¬¦ä¸²è½¬æ¢ä¸º 0 (æ•´æ•°çš„é»˜è®¤å€¼)ã€‚ parse å°†å­—ç¬¦ä¸²è½¬æ¢ä¸ºå®ç° FromStr çš„ä»»ä½•å…¶ä»–ç±»å‹ï¼Œå¹¶åœ¨é”™è¯¯æ—¶è¿”å› Noneã€‚

let good_year_from_input = "1909";
let bad_year_from_input = "190blarg";
let good_year = good_year_from_input.parse().ok().unwrap_or_default();
let bad_year = bad_year_from_input.parse().ok().unwrap_or_default();

assert_eq!(1909, good_year);
assert_eq!(0, bad_year);
Run

impl<T: Deref> Option<T>

pub fn as_deref(&self) -> Option<&T::Target>

ä» Option<T> (æˆ– &Option<T>) è½¬æ¢ä¸º Option<&T::Target>ã€‚

å°†åŸå‹ Option ä¿ç•™åœ¨åŸä½ï¼Œåˆ›å»ºä¸€ä¸ªå¸¦æœ‰å¯¹åŸå‹ Option çš„å¼•ç”¨çš„æ–° Optionï¼Œå¹¶é€šè¿‡ Deref å¼ºåˆ¶æ‰è¡Œå…¶å†…å®¹ã€‚

Examples

let x: Option<String> = Some("hey".to_owned());
assert_eq!(x.as_deref(), Some("hey"));

let x: Option<String> = None;
assert_eq!(x.as_deref(), None);
Run

impl<T: DerefMut> Option<T>

pub fn as_deref_mut(&mut self) -> Option<&mut T::Target>

ä» Option<T> (æˆ– &mut Option<T>) è½¬æ¢ä¸º Option<&mut T::Target>ã€‚

å°†åŸå‹ Option ä¿ç•™åœ¨åŸä½ï¼Œåˆ›å»ºä¸€ä¸ªæ–°çš„ Optionï¼Œå…¶ä¸­åŒ…å«å¯¹å†…éƒ¨ç±»å‹çš„ Deref::Target ç±»å‹çš„å¼•ç”¨ã€‚

Examples

let mut x: Option<String> = Some("hey".to_owned());
assert_eq!(x.as_deref_mut().map(|x| {
    x.make_ascii_uppercase();
    x
}), Some("HEY".to_owned().as_mut_str()));
Run

impl<T, E> Option<Result<T, E>>

pub fn transpose(self) -> Result<Option<T>, E>

å°† Result çš„ Option è½¬æ¢ä¸º Option çš„ Resultã€‚

None å°†è¢«æ˜ å°„åˆ° Ok(None)ã€‚ Some(Ok(_)) Some(Err(_)) å°†è¢«æ˜ å°„åˆ° Ok(Some(_)) å’Œ Err(_)ã€‚

Examples

#[derive(Debug, Eq, PartialEq)]
struct SomeErr;

let x: Result<Option<i32>, SomeErr> = Ok(Some(5));
let y: Option<Result<i32, SomeErr>> = Some(Ok(5));
assert_eq!(x, y.transpose());
Run

impl<T> Option<Option<T>>

pub fn flatten(self) -> Option<T>

ä» Option<Option<T>> è½¬æ¢ä¸º Option<T>

Examples

åŸºæœ¬ç”¨æ³•:

let x: Option<Option<u32>> = Some(Some(6));
assert_eq!(Some(6), x.flatten());

let x: Option<Option<u32>> = Some(None);
assert_eq!(None, x.flatten());

let x: Option<Option<u32>> = None;
assert_eq!(None, x.flatten());
Run

å±•å¹³ä¸€æ¬¡åªèƒ½åˆ é™¤ä¸€å±‚åµŒå¥—:

let x: Option<Option<Option<u32>>> = Some(Some(Some(6)));
assert_eq!(Some(Some(6)), x.flatten());
assert_eq!(Some(6), x.flatten().flatten());
Run