Struct core::panic::PanicInfo

pub struct PanicInfo<'a> { /* private fields */ }

A struct providing information about a panic.

PanicInfo structure is passed to a panic hook set by the set_hook function.

Examples

use std::panic;

panic::set_hook(Box::new(|panic_info| {
    println!("panic occurred: {panic_info}");
}));

panic!("critical system failure");
Run

Implementations

impl<'a> PanicInfo<'a>

pub fn payload(&self) -> &(dyn Any + Send)

Returns the payload associated with the panic.

This will commonly, but not always, be a &'static str or String.

Examples

use std::panic;

panic::set_hook(Box::new(|panic_info| {
    if let Some(s) = panic_info.payload().downcast_ref::<&str>() {
        println!("panic occurred: {s:?}");
    } else {
        println!("panic occurred");
    }
}));

panic!("Normal panic");
Run

pub fn message(&self) -> Option<&Arguments<'_>>

🔬This is a nightly-only experimental API. (panic_info_message #66745)

If the panic! macro from the core crate (not from std) was used with a formatting string and some additional arguments, returns that message ready to be used for example with fmt::write

pub fn location(&self) -> Option<&Location<'_>>

Returns information about the location from which the panic originated, if available.

This method will currently always return Some, but this may change in future versions.

Examples

use std::panic;

panic::set_hook(Box::new(|panic_info| {
    if let Some(location) = panic_info.location() {
        println!("panic occurred in file '{}' at line {}",
            location.file(),
            location.line(),
        );
    } else {
        println!("panic occurred but can't get location information...");
    }
}));

panic!("Normal panic");
Run

pub fn can_unwind(&self) -> bool

🔬This is a nightly-only experimental API. (panic_can_unwind #92988)

Returns whether the panic handler is allowed to unwind the stack from the point where the panic occurred.

This is true for most kinds of panics with the exception of panics caused by trying to unwind out of a Drop implementation or a function whose ABI does not support unwinding.

It is safe for a panic handler to unwind even when this function returns false, however this will simply cause the panic handler to be called again.