Module core::panicking

🔬 This is a nightly-only experimental API. (core_panic)

对 libcore 的panic支持

Panic support for libcore

核心库不能定义panic，但它确实声明了panic。 这意味着允许 libcore 内部的函数发生panic，但为了有用，上游板条箱必须定义panic以供 libcore 使用。 当前panic的界面是：

The core library cannot define panicking, but it does declare panicking. This means that the functions inside of libcore are allowed to panic, but to be useful an upstream crate must define panicking for libcore to use. The current interface for panicking is:

fn panic_impl(pi: &core::panic::PanicInfo<'_>) -> !
Run

此定义允许使用任何常规消息进行panic，但不允许使用 `Box` 值失败。 （`PanicInfo` 只包含一个 `&(dyn Any + Send)`，为此我们在 `PanicInfo::internal_constructor` 中填充了一个虚拟值。）这样做的原因是 libcore 不允许分配。

This definition allows for panicking with any general message, but it does not allow for failing with a Box<Any> value. (PanicInfo just contains a &(dyn Any + Send), for which we fill in a dummy value in PanicInfo::internal_constructor.) The reason for this is that libcore is not allowed to allocate.

这个模块包含一些其他的panic函数，但这些只是编译器必要的语言项。 所有panic都通过这一功能汇集。 实际符号是通过 `#[panic_handler]` 属性声明的。

This module contains a few other panicking functions, but these are just the necessary lang items for the compiler. All panics are funneled through this one function. The actual symbol is declared through the #[panic_handler] attribute.

Functions

const_panic_fmtExperimental

This function is used instead of panic_fmt in const eval.

panicExperimental

The underlying implementation of libcore’s panic! macro when no formatting is used.

panic_displayExperimental

panic_fmtExperimental

The entry point for panicking with a formatted message.

panic_strExperimental