Module core::prelude::v1

The core prelude

This module is intended for users of libcore which do not link to libstd as well. This module is imported by default when #![no_std] is used in the same manner as the standard library's prelude.

Re-exports

pub use crate::marker::Copy;

pub use crate::marker::Copy;

pub use crate::marker::Send;

pub use crate::marker::Sized;

pub use crate::marker::Sync;

pub use crate::marker::Unpin;

pub use crate::ops::Drop;

pub use crate::ops::Fn;

pub use crate::ops::FnMut;

pub use crate::ops::FnOnce;

pub use crate::mem::drop;

pub use crate::clone::Clone;

pub use crate::clone::Clone;

pub use crate::cmp::PartialEq;

pub use crate::cmp::PartialEq;

pub use crate::cmp::PartialOrd;

pub use crate::cmp::PartialOrd;

pub use crate::cmp::Eq;

pub use crate::cmp::Eq;

pub use crate::cmp::Ord;

pub use crate::cmp::Ord;

pub use crate::convert::AsRef;

pub use crate::convert::AsMut;

pub use crate::convert::Into;

pub use crate::convert::From;

pub use crate::default::Default;

pub use crate::default::Default;

pub use crate::iter::Iterator;

pub use crate::iter::Extend;

pub use crate::iter::IntoIterator;

pub use crate::iter::DoubleEndedIterator;

pub use crate::iter::ExactSizeIterator;

pub use crate::option::Option;

pub use crate::option::Option::Some;

pub use crate::option::Option::None;

pub use crate::result::Result;

pub use crate::result::Result::Ok;

pub use crate::result::Result::Err;

pub use crate::fmt::macros::Debug;

pub use crate::hash::macros::Hash;

pub use crate::asm;

pub use crate::assert;

pub use crate::cfg;

pub use crate::column;

pub use crate::compile_error;

pub use crate::concat;

pub use crate::concat_idents;

pub use crate::env;

pub use crate::file;

pub use crate::format_args;

pub use crate::format_args_nl;

pub use crate::global_asm;

pub use crate::include;

pub use crate::include_bytes;

pub use crate::include_str;

pub use crate::line;

pub use crate::log_syntax;

pub use crate::module_path;

pub use crate::option_env;

pub use crate::stringify;

pub use crate::trace_macros;

pub use crate::macros::builtin::RustcDecodable;

pub use crate::macros::builtin::RustcEncodable;

pub use crate::macros::builtin::bench;

pub use crate::macros::builtin::global_allocator;

pub use crate::macros::builtin::test;

pub use crate::macros::builtin::test_case;