core::prelude

Module v1
·

The first version of the core prelude.

See the module-level documentation for more.

Re-exports

pub use crate::marker::Copy;

pub use crate::marker::Copy;

pub use crate::marker::Send;

pub use crate::marker::Sized;

pub use crate::marker::Sync;

pub use crate::marker::Unpin;

pub use crate::ops::Drop;

pub use crate::ops::Fn;

pub use crate::ops::FnMut;

pub use crate::ops::FnOnce;

pub use crate::ops::AsyncFn;

pub use crate::ops::AsyncFnMut;

pub use crate::ops::AsyncFnOnce;

pub use crate::mem::drop;

pub use crate::mem::align_of;

pub use crate::mem::align_of_val;

pub use crate::mem::size_of;

pub use crate::mem::size_of_val;

pub use crate::clone::Clone;

pub use crate::clone::Clone;

pub use crate::cmp::Eq;

pub use crate::cmp::Eq;

pub use crate::cmp::Ord;

pub use crate::cmp::Ord;

pub use crate::cmp::PartialEq;

pub use crate::cmp::PartialEq;

pub use crate::cmp::PartialOrd;

pub use crate::cmp::PartialOrd;

pub use crate::convert::AsMut;

pub use crate::convert::AsRef;

pub use crate::convert::From;

pub use crate::convert::Into;

pub use crate::default::Default;

pub use crate::default::Default;

pub use crate::iter::DoubleEndedIterator;

pub use crate::iter::ExactSizeIterator;

pub use crate::iter::Extend;

pub use crate::iter::IntoIterator;

pub use crate::iter::Iterator;

pub use crate::option::Option;

pub use crate::option::Option::None;

pub use crate::option::Option::Some;

pub use crate::result::Result;

pub use crate::result::Result::Err;

pub use crate::result::Result::Ok;

pub use crate::fmt::macros::Debug;

pub use crate::hash::macros::Hash;

pub use crate::assert;

pub use crate::assert_eq;

pub use crate::assert_ne;

pub use crate::cfg;

pub use crate::column;

pub use crate::compile_error;

pub use crate::concat;

pub use crate::debug_assert;

pub use crate::debug_assert_eq;

pub use crate::debug_assert_ne;

pub use crate::file;

pub use crate::format_args;

pub use crate::include;

pub use crate::include_bytes;

pub use crate::include_str;

pub use crate::line;

pub use crate::matches;

pub use crate::module_path;

pub use crate::option_env;

pub use crate::stringify;

pub use crate::todo;

pub use crate::try;Deprecated

pub use crate::unimplemented;

pub use crate::unreachable;

pub use crate::write;

pub use crate::writeln;

pub use self::ambiguous_macros_only::env;

pub use self::ambiguous_macros_only::panic;

pub use crate::cfg_select;Experimental

pub use crate::concat_bytes;Experimental

pub use crate::const_format_args;Experimental

pub use crate::log_syntax;Experimental

pub use crate::pattern_type;Experimental

pub use crate::trace_macros;Experimental

Macros

derefExperimental

Unstable placeholder for deref patterns.

type_ascribeExperimental

Unstable placeholder for type ascription.

Attribute Macros

derive

Attribute macro used to apply derive macros.

global_allocator

Attribute macro applied to a static to register it as a global allocator.

test

Attribute macro applied to a function to turn it into a unit test.

alloc_error_handlerExperimental

Attribute macro applied to a function to register it as a handler for allocation failure.

benchExperimental

Attribute macro applied to a function to turn it into a benchmark test.

cfg_accessibleExperimental

Keeps the item it’s applied to if the passed path is accessible, and removes it otherwise.

cfg_evalExperimental

Expands all #[cfg] and #[cfg_attr] attributes in the code fragment it’s applied to.

define_opaqueExperimental

Provide a list of type aliases and other opaque-type-containing type definitions to an item with a body. This list will be used in that body to define opaque types’ hidden types. Can only be applied to things that have bodies.

derive_constExperimental

Attribute macro used to apply derive macros for implementing traits in a const context.

eiiExperimental

Externally Implementable Item: Defines an attribute macro that can override the item this is applied to.

eii_declarationExperimental

Impl detail of EII

test_caseExperimental

An implementation detail of the #[test] and #[bench] macros.

unsafe_eiiExperimental

Unsafely Externally Implementable Item: Defines an unsafe attribute macro that can override the item this is applied to.