Module core::ptr

通过原始指针手动管理内存。

Manually manage memory through raw pointers.

另请参见指针原始类型。

See also the pointer primitive types.

Safety

此模块中的许多函数将原始指针作为参数并对其进行读取或写入。 为了安全起见，这些指针必须有效。 指针是否有效取决于它用于（读取或写入）的操作，以及访问的内存范围（即读取/写入的字节数）。 大多数函数使用 `*mut T` 和 `*const T` 来仅访问单个值，在这种情况下，文档会忽略大小并隐式假定它是 `size_of::()` 字节。

Many functions in this module take raw pointers as arguments and read from or write to them. For this to be safe, these pointers must be valid. Whether a pointer is valid depends on the operation it is used for (read or write), and the extent of the memory that is accessed (i.e., how many bytes are read/written). Most functions use *mut T and *const T to access only a single value, in which case the documentation omits the size and implicitly assumes it to be size_of::<T>() bytes.

有效性的确切规则尚未确定。 此时提供的保证非常少：

The precise rules for validity are not determined yet. The guarantees that are provided at this point are very minimal:

“空”指针永远无效，即使对于“大小为零”的访问也是如此。

A null pointer is never valid, not even for accesses of size zero.

为了使指针有效，指针必须是可解引用的，但并不总是足够的：从指针开始的给定大小的内存范围必须全部在单个分配对象的范围内。 请注意，在 Rust 中，每个（堆栈分配的）变量都被视为一个单独的分配对象。

For a pointer to be valid, it is necessary, but not always sufficient, that the pointer be dereferenceable: the memory range of the given size starting at the pointer must all be within the bounds of a single allocated object. Note that in Rust, every (stack-allocated) variable is considered a separate allocated object.

即使对于“大小为零”的操作，指针也不能指向已释放的内存，即，即使对于零大小的操作，释放也会使指针无效。 但是，将任何非零整数文字转换为指针对于零大小的访问都是有效的，即使某些内存恰好存在于该地址并被释放。 这对应于编写自己的分配器：分配零大小的对象并不是很难。 获取对零大小访问有效的指针的规范方法是“NonNull::dangling”。

Even for operations of size zero, the pointer must not be pointing to deallocated memory, i.e., deallocation makes pointers invalid even for zero-sized operations. However, casting any non-zero integer literal to a pointer is valid for zero-sized accesses, even if some memory happens to exist at that address and gets deallocated. This corresponds to writing your own allocator: allocating zero-sized objects is not very hard. The canonical way to obtain a pointer that is valid for zero-sized accesses is NonNull::dangling.

在用于在线程之间同步的“原子操作”的意义上，此模块中的函数执行的所有访问都是非原子的。 这意味着从不同线程对同一位置执行两次并发访问是未定义的行为，除非这两次访问都只从内存中读取。 请注意，这明确包括“read_volatile”和“write_volatile”：易失性访问不能用于线程间同步。

All accesses performed by functions in this module are non-atomic in the sense of atomic operations used to synchronize between threads. This means it is undefined behavior to perform two concurrent accesses to the same location from different threads unless both accesses only read from memory. Notice that this explicitly includes read_volatile and write_volatile: Volatile accesses cannot be used for inter-thread synchronization.

只要底层对象是活动的并且没有引用（只是原始指针）用于访问相同的内存，转换对指针的引用的结果就有效。

The result of casting a reference to a pointer is valid for as long as the underlying object is live and no reference (just raw pointers) is used to access the same memory.

这些公理，以及在指针算术中谨慎使用“偏移量”，足以在不安全的代码中正确实现许多有用的东西。 随着“别名”规则的确定，最终将提供更强有力的保证。 有关更多信息，请参阅“书”以及参考中专门讨论“未定义行为”的部分。

These axioms, along with careful use of offset for pointer arithmetic, are enough to correctly implement many useful things in unsafe code. Stronger guarantees will be provided eventually, as the aliasing rules are being determined. For more information, see the book as well as the section in the reference devoted to undefined behavior.

Alignment

上面定义的有效原始指针不一定正确对齐（其中“正确”对齐由指针类型定义，即`*const T`必须与`mem::align_of::()`对齐）。 但是，大多数函数都要求它们的参数正确对齐，并将在其文档中明确说明这一要求。 值得注意的例外是“read_unaligned”和“write_unaligned”。

Valid raw pointers as defined above are not necessarily properly aligned (where “proper” alignment is defined by the pointee type, i.e., *const T must be aligned to mem::align_of::<T>()). However, most functions require their arguments to be properly aligned, and will explicitly state this requirement in their documentation. Notable exceptions to this are read_unaligned and write_unaligned.

当一个函数需要正确对齐时，即使访问的大小为 0，即即使实际上没有触及内存，它也会这样做。 在这种情况下考虑使用“NonNull::dangling”。

When a function requires proper alignment, it does so even if the access has size 0, i.e., even if memory is not actually touched. Consider using NonNull::dangling in such cases.

Allocated object

对于一些操作，例如“偏移”或场投影（`expr.field`），“分配对象”的概念变得相关。 分配的对象是一个连续的内存区域。 分配对象的常见示例包括堆栈分配变量（每个变量是一个单独的分配对象）、堆分配（由全局分配器创建的每个分配都是一个单独的分配对象）和“静态”变量。

For several operations, such as offset or field projections (expr.field), the notion of an “allocated object” becomes relevant. An allocated object is a contiguous region of memory. Common examples of allocated objects include stack-allocated variables (each variable is a separate allocated object), heap allocations (each allocation created by the global allocator is a separate allocated object), and static variables.

Macros

addr_of

Create a const raw pointer to a place, without creating an intermediate reference.

addr_of_mut

Create a mut raw pointer to a place, without creating an intermediate reference.

Structs

DynMetadataExperimental

The metadata for a Dyn = dyn SomeTrait trait object type.

NonNull

*mut T but non-zero and covariant.

Traits

PointeeExperimental

Provides the pointer metadata type of any pointed-to type.

Functions

from_raw_partsExperimental

Forms a (possibly-wide) raw pointer from a data address and metadata.

from_raw_parts_mutExperimental

Performs the same functionality as from_raw_parts, except that a raw *mut pointer is returned, as opposed to a raw *const pointer.

metadataExperimental

Extract the metadata component of a pointer.

copy⚠

Copies count * size_of::<T>() bytes from src to dst. The source and destination may overlap.

copy_nonoverlapping⚠

Copies count * size_of::<T>() bytes from src to dst. The source and destination must not overlap.

drop_in_place⚠

Executes the destructor (if any) of the pointed-to value.

eq

Compares raw pointers for equality.

hash

Hash a raw pointer.

null

Creates a null raw pointer.

null_mut

Creates a null mutable raw pointer.

read⚠

Reads the value from src without moving it. This leaves the memory in src unchanged.

read_unaligned⚠

Reads the value from src without moving it. This leaves the memory in src unchanged.

read_volatile⚠

Performs a volatile read of the value from src without moving it. This leaves the memory in src unchanged.

replace⚠

Moves src into the pointed dst, returning the previous dst value.

slice_from_raw_parts

Forms a raw slice from a pointer and a length.

slice_from_raw_parts_mut

Performs the same functionality as slice_from_raw_parts, except that a raw mutable slice is returned, as opposed to a raw immutable slice.

swap⚠

Swaps the values at two mutable locations of the same type, without deinitializing either.

swap_nonoverlapping⚠

Swaps count * size_of::<T>() bytes between the two regions of memory beginning at x and y. The two regions must not overlap.

write⚠

Overwrites a memory location with the given value without reading or dropping the old value.

write_bytes⚠

Sets count * size_of::<T>() bytes of memory starting at dst to val.

write_unaligned⚠

Overwrites a memory location with the given value without reading or dropping the old value.

write_volatile⚠

Performs a volatile write of a memory location with the given value without reading or dropping the old value.

Trait aliases

ThinExperimental

Pointers to types implementing this trait alias are “thin”.