Struct core::ptr::DynMetadata

pub struct DynMetadata<Dyn: ?Sized> { /* private fields */ }

🔬 This is a nightly-only experimental API. (ptr_metadata #81513)

The metadata for a Dyn = dyn SomeTrait trait object type.

It is a pointer to a vtable (virtual call table) that represents all the necessary information to manipulate the concrete type stored inside a trait object. The vtable notably it contains:

type size

type alignment

a pointer to the type’s drop_in_place impl (may be a no-op for plain-old-data)

pointers to all the methods for the type’s implementation of the trait

Note that the first three are special because they’re necessary to allocate, drop, and deallocate any trait object.

It is possible to name this struct with a type parameter that is not a dyn trait object (for example DynMetadata<u64>) but not to obtain a meaningful value of that struct.

Implementations

impl<Dyn: ?Sized> DynMetadata<Dyn>

pub fn size_of(self) -> usize

🔬 This is a nightly-only experimental API. (ptr_metadata #81513)

Returns the size of the type associated with this vtable.

pub fn align_of(self) -> usize

🔬 This is a nightly-only experimental API. (ptr_metadata #81513)

Returns the alignment of the type associated with this vtable.

pub fn layout(self) -> Layout

🔬 This is a nightly-only experimental API. (ptr_metadata #81513)

Returns the size and alignment together as a Layout