Struct core::ptr::Unique

pub struct Unique<T: ?Sized> { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (unique #27730)

needs an RFC to flesh out design

A wrapper around a raw non-null *mut T that indicates that the possessor of this wrapper owns the referent. Useful for building abstractions like Box<T>, Vec<T>, String, and HashMap<K, V>.

Unlike *mut T, Unique<T> behaves "as if" it were an instance of T. It implements Send/Sync if T is Send/Sync. It also implies the kind of strong aliasing guarantees an instance of T can expect: the referent of the pointer should not be modified without a unique path to its owning Unique.

If you're uncertain of whether it's correct to use Unique for your purposes, consider using Shared, which has weaker semantics.

Unlike *mut T, the pointer must always be non-null, even if the pointer is never dereferenced. This is so that enums may use this forbidden value as a discriminant -- Option<Unique<T>> has the same size as Unique<T>. However the pointer may still dangle if it isn't dereferenced.

Unlike *mut T, Unique<T> is covariant over T. This should always be correct for any type which upholds Unique's aliasing requirements.

Methods

impl<T: Sized> Unique<T>

pub fn empty() -> Self

🔬 This is a nightly-only experimental API. (unique #27730)

Creates a new Unique that is dangling, but well-aligned.

This is useful for initializing types which lazily allocate, like Vec::new does.

impl<T: ?Sized> Unique<T>

pub const unsafe fn new_unchecked(ptr: *mut T) -> Self

🔬 This is a nightly-only experimental API. (unique #27730)

Creates a new Unique.

Safety

ptr must be non-null.

pub fn new(ptr: *mut T) -> Option<Self>

🔬 This is a nightly-only experimental API. (unique #27730)

Creates a new Unique if ptr is non-null.

pub fn as_ptr(self) -> *mut T

🔬 This is a nightly-only experimental API. (unique #27730)

Acquires the underlying *mut pointer.

pub unsafe fn as_ref(&self) -> &T

🔬 This is a nightly-only experimental API. (unique #27730)

Dereferences the content.

The resulting lifetime is bound to self so this behaves "as if" it were actually an instance of T that is getting borrowed. If a longer (unbound) lifetime is needed, use &*my_ptr.ptr().

pub unsafe fn as_mut(&mut self) -> &mut T

🔬 This is a nightly-only experimental API. (unique #27730)

Mutably dereferences the content.

The resulting lifetime is bound to self so this behaves "as if" it were actually an instance of T that is getting borrowed. If a longer (unbound) lifetime is needed, use &mut *my_ptr.ptr().