Enum core::result::Result

#[must_use = "this `Result` may be an `Err` variant, which should be handled"]
pub enum Result<T, E> {
    Ok(T),
    Err(E),
}

Result æ˜¯ä»£è¡¨æˆåŠŸ (Ok) æˆ–å¤±è´¥ (Err) çš„ç±»å‹ã€‚

æœ‰å…³è¯¦ç»†ä¿¡æ¯ï¼Œè¯·å‚è module documentationã€‚

Variants

Ok(T)

åŒ…å«æˆåŠŸå€¼

Err(E)

åŒ…å«é”™è¯¯å€¼

Implementations

impl<T, E> Result<T, E>

#[must_use = "if you intended to assert that this is ok, consider `.unwrap()` instead"]
pub const fn is_ok(&self) -> bool

å¦‚æœç»“æœä¸º Okï¼Œåˆ™è¿”å› trueã€‚

Examples

åŸºæœ¬ç”¨æ³•:

let x: Result<i32, &str> = Ok(-3);
assert_eq!(x.is_ok(), true);

let x: Result<i32, &str> = Err("Some error message");
assert_eq!(x.is_ok(), false);
Run

#[must_use = "if you intended to assert that this is err, consider `.unwrap_err()` instead"]
pub const fn is_err(&self) -> bool

å¦‚æœç»“æœä¸º Errï¼Œåˆ™è¿”å› trueã€‚

Examples

åŸºæœ¬ç”¨æ³•:

let x: Result<i32, &str> = Ok(-3);
assert_eq!(x.is_err(), false);

let x: Result<i32, &str> = Err("Some error message");
assert_eq!(x.is_err(), true);
Run

#[must_use]
pub fn contains<U>(&self, x: &U) -> bool where
U: PartialEq<T>,

ğŸ”¬ This is a nightly-only experimental API. (option_result_contains #62358)

å¦‚æœç»“æœæ˜¯åŒ…å«ç»™å®šå€¼çš„ Ok å€¼ï¼Œåˆ™è¿”å› trueã€‚

Examples

#![feature(option_result_contains)]

let x: Result<u32, &str> = Ok(2);
assert_eq!(x.contains(&2), true);

let x: Result<u32, &str> = Ok(3);
assert_eq!(x.contains(&2), false);

let x: Result<u32, &str> = Err("Some error message");
assert_eq!(x.contains(&2), false);
Run

#[must_use]
pub fn contains_err<F>(&self, f: &F) -> bool where
F: PartialEq<E>,

ğŸ”¬ This is a nightly-only experimental API. (result_contains_err #62358)

å¦‚æœç»“æœæ˜¯åŒ…å«ç»™å®šå€¼çš„ Err å€¼ï¼Œåˆ™è¿”å› trueã€‚

Examples

#![feature(result_contains_err)]

let x: Result<u32, &str> = Ok(2);
assert_eq!(x.contains_err(&"Some error message"), false);

let x: Result<u32, &str> = Err("Some error message");
assert_eq!(x.contains_err(&"Some error message"), true);

let x: Result<u32, &str> = Err("Some other error message");
assert_eq!(x.contains_err(&"Some error message"), false);
Run

pub fn ok(self) -> Option<T>

ä» Result<T, E> è½¬æ¢ä¸º Option<T>ã€‚

å°† self è½¬æ¢ä¸º Option<T>ï¼Œä½¿ç”¨ selfï¼Œå¹¶ä¸¢å¼ƒé”™è¯¯ (å¦‚æœæœ‰)ã€‚

Examples

åŸºæœ¬ç”¨æ³•:

let x: Result<u32, &str> = Ok(2);
assert_eq!(x.ok(), Some(2));

let x: Result<u32, &str> = Err("Nothing here");
assert_eq!(x.ok(), None);
Run

pub fn err(self) -> Option<E>

ä» Result<T, E> è½¬æ¢ä¸º Option<E>ã€‚

å°† self è½¬æ¢ä¸º Option<E>ï¼Œä½¿ç”¨ selfï¼Œå¹¶ä¸¢å¼ƒæˆåŠŸå€¼ (å¦‚æœæœ‰)ã€‚

Examples

åŸºæœ¬ç”¨æ³•:

let x: Result<u32, &str> = Ok(2);
assert_eq!(x.err(), None);

let x: Result<u32, &str> = Err("Nothing here");
assert_eq!(x.err(), Some("Nothing here"));
Run

pub const fn as_ref(&self) -> Result<&T, &E>

ä» &Result<T, E> è½¬æ¢ä¸º Result<&T, &E>ã€‚

äºç”Ÿä¸€ä¸ªæ–°çš„ Resultï¼Œå…¶ä¸­åŒ…å«å¯¹åŸå‹å¼•ç”¨çš„å¼•ç”¨ï¼Œå¹¶å°†åŸå‹ä¿ç•™åœ¨åŸå¤„ã€‚

Examples

åŸºæœ¬ç”¨æ³•:

let x: Result<u32, &str> = Ok(2);
assert_eq!(x.as_ref(), Ok(&2));

let x: Result<u32, &str> = Err("Error");
assert_eq!(x.as_ref(), Err(&"Error"));
Run

pub fn as_mut(&mut self) -> Result<&mut T, &mut E>

ä» &mut Result<T, E> è½¬æ¢ä¸º Result<&mut T, &mut E>ã€‚

Examples

åŸºæœ¬ç”¨æ³•:

fn mutate(r: &mut Result<i32, i32>) {
    match r.as_mut() {
        Ok(v) => *v = 42,
        Err(e) => *e = 0,
    }
}

let mut x: Result<i32, i32> = Ok(2);
mutate(&mut x);
assert_eq!(x.unwrap(), 42);

let mut x: Result<i32, i32> = Err(13);
mutate(&mut x);
assert_eq!(x.unwrap_err(), 0);
Run

pub fn map<U, F: FnOnce(T) -> U>(self, op: F) -> Result<U, E>

é€šè¿‡å¯¹åŒ…å«çš„ Ok å€¼åº”ç”¨å‡½æ•°ï¼Œå°† Err å€¼ Maps è½¬æ¢ä¸º Result<U, E>ï¼Œè€Œä¿æŒ Err å€¼ä¸å˜ã€‚

è¯¥å‡½æ•°å¯ç”¨äºç»„åˆä¸¤ä¸ªå‡½æ•°çš„ç»“æœã€‚

Examples

åœ¨å­—ç¬¦ä¸²çš„æ¯ä¸€è¡Œä¸Šå°†æ•°å­—ä¹˜ä»¥ 2 æ¥æ‰“å°æ•°å­—ã€‚

let line = "1\n2\n3\n4\n";

for num in line.lines() {
    match num.parse::<i32>().map(|i| i * 2) {
        Ok(n) => println!("{}", n),
        Err(..) => {}
    }
}
Run

pub fn map_or<U, F: FnOnce(T) -> U>(self, default: U, f: F) -> U

è¿”å›æä¾›çš„é»˜è®¤å€¼ (å¦‚æœ Err)ï¼Œæˆ–è€…å°†å‡½æ•°åº”ç”¨äºåŒ…å«çš„å€¼ (å¦‚æœ Ok)ï¼Œ

æ€¥åˆ‡åœ°è¯„ä¼°ä¼ é€’ç»™ map_or çš„å‚æ•°; å¦‚æœè¦ä¼ é€’å‡½æ•°è°ƒç”¨çš„ç»“æœï¼Œå»ºè®®ä½¿ç”¨ map_or_elseï¼Œå®ƒæ˜¯å»¶è¿Ÿè®¡ç®—çš„ã€‚

Examples

let x: Result<_, &str> = Ok("foo");
assert_eq!(x.map_or(42, |v| v.len()), 3);

let x: Result<&str, _> = Err("bar");
assert_eq!(x.map_or(42, |v| v.len()), 42);
Run

pub fn map_or_else<U, D: FnOnce(E) -> U, F: FnOnce(T) -> U>(
self,
default: D,
f: F
) -> U

Maps é€šè¿‡å°†å›é€€å‡½æ•°åº”ç”¨äºåŒ…å«çš„ Err å€¼ï¼Œæˆ–å°†é»˜è®¤å‡½æ•°åº”ç”¨äºåŒ…å«çš„ Ok å€¼ï¼Œå°† Result<T, E> è½¬æ¢ä¸º Uã€‚

æ­¤å‡½æ•°å¯ç”¨äºåœ¨å¤„ç†é”™è¯¯æ—¶è£å‹æˆåŠŸçš„ç»“æœã€‚

Examples

åŸºæœ¬ç”¨æ³•:

let k = 21;

let x : Result<_, &str> = Ok("foo");
assert_eq!(x.map_or_else(|e| k * 2, |v| v.len()), 3);

let x : Result<&str, _> = Err("bar");
assert_eq!(x.map_or_else(|e| k * 2, |v| v.len()), 42);
Run

pub fn map_err<F, O: FnOnce(E) -> F>(self, op: O) -> Result<T, F>

é€šè¿‡å¯¹åŒ…å«çš„ Err å€¼åº”ç”¨å‡½æ•°ï¼Œå°† Ok å€¼ Maps è½¬æ¢ä¸º Result<T, F>ï¼Œè€Œä¿æŒ Ok å€¼ä¸å˜ã€‚

æ­¤å‡½æ•°å¯ç”¨äºåœ¨å¤„ç†é”™è¯¯æ—¶ä¼ é€’æˆåŠŸçš„ç»“æœã€‚

Examples

åŸºæœ¬ç”¨æ³•:

fn stringify(x: u32) -> String { format!("error code: {}", x) }

let x: Result<u32, u32> = Ok(2);
assert_eq!(x.map_err(stringify), Ok(2));

let x: Result<u32, u32> = Err(13);
assert_eq!(x.map_err(stringify), Err("error code: 13".to_string()));
Run

pub fn iter(&self) -> Iter<'_, T>

è¿”å›å¯èƒ½åŒ…å«çš„å€¼çš„è¿­ä»£å™¨ã€‚

å¦‚æœç»“æœä¸º Result::Okï¼Œåˆ™è¿­ä»£å™¨å°†äºç”Ÿä¸€ä¸ªå€¼ï¼Œå¦åˆ™å°†ä¸äºç”Ÿä»»ä½•å€¼ã€‚

Examples

åŸºæœ¬ç”¨æ³•:

let x: Result<u32, &str> = Ok(7);
assert_eq!(x.iter().next(), Some(&7));

let x: Result<u32, &str> = Err("nothing!");
assert_eq!(x.iter().next(), None);
Run

pub fn iter_mut(&mut self) -> IterMut<'_, T>

è¿”å›å¯èƒ½åŒ…å«çš„å€¼çš„å¯å˜è¿­ä»£å™¨ã€‚

å¦‚æœç»“æœä¸º Result::Okï¼Œåˆ™è¿­ä»£å™¨å°†äºç”Ÿä¸€ä¸ªå€¼ï¼Œå¦åˆ™å°†ä¸äºç”Ÿä»»ä½•å€¼ã€‚

Examples

åŸºæœ¬ç”¨æ³•:

let mut x: Result<u32, &str> = Ok(7);
match x.iter_mut().next() {
    Some(v) => *v = 40,
    None => {},
}
assert_eq!(x, Ok(40));

let mut x: Result<u32, &str> = Err("nothing!");
assert_eq!(x.iter_mut().next(), None);
Run

pub fn and<U>(self, res: Result<U, E>) -> Result<U, E>

å¦‚æœç»“æœä¸º Okï¼Œåˆ™è¿”å› res; å¦åˆ™ï¼Œè¿”å› self çš„ Err å€¼ã€‚

Examples

åŸºæœ¬ç”¨æ³•:

let x: Result<u32, &str> = Ok(2);
let y: Result<&str, &str> = Err("late error");
assert_eq!(x.and(y), Err("late error"));

let x: Result<u32, &str> = Err("early error");
let y: Result<&str, &str> = Ok("foo");
assert_eq!(x.and(y), Err("early error"));

let x: Result<u32, &str> = Err("not a 2");
let y: Result<&str, &str> = Err("late error");
assert_eq!(x.and(y), Err("not a 2"));

let x: Result<u32, &str> = Ok(2);
let y: Result<&str, &str> = Ok("different result type");
assert_eq!(x.and(y), Ok("different result type"));
Run

pub fn and_then<U, F: FnOnce(T) -> Result<U, E>>(self, op: F) -> Result<U, E>

å¦‚æœç»“æœä¸º Okï¼Œåˆ™è°ƒç”¨ opï¼Œå¦åˆ™è¿”å› self çš„ Err å€¼ã€‚

è¯¥å‡½æ•°å¯ç”¨äºåŸºäº Result å€¼çš„æåˆ¶æµã€‚

Examples

åŸºæœ¬ç”¨æ³•:

fn sq(x: u32) -> Result<u32, u32> { Ok(x * x) }
fn err(x: u32) -> Result<u32, u32> { Err(x) }

assert_eq!(Ok(2).and_then(sq).and_then(sq), Ok(16));
assert_eq!(Ok(2).and_then(sq).and_then(err), Err(4));
assert_eq!(Ok(2).and_then(err).and_then(sq), Err(2));
assert_eq!(Err(3).and_then(sq).and_then(sq), Err(3));
Run

pub fn or<F>(self, res: Result<T, F>) -> Result<T, F>

å¦‚æœç»“æœä¸º Errï¼Œåˆ™è¿”å› res; å¦åˆ™ï¼Œè¿”å› self çš„ Ok å€¼ã€‚

æ€¥åˆ‡åœ°è¯„ä¼°ä¼ é€’ç»™ or çš„å‚æ•°; å¦‚æœè¦ä¼ é€’å‡½æ•°è°ƒç”¨çš„ç»“æœï¼Œå»ºè®®ä½¿ç”¨ or_elseï¼Œå®ƒæ˜¯å»¶è¿Ÿè®¡ç®—çš„ã€‚

Examples

åŸºæœ¬ç”¨æ³•:

let x: Result<u32, &str> = Ok(2);
let y: Result<u32, &str> = Err("late error");
assert_eq!(x.or(y), Ok(2));

let x: Result<u32, &str> = Err("early error");
let y: Result<u32, &str> = Ok(2);
assert_eq!(x.or(y), Ok(2));

let x: Result<u32, &str> = Err("not a 2");
let y: Result<u32, &str> = Err("late error");
assert_eq!(x.or(y), Err("late error"));

let x: Result<u32, &str> = Ok(2);
let y: Result<u32, &str> = Ok(100);
assert_eq!(x.or(y), Ok(2));
Run

pub fn or_else<F, O: FnOnce(E) -> Result<T, F>>(self, op: O) -> Result<T, F>

å¦‚æœç»“æœä¸º Errï¼Œåˆ™è°ƒç”¨ opï¼Œå¦åˆ™è¿”å› self çš„ Ok å€¼ã€‚

è¯¥å‡½æ•°å¯ç”¨äºåŸºäºç»“æœå€¼çš„æåˆ¶æµã€‚

Examples

åŸºæœ¬ç”¨æ³•:

fn sq(x: u32) -> Result<u32, u32> { Ok(x * x) }
fn err(x: u32) -> Result<u32, u32> { Err(x) }

assert_eq!(Ok(2).or_else(sq).or_else(sq), Ok(2));
assert_eq!(Ok(2).or_else(err).or_else(sq), Ok(2));
assert_eq!(Err(3).or_else(sq).or_else(err), Ok(9));
assert_eq!(Err(3).or_else(err).or_else(err), Err(3));
Run

pub fn unwrap_or(self, default: T) -> T

è¿”å›åŒ…å«çš„ Ok å€¼æˆ–æä¾›çš„é»˜è®¤å€¼ã€‚

æ€¥åˆ‡åœ°è¯„ä¼°ä¼ é€’ç»™ unwrap_or çš„å‚æ•°; å¦‚æœè¦ä¼ é€’å‡½æ•°è°ƒç”¨çš„ç»“æœï¼Œå»ºè®®ä½¿ç”¨ unwrap_or_elseï¼Œå®ƒæ˜¯æƒ°æ€æ±‚å€¼çš„ã€‚

Examples

åŸºæœ¬ç”¨æ³•:

let default = 2;
let x: Result<u32, &str> = Ok(9);
assert_eq!(x.unwrap_or(default), 9);

let x: Result<u32, &str> = Err("error");
assert_eq!(x.unwrap_or(default), default);
Run

pub fn unwrap_or_else<F: FnOnce(E) -> T>(self, op: F) -> T

è¿”å›åŒ…å«çš„ Ok å€¼æˆ–ä»é—­åŒ…ä¸­è®¡ç®—å¾—å‡ºã€‚

Examples

åŸºæœ¬ç”¨æ³•:

fn count(x: &str) -> usize { x.len() }

assert_eq!(Ok(2).unwrap_or_else(count), 2);
assert_eq!(Err("foo").unwrap_or_else(count), 3);
Run

pub unsafe fn unwrap_unchecked(self) -> T

ğŸ”¬ This is a nightly-only experimental API. (option_result_unwrap_unchecked #81383)

newly added

è¿”å›åŒ…å« self å€¼çš„åŒ…å«çš„ Ok å€¼ï¼Œè€Œä¸æ£€æŸ¥è¯¥å€¼æ˜¯å¦ä¸æ˜¯ Errã€‚

Safety

åœ¨ Err ä¸Šè°ƒç”¨æ­¤æ–¹æ³•æ˜¯ [undefined è¡Œä¸º]ã€‚

Examples

#![feature(option_result_unwrap_unchecked)]
let x: Result<u32, &str> = Ok(2);
assert_eq!(unsafe { x.unwrap_unchecked() }, 2);
Run

#![feature(option_result_unwrap_unchecked)]
let x: Result<u32, &str> = Err("emergency failure");
unsafe { x.unwrap_unchecked(); } // æœªå®šä¹‰çš„è¡Œä¸º!
Run

pub unsafe fn unwrap_err_unchecked(self) -> E

ğŸ”¬ This is a nightly-only experimental API. (option_result_unwrap_unchecked #81383)

newly added

è¿”å›åŒ…å« self å€¼çš„åŒ…å«çš„ Err å€¼ï¼Œè€Œä¸æ£€æŸ¥è¯¥å€¼æ˜¯å¦ä¸æ˜¯ Okã€‚

Safety

åœ¨ Ok ä¸Šè°ƒç”¨æ­¤æ–¹æ³•æ˜¯ [undefined è¡Œä¸º]ã€‚

Examples

#![feature(option_result_unwrap_unchecked)]
let x: Result<u32, &str> = Ok(2);
unsafe { x.unwrap_err_unchecked() }; // æœªå®šä¹‰çš„è¡Œä¸º!
Run

#![feature(option_result_unwrap_unchecked)]
let x: Result<u32, &str> = Err("emergency failure");
assert_eq!(unsafe { x.unwrap_err_unchecked() }, "emergency failure");
Run

impl<T: Copy, E> Result<&T, E>

pub fn copied(self) -> Result<T, E>

ğŸ”¬ This is a nightly-only experimental API. (result_copied #63168)

newly added

é€šè¿‡å¤åˆ¶ Ok éƒ¨ä»¶çš„å†…å®¹ï¼Œå°† Result<&T, E> çš„ Maps è½¬æ¢ä¸º Result<T, E>ã€‚

Examples

#![feature(result_copied)]
let val = 12;
let x: Result<&i32, i32> = Ok(&val);
assert_eq!(x, Ok(&12));
let copied = x.copied();
assert_eq!(copied, Ok(12));
Run

impl<T: Copy, E> Result<&mut T, E>

pub fn copied(self) -> Result<T, E>

ğŸ”¬ This is a nightly-only experimental API. (result_copied #63168)

newly added

é€šè¿‡å¤åˆ¶ Ok éƒ¨ä»¶çš„å†…å®¹ï¼Œå°† Result<&mut T, E> çš„ Maps è½¬æ¢ä¸º Result<T, E>ã€‚

Examples

#![feature(result_copied)]
let mut val = 12;
let x: Result<&mut i32, i32> = Ok(&mut val);
assert_eq!(x, Ok(&mut 12));
let copied = x.copied();
assert_eq!(copied, Ok(12));
Run

impl<T: Clone, E> Result<&T, E>

pub fn cloned(self) -> Result<T, E>

ğŸ”¬ This is a nightly-only experimental API. (result_cloned #63168)

newly added

é€šè¿‡å…‹éš† Ok éƒ¨åˆ†çš„å†…å®¹ï¼Œå°† Result<&T, E> Maps è½¬æ¢ä¸º Result<T, E>ã€‚

Examples

#![feature(result_cloned)]
let val = 12;
let x: Result<&i32, i32> = Ok(&val);
assert_eq!(x, Ok(&12));
let cloned = x.cloned();
assert_eq!(cloned, Ok(12));
Run

impl<T: Clone, E> Result<&mut T, E>

pub fn cloned(self) -> Result<T, E>

ğŸ”¬ This is a nightly-only experimental API. (result_cloned #63168)

newly added

é€šè¿‡å…‹éš† Ok éƒ¨åˆ†çš„å†…å®¹ï¼Œå°† Result<&mut T, E> Maps è½¬æ¢ä¸º Result<T, E>ã€‚

Examples

#![feature(result_cloned)]
let mut val = 12;
let x: Result<&mut i32, i32> = Ok(&mut val);
assert_eq!(x, Ok(&mut 12));
let cloned = x.cloned();
assert_eq!(cloned, Ok(12));
Run

impl<T, E: Debug> Result<T, E>

pub fn expect(self, msg: &str) -> T

è¿”å›åŒ…å« self å€¼çš„åŒ…å«çš„ Ok å€¼ã€‚

Panics

Panics (å¦‚æœå€¼ä¸º Err)ï¼Œå…¶ä¸­ panic æ¶ˆæ¯åŒ…æ‹¬ä¼ é€’çš„æ¶ˆæ¯ä»¥åŠ Err çš„å†…å®¹ã€‚

Examples

åŸºæœ¬ç”¨æ³•:

let x: Result<u32, &str> = Err("emergency failure");
x.expect("Testing expect"); // `Testing expect: emergency failure` çš„ panics
Run

pub fn unwrap(self) -> T

è¿”å›åŒ…å« self å€¼çš„åŒ…å«çš„ Ok å€¼ã€‚

ç”±äºæ­¤å‡½æ•°å¯èƒ½ä¸º panicï¼Œå› æ­¤é€šå¸¸ä¸å»ºè®®ä½¿ç”¨è¯¥å‡½æ•°ã€‚ ç›¸åï¼Œæ›´å–œæ¬¢ä½¿ç”¨æ¨¡å¼åŒ¹é…å¹¶æ˜¾å¼å¤„ç† Err å¤å°å†™ï¼Œæˆ–è€…è°ƒç”¨ unwrap_orï¼Œunwrap_or_else æˆ– unwrap_or_defaultã€‚

Panics

å¦‚æœè¯¥å€¼ä¸º Errï¼Œåˆ™ä¸º Panicsï¼Œå¹¶ç”± Err çš„å€¼æä¾› panic æ¶ˆæ¯ã€‚

Examples

åŸºæœ¬ç”¨æ³•:

let x: Result<u32, &str> = Ok(2);
assert_eq!(x.unwrap(), 2);
Run

let x: Result<u32, &str> = Err("emergency failure");
x.unwrap(); // `emergency failure` çš„ panics
Run

impl<T: Debug, E> Result<T, E>

pub fn expect_err(self, msg: &str) -> E

è¿”å›åŒ…å« self å€¼çš„åŒ…å«çš„ Err å€¼ã€‚

Panics

Panics (å¦‚æœå€¼ä¸º Ok)ï¼Œå…¶ä¸­ panic æ¶ˆæ¯åŒ…æ‹¬ä¼ é€’çš„æ¶ˆæ¯ä»¥åŠ Ok çš„å†…å®¹ã€‚

Examples

åŸºæœ¬ç”¨æ³•:

let x: Result<u32, &str> = Ok(10);
x.expect_err("Testing expect_err"); // `Testing expect_err: 10` çš„ panics
Run

pub fn unwrap_err(self) -> E

è¿”å›åŒ…å« self å€¼çš„åŒ…å«çš„ Err å€¼ã€‚

Panics

Panicsï¼Œå¦‚æœè¯¥å€¼ä¸º Okï¼Œåˆ™ç”± Ok çš„å€¼æä¾›è‡ªå®šä¹‰ panic æ¶ˆæ¯ã€‚

Examples

let x: Result<u32, &str> = Ok(2);
x.unwrap_err(); // `2` çš„ panics
Run

let x: Result<u32, &str> = Err("emergency failure");
assert_eq!(x.unwrap_err(), "emergency failure");
Run

impl<T: Default, E> Result<T, E>

pub fn unwrap_or_default(self) -> T

è¿”å›åŒ…å«çš„ Ok å€¼æˆ–é»˜è®¤å€¼

ç„¶åä½¿ç”¨ self å‚æ•°ï¼Œå¦‚æœä½¿ç”¨ Okï¼Œåˆ™è¿”å›åŒ…å«çš„å€¼ï¼Œå¦‚æœä½¿ç”¨ Errï¼Œåˆ™è¿”å›è¯¥ç±»å‹çš„é»˜è®¤å€¼ã€‚

Examples

å°†å­—ç¬¦ä¸²è½¬æ¢ä¸ºæ•´æ•°ï¼Œå°†æ ¼å¼ä¸æ­£ç¡®çš„å­—ç¬¦ä¸²è½¬æ¢ä¸º 0 (æ•´æ•°çš„é»˜è®¤å€¼)ã€‚ parse å°†å­—ç¬¦ä¸²è½¬æ¢ä¸ºå®ç° FromStr çš„ä»»ä½•å…¶ä»–ç±»å‹ï¼Œå¹¶åœ¨é”™è¯¯æ—¶è¿”å› Errã€‚

let good_year_from_input = "1909";
let bad_year_from_input = "190blarg";
let good_year = good_year_from_input.parse().unwrap_or_default();
let bad_year = bad_year_from_input.parse().unwrap_or_default();

assert_eq!(1909, good_year);
assert_eq!(0, bad_year);
Run

impl<T, E: Into<!>> Result<T, E>

pub fn into_ok(self) -> T

ğŸ”¬ This is a nightly-only experimental API. (unwrap_infallible #61695)

newly added

è¿”å›åŒ…å«çš„ Ok å€¼ï¼Œä½†ä¸è¿”å› panicsã€‚

ä¸ unwrap ä¸åŒï¼Œå·²çŸ¥è¯¥æ–¹æ³•æ°¸è¿œä¸ä¼šå¯¹å…¶å®ç°çš„ç»“æœç±»å‹è¿›è¡Œ panic çš„å¤„ç†ã€‚ å› æ­¤ï¼Œå®ƒå¯ä»¥ä»£æ›¿ unwrap ç”¨ä½œå¯ç»´æŠ¤æ€ä¿æŠ¤æªæ–½ï¼Œå¦‚æœä»¥åå°† Result çš„é”™è¯¯ç±»å‹æ›´æ”¹ä¸ºå®é™…å¯èƒ½å‘ç”Ÿçš„é”™è¯¯ï¼Œå®ƒå°†æ— æ³•ç¼–è¯‘ã€‚

Examples

åŸºæœ¬ç”¨æ³•:

fn only_good_news() -> Result<String, !> {
    Ok("this is fine".into())
}

let s: String = only_good_news().into_ok();
println!("{}", s);
Run

impl<T: Into<!>, E> Result<T, E>

pub fn into_err(self) -> E

ğŸ”¬ This is a nightly-only experimental API. (unwrap_infallible #61695)

newly added

è¿”å›åŒ…å«çš„ Err å€¼ï¼Œä½†ä»ä¸è¿”å› panicsã€‚

ä¸ unwrap_err ä¸åŒï¼Œå·²çŸ¥æ­¤æ–¹æ³•æ°¸è¿œä¸ä¼šåœ¨å…¶å®ç°çš„ç»“æœç±»å‹ä¸Šä½¿ç”¨ panicã€‚ å› æ­¤ï¼Œå®ƒå¯ä»¥ä»£æ›¿ unwrap_err ç”¨ä½œå¯ç»´æŠ¤æ€ä¿éšœï¼Œå¦‚æœ Result çš„ ok ç±»å‹ç¨åæ›´æ”¹ä¸ºå®é™…å¯ä»¥å‘ç”Ÿçš„ç±»å‹ï¼Œåˆ™å°†æ— æ³•ç¼–è¯‘ã€‚

Examples

åŸºæœ¬ç”¨æ³•:

fn only_bad_news() -> Result<!, String> {
    Err("Oops, it failed".into())
}

let error: String = only_bad_news().into_err();
println!("{}", error);
Run

impl<T: Deref, E> Result<T, E>

pub fn as_deref(&self) -> Result<&T::Target, &E>

ä» Result<T, E> (æˆ– &Result<T, E>) è½¬æ¢ä¸º Result<&<T as Deref>::Target, &E>ã€‚

é€šè¿‡ Deref å¼ºåˆ¶è½¬æ¢åŸå‹ Result çš„ Ok æˆå‘˜ï¼Œå¹¶è¿”å›æ–°çš„ Resultã€‚

Examples

let x: Result<String, u32> = Ok("hello".to_string());
let y: Result<&str, &u32> = Ok("hello");
assert_eq!(x.as_deref(), y);

let x: Result<String, u32> = Err(42);
let y: Result<&str, &u32> = Err(&42);
assert_eq!(x.as_deref(), y);
Run

impl<T: DerefMut, E> Result<T, E>

pub fn as_deref_mut(&mut self) -> Result<&mut T::Target, &mut E>

ä» Result<T, E> (æˆ– &mut Result<T, E>) è½¬æ¢ä¸º Result<&mut <T as DerefMut>::Target, &mut E>ã€‚

é€šè¿‡ DerefMut å¼ºåˆ¶è½¬æ¢åŸå‹ Result çš„ Ok æˆå‘˜ï¼Œå¹¶è¿”å›æ–°çš„ Resultã€‚

Examples

let mut s = "HELLO".to_string();
let mut x: Result<String, u32> = Ok("hello".to_string());
let y: Result<&mut str, &mut u32> = Ok(&mut s);
assert_eq!(x.as_deref_mut().map(|x| { x.make_ascii_uppercase(); x }), y);

let mut i = 42;
let mut x: Result<String, u32> = Err(42);
let y: Result<&mut str, &mut u32> = Err(&mut i);
assert_eq!(x.as_deref_mut().map(|x| { x.make_ascii_uppercase(); x }), y);
Run

impl<T, E> Result<Option<T>, E>

pub fn transpose(self) -> Option<Result<T, E>>

å°† Option çš„ Result è½¬æ¢ä¸º Result çš„ Optionã€‚

Ok(None) å°†è¢«æ˜ å°„åˆ° Noneã€‚ Ok(Some(_)) Err(_) å°†è¢«æ˜ å°„åˆ° Some(Ok(_)) å’Œ Some(Err(_))ã€‚

Examples

#[derive(Debug, Eq, PartialEq)]
struct SomeErr;

let x: Result<Option<i32>, SomeErr> = Ok(Some(5));
let y: Option<Result<i32, SomeErr>> = Some(Ok(5));
assert_eq!(x.transpose(), y);
Run

impl<T, E> Result<Result<T, E>, E>

pub fn flatten(self) -> Result<T, E>

ğŸ”¬ This is a nightly-only experimental API. (result_flattening #70142)

ä» Result<Result<T, E>, E> è½¬æ¢ä¸º Result<T, E>

Examples

åŸºæœ¬ç”¨æ³•:

#![feature(result_flattening)]
let x: Result<Result<&'static str, u32>, u32> = Ok(Ok("hello"));
assert_eq!(Ok("hello"), x.flatten());

let x: Result<Result<&'static str, u32>, u32> = Ok(Err(6));
assert_eq!(Err(6), x.flatten());

let x: Result<Result<&'static str, u32>, u32> = Err(6);
assert_eq!(Err(6), x.flatten());
Run

å±•å¹³ä¸€æ¬¡åªèƒ½åˆ é™¤ä¸€å±‚åµŒå¥—:

#![feature(result_flattening)]
let x: Result<Result<Result<&'static str, u32>, u32>, u32> = Ok(Ok(Ok("hello")));
assert_eq!(Ok(Ok("hello")), x.flatten());
assert_eq!(Ok("hello"), x.flatten().flatten());
Run

impl<T> Result<T, T>

pub const fn into_ok_or_err(self) -> T

ğŸ”¬ This is a nightly-only experimental API. (result_into_ok_or_err #82223)

newly added

å¦‚æœ self æ˜¯ Okï¼Œåˆ™è¿”å› Ok å€¼; å¦‚æœ self æ˜¯ Errï¼Œåˆ™è¿”å› Err å€¼ã€‚

æ¢å¥è¯è¯´ï¼Œæ­¤å‡½æ•°è¿”å› Result<T, T> çš„å€¼ (T)ï¼Œè€Œä¸ç®¡ç»“æœæ˜¯ Ok è¿˜æ˜¯ Errã€‚

ä¸ Atomic*::compare_exchange æˆ– slice::binary_search ç­‰ API ç»“åˆä½¿ç”¨æ—¶ï¼Œæ­¤åŠŸèƒ½å¾ˆæœ‰ç”¨ï¼Œä½†ä»…åœ¨æ‚¨ä¸å…³å¿ƒç»“æœæ˜¯å¦ä¸º Ok çš„æƒ…å†µä¸‹æ‰æœ‰ç”¨ã€‚

Examples

#![feature(result_into_ok_or_err)]
let ok: Result<u32, u32> = Ok(3);
let err: Result<u32, u32> = Err(4);

assert_eq!(ok.into_ok_or_err(), 3);
assert_eq!(err.into_ok_or_err(), 4);
Run