Module core::sliceStable

Slice management and manipulation

For more details std::slice.

Primitive Types

slice
Slice management and manipulation

Modules

bytes
Operations on [u8].

raw
Unsafe operations

Structs

Chunks
An iterator over a vector in (non-overlapping) chunks (size elements at a time).

Items
Immutable slice iterator

MutChunks
An iterator over a vector in (non-overlapping) mutable chunks (size elements at a time). When the vector len is not evenly divided by the chunk size, the last slice of the iteration will be the remainder.

MutItems
Mutable slice iterator

MutSplits
An iterator over the subslices of the vector which are separated by elements that match pred.

Splits
An iterator over the slices of a vector separated by elements that match a predicate function.

SplitsN
An iterator over the slices of a vector separated by elements that match a predicate function, splitting at most a fixed number of times.

Windows
An iterator over the (overlapping) slices of length size within a vector.

Enums

BinarySearchResult
The result of calling binary_search.

Traits

ImmutableOrdSlice
Extension methods for vectors containing Ord elements.

ImmutablePartialEqSlice
Extension methods for vectors contain PartialEq elements.

ImmutableSlice
Extension methods for immutable slices.

MutableCloneableSlice
Trait for &[T] where T is Cloneable

MutableSlice
Extension methods for vectors such that their elements are mutable.

Slice
Any vector that can be represented as a slice.

Functions

mut_ref_slice
Converts a pointer to A into a slice of length 1 (without copying).

ref_slice
Converts a pointer to A into a slice of length 1 (without copying).