Module core::slice

Slice management and manipulation

For more details std::slice.

Reexports

use clone::Clone;

use cmp::{Ordering, PartialEq, PartialOrd, Eq, Ord};

use cmp::Ordering::{Less, Equal, Greater};

use cmp;

use default::Default;

use intrinsics::assume;

use iter::*;

use ops::{FnMut, self, Index};

use ops::RangeFull;

use option::Option;

use option::Option::{None, Some};

use result::Result;

use result::Result::{Ok, Err};

use ptr;

use mem;

use marker::{Send, Sync, self};

use raw::Repr;

use raw::Slice as RawSlice;

Modules

bytes 	 [Deprecated]
Operations on [u8].

Structs

Chunks
An iterator over a slice in (non-overlapping) chunks (size elements at a time).

ChunksMut
An iterator over a slice in (non-overlapping) mutable chunks (size elements at a time). When the slice len is not evenly divided by the chunk size, the last slice of the iteration will be the remainder.

GenericSplitN
An private iterator over subslices separated by elements that match a predicate function, splitting at most a fixed number of times.

Iter
Immutable slice iterator

IterMut
Mutable slice iterator.

RSplitN
An iterator over subslices separated by elements that match a predicate function, limited to a given number of splits, starting from the end of the slice.

RSplitNMut
An iterator over subslices separated by elements that match a predicate function, limited to a given number of splits, starting from the end of the slice.

Split
An iterator over subslices separated by elements that match a predicate function.

SplitMut
An iterator over the subslices of the vector which are separated by elements that match pred.

SplitN
An iterator over subslices separated by elements that match a predicate function, limited to a given number of splits.

SplitNMut
An iterator over subslices separated by elements that match a predicate function, limited to a given number of splits.

Windows
An iterator over overlapping subslices of length size.

Traits

SliceExt 	 [Unstable]
Extension methods for slices.

SplitIter
An internal abstraction over the splitting iterators, so that splitn, splitn_mut etc can be implemented once.

Functions

from_raw_parts
Forms a slice from a pointer and a length.

from_raw_parts_mut
Performs the same functionality as from_raw_parts, except that a mutable slice is returned.

size_from_ptr

slice_index_len_fail

slice_index_order_fail