Function core::slice::from_raw_parts_mut

pub unsafe fn from_raw_parts_mut<'a, T>(data: *mut T, len: usize) -> &'a mut [T]

执行与“from_raw_parts”相同的功能，除了返回一个可变切片。

Performs the same functionality as from_raw_parts, except that a mutable slice is returned.

Safety

如果违反以下任何条件，则行为未定义：

Behavior is undefined if any of the following conditions are violated:

data 必须对 len * mem::size_of::() 多个字节的读取和写入都有效，并且必须正确对齐。 这尤其意味着：

data must be valid for both reads and writes for len * mem::size_of::<T>() many bytes, and it must be properly aligned. This means in particular:

该切片的整个内存范围必须包含在单个分配的对象中！ 切片永远不能跨越多个分配的对象。

The entire memory range of this slice must be contained within a single allocated object! Slices can never span across multiple allocated objects.

`data` 必须是非空的并且即使对于零长度切片也是对齐的。 原因之一是枚举布局优化可能依赖于对齐的引用（包括任意长度的切片）和非空值，以将它们与其他数据区分开来。 您可以使用“NonNull::dangling()”获取可用作零长度切片的“数据”的指针。

data must be non-null and aligned even for zero-length slices. One reason for this is that enum layout optimizations may rely on references (including slices of any length) being aligned and non-null to distinguish them from other data. You can obtain a pointer that is usable as data for zero-length slices using NonNull::dangling().

data 必须指向 len 个连续正确初始化的 T 类型值。

data must point to len consecutive properly initialized values of type T.

返回的 slice 引用的内存在生命周期 'a. 禁止读写访问。

The memory referenced by the returned slice must not be accessed through any other pointer (not derived from the return value) for the duration of lifetime 'a. Both read and write accesses are forbidden.

切片的总大小 len * mem::size_of::() 不得大于 isize::MAX。 请参阅pointer::offset 的安全文档。

The total size len * mem::size_of::<T>() of the slice must be no larger than isize::MAX. See the safety documentation of pointer::offset.