core::slice::
Module memchr

🔬This is a nightly-only experimental API. (slice_internals)

Pure Rust memchr implementation, taken from rust-memchr

Functions

memchrExperimental

Returns the first index matching the byte x in text.

memrchrExperimental

Returns the last index matching the byte x in text.