Trait core::str::pattern::ReverseSearcher

pub unsafe trait ReverseSearcher<'a>: Searcher<'a> {
    fn next_back(&mut self) -> SearchStep;

    fn next_match_back(&mut self) -> Option<(usize, usize)> { ... }

    fn next_reject_back(&mut self) -> Option<(usize, usize)> { ... }
}

🔬 This is a nightly-only experimental API. (pattern #27721)

API not fully fleshed out and ready to be stabilized

字符串模式的反向搜索器。

A reverse searcher for a string pattern.

此Trait提供了从字符串的背面（右侧）开始搜索模式的非重叠匹配的方法。

This trait provides methods for searching for non-overlapping matches of a pattern starting from the back (right) of a string.

如果模式支持从后面搜索它，它将由 Pattern Trait的相关 Searcher 类型实现。

It will be implemented by associated Searcher types of the Pattern trait if the pattern supports searching for it from the back.

此Trait返回的索引范围不需要与反向正向搜索的索引范围完全匹配。

The index ranges returned by this trait are not required to exactly match those of the forward search in reverse.

关于这个 trait 被标记为不安全的原因，请参见它们的父 trait Searcher。

For the reason why this trait is marked unsafe, see them parent trait Searcher.

Required methods

fn next_back(&mut self) -> SearchStep

🔬 This is a nightly-only experimental API. (pattern #27721)

API not fully fleshed out and ready to be stabilized

从后面开始执行下一个搜索步骤。

Performs the next search step starting from the back.

如果 haystack[a..b] 匹配模式，则返回 Match(a, b)。

Returns Match(a, b) if haystack[a..b] matches the pattern.

如果 haystack[a..b] 无法匹配模式，即使是部分匹配，则返回 Reject(a, b)。

Returns Reject(a, b) if haystack[a..b] can not match the pattern, even partially.

如果 haystack 的每个字节都被访问过，则返回 Done

Returns Done if every byte of the haystack has been visited

直到 Done 的 Match 和 Reject 值流将包含相邻、不重叠、覆盖整个 haystack 并且位于 utf8 边界上的索引范围。

The stream of Match and Reject values up to a Done will contain index ranges that are adjacent, non-overlapping, covering the whole haystack, and laying on utf8 boundaries.

匹配结果需要包含整个匹配模式，但是拒绝结果可能会分成任意多个相邻片段。 两个范围的长度都可能为零。

A Match result needs to contain the whole matched pattern, however Reject results may be split up into arbitrary many adjacent fragments. Both ranges may have zero length.

例如，模式 "aaa" 和 haystack "cbaaaaab" 可能产生流 [Reject(7, 8), Match(4, 7), Reject(1, 4), Reject(0, 1)]。

As an example, the pattern "aaa" and the haystack "cbaaaaab" might produce the stream [Reject(7, 8), Match(4, 7), Reject(1, 4), Reject(0, 1)].

Provided methods

fn next_match_back(&mut self) -> Option<(usize, usize)>

🔬 This is a nightly-only experimental API. (pattern #27721)

API not fully fleshed out and ready to be stabilized

查找下一个匹配结果。 请参阅 next_back()。

Finds the next Match result. See next_back().

fn next_reject_back(&mut self) -> Option<(usize, usize)>

🔬 This is a nightly-only experimental API. (pattern #27721)

API not fully fleshed out and ready to be stabilized

查找下一个拒绝结果。 请参阅 next_back()。

Finds the next Reject result. See next_back().