Module core::sync::atomic

原子类型

Atomic types

原子类型提供线程之间的原始共享内存通信，并且是其他并发类型的构建块。

Atomic types provide primitive shared-memory communication between threads, and are the building blocks of other concurrent types.

该模块定义了选择数量的基本类型的原子版本，包括 AtomicBool、AtomicIsize、AtomicUsize、AtomicI8、AtomicU16 等。原子类型表示操作，当正确使用时，可以在线程之间同步更新。

This module defines atomic versions of a select number of primitive types, including AtomicBool, AtomicIsize, AtomicUsize, AtomicI8, AtomicU16, etc. Atomic types present operations that, when used correctly, synchronize updates between threads.

每个方法都采用一个 Ordering，它表示该操作的内存屏障强度。 这些排序与 C++20 原子排序相同。 有关更多信息，请参阅 nomicon。

Each method takes an Ordering which represents the strength of the memory barrier for that operation. These orderings are the same as the C++20 atomic orderings. For more information see the nomicon.

原子变量在线程之间共享是安全的（它们实现了 Sync），但它们本身并不提供共享机制并遵循 Rust 的线程模型。 共享原子变量的最常见方法是将其放入 Arc（原子引用计数的共享指针）中。

Atomic variables are safe to share between threads (they implement Sync) but they do not themselves provide the mechanism for sharing and follow the threading model of Rust. The most common way to share an atomic variable is to put it into an Arc (an atomically-reference-counted shared pointer).

原子类型可以存储在静态变量中，使用诸如 AtomicBool::new 之类的常量初始化器进行初始化。 原子静态通常用于惰性全局初始化。

Atomic types may be stored in static variables, initialized using the constant initializers like AtomicBool::new. Atomic statics are often used for lazy global initialization.

Portability

如果可用，则此模块中的所有原子类型都保证是无锁的。 这意味着它们不会在内部获取全局互斥锁。 原子类型和操作不保证无需等待。 这意味着像 fetch_or 这样的操作可以通过比较和交换循环来实现。

All atomic types in this module are guaranteed to be lock-free if they’re available. This means they don’t internally acquire a global mutex. Atomic types and operations are not guaranteed to be wait-free. This means that operations like fetch_or may be implemented with a compare-and-swap loop.

原子操作可以用更大尺寸的原子在指令层实现。 例如，某些平台使用 4 字节原子指令来实现 AtomicI8。 请注意，这种模拟不应该对代码的正确性产生影响，它只是需要注意的事情。

Atomic operations may be implemented at the instruction layer with larger-size atomics. For example some platforms use 4-byte atomic instructions to implement AtomicI8. Note that this emulation should not have an impact on correctness of code, it’s just something to be aware of.

此模块中的原子类型可能并非在所有平台上都可用。 然而，这里的原子类型都是广泛可用的，并且通常可以依赖于现有的。 一些值得注意的例外是：

The atomic types in this module might not be available on all platforms. The atomic types here are all widely available, however, and can generally be relied upon existing. Some notable exceptions are:

具有 32 位指针的 PowerPC 和 MIPS 平台没有 AtomicU64 或 AtomicI64 类型。

PowerPC and MIPS platforms with 32-bit pointers do not have AtomicU64 or AtomicI64 types.

ARM 平台如 armv5te 不支持 Linux 只提供加载和存储操作，不支持比较和交换 (CAS) 操作，例如 swap、fetch_add 等。另外在 Linux 上，这些 CAS 操作是通过操作系统支持实现的 ，这可能会带来性能损失。

ARM platforms like armv5te that aren’t for Linux only provide load and store operations, and do not support Compare and Swap (CAS) operations, such as swap, fetch_add, etc. Additionally on Linux, these CAS operations are implemented via operating system support, which may come with a performance penalty.

带有 thumbv6m 的 ARM 目标仅提供加载和存储操作，不支持比较和交换 (CAS) 操作，例如 swap、fetch_add 等。

ARM targets with thumbv6m only provide load and store operations, and do not support Compare and Swap (CAS) operations, such as swap, fetch_add, etc.

请注意，将来可能会添加也不支持某些原子操作的平台。 最大可移植代码需要注意使用了哪些原子类型。 AtomicUsize 和 AtomicIsize 通常是最便携的，但即便如此，它们也不是随处可用。 作为参考，std 库需要指针大小的原子，尽管核心不需要。

Note that future platforms may be added that also do not have support for some atomic operations. Maximally portable code will want to be careful about which atomic types are used. AtomicUsize and AtomicIsize are generally the most portable, but even then they’re not available everywhere. For reference, the std library requires pointer-sized atomics, although core does not.

目前，您需要使用 #[cfg(target_arch)] 主要用于有条件地编译带有原子的代码。 还有一个不稳定的#[cfg(target_has_atomic)]，将来可能会稳定下来。

Currently you’ll need to use #[cfg(target_arch)] primarily to conditionally compile in code with atomics. There is an unstable #[cfg(target_has_atomic)] as well which may be stabilized in the future.

Examples

一个简单的自旋锁：

A simple spinlock:

use std::sync::Arc;
use std::sync::atomic::{AtomicUsize, Ordering};
use std::{hint, thread};

fn main() {
    let spinlock = Arc::new(AtomicUsize::new(1));

    let spinlock_clone = Arc::clone(&spinlock);
    let thread = thread::spawn(move|| {
        spinlock_clone.store(0, Ordering::SeqCst);
    });

    // Wait for the other thread to release the lock
    while spinlock.load(Ordering::SeqCst) != 0 {
        hint::spin_loop();
    }

    if let Err(panic) = thread.join() {
        println!("Thread had an error: {:?}", panic);
    }
}
Run

保持活动线程的全局计数：

Keep a global count of live threads:

use std::sync::atomic::{AtomicUsize, Ordering};

static GLOBAL_THREAD_COUNT: AtomicUsize = AtomicUsize::new(0);

let old_thread_count = GLOBAL_THREAD_COUNT.fetch_add(1, Ordering::SeqCst);
println!("live threads: {}", old_thread_count + 1);
Run

Structs

AtomicBool

A boolean type which can be safely shared between threads.

AtomicI8

An integer type which can be safely shared between threads.

AtomicI16

An integer type which can be safely shared between threads.

AtomicI32

An integer type which can be safely shared between threads.

AtomicI64

An integer type which can be safely shared between threads.

AtomicIsize

An integer type which can be safely shared between threads.

AtomicPtr

A raw pointer type which can be safely shared between threads.

AtomicU8

An integer type which can be safely shared between threads.

AtomicU16

An integer type which can be safely shared between threads.

AtomicU32

An integer type which can be safely shared between threads.

AtomicU64

An integer type which can be safely shared between threads.

AtomicUsize

An integer type which can be safely shared between threads.

Enums

Ordering

Atomic memory orderings

Constants

ATOMIC_I8_INITDeprecatedExperimental

An atomic integer initialized to 0.

ATOMIC_I16_INITDeprecatedExperimental

An atomic integer initialized to 0.

ATOMIC_I32_INITDeprecatedExperimental

An atomic integer initialized to 0.

ATOMIC_I64_INITDeprecatedExperimental

An atomic integer initialized to 0.

ATOMIC_U8_INITDeprecatedExperimental

An atomic integer initialized to 0.

ATOMIC_U16_INITDeprecatedExperimental

An atomic integer initialized to 0.

ATOMIC_U32_INITDeprecatedExperimental

An atomic integer initialized to 0.

ATOMIC_U64_INITDeprecatedExperimental

An atomic integer initialized to 0.

ATOMIC_BOOL_INITDeprecated

An AtomicBool initialized to false.

ATOMIC_ISIZE_INITDeprecated

An atomic integer initialized to 0.

ATOMIC_USIZE_INITDeprecated

An atomic integer initialized to 0.

Functions

compiler_fence

A compiler memory fence.

fence

An atomic fence.

spin_loop_hintDeprecated

Signals the processor that it is inside a busy-wait spin-loop (“spin lock”).