Enum core::sync::atomic::Ordering

#[non_exhaustive]
pub enum Ordering {
    Relaxed,
    Release,
    Acquire,
    AcqRel,
    SeqCst,
}

原子内存排序

Atomic memory orderings

内存排序指定原子操作同步内存的方式。 在其最弱的 Ordering::Relaxed 中，仅同步操作直接接触的内存。 另一方面，Ordering::SeqCst 操作的存储-加载对同步其他内存，同时额外保留所有线程中此类操作的总顺序。

Memory orderings specify the way atomic operations synchronize memory. In its weakest Ordering::Relaxed, only the memory directly touched by the operation is synchronized. On the other hand, a store-load pair of Ordering::SeqCst operations synchronize other memory while additionally preserving a total order of such operations across all threads.

Rust 的内存顺序与 C++20 相同。

Rust’s memory orderings are the same as those of C++20.

有关更多信息，请参阅 nomicon。

For more information see the nomicon.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Relaxed

没有排序约束，只有原子操作。

No ordering constraints, only atomic operations.

对应于 C++20 中的 memory_order_relaxed。

Corresponds to memory_order_relaxed in C++20.

Release

当与 store 结合使用时，所有先前的操作都在使用 Acquire（或更强）排序加载此值之前排序。 特别是，所有先前的写入对执行该值的获取（或更强）加载的所有线程都是可见的。

When coupled with a store, all previous operations become ordered before any load of this value with Acquire (or stronger) ordering. In particular, all previous writes become visible to all threads that perform an Acquire (or stronger) load of this value.

请注意，将此排序用于组合加载和存储的操作会导致轻松加载操作！

Notice that using this ordering for an operation that combines loads and stores leads to a Relaxed load operation!

此排序仅适用于可以执行存储的操作。

This ordering is only applicable for operations that can perform a store.

对应于 C++20 中的 memory_order_release。

Corresponds to memory_order_release in C++20.

Acquire

当与加载相结合时，如果加载的值是由具有发布（或更强）排序的存储操作写入的，则所有后续操作都在该存储之后排序。 特别是，所有后续加载都将看到在存储之前写入的数据。

When coupled with a load, if the loaded value was written by a store operation with Release (or stronger) ordering, then all subsequent operations become ordered after that store. In particular, all subsequent loads will see data written before the store.

请注意，将此排序用于组合加载和存储的操作会导致轻松存储操作！

Notice that using this ordering for an operation that combines loads and stores leads to a Relaxed store operation!

此排序仅适用于可以执行加载的操作。

This ordering is only applicable for operations that can perform a load.

对应于 C++20 中的 memory_order_acquire。

Corresponds to memory_order_acquire in C++20.

AcqRel

同时具有 Acquire 和 Release 的效果：对于负载，它使用 Acquire 排序。 对于商店，它使用发布顺序。

Has the effects of both Acquire and Release together: For loads it uses Acquire ordering. For stores it uses the Release ordering.

请注意，在 compare_and_swap 的情况下，操作可能最终没有执行任何存储，因此它只是 Acquire ordering。 但是，AcqRel 永远不会执行宽松访问。

Notice that in the case of compare_and_swap, it is possible that the operation ends up not performing any store and hence it has just Acquire ordering. However, AcqRel will never perform Relaxed accesses.

此排序仅适用于结合加载和存储的操作。

This ordering is only applicable for operations that combine both loads and stores.

对应于 C++20 中的 memory_order_acq_rel。

Corresponds to memory_order_acq_rel in C++20.

SeqCst

就像 Acquire/Release/AcqRel（分别用于加载、存储和加载存储操作）一样，额外保证所有线程以相同的顺序查看所有顺序一致的操作。

Like Acquire/Release/AcqRel (for load, store, and load-with-store operations, respectively) with the additional guarantee that all threads see all sequentially consistent operations in the same order.

对应于 C++20 中的 memory_order_seq_cst。

Corresponds to memory_order_seq_cst in C++20.