Function core::sync::atomic::compiler_fence

pub fn compiler_fence(order: Ordering)

编译器内存栅栏。

A compiler memory fence.

`compiler_fence` 不会发出任何机器代码，但会限制编译器允许执行的内存重新排序的种类。 具体来说，根据给定的“排序”语义，可能不允许编译器将调用之前或之后的读取或写入移动到调用 `compiler_fence` 的另一端。 请注意，它不会阻止硬件进行此类重新排序。 这在单线程执行上下文中不是问题，但是当其他线程可能同时修改内存时，需要更强的同步原语，例如“fence”。

compiler_fence does not emit any machine code, but restricts the kinds of memory re-ordering the compiler is allowed to do. Specifically, depending on the given Ordering semantics, the compiler may be disallowed from moving reads or writes from before or after the call to the other side of the call to compiler_fence. Note that it does not prevent the hardware from doing such re-ordering. This is not a problem in a single-threaded, execution context, but when other threads may modify memory at the same time, stronger synchronization primitives such as fence are required.

不同排序语义阻止的重新排序是：

The re-ordering prevented by the different ordering semantics are:

使用“SeqCst”，不允许在这一点上重新排序读取和写入。

with SeqCst, no re-ordering of reads and writes across this point is allowed.

使用“Release”，前面的读取和写入不能移到后面的写入之后。

with Release, preceding reads and writes cannot be moved past subsequent writes.

使用“Acquire”，后续的读取和写入不能移动到前面的读取之前。

with Acquire, subsequent reads and writes cannot be moved ahead of preceding reads.

使用“AcqRel”，以上两个规则都被强制执行。

with AcqRel, both of the above rules are enforced.

`compiler_fence` 通常只用于防止线程与自身竞争。 也就是说，如果给定线程正在执行一段代码，然后被中断，并开始在其他地方执行代码（同时仍在同一个线程中，并且在概念上仍然在同一个核心上）。 在传统程序中，这只能在注册信号处理程序时发生。 在更底层的代码中，在处理中断、实现带抢占的绿色线程等时也会出现这种情况。鼓励好奇的读者阅读 Linux 内核对“内存屏障”的讨论。

compiler_fence is generally only useful for preventing a thread from racing with itself. That is, if a given thread is executing one piece of code, and is then interrupted, and starts executing code elsewhere (while still in the same thread, and conceptually still on the same core). In traditional programs, this can only occur when a signal handler is registered. In more low-level code, such situations can also arise when handling interrupts, when implementing green threads with pre-emption, etc. Curious readers are encouraged to read the Linux kernel’s discussion of memory barriers.

Panics

如果 `order` 是“Relaxed”，则会出现panic。

Panics if order is Relaxed.

Examples

如果没有 `compiler_fence`，即使所有事情都发生在单个线程中，以下代码中的 `assert_eq!` 也不能保证成功。 要了解原因，请记住编译器可以自由地将存储交换为 `IMPORTANT_VARIABLE` 和 `IS_READY`，因为它们都是 `Ordering::Relaxed`。 如果是这样，并且在 `IS_READY` 更新后立即调用信号处理程序，则信号处理程序将看到 `IS_READY=1`，但 `IMPORTANT_VARIABLE=0`。 使用 `compiler_fence` 可以解决这种情况。

Without compiler_fence, the assert_eq! in following code is not guaranteed to succeed, despite everything happening in a single thread. To see why, remember that the compiler is free to swap the stores to IMPORTANT_VARIABLE and IS_READY since they are both Ordering::Relaxed. If it does, and the signal handler is invoked right after IS_READY is updated, then the signal handler will see IS_READY=1, but IMPORTANT_VARIABLE=0. Using a compiler_fence remedies this situation.

use std::sync::atomic::{AtomicBool, AtomicUsize};
use std::sync::atomic::Ordering;
use std::sync::atomic::compiler_fence;

static IMPORTANT_VARIABLE: AtomicUsize = AtomicUsize::new(0);
static IS_READY: AtomicBool = AtomicBool::new(false);

fn main() {
    IMPORTANT_VARIABLE.store(42, Ordering::Relaxed);
    // prevent earlier writes from being moved beyond this point
    compiler_fence(Ordering::Release);
    IS_READY.store(true, Ordering::Relaxed);
}

fn signal_handler() {
    if IS_READY.load(Ordering::Relaxed) {
        assert_eq!(IMPORTANT_VARIABLE.load(Ordering::Relaxed), 42);
    }
}
Run