Enum core::task::Poll

pub enum Poll<T> {
    Ready(T),
    Pending,
}

Indicates whether a value is available or if the current task has been scheduled to receive a wakeup instead.

Variants

Ready(T)

Represents that a value is immediately ready.

Pending

Represents that a value is not ready yet.

When a function returns Pending, the function must also ensure that the current task is scheduled to be awoken when progress can be made.

Implementations

impl<T> Poll<T>

pub fn map<U, F>(self, f: F) -> Poll<U>where
F: FnOnce(T) -> U,

Maps a Poll<T> to Poll<U> by applying a function to a contained value.

Examples

Converts a Poll<String> into a Poll<usize>, consuming the original:

let poll_some_string = Poll::Ready(String::from("Hello, World!"));
// `Poll::map` takes self *by value*, consuming `poll_some_string`
let poll_some_len = poll_some_string.map(|s| s.len());

assert_eq!(poll_some_len, Poll::Ready(13));
Run

pub const fn is_ready(&self) -> bool

Returns true if the poll is a Poll::Ready value.

Examples

let x: Poll<u32> = Poll::Ready(2);
assert_eq!(x.is_ready(), true);

let x: Poll<u32> = Poll::Pending;
assert_eq!(x.is_ready(), false);
Run

pub const fn is_pending(&self) -> bool

Returns true if the poll is a Pending value.

Examples

let x: Poll<u32> = Poll::Ready(2);
assert_eq!(x.is_pending(), false);

let x: Poll<u32> = Poll::Pending;
assert_eq!(x.is_pending(), true);
Run

pub fn ready(self) -> Ready<T>

🔬This is a nightly-only experimental API. (poll_ready #89780)

Extracts the successful type of a Poll<T>.

When combined with the ? operator, this function will propagate any Poll::Pending values to the caller, and extract the T from Poll::Ready.

Examples

#![feature(poll_ready)]

use std::task::{Context, Poll};
use std::future::{self, Future};
use std::pin::Pin;

pub fn do_poll(cx: &mut Context<'_>) -> Poll<()> {
    let mut fut = future::ready(42);
    let fut = Pin::new(&mut fut);

    let num = fut.poll(cx).ready()?;
    // ... use num

    Poll::Ready(())
}
Run

impl<T, E> Poll<Result<T, E>>

pub fn map_ok<U, F>(self, f: F) -> Poll<Result<U, E>>where
F: FnOnce(T) -> U,

Maps a Poll<Result<T, E>> to Poll<Result<U, E>> by applying a function to a contained Poll::Ready(Ok) value, leaving all other variants untouched.

This function can be used to compose the results of two functions.

Examples

let res: Poll<Result<u8, _>> = Poll::Ready("12".parse());
let squared = res.map_ok(|n| n * n);
assert_eq!(squared, Poll::Ready(Ok(144)));
Run

pub fn map_err<U, F>(self, f: F) -> Poll<Result<T, U>>where
F: FnOnce(E) -> U,

Maps a Poll::Ready<Result<T, E>> to Poll::Ready<Result<T, F>> by applying a function to a contained Poll::Ready(Err) value, leaving all other variants untouched.

This function can be used to pass through a successful result while handling an error.

Examples

let res: Poll<Result<u8, _>> = Poll::Ready("oops".parse());
let res = res.map_err(|_| 0_u8);
assert_eq!(res, Poll::Ready(Err(0)));
Run

impl<T, E> Poll<Option<Result<T, E>>>

pub fn map_ok<U, F>(self, f: F) -> Poll<Option<Result<U, E>>>where
F: FnOnce(T) -> U,

Maps a Poll<Option<Result<T, E>>> to Poll<Option<Result<U, E>>> by applying a function to a contained Poll::Ready(Some(Ok)) value, leaving all other variants untouched.

This function can be used to compose the results of two functions.

Examples

let res: Poll<Option<Result<u8, _>>> = Poll::Ready(Some("12".parse()));
let squared = res.map_ok(|n| n * n);
assert_eq!(squared, Poll::Ready(Some(Ok(144))));
Run

pub fn map_err<U, F>(self, f: F) -> Poll<Option<Result<T, U>>>where
F: FnOnce(E) -> U,

Maps a Poll::Ready<Option<Result<T, E>>> to Poll::Ready<Option<Result<T, F>>> by applying a function to a contained Poll::Ready(Some(Err)) value, leaving all other variants untouched.

This function can be used to pass through a successful result while handling an error.

Examples

let res: Poll<Option<Result<u8, _>>> = Poll::Ready(Some("oops".parse()));
let res = res.map_err(|_| 0_u8);
assert_eq!(res, Poll::Ready(Some(Err(0))));
Run