Struct core::task::Waker

#[repr(transparent)]
pub struct Waker { /* private fields */ }

A Waker is a handle for waking up a task by notifying its executor that it is ready to be run.

This handle encapsulates a RawWaker instance, which defines the executor-specific wakeup behavior.

The typical life of a Waker is that it is constructed by an executor, wrapped in a Context, then passed to Future::poll(). Then, if the future chooses to return Poll::Pending, it must also store the waker somehow and call Waker::wake() when the future should be polled again.

Implements Clone, Send, and Sync; therefore, a waker may be invoked from any thread, including ones not in any way managed by the executor. For example, this might be done to wake a future when a blocking function call completes on another thread.

Implementations

impl Waker

pub fn wake(self)

Wake up the task associated with this Waker.

As long as the executor keeps running and the task is not finished, it is guaranteed that each invocation of wake() (or wake_by_ref()) will be followed by at least one poll() of the task to which this Waker belongs. This makes it possible to temporarily yield to other tasks while running potentially unbounded processing loops.

Note that the above implies that multiple wake-ups may be coalesced into a single poll() invocation by the runtime.

Also note that yielding to competing tasks is not guaranteed: it is the executor’s choice which task to run and the executor may choose to run the current task again.

pub fn wake_by_ref(&self)

Wake up the task associated with this Waker without consuming the Waker.

This is similar to wake(), but may be slightly less efficient in the case where an owned Waker is available. This method should be preferred to calling waker.clone().wake().

pub fn will_wake(&self, other: &Waker) -> bool

Returns true if this Waker and another Waker would awake the same task.

This function works on a best-effort basis, and may return false even when the Wakers would awaken the same task. However, if this function returns true, it is guaranteed that the Wakers will awaken the same task.

This function is primarily used for optimization purposes.

pub unsafe fn from_raw(waker: RawWaker) -> Waker

Creates a new Waker from RawWaker.

The behavior of the returned Waker is undefined if the contract defined in RawWaker’s and RawWakerVTable’s documentation is not upheld. Therefore this method is unsafe.

pub fn as_raw(&self) -> &RawWaker

🔬This is a nightly-only experimental API. (waker_getters #87021)

Get a reference to the underlying RawWaker.